% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakAnnotation.R
\name{addArchRAnnotations}
\alias{addArchRAnnotations}
\title{Add ArchR annotations to an ArchRProject}
\usage{
addArchRAnnotations(ArchRProj = NULL, db = "ArchR",
  collection = "EncodeTFBS", name = collection, force = FALSE)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{db}{A character indicating which database or a path to a database to use for peak annotation. Options include ArchR, LOLA, and a valid path to a file of class \code{ArchRAnno}.}

\item{collection}{A character indicating which collection within the database to collect for annotation.
For ArchR, options "ATAC", "EncodeTFBS", "CistromeTFBS", or "Codex".
For LOLA, options include "EncodeTFBS" "CistromeTFBS", "CistromeEpigenome", "Codex", or "SheffieldDnase".
If supplying a custom ArchRAnno please use a valid collection.}

\item{name}{The name of \code{peakAnnotation} object to be stored as in \code{ArchRProject}.}

\item{force}{A boolean value indicating whether to force the \code{peakAnnotation} object indicated by \code{name} to be overwritten if it already exist in the given \code{ArchRProject}.}
}
\description{
This function adds information about which peaks in the ArchR database contain input regions to a given ArchRProject. For each peak, a binary value is stored indicating whether each region is observed within the peak region.
}
