% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalDefaults.R
\name{addArchRChrPrefix}
\alias{addArchRChrPrefix}
\title{Add a globally-applied requirement for chromosome prefix}
\usage{
addArchRChrPrefix(chrPrefix = TRUE)
}
\arguments{
\item{chrPrefix}{A boolean describing the requirement of chromosomes to have a "chr" prefix.}
}
\description{
This function will set the default requirement of chromosomes to have a "chr" prefix.
}
\examples{

# Add ArchR Chr Prefix
addArchRChrPrefix()

}
