% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoggerUtils.R
\name{addArchRDebugging}
\alias{addArchRDebugging}
\title{Set ArchR Debugging}
\usage{
addArchRDebugging(debug = FALSE)
}
\arguments{
\item{debug}{A boolean describing whether to use logging with ArchR.}
}
\description{
This function will set ArchR Debugging which will save an RDS if an error is encountered.
}
