% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalDefaults.R
\name{addArchRH5Level}
\alias{addArchRH5Level}
\title{Add a globally-applied compression level for h5 files}
\usage{
addArchRH5Level(level = 0)
}
\arguments{
\item{level}{The default compression level to be used for h5 file execution across all ArchR functions.}
}
\description{
This function will set the default compression level to be used for h5 file execution across all ArchR functions.
}
\examples{

# Add ArchR H5 Compression level
addArchRH5Level()

}
