% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoggerUtils.R
\name{addArchRLogging}
\alias{addArchRLogging}
\title{Set ArchR Logging}
\usage{
addArchRLogging(useLogs = TRUE)
}
\arguments{
\item{useLogs}{A boolean describing whether to use logging with ArchR.}
}
\description{
This function will set ArchR logging
}
\examples{

# Add ArchR Logging
addArchRLogging()

}
