% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoggerUtils.R
\name{addArchRVerbose}
\alias{addArchRVerbose}
\title{Set ArchR Verbosity for Log Messaging}
\usage{
addArchRVerbose(verbose = TRUE)
}
\arguments{
\item{verbose}{A boolean describing whether to printMessages in addition to logging with ArchR.}
}
\description{
This function will set ArchR logging verbosity.
}
