% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{addCellColData}
\alias{addCellColData}
\title{Add information to cellColData in ArchRProject}
\usage{
addCellColData(ArchRProj, data = NULL, name = NULL,
  cells = getCellNames(ArchRProj), force = FALSE, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{data}{data to add to cellColData}

\item{name}{new column name in cellColData if already exists set force = TRUE to override}

\item{cells}{names of cells corresponding to data}

\item{force}{if name already exists in cellColData set force = TRUE to override}

\item{...}{additional args}
}
\description{
This function adds new data to cellColData in ArchRProject
}
