% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixCNV.R
\name{addCNVMatrix}
\alias{addCNVMatrix}
\title{Add TileMatrix to Arrows/ArchRProject}
\usage{
addCNVMatrix(input, chromSizes = getChromSizes(input),
  blacklist = getBlacklist(input), genome = getGenome(input),
  windowSize = 1e+07, stepSize = 2e+06, excludeChr = c("chrM",
  "chrY"), threads = 1, parallelParam = NULL, force = FALSE, ...)
}
\arguments{
\item{input}{ArchRProject or ArrowFiles}

\item{chromSizes}{chromomosome sizes used for identifying number of tiles to count}

\item{windowSize}{size for each window to break up each chromosome}

\item{excludeChr}{exclude chromosomes from this analysis}

\item{threads}{number of threads}

\item{parallelParam}{parallel parameters for batch style execution}

\item{force}{force overwriting previous TileMatrix in ArrowFile}

\item{binarize}{save as a Sparse.Binary.Matrix or Sparse.Integer.Matrix}
}
\description{
This function for each sample will independently compute counts for each tile
per cell in the Arrow File
}
