% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiModal.R
\name{addCombinedDims}
\alias{addCombinedDims}
\title{Combine two or more modalities dimensionality reductions.}
\usage{
addCombinedDims(
  ArchRProj = NULL,
  name = "CombinedDims",
  reducedDims = NULL,
  dimWeights = NULL,
  dimsToUse = NULL,
  scaleDims = NULL,
  scaleBy = NULL,
  corCutOff = 0.75
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{name}{The name for the combinedDims to be stored as.}

\item{reducedDims}{The name of the \code{reducedDims} objects (i.e. "IterativeLSI") to use from the designated \code{ArchRProject}.}

\item{dimWeights}{A vector of weights to be used to weight each dimensionality reduction when combining.}

\item{dimsToUse}{A vector containing the dimensions from the \code{reducedDims} objects to use.}

\item{scaleDims}{A boolean value that indicates whether to z-score the reduced dimensions. The default is set to \code{NULL}, and will scale the dimensions
based on the value of \code{scaleDims} when the \code{reducedDims} were originally created during dimensionality reduction. This idea was introduced by Timothy Stuart.}

\item{scaleBy}{A character string indicating if the reduced dimensions should be scaled in either the row direction (default) or the column direction when \code{scaleDims = TRUE}.
In the case of SVD matrix, the default is to perform scaling for each cell, rather than on the components as in the \code{signac::RunSVD} implementation.
You can use \code{scaleBy = "column"} to perform scaling for each component. Like, \code{scaleDims}, the saved value of \code{scaleBy} will be used if set to \code{scaleBy = NULL}.}

\item{corCutOff}{A numeric cutoff for the correlation of each dimension to the sequencing depth. If the dimension has a correlation to
sequencing depth that is greater than the \code{corCutOff}, it will be excluded from analysis.}
}
\description{
This function will combine two or more modalities dimensionality reductions into a single reduction.
}
