% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoubletsScores.R
\name{addDemuxletResults}
\alias{addDemuxletResults}
\title{Add Demuxlet Results to ArchR Project}
\usage{
addDemuxletResults(ArchRProj, bestFiles, sampleNames)
}
\arguments{
\item{ArchRProj}{ArchR Project}

\item{bestFiles}{paths to best files}

\item{sampleNames}{sampleNames corresponding to best files (match those in ArchRProj)}
}
\description{
This function will read in .best file output from demuxlet and add the
classifications into the cellColData for the ArchR Project
}
