% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixDeviations.R
\name{addDeviationsMatrix}
\alias{addDeviationsMatrix}
\title{Add DeviationsMatrix to Arrow Files in ArchRProject}
\usage{
addDeviationsMatrix(ArchRProj, annotations = NULL, matches = NULL,
  bdgPeaks = getBdgPeaks(ArchRProj), matrixName = NULL, out = c("z",
  "deviations"), binarize = FALSE, threads = 1, parallelParam = NULL,
  force = FALSE, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{annotations}{annotaions name stored in ArchRProject}

\item{matrixName}{matrixName to be stored as in Arrow Files}

\item{out}{save ouptut matrices deviations and/or z}

\item{binarize}{binarize peaks prior to computing deviations}

\item{threads}{number of threads for parallel execution}

\item{parallelParam}{parallel parameters for batch style execution}

\item{force}{force overwriting previous TileMatrix in ArrowFile}
}
\description{
This function for each sample will independently compute counts for each tile
per cell and then infer gene activity scores.
}
