% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoubletsScores.R
\name{addDoubletScores}
\alias{addDoubletScores}
\title{Add Doublet Scores to a collection of ArrowFiles or an ArchRProject}
\usage{
addDoubletScores(
  input = NULL,
  useMatrix = "TileMatrix",
  k = 10,
  nTrials = 5,
  dimsToUse = 1:30,
  LSIMethod = 1,
  scaleDims = FALSE,
  scaleBy = "row",
  corCutOff = 0.75,
  knnMethod = "UMAP",
  UMAPParams = list(n_neighbors = 40, min_dist = 0.4, metric = "euclidean", verbose =
    FALSE),
  LSIParams = list(outlierQuantiles = NULL, filterBias = FALSE),
  outDir = getOutputDirectory(input),
  threads = getArchRThreads(),
  force = FALSE,
  parallelParam = NULL,
  verbose = TRUE,
  logFile = createLogFile("addDoubletScores")
)
}
\arguments{
\item{input}{An \code{ArchRProject} object or a character vector containing the paths to the ArrowFiles to be used.}

\item{useMatrix}{The name of the matrix to be used for performing doublet identification analyses. Options include "TileMatrix" and "PeakMatrix".}

\item{k}{The number of cells neighboring a simulated doublet to be considered as putative doublets.}

\item{nTrials}{The number of times to simulate nCell (number of cells in the sample) doublets to use for doublet simulation when calculating doublet scores.}

\item{dimsToUse}{A vector containing the dimensions from the \code{reducedDims} object to use in clustering.}

\item{LSIMethod}{A number or string indicating the order of operations in the TF-IDF normalization.
Possible values are: 1 or "tf-logidf", 2 or "log(tf-idf)", and 3 or "logtf-logidf".}

\item{scaleDims}{A boolean value that indicates whether to z-score the reduced dimensions during the LSI method performed for doublet determination.
The default is to NOT perform scaling on the SVD matrix. You can use \code{scaleDims = TRUE} to turn ON scaling.}

\item{scaleBy}{A character string indicating if the reduced dimensions should be scaled in either the row direction (default) or the column direction when \code{scaleDims = TRUE}.
In the case of SVD matrix, the default is to perform scaling for each cell, rather than on the components as in the \code{signac::RunSVD} implementation.
You can use \code{scaleBy = "column"} to perform scaling for each component.}

\item{corCutOff}{A numeric cutoff for the correlation of each dimension to the sequencing depth. If the dimension has a correlation
to sequencing depth that is greater than the \code{corCutOff}, it will be excluded from analysis.}

\item{knnMethod}{The name of the dimensionality reduction method to be used for k-nearest neighbors calculation. Possible values are "UMAP" or "LSI".}

\item{UMAPParams}{The list of parameters to pass to the UMAP function if "UMAP" is designated to \code{knnMethod}. See the function \code{umap} in the uwot package.}

\item{LSIParams}{The list of parameters to pass to the \code{IterativeLSI()} function. See \code{IterativeLSI()}.}

\item{outDir}{The relative path to the output directory for relevant plots/results from doublet identification.}

\item{threads}{The number of threads to be used for parallel computing.}

\item{force}{If the UMAP projection is not accurate (when R < 0.8 for the reprojection of the training data - this occurs when you
have a very homogenous population of cells), setting \code{force=FALSE} will return -1 for all doubletScores and doubletEnrichments. If you would like to
override this (not recommended!), you can bypass this warning by setting \code{force=TRUE}.}

\item{parallelParam}{A list of parameters to be passed for biocparallel/batchtools parallel computing.}

\item{verbose}{A boolean value that determines whether standard output is printed.}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
For each sample in the ArrowFiles or ArchRProject provided, this function will independently assign inferred doublet information
to each cell. This allows for removing strong heterotypic doublet-based clusters downstream. A doublet results from a droplet that
contained two cells, causing the ATAC-seq data to be a mixture of the signal from each cell.
}
\examples{

# Get Test ArchR Project
proj <- getTestProject()

# Add Doublet Scores for Small Project
proj <- addDoubletScores(proj, dimsToUse = 1:5, LSIParams = list(dimsToUse = 1:5, varFeatures=1000, iterations = 2))

}
