% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{addFeatureCounts}
\alias{addFeatureCounts}
\title{This function will add total counts of scATAC cells in provided features into ArchRProject.}
\usage{
addFeatureCounts(ArchRProj = NULL, features = NULL, name = NULL,
  addRatio = TRUE, threads = getArchRThreads())
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}
}
\description{
This function will add total counts of scATAC cells in provided features into ArchRProject.
}
