% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixFeatures.R
\name{addFeatureMatrix}
\alias{addFeatureMatrix}
\title{Add FeatureMatrix to Arrows/ArchRProject}
\usage{
addFeatureMatrix(input, features = NULL, matrixName = "FeatureMatrix",
  ceiling = Inf, binarize = FALSE, threads = 1,
  parallelParam = NULL, force = FALSE, ...)
}
\arguments{
\item{input}{ArchRProject or ArrowFiles}

\item{features}{GRanges to count for each cell}

\item{matrixName}{matrix output name in ArrowFiles cannot be a protected matrix name}

\item{ceiling}{ceiling for the number of counts per feature}

\item{binarize}{binarize matrix}

\item{threads}{number of threads}

\item{parallelParam}{parallel parameters for batch style execution}

\item{force}{force overwriting previous TileMatrix in ArrowFile}
}
\description{
This function for each sample will independently compute counts for each feature
per cell in the Arrow File
}
