% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixGeneExpression.R
\name{addGeneExpressionMatrix}
\alias{addGeneExpressionMatrix}
\title{Add Gene Expression Matrix to ArrowFiles or an ArchRProject}
\usage{
addGeneExpressionMatrix(
  input = NULL,
  seRNA = NULL,
  chromSizes = getChromSizes(input),
  excludeChr = c("chrM", "chrY"),
  scaleTo = 10000,
  verbose = TRUE,
  threads = getArchRThreads(),
  parallelParam = NULL,
  strictMatch = FALSE,
  force = TRUE,
  logFile = createLogFile("addGeneExpressionMatrix")
)
}
\arguments{
\item{input}{An \code{ArchRProject} object or character vector of ArrowFiles.}

\item{seRNA}{A a scRNA-seq \code{SummarizedExperiment} (cell x gene) to be integrated with the scATAC-seq data.
Cell names from this object much match those of the cell names in the ArrowFiles/ArchRProject. We will add support shortly
for Seurat Objects (see \code{Seurat::as.SingleCellExperiment}). The provided values \code{MUST} be in counts (integer), not log transformed.}

\item{chromSizes}{A GRanges object of the chromosome lengths. See \code{getChromSizes} for more info.}

\item{excludeChr}{A character vector containing the \code{seqnames} of the chromosomes that should be excluded from this analysis.}

\item{scaleTo}{Each column in the calculated gene score matrix will be normalized to a column sum designated by \code{scaleTo}.}

\item{verbose}{A boolean describing whether to print to console messages of progress.}

\item{threads}{The number of threads to be used for parallel computing.}

\item{parallelParam}{A list of parameters to be passed for biocparallel/batchtools parallel computing.}

\item{strictMatch}{A boolean value indicating whether every cell in \code{input} must be represented in \code{seRNA}. If set to \code{FALSE},
this and this \code{GeneExpressionMatrix} is used for certain downstream analyses such as \code{addIterativeLSI()}, then errors may occur
because not all cells will have relevant information.}

\item{force}{A boolean value indicating whether to force the matrix indicated by \code{matrixName} to be overwritten if it already exist in the given \code{input}.}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
This function, for each sample, will add gene expression values from a paired scATAC-seq + scRNA-seq
multi modal assay to the ArrowFiles or ArchRProject.
}
