% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeLSI.R
\name{addHarmony}
\alias{addHarmony}
\title{Add Harmony Batch Correction Reduced Dims to ArchRProject JJJ}
\usage{
addHarmony(ArchRProj = NULL, reducedDims = "IterativeLSI",
  dimsToUse = NULL, scaleDims = NULL, corCutOff = 0.75,
  name = "Harmony", groupBy = "Sample", verbose = TRUE,
  force = FALSE, ...)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object containing the dimensionality reduction matrix passed by \code{reducedDims}.}

\item{reducedDims}{The name of the \code{reducedDims} object (i.e. "IterativeLSI") to retrieve from the designated \code{ArchRProject}.}

\item{dimsToUse}{A vector containing the dimensions from the \code{reducedDims} object to use in clustering.}

\item{scaleDims}{A boolean describing whether to z-score the reduced dimensions for each cell. This is useful for minimizing the contribution of strong biases
(dominating early PCs) and lowly abundant populations. However, this may lead to stronger sample-specific biases since it is over-weighting latent PCs.
If \code{NULL} this will scale the dimensions depending on if this were set true when the \code{reducedDims} were created by the dimensionality reduction method.
This idea was introduced by Timothy Stuart.}

\item{corCutOff}{A numeric cutoff for the correlation of each dimension to the sequencing depth. If the dimension has a correlation to sequencing depth that is greater than the \code{corCutOff}, it will be excluded from analysis.}

\item{name}{The column name of the cluster label column to be added to \code{cellColData} if \code{input} is an \code{ArchRProject} object.}

\item{verbose}{A boolean value indicating whether to use verbose output during execution of this function. Can be set to FALSE for a cleaner output.}

\item{force}{A boolean value that indicates whether or not to overwrite data in a given column when the value passed to \code{name} already exists as a column name in \code{cellColData}.}

\item{...}{Additional arguments to be provided to harmony::HarmonyMatrix}
}
\description{
This function will identify clusters from a reduced dimensions object in an ArchRProject or from a supplied reduced dimensions matrix.
}
