% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{addImputeWeights}
\alias{addImputeWeights}
\title{QQQ}
\usage{
addImputeWeights(ArchRProj = NULL, reducedDims = "IterativeLSI",
  dimsToUse = NULL, td = 3, ka = 4, sampleCells = max(5000,
  floor(nCells(ArchRProj)/10)), k = 15, epsilon = 1, ...)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{reducedDims}{QQQ The name of the \code{reducedDims} object to retrieve from the designated \code{ArchRProject}. Options include QQQ. QQQ Not required if input is a matrix.}

\item{dimsToUse}{QQQ A vector containing the dimensions from the \code{reducedDims} object to use in clustering.}

\item{td}{diffusion time (number of iterations) for Magic}

\item{ka}{kNN autotune parameter for Magic}

\item{sampleCells}{number of cells to sample per block of estimated imputation matrix}

\item{k}{number of nearest neighbors to use for Magic}

\item{epsilon}{a value for the standard deviation of the kernel for Magic}

\item{...}{additional params}
}
\description{
This function QQQ
}
