% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleScore.R
\name{addModuleScore}
\alias{addModuleScore}
\title{Add Module Scores to an ArchRProject}
\usage{
addModuleScore(
  ArchRProj = NULL,
  useMatrix = NULL,
  name = "Module",
  features = NULL,
  nBin = 25,
  nBgd = 100,
  seed = 1,
  threads = getArchRThreads(),
  logFile = createLogFile("addModuleScore")
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{seed}{A number to be used as the seed for random number generation. It is recommended to keep track of the seed used so that you can
reproduce results downstream.}

\item{threads}{The number of threads to be used for parallel computing.}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
This function computes imputations weights that describe each cell as a linear combination of many cells based on a MAGIC diffusion matrix.
}
\details{
RRR
}
