% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trajectory.R
\name{addMonocleTrajectory}
\alias{addMonocleTrajectory}
\title{Add a Monocle Trajectory to an ArchR Project #NEW}
\usage{
addMonocleTrajectory(
  ArchRProj = NULL,
  name = "Trajectory",
  useGroups = NULL,
  groupBy = "Clusters",
  monocleCDS = NULL,
  force = FALSE
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{name}{A string indicating the name of the fitted trajectory to be added in \code{cellColData}.}

\item{useGroups}{The cell groups to be used for creating trajectory analysis.}

\item{groupBy}{A string indicating the column name from \code{cellColData} that contains the cell group definitions used in
\code{useGroups} to constrain trajectory analysis.}

\item{monocleCDS}{A monocle CDS object created from \code{getMonocleTrajectories}.}

\item{force}{A boolean value indicating whether to force the trajactory indicated by \code{name} to be overwritten if it already exists in the given \code{ArchRProject}.}
}
\description{
This function will add a trajectory from a monocle CDS created from \code{getMonocleTrajectories} to an
ArchRProject.
}
