% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{addMotifAnnotations}
\alias{addMotifAnnotations}
\title{Add Motif Annotations to ArchRProject}
\usage{
addMotifAnnotations(ArchRProj = NULL, motifSet = "JASPAR2018",
  name = "Motif", species = NULL, collection = "CORE",
  cutOff = 5e-05, w = 7, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{motifSet}{motifSet JASPAR : JASPAR2016, JASPAR2018; chromVARmotifs : human, mouse, encode, homer}

\item{name}{of annotations to store as in ArchRProject}

\item{species}{species relevant to dataset (default will guess based on getGenome)}

\item{collection}{JASPAR collection (default = CORE)}

\item{cutOff}{pvalue cutoff for motif search (see motimatchr)}

\item{w}{width to consider for motif (see motimatchr)}

\item{...}{additional args}
}
\description{
This function adds motif postions and matches to an ArchRProject
}
