% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakAnnotation.R
\name{addPeakAnnotations}
\alias{addPeakAnnotations}
\title{Add peakAnnotations to an ArchRProject}
\usage{
addPeakAnnotations(ArchRProj = NULL, regions = NULL, name = "Region",
  force = FALSE, ...)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{regions}{The name of peakAnnotations to be stored as in \code{ArchRProject}}

\item{name}{The name of peakAnnotations to be stored as in \code{ArchRProject}}

\item{force}{force creation if peakAnnotation with same name already exists}

\item{...}{additional args}
}
\description{
This function adds information about which peaks contain motifs to a given ArchRProject. For each peak, a binary value is stored indicating whether each motif is observed within the peak region.
}
