% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixFeatures.R
\name{addPeakMatrix}
\alias{addPeakMatrix}
\title{Add PeakMatrix to Arrows in ArchRProject}
\usage{
addPeakMatrix(ArchRProj, ceiling = 4, binarize = FALSE,
  parallelParam = NULL, threads = 1, force = FALSE, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{ceiling}{ceiling for the number of counts per feature}

\item{binarize}{binarize matrix}

\item{parallelParam}{parallel parameters for batch style execution}

\item{threads}{number of threads}

\item{force}{force overwriting previous TileMatrix in ArrowFile}
}
\description{
This function for each sample will independently compute counts for each peak
per cell in the Arrow File
}
