% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{addPeakSet}
\alias{addPeakSet}
\title{Add PeakSet to ArchRProject}
\usage{
addPeakSet(ArchRProj, peakSet, force = FALSE, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{peakSet}{peakSet as a GRanges}

\item{force}{force overriding peakSet in ArchRProject}

\item{...}{additional args}
}
\description{
This function adds a peakSet to an ArchRProject
}
