% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{addProjectSummary}
\alias{addProjectSummary}
\title{Add projectSummary tp ArchRProject}
\usage{
addProjectSummary(ArchRProj, name, summary, ...)
}
\arguments{
\item{ArchRProj}{An `ArchRProject` object.}

\item{name}{QQQ The name of the summary information to add to the `ArchRProject` object.}

\item{summary}{QQQ A vector to add as summary information to the `ArchRProject` object.}

\item{...}{additional args}
}
\description{
This function adds info to the projectSummary from an ArchRProject
}
