% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{addSampleColData}
\alias{addSampleColData}
\title{Add information to sampleColData in ArchRProject}
\usage{
addSampleColData(ArchRProj, data = NULL, name = NULL,
  samples = rownames(sampleColData(ArchRProj)), force = FALSE)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{data}{data to add to sampleColData}

\item{name}{new column name in sampleColData if already exists set force = TRUE to override}

\item{force}{if name already exists in sampleColData set force = TRUE to override}

\item{cells}{names of samples corresponding to data}

\item{...}{additional args}
}
\description{
This function adds new data to sampleColData in ArchRProject
}
