% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesUtils.R
\name{addSeqLengths}
\alias{addSeqLengths}
\title{Add Seqlengths to genomic ranges}
\usage{
addSeqLengths(gr, genome)
}
\arguments{
\item{gr}{see validGRanges}

\item{genome}{see validBSgenome}
}
\description{
Add Seqlengths to genomic ranges
}
