% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesUtils.R
\name{addSeqLengthsGR}
\alias{addSeqLengthsGR}
\title{Adds seqlength information to the seqnames of a Genomic Ranges object}
\usage{
addSeqLengthsGR(gr = NULL, genome = NULL)
}
\arguments{
\item{gr}{A \code{GRanges} object.}

\item{genome}{The name of a valid genome (for example "hg38", "hg19", or "mm10"). See \code{validBSgenome()}.}
}
\description{
This function adds seqlength information for each of the seqnames in the provided Genomic Ranges object.
}
