% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trajectory.R
\name{addSlingShotTrajectories}
\alias{addSlingShotTrajectories}
\title{Add a Slingshot Trajectories to an ArchR Project #NEW}
\usage{
addSlingShotTrajectories(
  ArchRProj = NULL,
  name = "SlingShot",
  useGroups = NULL,
  principalGroup = NULL,
  groupBy = NULL,
  embedding = NULL,
  reducedDims = NULL,
  force = FALSE,
  seed = 1
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{name}{A string indicating the name of the fitted trajectory to be added in \code{cellColData}.}

\item{useGroups}{A character vector that is used to select a subset of groups by name from the designated \code{groupBy} column
in \code{cellColData}. This limits the groups used to identify trajectories.}

\item{principalGroup}{The principal group which represents the group that will be the starting point for all trajectories.}

\item{groupBy}{A string indicating the column name from \code{cellColData} that contains the cell group definitions used in
\code{useGroups} to constrain trajectory analysis.}

\item{embedding}{A string indicating the name of the \code{embedding} object from the \code{ArchRProject} that should be used for trajectory analysis.}

\item{reducedDims}{A string indicating the name of the \code{reducedDims} object from the \code{ArchRProject} that should be used for trajectory analysis. \code{embedding} must equal NULL to use.}

\item{force}{A boolean value indicating whether to force the trajactory indicated by \code{name} to be overwritten if it already exists in the given \code{ArchRProject}.}

\item{seed}{A number to be used as the seed for random number generation for trajectory creation.}
}
\description{
This function will fit a supervised trajectory in a lower dimensional space that
can then be used for downstream analyses.
}
\examples{

# Get Test ArchR Project
proj <- getTestProject()

#Add SlingShot Trajectory
proj <- addSlingShotTrajectories(
  ArchRProj = proj,
  name = "Trajectory_SlingShot", 
  useGroups = c("C1", "C2", "C3"), 
  principalGroup = "C1", 
  groupBy = "Clusters", 
  embedding = "UMAP"
)

}
