% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixTiles.R
\name{addTileMatrix}
\alias{addTileMatrix}
\title{Add TileMatrix to Arrows/ArchRProject}
\usage{
addTileMatrix(input, chromSizes = getChromSizes(input), tileSize = 500,
  binarize = TRUE, excludeChr = c("chrM", "chrY"), threads = 1,
  parallelParam = NULL, force = FALSE, ...)
}
\arguments{
\item{input}{ArchRProject or ArrowFiles}

\item{chromSizes}{chromomosome sizes used for identifying number of tiles to count}

\item{tileSize}{size for each tile to break up each chromosome}

\item{binarize}{save as a Sparse.Binary.Matrix or Sparse.Integer.Matrix}

\item{excludeChr}{exclude chromosomes from this analysis}

\item{threads}{number of threads}

\item{parallelParam}{parallel parameters for batch style execution}

\item{force}{force overwriting previous TileMatrix in ArrowFile}
}
\description{
This function for each sample will independently compute counts for each tile
per cell in the Arrow File
}
