% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trajectory.R
\name{addTrajectory}
\alias{addTrajectory}
\title{Add Supervised Trajectory to an ArchR Project}
\usage{
addTrajectory(ArchRProj, name = "Trajectory", trajectory = NULL,
  groupBy = "Clusters", reducedDims = "IterativeLSI",
  preFilter = 0.1, postFilter = 0.1, dof = 250, spar = 1,
  force = FALSE, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{name}{name of fitted trajectory to be added in cellColData}

\item{trajectory}{trajectory of groups to constrain supervised fitting (in order)}

\item{groupBy}{initial group column in cellColData to constrain supervised fit}

\item{reducedDims}{name of reduced dimensions used for distance computation}

\item{preFilter}{pre filtering quantile for supervised trajectory fit}

\item{postFilter}{post filtering quantile for supervised trajectory fit}

\item{dof}{degrees of freedom}

\item{spar}{sparsity}

\item{force}{force addition into ArchRProject if the column name exists}

\item{...}{additional args}

\item{name}{name of column in cellColData or Feature in Array in Arrows}
}
\description{
This function will fit a supervised trajectory in a lower dimensional space that 
can then be used for downstream analyses.
}
