% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Embedding.R
\name{addUMAP}
\alias{addUMAP}
\title{Add a UMAP embedding of a reduced dimensions object to an ArchRProject JJJ}
\usage{
addUMAP(ArchRProj = NULL, reducedDims = "IterativeLSI",
  name = "UMAP", nNeighbors = 40, minDist = 0.4, metric = "cosine",
  dimsToUse = NULL, scaleDims = NULL, corCutOff = 0.75,
  saveModel = FALSE, verbose = TRUE, seed = 1, force = FALSE,
  threads = max(floor(getArchRThreads()/2), 1), ...)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{reducedDims}{The name of the \code{reducedDims} object (i.e. "IterativeLSI") to use from the designated \code{ArchRProject}.}

\item{name}{The name for the UMAP embedding to be stored as in the given \code{ArchRProject} object.}

\item{nNeighbors}{An integer describing the number of nearest neighbors to compute a \code{umap}. This argument is passed to \code{n_neighbors} in \code{uwot::umap}.}

\item{minDist}{A numeric describing how tightly the \code{umap} is allowed to pack points together. This argument is passed to \code{min_dist} in \code{uwot::umap}. For more info on this see https://jlmelville.github.io/uwot/abparams.html.}

\item{metric}{A numeric describing how distance is computed in the \code{reducedDims} to compute a \code{umap}. This argument is passed to \code{metric} in \code{uwot::umap}.}

\item{dimsToUse}{A vector containing the dimensions from the \code{reducedDims} object to use in computing the embedding.}

\item{scaleDims}{A boolean describing whether to z-score the reduced dimensions for each cell. This is useful for minimizing the contribution of strong biases
(dominating early PCs) and lowly abundant populations. However, this may lead to stronger sample-specific biases since it is over-weighting latent PCs.
If \code{NULL} this will scale the dimensions depending on if this were set true when the \code{reducedDims} were created by the dimensionality reduction method.
This idea was introduced by Timothy Stuart.}

\item{corCutOff}{A numeric cutoff for the correlation of each dimension to the sequencing depth. If the dimension has a correlation to sequencing depth that is greater than the \code{corCutOff}, it will be excluded from analysis.}

\item{saveModel}{A boolean value indicating whether to save the UMAP model for downstream usage such as projection of data into the UMAP embedding.}

\item{verbose}{A boolean value that indicates whether printing UMAP output.}

\item{seed}{A number to be used as the seed for random number generation. It is recommended to keep track of the seed used so that you can reproduce results downstream.}

\item{force}{A boolean value that indicates whether to overwrite the relevant data in the \code{ArchRProject} object if the embedding named by \code{name} already exists.}

\item{threads}{The number of threads to be used for parallel computing.}

\item{...}{Additional Params to add to \code{uwot::umap}}
}
\description{
This function will compute a UMAP embedding and add it to an ArchRProject.
}
