% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZZ-Trajectory.R
\name{alignCellsToTrajectory}
\alias{alignCellsToTrajectory}
\title{Get Optimal Plotting Windows}
\usage{
alignCellsToTrajectory(mat, groups, trajectory, filterInitial = 0.1,
  finterFinal = 0.25, dof = 250, spar = 1)
}
\arguments{
\item{rle}{rle}
}
\description{
Get Optimal Plotting Windows
}
