% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part3_functions.R
\name{arch_to_adjacency}
\alias{arch_to_adjacency}
\title{Convert Architecture(s) back to Adjacency Matrices}
\usage{
arch_to_adjacency(arch, .id = NULL)
}
\arguments{
\item{arch}{data.frame of architecture(s) form}

\item{.id}{name of architecture unique id column. (Typically is \code{"id"}.) Default is NULL.}
}
\description{
Intended to be used on data.frames generated by \code{enumerate_adjacency()}.
}
\examples{
a = tribble(
 ~id, ~d1_a1_a2,  ~d1_a2_a1,
 1, 1,          0,
 2, 1,          1
)
arch_to_adjacency(arch = a, .id = "id")
}
\author{
Tim Fraser
}
