% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archeR.R
\docType{package}
\name{archeR}
\alias{archeR}
\alias{archeR-package}
\title{archeR: A package for discovering different architectures of sequence elements.}
\description{
Given a set of DNA sequences, the \code{archeR} package enables discovery of
clusters of sequence with characteristic sequence architectures.
}
\details{
The archeR package provides three categories of important functions:
related to data preparation, performing non-negative matrix factorization,
performing clustering, and visualization-related functions.
}
\section{Data preparation functions}{

The foo functions ...
}

\section{Performing NMF}{
NA
}

\section{Performing clustering}{
NA
}

\section{Visualizations}{
NA
}

