% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archR.R
\docType{package}
\name{archR}
\alias{archR}
\alias{archR-package}
\title{archR: A package for discovering different architectures of sequence elements.}
\description{
Given a set of DNA sequences, the \code{archR} package enables discovery of
clusters of sequence with characteristic sequence architectures.
}
\details{
The archR package provides three categories of important functions:
related to data preparation, performing non-negative matrix factorization,
performing clustering, and visualization-related functions.
}
\section{Data preparation functions}{

The foo functions ...
}

\section{Performing NMF}{
NA
}

\section{Performing clustering}{
NA
}

\section{Visualizations}{
NA
}

