% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archR_slurm_call.R
\name{archR_get_job_status}
\alias{archR_get_job_status}
\title{Date: 2020-06-27
Author: Sarvesh Nikumbh}
\usage{
archR_get_job_status(slr_job)
}
\arguments{
\item{slr_job}{A \code{slurm_job} object.}
}
\value{
A list with three elements: \code{completed} is a logical value
  indicating if all portions of the job have completed or stopped, \code{queue}
  contains the information on job elements still in queue, and
  \code{log} contains the console/error logs.
}
\description{
This is an adapted version of the slurm_call function from rslurm package.
Changes implemented:
1. Access custom directory created at custom location specified by the user
}
\details{
Get the status of a Slurm job

This function returns the completion status of a Slurm job, its queue status
if any and log outputs.

The \code{queue} element of the output is a data frame matching the output
of the Slurm \code{squeue} command for that job; it will only indicate portions
of job that are running or in queue. The \code{log} element is a
vector of the contents of console/error output files for each node where the
job is running.
}
