% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorPalettes.R
\docType{data}
\name{ArchR_palettes}
\alias{ArchR_palettes}
\title{List of palettes to be used in plots}
\format{An object of class \code{list} of length 25.}
\usage{
ArchR_palettes
}
\description{
List of palettes to be used in plots
}
\keyword{datasets}
