% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRHeatmap.R
\name{ArchRHeatmap}
\alias{ArchRHeatmap}
\title{Plot Nice Lookng Heatmap Using Complex Heatmap}
\usage{
ArchRHeatmap(
  mat = NULL,
  scale = FALSE,
  limits = c(min(mat), max(mat)),
  colData = NULL,
  color = paletteContinuous(set = "solarExtra", n = 100),
  clusterCols = TRUE,
  clusterRows = FALSE,
  labelCols = FALSE,
  labelRows = FALSE,
  colorMap = NULL,
  useRaster = TRUE,
  rasterQuality = 5,
  split = NULL,
  fontSizeRows = 10,
  fontSizeCols = 10,
  fontSizeLabels = 8,
  colAnnoPerRow = 4,
  showRowDendrogram = FALSE,
  showColDendrogram = FALSE,
  customRowLabel = NULL,
  customRowLabelIDs = NULL,
  customColLabel = NULL,
  customColLabelIDs = NULL,
  customLabelWidth = 0.75,
  rasterDevice = "png",
  padding = 45,
  borderColor = NA,
  draw = TRUE,
  name = "Heatmap"
)
}
\arguments{
\item{mat}{A matrix to plot heatmap from.}

\item{scale}{A boolean whether to convert to row Z-scores}

\item{limits}{A vector of two values describing min and mix limits to plot}

\item{colData}{A DataFrame matching the columns of the input matrix to overlay on the columns of the heatmap}

\item{color}{A palette to use for heatmap continuous color scheme. See \code{paletteContinuous}.}

\item{clusterCols}{A boolean describing whether to cluster columns for heatmap.}

\item{clusterRows}{A boolean describing whether to cluster rows for heatmap.}

\item{colorMap}{A list of color mappings matching the column names in colData.}

\item{useRaster}{A boolean whether to use rastering when plotting heatmap.}

\item{rasterQuality}{Raster resolution of raster. Default is set to 5 higher numbers increase resolution.}

\item{split}{A vector of groupings that will split the rows of heatmap (must be equal to nrow(mat)).}

\item{fontSizeRows}{A numeric value representing the font size for rownames.}

\item{fontSizeCols}{A numeric value representing the font size for colnames.}

\item{fontSizeLabels}{A numeric value representing the font size for labels.}

\item{colAnnoPerRow}{An integer value describing the number of column annotations per row in the legend.}

\item{showRowDendrogram}{A boolean whether to show row dendrogram in heatmap.}

\item{showColDendrogram}{A boolean whether to show column dendrogram in heatmap.}

\item{customRowLabel}{A vector of indices to custom label from rows.}

\item{customRowLabelIDs}{A vector of custom labels to overlay. Should match length of \code{customRowLabel}.}

\item{customColLabel}{A vector of indices to custom label from columns.}

\item{customColLabelIDs}{A vector of custom labels to overlay. Should match length of \code{customColLabel}.}

\item{customLabelWidth}{A numeric describing the width of the column labels.}

\item{rasterDevice}{Which device to use for rastering see \code{ComplexHeatmap}.}

\item{padding}{A numeric (in cm) to pad the heatmap ie adding white space so that the final plot doesnt cutoff.}

\item{borderColor}{A character representing the border color for each cell in the heatmap.}

\item{draw}{A boolean whether to draw the heatmap immediately. If FALSE this will return a ComplexHeatmap object.}

\item{name}{A character that will appear above color bar legend in heatmap.}
}
\description{
Plot Nice Lookng Heatmap Using Complex Heatmap
}
