% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archR_auxiliary_functionsII.R
\name{archRSetConfig}
\alias{archRSetConfig}
\title{Set archR run configuration}
\usage{
archRSetConfig(innerChunkSize = 500, kMin = 2, kMax = 8,
  cvFolds = 5, parallelize = FALSE, nCoresUse = NA,
  nIterationsUse = 500, seedVal = 10208090, alphaBase = 0,
  alphaPow = 1, minSeqs = 25, modSelLogFile = "log.txt",
  flags = list(debugFlag = FALSE, timeFlag = FALSE, verboseFlag = TRUE,
  plotVerboseFlag = FALSE))
}
\arguments{
\item{innerChunkSize}{Numeric. Specify the size of the inner chunks of
sequences.}

\item{kMin}{Numeric. Specify the minimum of the range of values to be tested
for number of NMF basis vectors.}

\item{kMax}{Numeric. Specify the maximum of the range of values to be tested
for number of NMF basis vectors.}

\item{cvFolds}{Numeric. Specify the number of cross-validation folds used for
model selection.}

\item{parallelize}{Logical. Specify whether to parallelize the procedure.}

\item{nCoresUse}{The number of cores to be used when `parallelize` is set to
TRUE.}

\item{nIterationsUse}{Specify the number of bootstrapped iterations to be
performed with NMF.}

\item{seedVal}{Specify the seed value for reproducibility.}

\item{alphaBase, alphaPow}{Specify the base value and the power for computing
'alpha' in performing model selection for NMF. alpha = alphaBase^alphaPow.
Alpha specifies the regularization for NMF. Default: 0 and 1 respectively.}

\item{minSeqs}{Specify the minimum number of sequences, such that any
cluster/chunk of size less than or equal to it will not further
processed/clustered.}

\item{modSelLogFile}{Specify a name for the file where model selection logs
will be wrtten.}

\item{flags}{List with four Logical elements as detailed.
\describe{
  \item{debugFlag}{Whether debug information for the run is printed}
  \item{verboseFlag}{Whether verbose information for the run is printed}
  \item{plotVerboseFlag}{Whether verbose plotting is performed for the run}
  \item{timeFlag}{Whether timing information is printed for the run}
}}
}
\value{
a list with all params for archR set
}
\description{
This function sets the configuration for `archR`.
}
