% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part5_functions.R
\name{as_bit}
\alias{as_bit}
\title{Convert Integers to Bits as vectors.}
\usage{
as_bit(x, n, simplify = TRUE)
}
\arguments{
\item{x}{vector of integer(s) to be transformed}

\item{n}{length of bitstring to represent them as.}

\item{simplify}{(logical) should the bitstrings, if multiple, be concatenated into a matrix, or returned as a list?}
}
\description{
Vectorized version of \code{decimal2binary()} from \code{GA} package.
}
\details{
PART 5 FUNCTIONS
This script contains a series of functions that support the genetic algorithm for optimization of enumerated architectures.
Still under development. You are welcome to suggest new functions.

Some of these are internal functions, while others are functions intended to be used by actual R users.
If you encounter an issue with any of these functions, please tag it in a Github Issue.
}
\examples{
as_bit(x = c(0,2,3), n = c(1,2,3), simplify = FALSE)
}
\author{
Tim Fraser
}
