% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{$.ArchRProject}
\alias{$.ArchRProject}
\title{Accessing cellColData directly from dollar.sign accessor}
\usage{
\method{$}{ArchRProject}(x, i, ...)
}
\description{
This function will allow direct access to cellColData with a \code{$} accessor.
}
