% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{$<-.ArchRProject}
\alias{$<-.ArchRProject}
\title{Add directly to cellColData directly from dollar.sign accessor}
\usage{
\method{$}{ArchRProject}(x, i) <- value
}
\description{
This function will allow adding directly to cellColData with a \code{$} accessor.
}
