% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_clusters.R
\name{choose_clusters}
\alias{choose_clusters}
\title{Find the appropriate numebr of clusters for the given data.}
\usage{
choose_clusters(givenMat, distMethod = "euclidean",
  clustMethod = "kmeans", nCluster_vals_test = seq(3, 5))
}
\arguments{
\item{givenMat}{A data matrix representing sequences along columns and their
features aling the rows.}

\item{distMethod}{The distance method to be used. Default is \code{euclidean}.
Takes all values that can be passed to the \code{dist} method from the
\code{stats} package.}

\item{clustMethod}{The method used for clustering the data. Default is
\code{kmeans}. The other value supported currently in \code{hclust}.}

\item{nCluster_vals_test}{The different number of clusters to be tested.
Default is \code{\(3,\ldots,5\)}.}
}
\value{
The number of clusters most suitable for the given data.
}
\description{
Tests for different clusterings (different number of clusters)
of the given data. The average silhoutte value is used for choosing the best
suitable number of clusters.
}
\examples{


}
