% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archR_auxiliary_functionsI.R
\name{collate_clusters}
\alias{collate_clusters}
\title{Collate sequence IDs from an existing clustering according to a new,
given clustering}
\usage{
collate_clusters(clustering, orig_clust_assign)
}
\arguments{
\item{clustering}{A list giving clustering of factors, i.e., the indices of
factors in each cluster}

\item{orig_clust_assign}{A list of sequence IDs in existing clusters.}
}
\value{
A list with sequence IDs collated by the specified clustering
}
\description{
Collate sequence IDs from an existing clustering according to a new,
given clustering
}
