% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archR_auxiliary_functionsI.R
\name{collect_cluster_labels}
\alias{collect_cluster_labels}
\title{Get cluster labels for chosen level/iteration}
\usage{
collect_cluster_labels(given_seqsClustLabels, chooseLevel = 1)
}
\arguments{
\item{given_seqsClustLabels}{from archR result object}

\item{chooseLevel}{choose a level/iteration. This value is number of
iterations + 1}
}
\value{
A numeric vector of the same size as seqsClustLabels, with labels
only up to the chosen iteration
}
\description{
Given a seqsClustLabels, collect cluster labels for sequences at
 the chosen iteration/level.
}
