% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZZ-CoAccessibility.R
\name{computeCoAccessibility}
\alias{computeCoAccessibility}
\title{Extend Filter then Normalize Scores for Summits}
\usage{
computeCoAccessibility(ArchRProject, k = 100, knnIteration = 5000,
  overlapCutoff = 0.8, seed = 1, knnMethod = "nabor",
  maxDist = 250000, ...)
}
\description{
Extend Filter then Normalize Scores for Summits
}
