% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{computeKNN}
\alias{computeKNN}
\title{QQQ Should this be exported?}
\usage{
computeKNN(data, query = NULL, k = 50, method = NULL,
  includeSelf = FALSE, ...)
}
\description{
QQQ Should this be exported?
}
