% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentSemanticIndexing.R
\name{computeLSI}
\alias{computeLSI}
\title{Compute LSI}
\usage{
computeLSI(mat, nDimensions = 50, binarize = TRUE, seed = 1,
  verbose = TRUE, tstart = NULL, ...)
}
\arguments{
\item{mat}{sparseMatrix (dgcMatrix) for LSI}

\item{nDimensions}{number of LSI dimensions to compute}

\item{binarize}{binarize matrix prior to LSI}

\item{seed}{seed for analysis}

\item{verbose}{verbose}

\item{tstart}{time stamp to pass}

\item{...}{additional args}
}
\description{
This function will compute a LSI transform (TF-IDF followed by SVD)
}
