% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperUtils.R
\name{confusionMatrix}
\alias{confusionMatrix}
\title{Create a Confusion Matrix based on two value vectors}
\usage{
confusionMatrix(i = NULL, j = NULL)
}
\arguments{
\item{i}{A character/numeric value vector to see concordance with j.}

\item{j}{A character/numeric value vector to see concordance with i.}
}
\description{
This function creates a confusion matrix based on two value vectors.
}
\examples{

# Get Test ArchR Project
proj <- getTestProject()

# Overlap of Clusters and CellType
confusionMatrix(proj$Clusters, proj$CellType)

# Overlap of Cell Type and RNA Predict
confusionMatrix(proj$CellType, proj$predictedGroup_Un)

}
