% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part4_functions.R
\name{connectivity}
\alias{connectivity}
\title{Get total connectivity of decision i}
\usage{
connectivity(
  data,
  decision_i = "d3",
  decisions = c("d1", "d2", "d3"),
  metric = "m1"
)
}
\arguments{
\item{data}{data.frame}

\item{decision_i}{string: name of decision variable}

\item{decisions}{vector: names of decision variables under consideration}

\item{metric}{string: name of performance metric under evaluation}
}
\description{
Get total connectivity of decision i
}
