% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part4_functions.R
\name{connectivity_ij}
\alias{connectivity_ij}
\title{Get total connectivity of decision i given decision j}
\usage{
connectivity_ij(data, decision_i, decision_j = "d2", metric = "m1")
}
\arguments{
\item{data}{data.frame}

\item{decision_i}{string: name of decision variable i}

\item{decision_j}{string: name of decision variable j}

\item{metric}{string: name of performance metric under evaluation}
}
\description{
Get total connectivity of decision i given decision j
}
