% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesUtils.R
\name{constructGRanges}
\alias{constructGRanges}
\title{Construct GRanges seqnames start end accounting for ends before starts (adding strandedness)}
\usage{
constructGRanges(seqnames, start, end, ignore.strand = TRUE)
}
\arguments{
\item{seqnames}{seqnames of GRanges}

\item{start}{start of GRanges}

\item{end}{end of GRanges}

\item{ignore.strand}{ignore strandedness for overlaps}
}
\description{
Construct GRanges seqnames start end accounting for ends before starts (adding strandedness)
}
