% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrativeAnalysis.R
\name{correlateTrajectories}
\alias{correlateTrajectories}
\title{Correlate Trajectories}
\usage{
correlateTrajectories(
  seTrajectory1 = NULL,
  seTrajectory2 = NULL,
  corCutOff = 0.5,
  varCutOff1 = 0.8,
  varCutOff2 = 0.8,
  removeFromName1 = c("underscore", "dash"),
  removeFromName2 = c("underscore", "dash"),
  useRanges = FALSE,
  fix1 = "center",
  fix2 = "start",
  maxDist = 250000,
  log2Norm1 = TRUE,
  log2Norm2 = TRUE,
  force = FALSE,
  logFile = createLogFile("correlateTrajectories")
)
}
\arguments{
\item{corCutOff}{A numeric cutoff for the correlation of each dimension to the sequencing depth. If the dimension has a correlation to
sequencing depth that is greater than the \code{corCutOff}, it will be excluded from analysis.}

\item{removeFromName1}{A character vector describing how to filter names in matrix 1.
Options include "underscore", "dash", "numeric". The string portion prior to these will be kept.}

\item{removeFromName2}{A character vector describing how to filter names in matrix 2.
Options include "underscore", "dash", "numeric". The string portion prior to these will be kept.}

\item{log2Norm1}{A boolean describing whether to log2 normalize matrix 1.}

\item{log2Norm2}{A boolean describing whether to log2 normalize matrix 2.}

\item{ArchRProj}{An \code{ArchRProject} object.}

\item{useMatrix1}{A character describing the first matrix to use. See \code{getAvailableMatrices} for valid options.}

\item{useMatrix2}{A character describing the second matrix to use. See \code{getAvailableMatrices} for valid options.}

\item{useSeqnames1}{A character vector describing which seqnames to use in matrix 1.}

\item{useSeqnames2}{A character vector describing which seqnames to use in matrix 2.}

\item{reducedDims}{The name of the \code{reducedDims} object (i.e. "IterativeLSI") to use from the designated \code{ArchRProject}.}

\item{dimsToUse}{A vector containing the dimensions from the \code{reducedDims} object to use in computing the embedding.}

\item{scaleDims}{A boolean value that indicates whether to z-score the reduced dimensions for each cell. This is useful for minimizing
the contribution of strong biases (dominating early PCs) and lowly abundant populations. However, this may lead to stronger sample-specific
biases since it is over-weighting latent PCs. If set to \code{NULL} this will scale the dimensions based on the value of \code{scaleDims} when the
\code{reducedDims} were originally created during dimensionality reduction. This idea was introduced by Timothy Stuart.}

\item{k}{An \code{ArchRProject} object.}

\item{knnIteration}{An \code{ArchRProject} object.}

\item{overlapCutoff}{An \code{ArchRProject} object.}

\item{seed}{An \code{ArchRProject} object.}

\item{threads}{The number of threads to be used for parallel computing.}
}
\description{
This function will correlate 2 trajectory matrices from getTrajectory.
}
