% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part3_functions.R
\name{count_ds}
\alias{count_ds}
\title{Count N for Downselecting}
\usage{
count_ds(k, n)
}
\arguments{
\item{k}{number of items to choose}

\item{n}{total number of items available to choose from}
}
\description{
Function to count total number of ways you can select \code{k} items from \code{n} items.
}
\note{
count_ds(k = 2, n = 10)
}
\author{
Tim Fraser
}
