% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArrow.R
\name{createArrowFiles}
\alias{createArrowFiles}
\title{Create Arrow Files}
\usage{
createArrowFiles(inputFiles = NULL, sampleNames = NULL,
  outputNames = paste0("./", sampleNames), geneAnno = NULL,
  genomeAnno = NULL, filterFrags = 1000, filterTSS = 4,
  removeFilteredCells = TRUE, minFrags = 500,
  outDir = "QualityControl", nucLength = 147, TSSParams = list(),
  excludeChr = c("chrM", "chrY"), nChunk = 5, bcTag = "qname",
  gsubExpression = NULL, bamFlag = NULL, offsetPlus = 4,
  offsetMinus = -5, addTileMat = TRUE, TileMatParams = list(),
  addGeneScoreMat = TRUE, GeneScoreMatParams = list(), force = FALSE,
  threads = 1, parallelParam = NULL, verboseHeader = TRUE,
  verboseAll = FALSE, ...)
}
\arguments{
\item{inputFiles}{input files (tabixFile, bamFile or textFile)}

\item{sampleNames}{sample names corresponding to input files}

\item{outputNames}{output names prefix (ie PBMC -> PBMC.arrow)}

\item{geneAnno}{geneAnnotation input for TSS Scores etc.}

\item{genomeAnno}{genomeAnnotation input for ChromSizes Nucleotide Information etc.}

\item{filterFrags}{min fragments per cell to be filtered for analyses such as tileMat etc.}

\item{filterTSS}{min TSS Score per cell to be filtered for analyses such as tileMat etc.}

\item{removeFilteredCells}{remove fragments corresponding to cells pass filterFrags and filterTSS}

\item{minFrags}{min fragments per cell to be immediately filtered}

\item{outDir}{out directory for QC information from sample to be plotted / saved}

\item{nucLength}{nucleosome length for id'ing fragments as sub-, mono-, or multi-nucleosome spanning}

\item{TSSParams}{TSS parameters for computing TSS scores}

\item{excludeChr}{exclude these chromosomes from analysis downstream (does not apply to fragments)}

\item{nChunk}{number of chunks per chromosome when reading in input files}

\item{bcTag}{barcode tag location in bam file (see ScanBam in Rsamtools)}

\item{bamFlag}{list of bam flags for reading in fragments from input files (see ScanBam in Rsamtools)}

\item{offsetPlus}{Tn5 offset of "+" stranded insertion (see Buenrostro 2013)}

\item{offsetMinus}{Tn5 offset of "-" stranded insertion (see Buenrostro 2013)}

\item{addTileMat}{addTileMatrix to ArrowFiles}

\item{TileMatParams}{additional parameters to pass to addTileMatrix (see addTileMatrix)}

\item{addGeneScoreMat}{addGeneScoreMatrix to ArrowFiles}

\item{GeneScoreMatParams}{additional parameters to pass to addGeneScoreMatrix (see addGeneScoreMatrix)}

\item{force}{force creation of arrow files if already exist}

\item{threads}{number threads for parallel execution}

\item{parallelParam}{parallel parameters for batch style execution}

\item{...}{additional args}
}
\description{
This function will create an Arrow Files from input files
for downstream analysis
}
