% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{createGeneAnnnotation}
\alias{createGeneAnnnotation}
\title{Create Gene Annotation for ArchR}
\usage{
createGeneAnnnotation(genome = NULL, TxDb = NULL, OrgDb = NULL,
  genes = NULL, exons = NULL, TSS = NULL)
}
\arguments{
\item{genome}{A string that specifies the genome (ie hg38, hg19, mm10, mm9).}

\item{TxDb}{A transcript database from Bioconductor which contains information for gene/transcript coordinates.}

\item{OrgDb}{An organism database from Bioconductor which contains information for gene/transcript symbols from ids.}

\item{genes}{A GRanges of gene coordinates (start to end). Needs to have a symbols column matching the exons symbols column.}

\item{exons}{A GRanges of gene exon coordinates. Needs to have a symbols column matching the genes symbols column}

\item{TSS}{A GRanges of transcription start sites (stranded) for computing TSS enrichment scores downstream.}
}
\description{
This function will create a gene annotation that can be used for createArrowFiles, ArchRProject, etc.
}
