% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoggerUtils.R
\name{createLogFile}
\alias{createLogFile}
\title{Create a Log File for ArchR}
\usage{
createLogFile(name = NULL, logDir = "ArchRLogs",
  useLogs = getArchRLogging())
}
\arguments{
\item{name}{A character string to add a more descriptive name in logFile}

\item{logDir}{A path to a directory where logFiles should be written to}
}
\description{
This function will create a log file for ArchR functions. If ArchRLogging is not TRUE
this function will return NULL.
}
