% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyNMF_model_selection_functions.R
\name{cv_model_select_pyNMF}
\alias{cv_model_select_pyNMF}
\title{Perform cross-validation and model selection}
\usage{
cv_model_select_pyNMF(X, param_ranges, kFolds = 5, parallel = FALSE,
  nCores = NA, seed_val = 10208090, logfile = "outfile.txt",
  set_verbose = 0)
}
\arguments{
\item{X}{The given data matrix}

\item{param_ranges}{An object holding the range of values for parameters
\code{k}, \code{alphaBase}, and \code{alphaPow}.}

\item{kFolds}{The number of cross-validation folds.}

\item{parallel}{Set to \code{1} if you want to parallelize, else \code{0}.}

\item{nCores}{If \code{parallel} is set to \code{1}}

\item{seed_val}{The seed to be set.}

\item{logfile}{The log file name.}

\item{set_verbose}{Default to \code{0} which will not print any messages, or
can be set to \code{1} which will print messages. Value passed to the
\code{get_q2_val} function}
}
\value{
A tibble of grid_search_results
}
\description{
Perform cross-validation and model selection
}
\examples{


}
