% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperUtils.R
\name{.ArchRLogo}
\alias{.ArchRLogo}
\title{This function returns ascii archr LOGO or arrow etc.}
\usage{
.ArchRLogo(ascii = "Logo")
}
\arguments{
\item{ascii}{logo, arrow, target}
}
\description{
This function returns ascii archr LOGO or arrow etc.
}
