% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperUtils.R
\name{.checkPath}
\alias{.checkPath}
\title{Check path for utility}
\usage{
.checkPath(u = NULL, path = NULL, error = TRUE)
}
\arguments{
\item{u}{utility that you want to check is in path}

\item{path}{check on top of path a custom path}

\item{error}{cause error if not in path}
}
\description{
Check path for utility
}
