% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{.DollarNames.ArchRProject}
\alias{.DollarNames.ArchRProject}
\title{Accessing cellColData directly from dollar.sign accessor}
\usage{
.DollarNames.ArchRProject(x, pattern = "")
}
\description{
This function will allow direct access to cellColData with a \code{$} accessor.
}
