% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperUtils.R
\name{.fileExtension}
\alias{.fileExtension}
\title{Get File Extension}
\usage{
.fileExtension(x)
}
\arguments{
\item{x}{character string refering to a file you want to get the extension from}
}
\description{
Get File Extension
}
