% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archR_auxiliary_functionsI.R
\name{.get_clusters_from_hc2}
\alias{.get_clusters_from_hc2}
\title{Get clusters from hclust object (improved version of previous ad-hoc
approach -- this allows singletons)}
\usage{
.get_clusters_from_hc2(
  hcObj,
  linkage = "complete",
  bound_scale = 0.75,
  pIter = 4,
  distMat,
  verbose = FALSE
)
}
\arguments{
\item{hcObj}{The hierachical clustering, hclust object}

\item{linkage}{String. Specify the linkage used -- works best for ward.D.}

\item{bound_scale}{Specify a threshold percentage. This will be used to 
compute the extent of the distance allowed between any two elements so that
they can be combined. See more explanation in the function code}

\item{pIter}{Numebr of iterations of joining procedures to be performed (See
while loop in the function code).}

\item{distMat}{Distance matrix as returned by \code{\link[stats]{dist}}.}

\item{verbose}{Specify TRUE or FALSE to control verbosity of output. Setting 
TRUE will print intermediate output enabling debugging.}
}
\description{
Get clusters from hclust object (improved version of previous ad-hoc
approach -- this allows singletons)
}
