% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archR_auxiliary_functionsI.R
\name{.get_clusters_from_hc_using_cutree}
\alias{.get_clusters_from_hc_using_cutree}
\title{Get clusters out of hierarchical clustering object using cutree. 
Used internally by archR.}
\usage{
.get_clusters_from_hc_using_cutree(
  hcObj,
  distMat,
  hStep = 0.05,
  parentChunks = NULL,
  keepSiblingsUncollated = FALSE,
  enableSwitchSilToCH = TRUE,
  minClusters = 2,
  distThreshold = 3,
  verbose = FALSE
)
}
\arguments{
\item{hcObj}{The hierarchical clustering object as returned by 
\code{\link[stats]{hclust}}.}

\item{distMat}{The distance matrix that was used by hclust, a 
\code{\link[stats]{dist}} object.}

\item{hStep}{Numeric. The step size used to increment height values for 
cutree. Default value is 0.05.}

\item{parentChunks}{List. Specify the factor numbers in the previous 
iteration of archR that factors in the current iteration resulted from. 
Default value is NULL.}

\item{keepSiblingsUncollated}{Logical. Specify TRUE if all clusters from a 
single parent chunk should not be collated, even when recommended so based on 
silhouette value computation. Default value FALSE. i.e. if they get collated, it is
 left as collated.}

\item{enableSwitchSilToCH}{Logical. When and if the clustering result has 
singletons, using the average silhouette value may not be reliable to decide 
on the optimum number of clusters. In this case, if this argument is TRUE, 
the Calinski-Harabasz index is used instead. Setting it to FALSE, the
switch is disabled, i.e., average silhouette value is used. Default is TRUE.}

\item{minClusters}{Integer. Specify the minimum number of clusters to be 
fetched from the HAC solution.}

\item{distThreshold}{Numeric. Specify a threshold of units of distance for 
considering any two elements as close enough to be merged in the same cluster.
The default value is specified for Euclidean distance.}

\item{verbose}{Logical. Specify TRUE for verbose output.}
}
\description{
Clusters from a hierarchical clustering object are obtained 
by using cutree at different heights of the tree. The optimum number of 
clusters are decided based on the average silhouette value of the clustering.
}
