% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_from_FASTA.R
\name{.one_hot_decode}
\alias{.one_hot_decode}
\title{One-hot decode}
\usage{
.one_hot_decode(oneHotEncodedSeqV)
}
\arguments{
\item{oneHotEncodedSeqV}{A single one-hot encoded sequence vector.}
}
\value{
The one-hot decoded sequence of ACGTs.
}
\description{
One-hot decode a given one-hot encoded DNA sequence.
}
