% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_from_FASTA.R
\name{.one_hot_encode_dinuc}
\alias{.one_hot_encode_dinuc}
\title{One-hot encode dinucleotide profiles}
\usage{
.one_hot_encode_dinuc(givenSeq)
}
\arguments{
\item{givenSeq}{A single sequence.}
}
\value{
The one-hot encoded sequence.
}
\description{
One-hot encode the dinucleotide profile of a given DNA sequence.
}
