% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_from_FASTA.R
\name{.one_hot_encode_sinuc}
\alias{.one_hot_encode_sinuc}
\title{One-hot encode}
\usage{
.one_hot_encode_sinuc(givenSeq)
}
\arguments{
\item{givenSeq}{A single DNA sequence as character/string.}
}
\value{
The one-hot encoded sequence.
}
\description{
One-hot encode a given DNA sequence.
}
