% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_matrix_of_acgt_image.R
\name{.seqs_to_mat}
\alias{.seqs_to_mat}
\title{Convert DNA sequences from DNAStringSet object to a numeric matrix}
\usage{
.seqs_to_mat(
  seqs,
  position_labels,
  annClusters = NULL,
  sinuc_or_dinuc = "sinuc"
)
}
\arguments{
\item{annClusters}{For later, when clusters are to be annotated}

\item{sinuc_or_dinuc}{For later, maybe}
}
\description{
This function converts a DNAStringSet object into a numeric
matrix
}
