% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperUtils.R
\name{.tempfile}
\alias{.tempfile}
\title{Check path for utility}
\usage{
.tempfile(pattern = "tmp", tmpdir = "tmp", fileext = "",
  addDOC = TRUE)
}
\arguments{
\item{u}{utility that you want to check is in path}

\item{path}{check on top of path a custom path}

\item{error}{cause error if not in path}
}
\description{
Check path for utility
}
