% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part3_functions.R
\name{enum}
\alias{enum}
\title{PART 3 FUNCTIONS
This script contains a series of functions that allow for enumeration.
Some of these are internal functions, while others are functions intended to be used by actual R users.
If you encounter an issue with any of these functions, please tag it in a Github Issue.}
\usage{
enum(i = 1, d, c)
}
\arguments{
\item{i}{...}

\item{d}{data.frame of decision-alternative-ids}

\item{c}{data.frame of constraints}
}
\description{
PART 3 FUNCTIONS
This script contains a series of functions that allow for enumeration.
Some of these are internal functions, while others are functions intended to be used by actual R users.
If you encounter an issue with any of these functions, please tag it in a Github Issue.
}
\author{
Tim Fraser
}
