% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part3_functions.R
\name{enumerate}
\alias{enumerate}
\title{Enumerate Your Decision Tree}
\usage{
enumerate(d = NULL, c = NULL)
}
\arguments{
\item{d}{data.frame of decision-alternative items}

\item{c}{data.frame of constraints}
}
\description{
Enumerate Your Decision Tree
}
\author{
Tim Fraser
}
