% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part3_functions.R
\name{enumerate_assignment}
\alias{enumerate_assignment}
\title{Enumerate Assignment Matrix into Architectures}
\usage{
enumerate_assignment(n, m, k = 1, n_alts = 2, .did = 1, .id = FALSE)
}
\arguments{
\item{n}{(integer) total number of groups in part a of the assignment matrix}

\item{m}{(integer) total number of groups in part b of the assignment matrix}

\item{k}{(integer) total number of items (a-b pairs) that can be selected per assignment matrix. If n_alts = 2, k is be length 1. If n_alts > 2, k must be length n_alts+1. By default \code{k} is \code{1}.}

\item{n_alts}{(integer, optional) total number of values that can be assigned to each item (eg. 2= c(0,1)). Default is \code{2}. (Experimental; don't change unless you've thought deeply about what you're doing!)}

\item{.did}{unique decision id.}

\item{.id}{(logical) Include a unique ID for each row? Defaults to FALSE.}
}
\description{
If n_alts = 2, k can have 1 value.
k\link{1} corresponds to when alternative = \link{1}
If n_alts = 3, k can have 2 values.
k\link{1} corresponds to when alternative = \link{1}
k\link{2} corresponds to when alternative = \link{2}
}
\examples{
enumerate_assignment(n = 3, m = 2, k = 5, n_alts = 2)
enumerate_assignment(n = 3, m = 2, k = c(3,2), n_alts = 3)

}
\author{
Tim Fraser, PhD
}
