% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part3_functions.R
\name{enumerate_ds}
\alias{enumerate_ds}
\title{Enumerate Downselected Decision}
\usage{
enumerate_ds(n, k, .id = FALSE, .did = 1)
}
\arguments{
\item{n}{total number of items available to choose from}

\item{k}{number of items to choose}

\item{.id}{(logical, optional) add a unique ID column to the end?}

\item{.did}{(integer, optional) Decision id. For example, if .did = 1, then columns will be \code{d1_1}, \code{d1_2}, \code{d1_[n]}...}
}
\description{
Enumerate Downselected Decision
}
\author{
Tim Fraser, PhD
}
