% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part3_functions.R
\name{enumerate_permutation}
\alias{enumerate_permutation}
\title{Enumerate Permuations into Architectures}
\source{
https://stackoverflow.com/questions/11095992/generating-all-distinct-permutations-of-a-list-in-r
}
\usage{
enumerate_permutation(n, k = NULL, .did = 1, .id = FALSE)
}
\arguments{
\item{n}{total number of items in the set for permuting}

\item{k}{number of items to choose. Must be less than or equal to \code{n}.}

\item{.did}{unique decision id.}

\item{.id}{(logical) Include a unique ID for each row? Defaults to FALSE.}
}
\description{
Enumerate Permuations into Architectures
}
\examples{
enumerate_permutation(n = 2)
enumerate_permutation(n = 3, k = 2, .did = 5)
}
