% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part3_functions.R
\name{enumerate_sf}
\alias{enumerate_sf}
\title{Enumerate Standard Form}
\usage{
enumerate_sf(n = c(2, 3), forbidden = NULL, .did = 1, .id = FALSE)
}
\arguments{
\item{n}{a vector describing the number of alternatives for each decision, sequentially.}

\item{forbidden}{Any forbidden possibilities. Not recommended. Easier to constrain using \code{dplyr::filter()} after the fact.}

\item{.did}{(integer, optional) unique ID of first decision. Used to make column names for sequential decisions. Defaults to 1.}

\item{.id}{(logical, optional) Whether or not to add a unique ID for each row. Defaults to FALSE.}
}
\description{
Enumerate Standard Form
}
\examples{
enumerate_sf(n = c(2, 3))
enumerate_sf(n = c(2, 3), forbidden = c(NA,2))
enumerate_sf(n = c(2, 3), .did = 4)

}
\author{
Tim Fraser
}
