% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part4_functions.R
\name{evaluate_tidy}
\alias{evaluate_tidy}
\title{Evaluate Architecture Metrics in a Tidyverse-Friendly Way}
\usage{
evaluate_tidy(par, d, vars = NULL, c = NULL, invalid = 0)
}
\arguments{
\item{par}{architecture(s), either as a vector of integers or a tibble of architectures}

\item{d}{data.frame of decision-alternative items}

\item{vars}{vector of metric names. Must be columns in \code{d}.}

\item{c}{(Optional) data.frame of constraints.}

\item{invalid}{value that an invalid architecture receives.}
}
\description{
Let's write an evaluate function that will, for any enumerated tidy data.frame,
join in your metrics and compute them, then return a final dataframe
}
\author{
Tim Fraser
}
