% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupExport.R
\name{exportMonocle3}
\alias{exportMonocle3}
\title{Export to Monocle3}
\usage{
exportMonocle3(
  ArchRProj = NULL,
  useMatrix = NULL,
  threads = getArchRThreads(),
  verbose = TRUE,
  binarize = T,
  logFile = createLogFile("exportMonocle3")
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{useMatrix}{The name of the matrix in the ArrowFiles. See getAvailableMatrices to see options}

\item{threads}{An integer specifying the number of threads for parallel.}

\item{verbose}{A boolean specifying to print messages during computation.}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
This function will return a monocle3 cell_data_set for a assay in a ArchRProject.
}
