% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesUtils.R
\name{extendGR}
\alias{extendGR}
\title{Extend regions from a Genomic Ranges object}
\usage{
extendGR(gr = NULL, upstream = NULL, downstream = NULL)
}
\arguments{
\item{gr}{A \code{GRanges} object.}

\item{upstream}{The number of basepairs upstream (5') to extend each region in \code{gr} in a strand-aware fashion.}

\item{downstream}{The number of basepairs downstream (3') to extend each region in \code{gr} in a strand-aware fashion.}
}
\description{
This function extends each region in a Genomic Ranges object by a designated upstream and downstream extension in a strand-aware fashion
}
\examples{

# Dummy GR
gr <- GRanges(
  seqnames = "chr1",
  ranges = IRanges(
    start = c(1, 4, 11),
    end = c(10, 12, 20)
  ),
  score = c(1, 2, 3)
)

# Non Overlapping
extendGR(gr, 1, 2)

}
