% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesUtils.R
\name{extendGRanges}
\alias{extendGRanges}
\title{Merge Genomic Ranges}
\usage{
extendGRanges(x, upstream, downstream)
}
\arguments{
\item{query}{see validGRanges}

\item{subject}{see validGRanges}

\item{ignore.strand}{ignore strandedness for overlaps}
}
\description{
Merge Genomic Ranges
}
