% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZZ-FeatureEnrichment.R
\name{featureEnrichment}
\alias{featureEnrichment}
\title{Feature Matches Over Representation}
\usage{
featureEnrichment(featureMatches, compare, background)
}
\arguments{
\item{compare}{vector of compare peaks idx which will be used for hypergeometric}

\item{background}{vector of background peaks idx which will be used for hypergeometric}

\item{genes}{motif similarity matrix used for labeling family info default is null}
}
\description{
This function takes in genes matches combined with a vector of peaks in target and peaks in bdg and returns hypergeometric minus log 10 pvalues
}
