% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterCells.R
\name{filterCells}
\alias{filterCells}
\title{Filter cells in an ArchRProject}
\usage{
filterCells(ArchRProj = NULL, filterList = NULL, ...)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{filterList}{A list of filters based on \code{cellColData} to apply when filtering cells.
Format should be a named list where the name corresponds to a column name in \code{cellColData} and the value corresponds to the filter criteria.
If numeric, a lower threshold is expected, below which cells are filtered (a higher threshold is optional and can be included using a numeric vector).
If a chatacter vector, only the rows in \code{cellColData} with \code{rownames} corresponding to the supplied character values are kept.
If a list or simpleList, the user can additionally supply filters that are applied to all samples or a subset of specified samples.
For example, to apply a filter to all samples: list("TSSEnrichment" = c(4,25)).
Similarly, to aply a filter only to specific samples: list("TSSEnrichment" = list("Sample1" = c(4, 25), "Sample2" = c(5, 25))).}

\item{...}{additional params}
}
\description{
This function returns an ArchRProject object that has been filtered to remove cells that do not pass the filter critera suppplied in filterList.
}
