% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterCells.R
\name{filterDoublets}
\alias{filterDoublets}
\title{Filter Doublets From an ArchR Project}
\usage{
filterDoublets(ArchRProj, cutEnrich = 1, cutScore = -Inf,
  filterRatio = 1, ...)
}
\arguments{
\item{ArchRProj}{an ArchR Project object}

\item{cutEnrich}{minimum cutoff for doubletEnrichment which represents number of simulated doublets nearest a cell over the expected if uniform.}

\item{cutScore}{minimum cutoff for doubletScore which represents -log10 binomial adjusted p-value.}

\item{filterRatio}{filter ratio for max number of inferred doublets to remove based on the number of cells PF. If there are 10,000 cells the maximum would be filterRatio * 10,000^2 / (1000 * 100).}

\item{...}{additional params}
}
\description{
This function wil filter doublets from an ArchRProject after addDoubletScores has been ran
}
