% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterCells.R
\name{filterPlot}
\alias{filterPlot}
\title{Filter Plot for cells in an ArchRProject}
\usage{
filterPlot(ArchRProj, filterList, sampleNames = NULL, ...)
}
\arguments{
\item{ArchRProj}{an ArchR Project object}

\item{filterList}{list of filters for filtering cells from cellColData (up to 2 will be plotted)}

\item{sampleNames}{specific samples to include}

\item{...}{additional params}
}
\description{
This function plots a list of filters to see which cells would pass filter.
}
