% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReproduciblePeakSet.R
\name{findMacs}
\alias{findMacs}
\title{Find the installed location of the MACS executable}
\usage{
findMacs(peakMethod = "Macs2")
}
\arguments{
\item{peakMethod}{The peak calling method to be used. Options include "Macs2" and "Macs3". Default is \code{"Macs2"}.}
}
\description{
This function attempts to find the path to the MACS2 or MACS3 executable by searching the path and python's pip.
}
\examples{

# Get Macs path
findMacs()

}
