% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReproduciblePeakSet.R
\name{findMacs2}
\alias{findMacs2}
\title{Find the installed location of the MACS2 executable}
\usage{
findMacs2()
}
\description{
This function attempts to find the path to the MACS2 executable by serting the path and python's pip.
}
\examples{

# Get Macs2
findMacs2()

}
