% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyNMF_model_selection_functions.R
\name{generate_folds}
\alias{generate_folds}
\title{Generate cross-validation data splits}
\usage{
generate_folds(Xdims, kFolds, seed_val)
}
\arguments{
\item{Xdims}{Dimensions of matrix data matrix X.}

\item{kFolds}{Number of cross-validation folds.}

\item{seed_val}{The seed value.}
}
\value{
A list of two elements: rowIDs and columnIDs for different cross-
validation folds.
}
\description{
Generate cross-validation data splits
}
\examples{

}
