% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part1_functions.R
\name{get_adjacency}
\alias{get_adjacency}
\title{Convert an edgelist table to an adjacency matrix}
\usage{
get_adjacency(edges, matrix = TRUE)
}
\arguments{
\item{edges}{(data.frame) Requires a 'from' (character), 'to' (character), and 'value' column.}

\item{matrix}{(logical) TRUE/FALSE. By default, format = TRUE, so outputs a matrix; outputs a table if FALSE}
}
\description{
By default, nodes is sourced from edges. (THIS MEANS WE'VE CENSORED ALL UNINVOLVED NODES.)
}
\note{
We use \verb{@importFrom} tags to mention function from external packages we need for this function, written like packagename function1 function2.
}
\author{
Tim Fraser, PhD
}
