% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part1_functions.R
\name{get_ahp}
\alias{get_ahp}
\title{Get Analytical Hierarchy Process Table}
\usage{
get_ahp(
  alternatives = archr::pairs_alts,
  criteria = archr::pairs_criteria,
  format = "wide"
)
}
\arguments{
\item{alternatives}{data.frame of pairwise comparison of alternatives}

\item{criteria}{data.frame of pairwise comparison of criteria}

\item{format}{"wide" or "long". Describes output format of AHP table.}
}
\description{
Provide a table of alternatives and a table of criteria,
and this will return your AHP output table.
}
\author{
Tim Fraser, PhD
}
