% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyNMF_model_selection_functions.R
\name{get_best_alpha}
\alias{get_best_alpha}
\title{Get the best perfing value of \eqn{\alpha}.}
\usage{
get_best_alpha(for_alpha, for_k, min_or_max = min)
}
\arguments{
\item{for_alpha}{grid_search_result copy for \eqn{\alpha}}

\item{for_k}{grid_search_search_result copy for K.}

\item{min_or_max}{Specify whether min or max is to be used as the condition to choose one when multiple values satisfy the threshold.}
}
\value{
The best performing value of \eqn{\alpha}.
}
\description{
Get the best perfing value of \eqn{\alpha}.
}
\examples{

}
