% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyNMF_model_selection_functions.R
\name{get_best_K}
\alias{get_best_K}
\title{Get the best performing value of K (number of factors in NMF)}
\usage{
get_best_K(x)
}
\arguments{
\item{x}{A tibble, grid_search_results, as returned by
\code{cv_model_select_pyNMF}}
}
\value{
A number The best performing value of K.
}
\description{
Get the best performing value of K (number of factors in NMF)
}
