% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archR_auxiliary_functionsI.R
\name{get_clBasVec}
\alias{get_clBasVec}
\alias{get_clBasVec_k}
\alias{get_clBasVec_m}
\alias{get_seqClLab}
\title{Get functions for archR result object}
\usage{
get_clBasVec(res, iter)

get_clBasVec_k(res, iter)

get_clBasVec_m(res, iter)

get_seqClLab(res, iter)
}
\arguments{
\item{res}{archR result object.}

\item{iter}{Choose the iteration of archR result to get from.}
}
\description{
Get functions for archR result object
}
\section{Functions}{
\itemize{
\item \code{get_clBasVec_k}: Number of basis vectors (clusters) at the 
selected iteration.

\item \code{get_clBasVec_m}: The basis vectors matrix at the chosen 
iteration. Note that eatures along rows.

\item \code{get_seqClLab}: The cluster IDs for each sequence. Note that 
order of sequences here is as per the input.
}}

\seealso{
\code{\link{seqs_str}}
}
