% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clusters.R
\name{get_clusters}
\alias{get_clusters}
\title{Get a specified number of clusters from the given data.}
\usage{
get_clusters(givenMat, clustMethod, nCluster)
}
\arguments{
\item{givenMat}{A data matrix representing sequences along columns and their
features aling the rows.}

\item{clustMethod}{The method used for clustering the data. Default is
\code{kmeans}. The other value supported currently in \code{hclust}.}

\item{nCluster}{The numebr of clusters to be extracted.}
}
\value{
A list containing the clustering result \code{clust_sol}, and the
reordered indices \code{reordering_idx}.
}
\description{
Get the specified number of clusters from the given data using
the chosen clusering method.
}
