% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part1_functions.R
\name{get_cycles}
\alias{get_cycles}
\title{Get all cycles in a graph starting and ending with a given root}
\usage{
get_cycles(g = NULL, root = 1, cutoff = 50, vars = NULL)
}
\arguments{
\item{g}{igraph directed graph object}

\item{root}{name of root vertex}

\item{cutoff}{Max length of simple paths to compute. Helpful for really big graphs.}

\item{vars}{A vector of 1 or more edge weight names from \code{g}. If \code{vars = NULL}, then skips edgewise edge weight computing.}
}
\description{
Get all cycles in a graph starting and ending with a given root
}
\author{
Tim Fraser
}
