% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part1_functions.R
\name{get_edges}
\alias{get_edges}
\title{Get edgelist from adjacency matrix}
\usage{
get_edges(a = NULL)
}
\arguments{
\item{a}{Adjaceny Matrix of n x n values, of \code{matrix} class in R. Optionally with column and row names.}
}
\description{
Get edgelist from adjacency matrix
}
\author{
Tim Fraser
}
