% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part1_functions.R
\name{get_graph}
\alias{get_graph}
\title{Get \code{igraph} Graph object}
\usage{
get_graph(a = NULL, edges = NULL, nodes = NULL)
}
\arguments{
\item{a}{Adjacency Matrix}

\item{edges}{Edgelist Data.frame with columns \code{from}, \code{to}, and \code{value}}

\item{nodes}{Nodelist Data.frame with column \code{name}}
}
\description{
Get an \code{igraph} formatted graph object
from an adjacency matrix \code{a}, an \code{edges} edgelist,
or an \code{edges} edgelist and a \code{nodes} nodelist.
}
\author{
Tim Fraser
}
