% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part1_functions.R
\name{get_lambda}
\alias{get_lambda}
\title{Get Lambda Statistic for an AHP table}
\usage{
get_lambda(data)
}
\arguments{
\item{data}{data.frame of output from \code{get_priority()}. Can handle priority statistics from a data.frame of priorities by alternatives, by criteria, or for both.}
}
\description{
Get Lambda Statistic for an AHP table
}
\author{
Tim Fraser
}
