% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part1_functions.R
\name{get_nodes}
\alias{get_nodes}
\title{Get nodes from edgelist or adjacency matrix}
\usage{
get_nodes(edges = NULL, a = NULL)
}
\arguments{
\item{edges}{Optionally, an edgelist where the first column is \code{from} vertices and the second column is \code{to} vertices.}

\item{a}{Optionally, an adjacency matrix}
}
\description{
Get nodes from edgelist or adjacency matrix
}
\author{
Tim Fraser
}
