% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_from_FASTA.R
\name{get_one_hot_encoded_seqs}
\alias{get_one_hot_encoded_seqs}
\title{Get One-Hot Encoded Sequences}
\usage{
get_one_hot_encoded_seqs(givenFastaSeqs, sinuc_or_dinuc = "sinuc")
}
\arguments{
\item{givenFastaSeqs}{List of sequences.}

\item{sinuc_or_dinuc}{'sinuc' or 'dinuc'}
}
\value{
One-hot encoded sequences.
}
\description{
Get the one-hot encoding representation of the given sequences.
}
