% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyNMF_model_selection_functions.R
\name{get_q2_threshold_by_k}
\alias{get_q2_threshold_by_k}
\title{Get the threshold value for selection of \eqn{\alpha} by looking at
cross-validation performance for K.}
\usage{
get_q2_threshold_by_k(model_selectK)
}
\arguments{
\item{model_selectK}{Cross-validation performance over K values.}
}
\value{
The \eqn{Q^2} threshold value.
}
\description{
Get the threshold value for selection of \eqn{\alpha} by looking at
cross-validation performance for K.
}
\examples{

}
