% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyNMF_model_selection_functions.R
\name{get_q2_using_py}
\alias{get_q2_using_py}
\title{Get the reconstruction accuracy, \eqn{Q^2}, for a given choice of parameter
values}
\usage{
get_q2_using_py(x, seed_val, verbose = 0)
}
\arguments{
\item{x}{The parameter values. Expected columns "k_vals", "alpha" and "fold".}

\item{seed_val}{The seed to be used.}

\item{verbose}{Default to \code{0} which will not print any messages, or can
be set to \code{1} which will print messages.}
}
\value{
\eqn{Q^2} value
}
\description{
Get the reconstruction accuracy, \eqn{Q^2}, for a given choice of parameter
values
}
\examples{


}
