% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part1_functions.R
\name{get_scores}
\alias{get_scores}
\title{Get Stakeholder Importance Scores}
\usage{
get_scores(g, root = 1, var = "value", cutoff = 50, square = FALSE)
}
\arguments{
\item{g}{igraph directed graph object}

\item{root}{node id for root (project node)}

\item{var}{variable for which to compute importance metrics}

\item{cutoff}{max path length to assess, to be passed to \code{get_cycles()}}

\item{square}{(logical) should you square the final edge that connects back to the project? Defaults to FALSE (no).}
}
\description{
Function to generate stakeholder important scores with regard to any project root.
}
