% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represent_matrix_of_acgt.R,
%   R/represent_matrix_of_acgt_with_ggplot2.R
\name{get_tick_labels}
\alias{get_tick_labels}
\title{Getter function for labels of x- or y-ticks.}
\usage{
get_tick_labels(givenVec, use_asis = T, at_every)

get_tick_labels(givenVec, use_asis = T, at_every)
}
\arguments{
\item{givenVec}{The vector of values for the x- or y-ticks.}

\item{use_asis}{Use the vector \emph{as is}, i.e., all given values will
be shown as ticks.}

\item{at_every}{Show tickmarks only at positions specified.}
}
\value{
A label vector that can be directly used for setting the tickmarks.

A label vector that can be directly used for setting the tickmarks.
}
\description{
A convenience function to provide the right set of x-tick labels for the image matrix representing the sequences.

A convenience function to provide the right set of x-tick labels
 for the image matrix representing the sequences.
}
