% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getAnnotation}
\alias{getAnnotation}
\title{Get annotation from an ArchRProject}
\usage{
getAnnotation(ArchRProj, name = NULL, ...)
}
\arguments{
\item{ArchRProj}{An `ArchRProject` object.}

\item{name}{QQQ The name of the annotation object to retrieve from the designated `ArchRProject`. Options include QQQ.}

\item{...}{additional args}
}
\description{
This function gets an annotation from a given ArchRProject.
}
