% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalDefaults.R
\name{getArchRChrPrefix}
\alias{getArchRChrPrefix}
\title{Get a globally-applied requirement for chromosome prefix}
\usage{
getArchRChrPrefix()
}
\description{
This function will get the default requirement of chromosomes to have a "chr" prefix.
}
