% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoggerUtils.R
\name{getArchRDebugging}
\alias{getArchRDebugging}
\title{Get ArchR Debugging}
\usage{
getArchRDebugging()
}
\description{
This function will get ArchR Debugging which will save an RDS if an error is encountered.
}
