% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getArchRGenome}
\alias{getArchRGenome}
\title{Get a globally defined genome to all ArchR functions. JJJ}
\usage{
getArchRGenome(geneAnnotation = FALSE, genomeAnnotation = FALSE)
}
\arguments{
\item{geneAnnotation}{Export geneAnnotation set with addArchRGenome. The geneAnnotation (see \code{createGeneAnnotation()}) to associate with the ArrowFiles. This is used downstream to calculate TSS Enrichment Scores etc.}

\item{genomeAnnotation}{Export genomeAnnotation set with addArchRGenome. The genomeAnnotation (see \code{createGenomeAnnotation()}) to associate with the ArrowFiles. This is used downstream to collect chromosome sizes and nucleotide information etc.}
}
\description{
This function will get the genome across all ArchR functions. Additionally can return geneAnnotation and genomeAnnotation associated with ArchRGenome if desired.
}
