% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalDefaults.R
\name{getArchRH5Level}
\alias{getArchRH5Level}
\title{Get globally-applied compression level for h5 files}
\usage{
getArchRH5Level()
}
\description{
This function will get the default compression level to be used for h5 file execution across all ArchR functions.
}
\examples{

# Get ArchR H5 Compression level
getArchRH5Level()

}
