% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalDefaults.R
\name{getArchRLB}
\alias{getArchRLB}
\title{Get the globally-applied load balancing setup}
\usage{
getArchRLB()
}
\description{
This function will get the load balancing setup when \code{mclapply} (parallelised version of lapply)
is used in ArchR functions.
}
\examples{

# Get load balancing setup
getArchRLB()

}
