% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalDefaults.R
\name{getArchRPathFormat}
\alias{getArchRPathFormat}
\title{Get the globally-applied outputDirectory path format}
\usage{
getArchRPathFormat()
}
\description{
This function will get the path format setup of \code{outputDirectory} used in ArchR functions.
If \code{TRUE}, absolute path is used, and relative path if the function returns \code{FALSE}.
}
\examples{

# Get path format setup
getArchRPathFormat()

}
