% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalDefaults.R
\name{getArchRThreads}
\alias{getArchRThreads}
\title{Get globally-applied number of threads to use for parallel computing.}
\usage{
getArchRThreads()
}
\description{
This function will get the number of threads to be used for parallel execution across all ArchR functions.
}
