% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectMethods.R
\name{getArrowFiles}
\alias{getArrowFiles}
\title{Get ArrowFiles from an ArchRProject}
\usage{
getArrowFiles(ArchRProj = NULL)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}
}
\description{
This function gets the names of all ArrowFiles associated with a given ArchRProject.
}
