% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectMethods.R
\name{getAvailableMatrices}
\alias{getAvailableMatrices}
\title{Get a list available matrices in the ArrowFiles storted in an ArchRProject}
\usage{
getAvailableMatrices(ArchRProj = NULL)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}
}
\description{
This function gets the available matrices from the ArrowFiles in a given ArchRProject object.
}
