% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixDeviations.R
\name{getBgdPeaks}
\alias{getBgdPeaks}
\title{Get Background Peaks from an ArchRProject}
\usage{
getBgdPeaks(
  ArchRProj = NULL,
  nIterations = 50,
  w = 0.1,
  binSize = 50,
  seed = 1,
  force = FALSE
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{nIterations}{The number of background peaks to sample. See \code{chromVAR::getBackgroundPeaks()}.}

\item{w}{The parameter controlling similarity measure of background peaks. See \code{chromVAR::getBackgroundPeaks()}.}

\item{binSize}{The precision with which the similarity is computed. See \code{chromVAR::getBackgroundPeaks()}.}

\item{seed}{A number to be used as the seed for random number generation. It is recommended to keep track of the seed used so that you can reproduce results downstream.}

\item{force}{A boolean value indicating whether to force the file indicated by \code{outFile} to be overwritten if it already exists.}
}
\description{
This function will get/compute background peaks controlling for total accessibility and GC-content from an ArchRProject.
}
