% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getCellColData}
\alias{getCellColData}
\title{Get cellColData from an ArchRProject}
\usage{
getCellColData(ArchRProj, select = NULL, drop = FALSE, ...)
}
\arguments{
\item{ArchRProj}{An ArchRProject object.}

\item{select}{A character vector of column names to select from cellColData if you would like to subset the returned data.}

\item{drop}{A boolean argument to indicate whether additional data.frame information should be dropped if selecting only a single column name.}

\item{...}{additional args}
}
\description{
This function gets the cellColData from a given ArchRProject.
}
