% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getExons}
\alias{getExons}
\title{Get the exons from an ArchRProject}
\usage{
getExons(ArchRProj, symbols = NULL, ...)
}
\arguments{
\item{ArchRProj}{An ArchRProject object.}

\item{...}{additional args}
}
\description{
This function gets the exons (in format QQQ) from the geneAnnotation of a given ArchRProject.
}
