% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Footprinting.R
\name{getFootprints}
\alias{getFootprints}
\title{Calculate footprints from an ArchRProject}
\usage{
getFootprints(
  ArchRProj = NULL,
  positions = NULL,
  plotName = "Plot-Footprints",
  groupBy = "Clusters",
  useGroups = NULL,
  flank = 250,
  minCells = 25,
  nTop = NULL,
  threads = getArchRThreads(),
  verbose = TRUE,
  logFile = createLogFile("getFootprints")
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{positions}{A \code{list} or \code{GenomicRangesList} of \code{GRanges} containing the positions to incorporate into the footprint. Each position should be stranded.}

\item{plotName}{The prefix to add to the file name for the output PDF file containing the footprint plots.}

\item{groupBy}{The name of the column in \code{cellColData} used in the \code{addGroupCoverages()} function for grouping multiple cells together.}

\item{useGroups}{A character vector that is used to select a subset of groups by name from the designated \code{groupBy} column in \code{cellColData}.
This limits the groups used to perform footprinting.}

\item{flank}{The number of basepairs from the position center (+/-) to consider as the flank.}

\item{minCells}{The minimum number of cells required in a given cell group to permit footprint generation.}

\item{nTop}{The number of genomic regions to consider. Only the top \code{nTop} genomic regions based on the "score" column in the \code{GRanges}
object will be considered for the footprint.}

\item{threads}{The number of threads to be used for parallel computing.}

\item{verbose}{A boolean value that determines whether standard output includes verbose sections.}

\item{pal}{The name of a custom palette from \code{ArchRPalettes} to use for plotting the lines corresponding to the footprints.}

\item{flankNorm}{The number of basepairs to consider at the edge of the flank region (+/-) to be used for footprint normalization.}

\item{smoothWindow}{The size in basepairs of the sliding window to be used for smoothing of the footprint signal.}

\item{normMethod}{The name of the normalization method to use to normalize the footprint relative to the Tn5 insertion bias. Options
include "none", "subtract", "divide". "Subtract" means subtracting the normalized Tn5 Bias. "Divide" means dividing the normalized Tn5 Bias.}

\item{inputSE}{Input a previous footprint Summarized Experiment (returned after running \code{plotFootprints}) to be plotted instead of
regenerating the footprinting information.}

\item{height}{The height in inches to be used for the output PDF.}

\item{width}{The width in inches to be used for the output PDF file.}

\item{addDOC}{A boolean variable that determines whether to add the date of creation to end of the PDF file name. This is useful for
preventing overwritting of old plots.}
}
\description{
This function will get footprints for all samples in a given ArchRProject or a properly-formatted Summarized Experiment
}
