% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrowRead.R
\name{getFragmentsFromProject}
\alias{getFragmentsFromProject}
\title{Get the fragments from an ArchRProject}
\usage{
getFragmentsFromProject(
  ArchRProj = NULL,
  subsetBy = NULL,
  cellNames = NULL,
  verbose = FALSE,
  logFile = createLogFile("getFragmentsFromProject")
)
}
\arguments{
\item{subsetBy}{A Genomic Ranges object to subset fragments by.}

\item{cellNames}{A character vector indicating the cell names of a subset of cells from which fragments whould be extracted.
This allows for extraction of fragments from only a subset of selected cells. By default, this function will extract all cells
from the provided ArrowFile using \code{getCellNames()}.}

\item{verbose}{A boolean value indicating whether to use verbose output during execution of this function. Can be set to \code{FALSE} for a cleaner output.}

\item{logFile}{The path to a file to be used for logging ArchR output.}

\item{ArchRProject}{An \code{ArchRProject} object to get fragments from.}
}
\description{
This function retrieves the fragments from a given ArchRProject as a GRangesList object.
}
