% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getGeneAnnotation}
\alias{getGeneAnnotation}
\title{Get geneAnnotation from ArchRProject}
\usage{
getGeneAnnotation(ArchRProj, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{...}{additional args}
}
\description{
This function gets geneAnnotation in ArchRProject
}
