% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getGenes}
\alias{getGenes}
\title{Get Genes from ArchRProject}
\usage{
getGenes(ArchRProj, symbols = NULL, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{symbols}{gene symbols to subset}

\item{...}{additional args}
}
\description{
This function gets genes from geneAnnotation in ArchRProject
}
