% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectMethods.R
\name{getGenome}
\alias{getGenome}
\title{Get the genome used by an ArchRProject}
\usage{
getGenome(ArchRProj = NULL)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}
}
\description{
This function gets the name of the genome from the genomeAnnotation used by a given ArchRProject.
}
\examples{

# Get Test ArchR Project
proj <- getTestProject()

# Get Genome in ArchRProj
getGenome(proj)

# Get Genome loaded globally
getGenome()

}
