% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getGenomeAnnotation}
\alias{getGenomeAnnotation}
\title{Get genomeAnnotation from ArchRProject}
\usage{
getGenomeAnnotation(ArchRProj, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{...}{additional args}
}
\description{
This function gets genomeAnnotation in ArchRProject
}
