% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupExport.R
\name{getGroupFragments}
\alias{getGroupFragments}
\title{Export Group Fragment Files}
\usage{
getGroupFragments(
  ArchRProj = NULL,
  groupBy = "Clusters",
  threads = getArchRThreads(),
  logFile = createLogFile("getGroupFragments")
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{groupBy}{A string that indicates how cells should be grouped. This string corresponds to one of the standard or
user-supplied \code{cellColData} metadata columns (for example, "Clusters"). Cells with the same value annotated in this metadata
column will be grouped together and their fragments exported to \code{outputDirectory}/GroupFragments.}

\item{threads}{An integer specifying the number of threads for parallel.}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
This function will group export fragment files for each group in an ArchRProject.
}
\examples{

# Get Test ArchR Project
proj <- getTestProject()

# Get Group BW
frags <- getGroupFragments(proj, groupBy = "Clusters")

}
