% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getGroupSummary}
\alias{getGroupSummary}
\title{Get the number of cells from an ArchRProject/ArrowFile JJJ}
\usage{
getGroupSummary(
  ArchRProj = NULL,
  groupBy = "Sample",
  select = "TSSEnrichment",
  summary = "median",
  na.rm = TRUE
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{select}{A character vector containing the column names to select from \code{cellColData}.}
}
\description{
This function gets number of cells from an ArchRProject or ArrowFile
}
