% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkerFeatures.R
\name{getMarkers}
\alias{getMarkers}
\title{Get Marker Features from a marker summarized experiment}
\usage{
getMarkers(seMarker = NULL, cutOff = "FDR <= 0.1 & Log2FC >= 0.5",
  n = NULL, returnGR = FALSE)
}
\arguments{
\item{seMarker}{A \code{SummarizedExperiment} object returned by \code{ArchR::markerFeatures()}.}

\item{cutOff}{A valid-syntax logical statement that defines which marker features from \code{seMarker}. \code{cutoff} can contain any of the \code{assayNames} from \code{seMarker}.}

\item{n}{An integer describing the maximum number of features to return per group.}

\item{returnGR}{A boolean describing whether to return as a \code{GRanges}. Only valid when \code{seMarker} is computed for a PeakMatrix.}
}
\description{
This function will identify Markers and return a List of Features or a GRangesList for each group of significant marker features.
}
