% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakAnnotation.R
\name{getMatches}
\alias{getMatches}
\title{Get peakAnnotation matches from an ArchRProject}
\usage{
getMatches(ArchRProj, name = NULL, annoName = NULL, ...)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{name}{The name of the annotation object (i.e. Motifs) to retrieve from the designated \code{ArchRProject}.}

\item{annoName}{The name of a specific annotation to subset within the peakAnnotation.}

\item{...}{additional args}
}
\description{
This function gets peakAnnotation matches from a given ArchRProject.
}
