% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrowRead.R
\name{getMatrixFromProject}
\alias{getMatrixFromProject}
\title{Get a data matrix stored in an ArchRProject RRR}
\usage{
getMatrixFromProject(
  ArchRProj = NULL,
  useMatrix = "GeneScoreMatrix",
  useSeqnames = NULL,
  verbose = TRUE,
  binarize = FALSE,
  threads = getArchRThreads(),
  logFile = createLogFile("getMatrixFromProject")
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object to get data matrix from.}

\item{useMatrix}{The name of the data matrix to retrieve from the given ArrowFile. Options include "TileMatrix", "GeneScoreMatrix", etc.}

\item{useSeqnames}{A character vector of chromosome names to be used to subset the data matrix being obtained.}

\item{verbose}{A boolean value indicating whether to use verbose output during execution of  this function. Can be set to FALSE for a cleaner output.}

\item{binarize}{A boolean value indicating whether the matrix should be binarized before return. This is often desired when working with insertion counts.}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
This function gets a given data matrix from an \code{ArchRProject}.
}
