% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getOutputDirectory}
\alias{getOutputDirectory}
\title{Get outputDirectory from an ArchRProject}
\usage{
getOutputDirectory(ArchRProj = NULL)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}
}
\description{
This function gets the outputDirectory from a given ArchRProject. If null this returns "QualityControl" directory.
}
