% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupExport.R
\name{getPBGroupSE}
\alias{getPBGroupSE}
\title{Export PseudoBulk Group Summarized Experiment}
\usage{
getPBGroupSE(
  ArchRProj = NULL,
  useMatrix = "GeneScoreMatrix",
  groupBy = "Clusters",
  divideN = TRUE,
  scaleTo = 10000,
  useLabels = TRUE,
  sampleLabels = "Sample",
  minCells = 40,
  maxCells = 500,
  minReplicates = 2,
  maxReplicates = 5,
  sampleRatio = 0.8,
  verbose = TRUE,
  threads = getArchRThreads(),
  logFile = createLogFile("getPBGroupSE")
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{useMatrix}{The name of the matrix in the ArrowFiles. See getAvailableMatrices to see options}

\item{groupBy}{The name of the column in \code{cellColData} to use for grouping cells together for summarizing.}

\item{divideN}{A boolean describing whether to divide by the number of cells.}

\item{scaleTo}{Depth normalize to this value if not NULL.}

\item{useLabels}{A boolean value indicating whether to use sample labels to create sample-aware subgroupings during as pseudo-bulk replicate generation.}

\item{sampleLabels}{The name of a column in \code{cellColData} to use to identify samples. In most cases, this parameter should be left as \code{NULL} and you
should only use this parameter if you do not want to use the default sample labels stored in \code{cellColData$Sample}. However, if your individual Arrow
files do not map to individual samples, then you should set this parameter to accurately identify your samples. This is the case in (for example)
multiplexing applications where cells from different biological samples are mixed into the same reaction and demultiplexed based on a lipid barcode or genotype.}

\item{minCells}{The minimum number of cells required in a given cell group to permit insertion coverage file generation.}

\item{maxCells}{The maximum number of cells to use during insertion coverage file generation.}

\item{minReplicates}{The minimum number of pseudo-bulk replicates to be generated.}

\item{maxReplicates}{The maximum number of pseudo-bulk replicates to be generated.}

\item{sampleRatio}{The fraction of the total cells that can be sampled to generate any given pseudo-bulk replicate.}

\item{verbose}{A boolean specifying to print messages during computation.}

\item{threads}{An integer specifying the number of threads for parallel.}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
This function will determine cell groups for pseudobulk, summarize and export a summarized experiment for a assay in a ArchRProject.
}
\examples{

# Get Test ArchR Project
proj <- getTestProject()

# Get Group SE
se <- getPBGroupSE(proj, useMatrix = "PeakMatrix", groupBy = "Clusters")

}
