% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrativeAnalysis.R
\name{getPeak2GeneLinks}
\alias{getPeak2GeneLinks}
\title{Get the peak-to-gene links from an ArchRProject}
\usage{
getPeak2GeneLinks(
  ArchRProj = NULL,
  corCutOff = 0.45,
  FDRCutOff = 1e-04,
  varCutOffATAC = 0.25,
  varCutOffRNA = 0.25,
  resolution = 1,
  returnLoops = TRUE
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{corCutOff}{A numeric describing the minimum numeric peak-to-gene correlation to return.}

\item{FDRCutOff}{A numeric describing the maximum numeric peak-to-gene false discovery rate to return.}

\item{varCutOffATAC}{A numeric describing the minimum variance quantile of the ATAC peak accessibility when selecting links.}

\item{varCutOffRNA}{A numeric describing the minimum variance quantile of the RNA gene expression when selecting links.}

\item{resolution}{A numeric describing the bp resolution to return loops as. This helps with overplotting of correlated regions.}

\item{returnLoops}{A boolean indicating to return the peak-to-gene links as a \code{GRanges} "loops" object designed for use with
the \code{ArchRBrowser()} or as an \code{ArchRBrowserTrack()}.}
}
\description{
This function obtains peak-to-gene links from an ArchRProject.
}
\examples{

# Get Test ArchR Project
proj <- getTestProject()

# Add P2G Links
proj <- addPeak2GeneLinks(proj, k = 20)

# Get P2G Links
p2g <- getPeak2GeneLinks(proj)

}
