% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakAnnotation.R
\name{getPeakAnnotation}
\alias{getPeakAnnotation}
\title{Get annotation from an ArchRProject}
\usage{
getPeakAnnotation(ArchRProj, name = NULL, ...)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{name}{QQQ The name of the annotation object to retrieve from the designated \code{ArchRProject}. Options include QQQ.}

\item{...}{additional args}
}
\description{
This function gets an annotation from a given ArchRProject.
}
