% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getPeakSet}
\alias{getPeakSet}
\title{Get PeakSet from ArchRProject}
\usage{
getPeakSet(ArchRProj, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{...}{additional args}
}
\description{
This function gets peakSet from an ArchRProject
}
