% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getProjectSummary}
\alias{getProjectSummary}
\title{Get projectSummary from ArchRProject}
\usage{
getProjectSummary(ArchRProj, returnSummary = FALSE, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{returnSummary}{return summary or just print}

\item{...}{additional args}
}
\description{
This function prints the projectSummary from an ArchRProject
}
