% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectMethods.R
\name{getReducedDims}
\alias{getReducedDims}
\title{Get dimensionality reduction information stored in an ArchRProject}
\usage{
getReducedDims(
  ArchRProj = NULL,
  reducedDims = "IterativeLSI",
  returnMatrix = TRUE,
  dimsToUse = NULL,
  scaleDims = NULL,
  scaleBy = NULL,
  corCutOff = 0.75
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{reducedDims}{The name of the \code{reducedDims} object (i.e. "IterativeLSI") to retrieve from the designated \code{ArchRProject}.}

\item{returnMatrix}{If set to "mat" or "matrix", the function will return the \code{reducedDims} object as a matrix with entries for
each individual cell. Otherwise, it will return the full \code{reducedDims} object.}

\item{dimsToUse}{A vector containing the dimensions (i.e. 1:30) to return from the \code{reducedDims} object.}

\item{scaleDims}{A boolean value that indicates whether to z-score the reduced dimensions. The default is set to \code{NULL}, and will scale the dimensions
based on the value of \code{scaleDims} when the \code{reducedDims} were originally created during dimensionality reduction. This idea was introduced by Timothy Stuart.}

\item{scaleBy}{A character string indicating if the reduced dimensions should be scaled in either the row direction (default) or the column direction when \code{scaleDims = TRUE}.
In the case of SVD matrix, the default is to perform scaling for each cell, rather than on the components as in the \code{signac::RunSVD} implementation.
You can use \code{scaleBy = "column"} to perform scaling for each component. Like, \code{scaleDims}, the saved value of \code{scaleBy} will be used if set to \code{scaleBy = NULL}.}

\item{corCutOff}{A numeric cutoff for the correlation of each dimension to the sequencing depth. If the dimension has a correlation
to sequencing depth that is greater than the \code{corCutOff}, it will be excluded.}
}
\description{
This function gets a dimensionality reduction object (i.e. UMAP, tSNE, etc) from a given ArchRProject.
}
\examples{

# Get Test ArchR Project
proj <- getTestProject()

# Get Iterative LSI
getReducedDims(proj, reducedDims = "IterativeLSI")

}
