% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getSampleColData}
\alias{getSampleColData}
\title{Get sampleColData from an ArchRProject}
\usage{
getSampleColData(ArchRProj, select = NULL, drop = FALSE, ...)
}
\arguments{
\item{ArchRProj}{An ArchRProject object.}

\item{select}{select a subset of column names from sampleColData}

\item{drop}{drop if selecting only one column name}

\item{...}{additional args}
}
\description{
This function gets the sampleColData from a given ArchRProject.
}
