% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectMethods.R
\name{getSeqnames}
\alias{getSeqnames}
\title{Get the seqnames that could be selected from a given data matrix within an ArchRProject}
\usage{
getSeqnames(ArchRProj = NULL, useMatrix = "GeneScoreMatrix")
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{useMatrix}{The name of the data matrix as stored in the ArrowFiles of the \code{ArchRProject}. Options include "TileMatrix", "GeneScoreMatrix", etc.}
}
\description{
This function will identify available seqnames from a given data matrix  (i.e. "GeneScoreMatrix", or "TileMatrix") and return
them for downstream plotting utilities.
}
\examples{

# Get Test ArchR Project
proj <- getTestProject()

# Get Seqnames in Matrix
getSeqnames(proj, useMatrix = "GeneScoreMatrix")

}
