% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputData.R
\name{getTestArrow}
\alias{getTestArrow}
\title{Get PBMC Small Test Arrow file}
\usage{
getTestArrow(version = 2)
}
\arguments{
\item{version}{version of test arrow to return}
}
\description{
V2 : This function will return a test arrow file in your cwd.
}
\examples{

# Get Test Arrow
arrow <- getTestArrow()

}
