% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputData.R
\name{getTestFragments}
\alias{getTestFragments}
\title{Get PBMC Small Test Fragments}
\usage{
getTestFragments(x)
}
\description{
This function will download fragments for a small PBMC test dataset (2k Cells) spanning chr1 and 2 (~20MB).
}
