% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputData.R
\name{getTestProject}
\alias{getTestProject}
\title{Get PBMC Small Test Project}
\usage{
getTestProject(version = 2)
}
\arguments{
\item{version}{version of test fragments to return}
}
\description{
V1 : This function will download an ArchRProject for a small PBMC test dataset.
V2 : This function will return an ArchRProject for a small PBMC test dataset in your cwd.
}
\examples{

# Get Test Project
proj <- getTestProject()

}
