% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trajectory.R
\name{getTrajectory}
\alias{getTrajectory}
\title{Get Supervised Trajectory from an ArchR Project}
\usage{
getTrajectory(ArchRProj, name = "Trajectory",
  useMatrix = "GeneScoreMatrix", groupEvery = 2, threads = 1,
  scaleTo = 10000, log2Norm = TRUE, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{name}{name of fitted trajectory in cellColData}

\item{useMatrix}{matrix to summarize across the trajectory}

\item{groupEvery}{group cells every x quantile}

\item{threads}{number of threads to use}

\item{scaleTo}{scale summarized matrix to}

\item{log2Norm}{log2 normalize the scaled summarized matrix}

\item{...}{additional args}
}
\description{
This function will fit get a supervised trajectory from an ArchRProject and aggregate signal
from a matrix and smooth across the trajectory
}
