% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectMethods.R
\name{getTSS}
\alias{getTSS}
\title{Get the transcription start sites of all genes in an ArchRProject}
\usage{
getTSS(ArchRProj = NULL)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}
}
\description{
This function gets the transcription start sites (TSSs) as a GRanges object of all genes from the geneAnnotation of a given ArchRProject.
}
\examples{

# Get Test ArchR Project
proj <- getTestProject()

# Get TSS in ArchRProj
getTSS(proj)

# Get TSS loaded globally
getTSS()

}
