% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getTutorialData}
\alias{getTutorialData}
\title{Get Tutorial Data For ArchR}
\usage{
getTutorialData(tutorial = "hematopoiesis", ...)
}
\arguments{
\item{tutorial}{Available tutorials are Hematopoiesis, PBMC, FreshFrozen}

\item{...}{additional args}
}
\description{
This function will download data for a given tutorial and return the inputFiles for ArchR
}
