% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getValidBarcodes}
\alias{getValidBarcodes}
\title{Get Valid Barcodes from 10x Cell Ranger output to pre-filter barcodes}
\usage{
getValidBarcodes(csvFiles, sampleNames, ...)
}
\arguments{
\item{csvFiles}{vector of filenames for reading in and identifying valid cell barcodes}

\item{sampleNames}{sample names associated with each individual csv file}

\item{...}{additional args}
}
\description{
This function will read in processed 10x cell ranger files and identify barcodes that are associated with a cell that passed QC.
}
