% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixDeviations.R
\name{getVarDeviations}
\alias{getVarDeviations}
\title{Get Variable Deviations across cells in ArchRProject.}
\usage{
getVarDeviations(ArchRProj, name = "MotifMatrix", plot = TRUE,
  n = 25)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{name}{name of DeviationsMatrix see addDeviationsMatrix}

\item{plot}{plot ranked variability for each annotation}

\item{n}{number of annotations to label with ggrepel}
}
\description{
This function will rank the variability of the deviations computed by ArchR and label the top variable annotations.
}
