% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgplotUtils.R
\name{ggAlignPlots}
\alias{ggAlignPlots}
\title{Align ggplot plots vertically or horizontally}
\usage{
ggAlignPlots(..., plotList = NULL, sizes = NULL, type = "v", draw = TRUE)
}
\arguments{
\item{...}{All additional arguments will be interpreted as \code{ggplot2} plot objects and used if and only if \code{plotList} is \code{NULL}}

\item{plotList}{A list of \code{ggplot2} plot objects to be aligned.}

\item{sizes}{A numeric vector or list of values indicating the relative size for each of the objects in \code{plotList} or supplied in \code{...}. If the plot is supplied in \code{...} the order is the same as the input in this function. If set to NULL all plots will be evenly distributed.}

\item{type}{A string indicating wheter vertical ("v") or horizontal ("h") alignment should be used for the multi-plot layout.}

\item{draw}{A boolean value indicating whether to draw the plot(s) (\code{TRUE}) or return a graphical object (\code{FALSE}).}
}
\description{
This function aligns ggplots vertically or horizontally
}
\examples{

# Create Random Data
m <- data.frame(x=matrix(rnorm(10, 2),ncol=1))
m$color <- sample(c("A", "B"), 10, replace = TRUE)

# Plot
p <- ggGroup(x = m$color, y = m$x)

# To PDF
pdf("test.pdf", width = 4, height = 7)
ggAlignPlots(p, p)
dev.off()

}
