% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgplotHelper.R
\name{ggGroup}
\alias{ggGroup}
\title{A ggplot-based ridge/violin plot wrapper function JJJ}
\usage{
ggGroup(x = NULL, y = NULL, xlabel = NULL, ylabel = NULL,
  groupOrder = NULL, groupSort = FALSE, size = 1, baseSize = 10,
  ridgeScale = 1, ratioYX = NULL, alpha = 1, title = "",
  pal = paletteDiscrete(values = x, set = "stallion"),
  addBoxPlot = FALSE, plotAs = "ridges", ...)
}
\arguments{
\item{x}{A character vector containing the categorical x-axis values for each y-axis value.}

\item{y}{A numeric vector containing the y-axis values for each point.}

\item{xlabel}{The label to plot for the x-axis.}

\item{ylabel}{The label to plot for the y-axis.}

\item{groupOrder}{A character vector indicating a custom order for plotting x-axis categorical values. Should contain all possible values of \code{x} in the desired order.}

\item{size}{The line width for boxplot/summary lines.}

\item{baseSize}{The base font (in points) size to use in the plot.}

\item{ratioYX}{The aspect ratio of the x and y axes on the plot.}

\item{title}{The title of the plot.}

\item{pal}{A named custom palette (see \code{paletteDiscrete()} and \code{ArchRPalettes}) for discrete coloring.}

\item{...}{Additional params for plotting}

\item{sampleRatio}{The fraction of the total number of points to be displayed over violins. A value of 0.1 would plot 10 percent of the total data points over the violin.}
}
\description{
This function is a wrapper around ggplot geom_density_ridges or geom_violin to allow for plotting group distribution plots in ArchR.
}
