% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgplotHelper.R
\name{ggHex}
\alias{ggHex}
\title{Ggplot Hexplot summary of points in a standardized manner}
\usage{
ggHex(x = NULL, y = NULL, color = NULL, pal = paletteContinuous(set
  = "solar_extra"), bins = 150, xlim = NULL, ylim = NULL,
  extend = 0.05, xlabel = "x", ylabel = "y", title = "",
  colorTitle = "values", baseSize = 6, ratioYX = 1, FUN = "median",
  addPoints = FALSE, ...)
}
\arguments{
\item{x}{x vector of data to be plot}

\item{y}{y vector of data to be plot}

\item{color}{color vector of values to be plot (must be same length as x,y)}

\item{pal}{custom palette option}

\item{bins}{number of bins for hexplot}

\item{xlim}{xlimits for plot}

\item{ylim}{ylimits for plot}

\item{extend}{extend limits by this proportion if not set by xlim or ylim}

\item{xlabel}{label for x-axis}

\item{ylabel}{label for y-axis}

\item{title}{title of plot}

\item{colorTitle}{title for legend corresponding to color}

\item{baseSize}{baseSize for fonts}

\item{ratioYX}{ratio of y to x in plot}

\item{FUN}{function for summarizing hexagons}

\item{...}{additional params to pass}
}
\description{
This function will plot x,y coordinates values summarized in hexagons in a standardized manner
}
