% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgplotHelper.R
\name{ggLine}
\alias{ggLine}
\title{GG Plot One to One Heatscatter}
\usage{
ggLine(x, y, color = NULL, discrete = TRUE, discreteSet = "stallion",
  continuousSet = "solar_extra", pal = NULL, size = 1, xlim = NULL,
  ylim = NULL, extend = 0.05, xlabel = "x", ylabel = "y",
  title = "", alpha = 1, baseSize = 6, ratioYX = 1,
  nullColor = "lightGrey")
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{pal}{continuous color palette to use}

\item{size}{geom_point size}

\item{xlabel}{xlabel}

\item{ylabel}{ylabel}

\item{title}{ggtitle}

\item{alpha}{geom_point alpha}

\item{baseSize}{base_font size}

\item{min}{xmin quantile [0,1]}

\item{max}{xmax quantile [0,1]}

\item{plot_n}{number of points to plot}

\item{kernel_n}{n for MASS::kde2d default = 100}

\item{plot_n}{number of points to plot}
}
\description{
GG Plot One to One Heatscatter
}
