% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgplotHelper.R
\name{ggOneToOne}
\alias{ggOneToOne}
\title{GG Plot One to One Heatscatter}
\usage{
ggOneToOne(x, y, nPlot = 100 * 10^3, nKernel = 100, size = 2,
  xlabel = "x", ylabel = "y", title = "Sample Correlation",
  min = 0.1, max = 0.9999, densityMax = 0.95, extend = 0.05,
  alpha = 1, baseSize = 12, pal = paletteContinuous(set = "viridis"))
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{nPlot}{number of points to plot}

\item{nKernel}{n for MASS::kde2d default = 100}

\item{size}{geom_point size}

\item{xlabel}{xlabel}

\item{ylabel}{ylabel}

\item{title}{ggtitle}

\item{min}{xmin quantile [0,1]}

\item{max}{xmax quantile [0,1]}

\item{alpha}{geom_point alpha}

\item{baseSize}{base_font size default is 12}

\item{pal}{continuous color palette to use}
}
\description{
GG Plot One to One Heatscatter
}
