% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgplotHelper.R
\name{ggViolin}
\alias{ggViolin}
\title{A ggplot-based violin plot wrapper function}
\usage{
ggViolin(x = NULL, y = NULL, xlabel = NULL, ylabel = NULL,
  xOrder = NULL, addPoints = FALSE, size = 1, baseSize = 6,
  ratioYX = NULL, sampleRatio = 0.1, title = "",
  pal = paletteDiscrete(values = x, set = "stallion"), ...)
}
\arguments{
\item{x}{A character vector containing the categorical x-axis values for each y-axis value.}

\item{y}{A numeric vector containing the y-axis values for each point.}

\item{xlabel}{The label to plot for the x-axis.}

\item{ylabel}{The label to plot for the y-axis.}

\item{xOrder}{A character vector indicating a custom order for plotting x-axis categorical values for plotting. Should contain all possible values of \code{x} in the desired order.}

\item{addPoints}{A boolean value indicating whether individual points should be added to the plot using \code{geom_quasirandom}.}

\item{size}{The line width for boxplot lines.}

\item{baseSize}{The base font size to use in the plot.}

\item{ratioYX}{The aspect ratio of the x and y axes on the plot.}

\item{sampleRatio}{sampling ratio for number of dots to be shown from original data over violins to prevent over-crowding. Default is set to 0.1.}

\item{title}{The title of the plot.}

\item{pal}{A custom palette for discrete coloring.}
}
\description{
A ggplot-based violin plot wrapper function
}
