% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperUtils.R
\name{\%bcin\%}
\alias{\%bcin\%}
\title{Generic matching function for S4Vector objects}
\usage{
x \%bcin\% table
}
\arguments{
\item{x}{An \code{S4Vector} object to search for in \code{table}.}

\item{table}{The set of \code{S4Vector} objects to serve as the base for the match function.}
}
\description{
This function provides a generic matching function for S4Vector objects primarily to avoid ambiguity.
}
\examples{

#Test
Rle(c("A", "B", "C")) \%bcin\% Rle(c("A", "C"))

}
