% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperUtils.R
\name{\%ni\%}
\alias{\%ni\%}
\title{Negated Value Matching}
\usage{
x \%ni\% table
}
\arguments{
\item{x}{x search within table}

\item{table}{to search x in}
}
\description{
This function is the reciprocal of %in%
See match funciton in base R
x %ni% table
}
