% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archR_helper_functions.R
\name{handle_chunk_w_NMF}
\alias{handle_chunk_w_NMF}
\title{Set archR run configuration}
\usage{
handle_chunk_w_NMF(innerChunkIdx, innerChunksColl, this_tss.seqs,
  globFactors, globClustAssignments, config)
}
\arguments{
\item{innerChunkIdx}{samarth}

\item{innerChunksColl}{samarth}

\item{this_tss.seqs}{samarth}

\item{globFactors}{samarth}

\item{globClustAssignments}{samarth}

\item{config}{samarth}
}
\value{
NMF result on the given chunk
}
\description{
samarth
}
