% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archR_auxiliary_functionsII.R
\name{handle_dir_creation}
\alias{handle_dir_creation}
\title{Handle directory creation}
\usage{
handle_dir_creation(
  givenODir,
  flags = list(debugFlag = FALSE, verboseFlag = FALSE, plotFlag = FALSE, timeFlag =
    FALSE)
)
}
\description{
Handle directory creation
}
