% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_entropy.R
\name{i}
\alias{i}
\title{i(x,y)}
\usage{
i(x, y, xy)
}
\arguments{
\item{x}{probability P(x)}

\item{y}{probability P(y)}

\item{xy}{probability P(x,y) = probability of event x and y occurring at the same time}
}
\description{
Mutual Information
}
