% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyClusters.R
\name{IdentifyClusters}
\alias{IdentifyClusters}
\title{Identify Clusters for Single Cell Data}
\usage{
IdentifyClusters(input, reducedDims = "IterativeLSI",
  name = "Clusters", sampleCells = NULL, seed = 1,
  method = "seurat", dimsToUse = NULL, knnAssign = 10,
  nOutlier = 20, verbose = TRUE, tstart = NULL, ...)
}
\arguments{
\item{input}{ArchRProject or matrix for cluster identification}

\item{reducedDims}{reducedDims of ArchRProject if provided}

\item{name}{name of cluster column if input is ArchRProject}

\item{seed}{seed}

\item{method}{supported methods are Seurat and LouvainJaccard}

\item{dimsToUse}{reduced dims to use}

\item{knnAssign}{number of nearest neighbors for assignment of outliers and estimation}

\item{nOutlier}{number of cells required for a cluster to be called if not then these will be considered an outlier}

\item{...}{arguments to provide Seurat::FindClusters or ArchR:::.clustLouvain (knn = 50, jaccard = TRUE)}
}
\description{
This function will identify clusters for single cell reduced dimensions supplied or from and ArchRProject
}
