% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{imputeMatrix}
\alias{imputeMatrix}
\title{Impute a matrix with impute weights}
\usage{
imputeMatrix(
  mat = NULL,
  imputeWeights = NULL,
  threads = getArchRThreads(),
  verbose = FALSE,
  logFile = createLogFile("imputeMatrix")
)
}
\arguments{
\item{mat}{a matrix or sparseMatrix of class dgCMatrix}

\item{imputeWeights}{imputeWeights from getImputeWeights(ArchRProj), see addImputeWeights for more details.}

\item{threads}{number of threads for parallel computing, see getArchRThreads for more information.}

\item{verbose}{print output during analysis.}

\item{logFile}{log file for storing important information.}
}
\description{
This function gets imputation weights from an ArchRProject to impute a numerical matrix
}
