% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalDefaults.R
\name{installExtraPackages}
\alias{installExtraPackages}
\title{Install extra packages used in ArchR that are not installed by default}
\usage{
installExtraPackages(force = FALSE)
}
\arguments{
\item{force}{If you want to force a reinstall of these pacakges.}
}
\description{
This function will install extra packages used in ArchR that are not installed by default.
}
