% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentSemanticIndexing.R
\name{IterativeLSI}
\alias{IterativeLSI}
\title{Compute Iterative LSI}
\usage{
IterativeLSI(ArchRProj = NULL, useMatrix = "TileMatrix",
  reducedDimsOut = "IterativeLSI", iterations = 3, dimsToUse = 1:25,
  binarize = TRUE, sampleCells = 5000, varFeatures = 50000,
  selectionMethod = "var", scaleTo = 10000, totalFeatures = 5e+05,
  filterQuantile = 0.99, saveIterations = TRUE,
  outDir = getOutputDirectory(ArchRProj), clusterParams = list(),
  runHarmony = FALSE, harmonyParams = list(), threads = 1,
  seed = 1, verboseHeader = TRUE, verboseAll = FALSE,
  force = FALSE, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{useMatrix}{use matrix for LSI clustering from Arrow}

\item{reducedDimsOut}{name of dimensionality reduction to be stored as}

\item{iterations}{number of LSI iterations to perform}

\item{dimsToUse}{number of dimensions to compute and use from LSI (TFIDF-SVD) for clustering}

\item{binarize}{binarize matrix prior to LSI}

\item{sampleCells}{number of cells to sample for LSI estimation}

\item{varFeatures}{number of variable features to use for LSI}

\item{selectionMethod}{selection method for variable features (var or vmr)}

\item{scaleTo}{scaleTo for Cluster Averages for variance calculation}

\item{totalFeatures}{number of features to consider (ranked by total number of counts) use for LSI}

\item{filterQuantile}{filter features for initial LSI that are above this quantile}

\item{saveIterations}{save LSI iterations as rds in the outDir}

\item{outDir}{output directory for saving LSI iterations}

\item{clusterParams}{additional params to pass to IdentifyClusters}

\item{runHarmony}{run harmony batch correction through the iterations}

\item{harmonyParams}{additional params to pass to harmony}

\item{threads}{number of threads for parallel execution}

\item{seed}{seed for analysis}

\item{verboseHeader}{verbose sections}

\item{verboseAll}{verbose sections and subsections}

\item{force}{verbose sections and subsections}

\item{...}{additional args}
}
\description{
This function will compute an iterative LSI dimensionality reduction
on an ArchRProject.
}
