% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesUtils.R
\name{keepFilteredChromosomes}
\alias{keepFilteredChromosomes}
\title{Filters unwanted chr mainly underscores}
\usage{
keepFilteredChromosomes(x, remove = c("chrM"), underscore = TRUE,
  standard = TRUE, pruning.mode = "coarse")
}
\arguments{
\item{x}{GRanges or something with seqlevels}

\item{remove}{remove vector}

\item{underscore}{remove all underscores?}

\item{standard}{keep standard chromosomes}
}
\description{
Filters unwanted chr mainly underscores
}
