% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{loadArchRProject}
\alias{loadArchRProject}
\title{Load Previous ArchRProject into R}
\usage{
loadArchRProject(path = "./", force = FALSE, showLogo = TRUE)
}
\arguments{
\item{path}{A character path to an \code{ArchRProject} directory that was previously saved using \code{saveArchRProject()}.}

\item{force}{A boolean value indicating whether missing optional \code{ArchRProject} components (i.e. peak annotations /
background peaks) should be ignored when re-normalizing file paths. If set to \code{FALSE} loading of the \code{ArchRProject}
will fail unless all components can be found.}

\item{showLogo}{A boolean value indicating whether to show the ascii ArchR logo after successful creation of an \code{ArchRProject}.}
}
\description{
This function will load a previously saved ArchRProject and re-normalize paths for usage.
}
\examples{

# Get Small PBMC Project Location
zipProj <- file.path(system.file("testdata", package="ArchR"), "PBSmall.zip")

# Copy to current directory
file.copy(zipProj, basename(zipProj), overwrite = TRUE)

# Unzip
unzip(basename(zipProj), overwrite = TRUE)

# Remove
file.remove(basename(zipProj))

# Load
loadArchRProject("PBSmall")

}
