% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sinuc_PWMs.R
\name{make_sinuc_PWMs}
\alias{make_sinuc_PWMs}
\title{Make a PWM-resembling matrix out of a given n-vector}
\usage{
make_sinuc_PWMs(givenMatrix, add_pseudo_counts = T, scale = T)
}
\arguments{
\item{givenMatrix}{Actually a vector that will be reshaped into a (PWM)
matrix of DNA sequences.}

\item{add_pseudo_counts}{Boolean, taking values TRUE/T or FALSE/F, specifying
whether or not pseudocounts are added to the matrix.}

\item{scale}{Boolean, taking values TRUE/T or FALSE/F, specifying whether or
not the matrix is scaled column-wise, i.e., all columns summed to 1.}
}
\value{
A (PWM) matrix with 4 rows corresponding to the 4 nucleotides (A, C,
G, T) and the relevant number of columns (i.e., number of elements in given
vector/4)
}
\description{
The given matrix (or simply a vector) is reshaped to have four rows for four nucleotides and a relevant number of columns.
}
