% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperUtils.R
\name{mapLabels}
\alias{mapLabels}
\title{Create a Confusion Matrix based on two value vectors RRR}
\usage{
mapLabels(labels = NULL, newLabels = NULL, oldLabels = names(newLabels))
}
\arguments{
\item{labels}{A character vector containing lables to map.}

\item{newLabels}{A character vector (same length as oldLabels) to map labels to from oldLabels.}

\item{oldLabels}{A character vector (same length as newLabels) to map labels from to newLabels}
}
\description{
This function creates a confusion matrix based on two value vectors.
}
