% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkerFeatures.R
\name{markerAnnoEnrich}
\alias{markerAnnoEnrich}
\title{QQQ}
\usage{
markerAnnoEnrich(seMarker = NULL, ArchRProj = NULL,
  annotations = NULL, matches = NULL,
  cutOff = "FDR <= 0.1 & Log2FC >= 0.5", background = "bgdPeaks", ...)
}
\arguments{
\item{seMarker}{QQQ A \code{SummarizedExperiment} object returned by \code{ArchR::markerFeatures()}.}

\item{ArchRProj}{An \code{ArchRProject} object.}

\item{annotations}{QQQ}

\item{matches}{QQQ}

\item{cutOff}{A valid-syntax logical statement that defines which marker features from \code{seMarker} will be plotted in the heatmap. \code{cutoff} can contain any of the \code{assayNames} from \code{seMarker}.}

\item{background}{QQQ}

\item{...}{additional args}
}
\description{
This function QQQ
}
