% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkerFeatures.R
\name{markerGR}
\alias{markerGR}
\title{Identify Marker Feature Ranges}
\usage{
markerGR(seMarker = NULL, cutOff = "FDR <= 0.1 & Log2FC >= 0.5")
}
\arguments{
\item{seMarker}{A \code{SummarizedExperiment} object returned by \code{ArchR::markerFeatures()}.}

\item{cutOff}{A valid-syntax logical statement that defines which marker features from \code{seMarker}. \code{cutoff} can contain any of the \code{assayNames} from \code{seMarker}.}
}
\description{
This function will identify Markers and return a GRangesList for each group of significant marker regions.
}
