% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkerFeatures.R
\name{markerRanges}
\alias{markerRanges}
\title{QQQ}
\usage{
markerRanges(seMarker, cutOff = "FDR <= 0.1 & Log2FC >= 0.5", ...)
}
\arguments{
\item{seMarker}{QQQ A \code{SummarizedExperiment} object returned by \code{ArchR::markerFeatures()}.}

\item{cutOff}{A valid-syntax logical statement that defines which marker features from \code{seMarker} will be plotted in the heatmap. \code{cutoff} can contain any of the \code{assayNames} from \code{seMarker}.}

\item{...}{additional args}
}
\description{
This function QQQ
}
