% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part4_functions.R
\name{me_ij}
\alias{me_ij}
\title{Get main effect of decision i given decision j}
\usage{
me_ij(
  data,
  decision_i,
  value_i,
  decision_j,
  value_j,
  metric = "m1",
  notj = FALSE
)
}
\arguments{
\item{data}{data.frame}

\item{decision_i}{string: name of decision variable i}

\item{value_i}{integer: value of decision variable i}

\item{decision_j}{string: name of decision variable j}

\item{value_j}{integer: value of decision variable j}

\item{metric}{string: name of performance metric under evaluation}

\item{notj}{logical: if true, assumes that decision j does not equal value_j. If false, assumes that decision j does equal value_j.}
}
\description{
Get main effect of decision i given decision j
}
