% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{nCells}
\alias{nCells}
\title{Get number of cells in ArchRProject}
\usage{
nCells(ArchRProj, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{...}{additional args}
}
\description{
This function gets number of cells in ArchRProject
}
