% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesUtils.R
\name{nonOverlappingGRanges}
\alias{nonOverlappingGRanges}
\title{Instead of counting overlaps get columns like max score or etc in query}
\usage{
nonOverlappingGRanges(gr, by = "score", decreasing = TRUE,
  verbose = FALSE)
}
\arguments{
\item{decreasing}{for order}

\item{query}{granges query}

\item{subject}{granges subject}

\item{colname}{mcols(gr)[[colname]] cannot be null}
}
\description{
Instead of counting overlaps get columns like max score or etc in query
}
