% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part3_functions.R
\name{one_arch_to_assignment}
\alias{one_arch_to_assignment}
\title{Convert One Architecture to Assignment Matrix}
\usage{
one_arch_to_assignment(data)
}
\arguments{
\item{data}{data.frame of edges, containing the fields \code{from}, \code{to}, and \code{value}, where \code{from} and \code{to} are \code{integers}.}
}
\description{
Helper function for \code{arch_to_assignment()}. Called in a loop.
}
\author{
Tim Fraser, PhD
}
