% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_from_FASTA.R
\name{one_hot_encode}
\alias{one_hot_encode}
\title{One-hot encode}
\usage{
one_hot_encode(givenSeq)
}
\arguments{
\item{givenSeq}{A single sequence.}
}
\value{
The one-hot encoded sequence.
}
\description{
One-hot encode a given DNA sequence.
}
