% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorPalettes.R
\name{paletteContinuous}
\alias{paletteContinuous}
\title{Continuous Color Palette}
\usage{
paletteContinuous(set = "solar_extra", n = 256, reverse = FALSE,
  returnStructure = FALSE)
}
\arguments{
\item{set}{continuous palette name or number}

\item{n}{number for gradient}

\item{reverse}{return reversed values}

\item{returnStructure}{return structure palette}
}
\description{
Continuous Color Palette
}
