% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part4_functions.R
\name{pareto}
\alias{pareto}
\title{Find Pareto Front using Nondominated Sorting}
\source{
Matlab implementation by Nozomi Hitomi, 11/17/2015
}
\usage{
pareto(x = NULL, y = NULL, m = NULL)
}
\arguments{
\item{matrix}{an \code{n}x\code{m} matrix with \code{m} objectives and \code{n} solutions}
}
\description{
Generates a \code{n}x1 logical vector with \code{TRUE} if the solution lies on the pareto front.
}
\author{
Tim Fraser (adapted from Nozomi Hitomi)
}
