% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part4_functions.R
\name{pareto_rank}
\alias{pareto_rank}
\title{Find Pareto Rank using Nondominated Sorting}
\source{
Adapted from Matlab implementation by Nozomi Hitomi, 11/17/2015
}
\usage{
pareto_rank(x = NULL, y = NULL, m = NULL)
}
\arguments{
\item{matrix}{an \code{n}x\code{m} matrix with \code{m} objectives and \code{n} solutions}
}
\description{
Generates a \code{n}x1 logical vector of parento ranks.
}
\author{
Tim Fraser (adapted from Nozomi Hitomi)
}
