% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrativeAnalysis.R
\name{peak2GeneHeatmap}
\alias{peak2GeneHeatmap}
\title{Plot Peak2Gene Heatmap from an ArchRProject}
\usage{
peak2GeneHeatmap(
  ArchRProj = NULL,
  corCutOff = 0.45,
  FDRCutOff = 1e-04,
  k = 25,
  nPlot = 25000,
  limitsATAC = c(-2, 2),
  limitsRNA = c(-2, 2),
  groupBy = "Clusters",
  palGroup = NULL,
  palATAC = paletteContinuous("solarExtra"),
  palRNA = paletteContinuous("blueYellow"),
  verbose = TRUE,
  logFile = createLogFile("peak2GeneHeatmap")
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{corCutOff}{A numeric describing the minimum numeric peak-to-gene correlation to return.}

\item{FDRCutOff}{A numeric describing the maximum numeric peak-to-gene false discovery rate to return.}

\item{k}{An integer describing the number of k-means clusters to group peak-to-gene links prior to plotting heatmaps.}

\item{nPlot}{An integer describing the maximum number of peak-to-gene links to plot in heatmap.}

\item{limitsATAC}{An integer describing the maximum number of peak-to-gene links to plot in heatmap.}

\item{limitsRNA}{An integer describing the maximum number of peak-to-gene links to plot in heatmap.}

\item{groupBy}{The name of the column in \code{cellColData} to use for labeling KNN groupings. The maximum group appeared in the KNN groupings is used.}

\item{palGroup}{A color palette describing the colors in \code{groupBy}. For example, if groupBy = "Clusters" try paletteDiscrete(ArchRProj$Clusters) for a color palette.}

\item{palATAC}{A color palette describing the colors to be used for the ATAC heatmap. For example, paletteContinuous("solarExtra").}

\item{palRNA}{A color palette describing the colors to be used for the RNA heatmap. For example, paletteContinuous("blueYellow").}

\item{verbose}{A}

\item{logFile}{A}
}
\description{
This function plots side by side heatmaps of linked ATAC and Gene regions from \code{addPeak2GeneLinks}.
}
