% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakAnnotation.R
\name{peakAnnoEnrichment}
\alias{peakAnnoEnrichment}
\title{Peak Annotation Hypergeometric Enrichment in Marker Peaks.}
\usage{
peakAnnoEnrichment(
  seMarker = NULL,
  ArchRProj = NULL,
  peakAnnotation = NULL,
  matches = NULL,
  cutOff = "FDR <= 0.1 & Log2FC >= 0.5",
  background = "all",
  ...
)
}
\arguments{
\item{seMarker}{A \code{SummarizedExperiment} object returned by \code{markerFeatures()}.}

\item{ArchRProj}{An \code{ArchRProject} object.}

\item{peakAnnotation}{A \code{peakAnnotation} object in the provided \code{ArchRProject} to be used for hypergeometric testing.}

\item{matches}{A custom \code{peakAnnotation} matches object used as input for the hypergeometric test. See \code{motifmatchr::matchmotifs()} for additional information.}

\item{cutOff}{A valid-syntax logical statement that defines which marker features from \code{seMarker} to use. \code{cutoff} can contain any of the \code{assayNames} from \code{seMarker}.}

\item{background}{A string that indicates whether to use a background set of matched peaks to compare against ("bgdPeaks") or all peaks ("all").}

\item{...}{additional args}
}
\description{
This function will perform hypergeometric enrichment of a given peak annotation within the defined marker peaks.
}
