% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_arch_for_clusters.R
\name{plot_arch_for_clusters}
\alias{plot_arch_for_clusters}
\title{Plot the sequence architectures of the sequence clusters.}
\usage{
plot_arch_for_clusters(givenSamplesMat, givenFeaturesMatrix, nCluster,
  clustering_sol, seqs, positions)
}
\arguments{
\item{givenSamplesMat}{The samples matrix resulting for NMF.}

\item{givenFeaturesMatrix}{The features matrix resulting from NMF.}

\item{nCluster}{The number of clusters.}

\item{clustering_sol}{The clustering solution as returned by
\code{get_clusters} function.}

\item{seqs}{May not be needed.}

\item{positions}{Labels of the positions in the sequences. Passed on to function
\code{viz_all_factors_in_combined_heatmaps_seqlogos}}
}
\value{

}
\description{
Plot the sequence architectures of the sequence clusters.
}
\examples{

}
