% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_arch_for_clusters.R
\name{plot_arch_for_clusters_new}
\alias{plot_arch_for_clusters_new}
\title{Plot cluster architectures as sequence logos.}
\usage{
plot_arch_for_clusters_new(
  tss.seqs_raw,
  list_of_elements,
  position_labels,
  xt_freq = 1,
  PDFfname = "archR_sequence_architectures.pdf"
)
}
\arguments{
\item{tss.seqs_raw}{Sequences as a DNAStringSet}

\item{list_of_elements}{Cluster elements as a list}

\item{position_labels}{Labels for sequence positions, should be of same
length as length of the sequences}

\item{xt_freq}{Frequency of x-axis ticks}

\item{PDFfname}{Specify a filename to be saved as PDF}
}
\description{
Given a collection of FASTA sequences as a DNAStringSet object,
and the clusters information, this function plots the architectures for all
clusters. If a name for the PDF file is provided, the resulting set of
architecture sequence logos are saved as a multi-page PDF.
}
