% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyNMF_model_selection_functions.R
\name{plot_cv_Alpha}
\alias{plot_cv_Alpha}
\title{Plot cross-validation performance of alpha values}
\usage{
plot_cv_Alpha(averages, threshold = 0)
}
\arguments{
\item{averages}{The cross-validation averages}

\item{threshold}{The threshold to be applied when choosing the best
performing value}
}
\value{
A ggplot object so you can simply call \code{print} or \code{save}
on it later.
}
\description{
Plot cross-validation performance of alpha values
}
\examples{

}
