% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyNMF_model_selection_functions.R
\name{plot_cv_K}
\alias{plot_cv_K}
\title{Title}
\usage{
plot_cv_K(averages)
}
\arguments{
\item{averages}{The average of performance values for different combinations
in grid_search}
}
\value{
A ggplot object so you can simply call \code{print} or \code{save}
on it later.
}
\description{
Title
}
\examples{

}
