% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ggheatmap.R
\name{plot_ggheatmap}
\alias{plot_ggheatmap}
\title{Plot a heatmap using ggplot2.}
\usage{
plot_ggheatmap(pwmMat, position_labels = NULL, savePDFfilename = NULL)
}
\arguments{
\item{pwmMat}{Matrix (usually a PWM, but can be non-normalized/any matrix)
to be represented as a heatmap.}

\item{position_labels}{Labels of the positions in the sequences.}

\item{savePDFfilename}{Name of the file which will be saved as PDF.}
}
\value{
A ggplot object so you can simply call \code{print} or \code{save}
on it later. If \code{savePDFfilename} is given, it is also saved and the
\code{ggplot} object returned.
}
\description{
Plot a heatmap using ggplot2.
}
\examples{


}
