% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ggheatmap.R
\name{plot_ggseqlogo}
\alias{plot_ggseqlogo}
\title{Visualize a given (PWM) matrix as a sequence logo.}
\usage{
plot_ggseqlogo(pwm_mat, method = "custom", pos_lab = NULL, pdf_name = NULL)
}
\arguments{
\item{pwm_mat}{Matrix (usually a PWM, but can be any non-normalized matrix) 
to be represented as a sequence logo.}

\item{method}{For \code{ggseqlogo}; either of 'custom', 'bits', or
'probability'. Default is 'custom'.}

\item{pos_lab}{Labels for sequence positions, should be of same
length as that of the sequences. Default value is NULL, when the 
positions are labeled from 1 to the length of the sequences.}

\item{pdf_name}{Name of the file which will be saved as PDF.}
}
\value{
A ggplot2 object so you can simply call \code{print} or \code{save}
on it later. If \code{pdf_name} is given, it is also saved in addition to 
returning the ggplot object.
}
\description{
Visualize a given (PWM) matrix as a sequence logo.
}
\examples{

res <- readRDS(system.file("extdata", "example_archRresult.rds", 
         package = "archR", mustWork = TRUE))

pwm <- archR::make_dinuc_PWMs(get_clBasVec_m(res,iter=1)[,1], 
                        add_pseudo_counts = FALSE)

plot_ggseqlogo(pwm_mat = pwm)

}
\seealso{
\code{\link{plot_ggheatmap}} for plotting PWMs as heatmaps,
\code{\link{plot_ggseqlogo_of_seqs}} for visualizing a collection of 
sequences by their sequence logo.

Other visualization functions: 
\code{\link{plot_ggheatmap}()},
\code{\link{viz_bas_vec_heatmap_seqlogo}()},
\code{\link{viz_seqs_acgt_mat_from_seqs}()}
}
\concept{visualization functions}
