% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRBrowser.R
\name{plotBrowserTrack}
\alias{plotBrowserTrack}
\title{Plot an ArchR Region Track}
\usage{
plotBrowserTrack(
  ArchRProj = NULL,
  region = NULL,
  groupBy = "Clusters",
  useGroups = NULL,
  plotSummary = c("bulkTrack", "featureTrack", "loopTrack", "geneTrack"),
  sizes = c(10, 1.5, 3, 4),
  features = getPeakSet(ArchRProj),
  loops = getCoAccessibility(ArchRProj),
  geneSymbol = NULL,
  useMatrix = NULL,
  log2Norm = TRUE,
  upstream = 50000,
  downstream = 50000,
  tileSize = 250,
  minCells = 25,
  normMethod = "ReadsInTSS",
  threads = getArchRThreads(),
  ylim = NULL,
  pal = NULL,
  baseSize = 7,
  scTileSize = 0.5,
  scCellsMax = 100,
  borderWidth = 0.4,
  tickWidth = 0.4,
  facetbaseSize = 7,
  geneAnnotation = getGeneAnnotation(ArchRProj),
  title = "",
  verbose = TRUE,
  logFile = createLogFile("plotBrowserTrack")
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{region}{A \code{GRanges} region that indicates the region to be plotted. If more than one region exists in the \code{GRanges} object,
all will be plotted. If no region is supplied, then the \code{geneSymbol} argument can be used to center the plot window at the
transcription start site of the supplied gene.}

\item{groupBy}{A string that indicates how cells should be grouped. This string corresponds to one of the standard or
user-supplied \code{cellColData} metadata columns (for example, "Clusters"). Cells with the same value annotated in this metadata
column will be grouped together and the average signal will be plotted.}

\item{useGroups}{A character vector that is used to select a subset of groups by name from the designated \code{groupBy} column in
\code{cellColData}. This limits the groups to be plotted.}

\item{plotSummary}{A character vector containing the features to be potted. Possible values include "bulkTrack" (the ATAC-seq signal),
"scTrack" (scATAC-seq signal), "featureTrack" (i.e. the peak regions), "geneTrack" (line diagrams of genes with introns and exons shown.
Blue-colored genes are on the minus strand and red-colored genes are on the plus strand), and "loopTrack" (links between a peak and a gene).}

\item{sizes}{A numeric vector containing up to 3 values that indicate the sizes of the individual components passed to \code{plotSummary}.
The order must be the same as \code{plotSummary}.}

\item{features}{A \code{GRanges} object containing the "features" to be plotted via the "featureTrack". This should be thought of as a
bed track. i.e. the set of peaks obtained using \verb{getPeakSet(ArchRProj))}.}

\item{loops}{A \code{GRanges} object containing the "loops" to be plotted via the "loopTrack".
This \code{GRanges} object start represents the center position of one loop anchor and the end represents the center position of another loop anchor.
A "loopTrack" draws an arc between two genomic regions that show some type of interaction. This type of track can be used
to display chromosome conformation capture data or co-accessibility links obtained using \code{getCoAccessibility()}.}

\item{geneSymbol}{If \code{region} is not supplied, plotting can be centered at the transcription start site corresponding to the gene symbol(s) passed here.}

\item{useMatrix}{If supplied geneSymbol, one can plot the corresponding GeneScores/GeneExpression within this matrix. I.E. "GeneScoreMatrix"}

\item{log2Norm}{If supplied geneSymbol, Log2 normalize the corresponding GeneScores/GeneExpression matrix before plotting.}

\item{upstream}{The number of basepairs upstream of the transcription start site of \code{geneSymbol} to extend the plotting window.
If \code{region} is supplied, this argument is ignored.}

\item{downstream}{The number of basepairs downstream of the transcription start site of \code{geneSymbol} to extend the plotting window.
If \code{region} is supplied, this argument is ignored.}

\item{tileSize}{The numeric width of the tile/bin in basepairs for plotting ATAC-seq signal tracks. All insertions in a single bin will be summed.}

\item{minCells}{The minimum number of cells contained within a cell group to allow for this cell group to be plotted. This argument can be
used to exclude pseudo-bulk replicates generated from low numbers of cells.}

\item{normMethod}{The name of the column in \code{cellColData} by which normalization should be performed. The recommended and default value
is "ReadsInTSS" which simultaneously normalizes tracks based on sequencing depth and sample data quality.}

\item{threads}{The number of threads to use for parallel execution.}

\item{ylim}{The numeric quantile y-axis limit to be used for for "bulkTrack" plotting. This should be expressed as \verb{c(lower limit, upper limit)} such as \code{c(0,0.99)}. If not provided, the y-axis limit will be c(0, 0.999).}

\item{pal}{A custom palette (see \code{paletteDiscrete} or \code{ArchRPalettes}) used to override coloring for groups.}

\item{baseSize}{The numeric font size to be used in the plot. This applies to all plot labels.}

\item{scTileSize}{The width of the tiles in scTracks. Larger numbers may make cells overlap more. Default is 0.5 for about 100 cells.}

\item{scCellsMax}{The maximum number of cells for scTracks.}

\item{borderWidth}{The numeric line width to be used for plot borders.}

\item{tickWidth}{The numeric line width to be used for axis tick marks.}

\item{facetbaseSize}{The numeric font size to be used in the facets (gray boxes used to provide track labels) of the plot.}

\item{geneAnnotation}{The \code{geneAnnotation} object to be used for plotting the "geneTrack" object. See \code{createGeneAnnotation()} for more info.}

\item{title}{The title to add at the top of the plot next to the plot's genomic coordinates.}

\item{verbose}{A boolean value that determines whether standard output should be printed.}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
This function will plot the coverage at an input region in the style of a browser track. It allows for normalization of the signal
which enables direct comparison across samples. Note that the genes displayed in these plots are derived from your \code{geneAnnotation}
(i.e. the \code{BSgenome} object you used) so they may not match other online genome browsers that use different gene annotations.
}
