% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Footprinting.R
\name{plotFootprint}
\alias{plotFootprint}
\title{Plot Group Footprints}
\usage{
plotFootprint(input = NULL, positions = NULL, groupBy = "Clusters",
  useGroups = NULL, pal = NULL, flank = 250, flankNorm = 50,
  smoothWindow = 10, nTop = NULL, normMethod = "none",
  threads = 16, verboseHeader = TRUE, verboseAll = FALSE, ...)
}
\arguments{
\item{input}{ArchRProject or previous Footprint Summarized Experiment}

\item{positions}{sample names corresponding to input files}

\item{groupBy}{output names prefix (ie PBMC -> PBMC.arrow)}

\item{useGroups}{geneAnnotation input for TSS Scores etc.}

\item{pal}{genomeAnnotation input for ChromSizes Nucleotide Information etc.}

\item{flank}{min fragments per cell to be filtered for analyses such as tileMat etc.}

\item{flankNorm}{min TSS Score per cell to be filtered for analyses such as tileMat etc.}

\item{smoothWindow}{remove fragments corresponding to cells pass filterFrags and filterTSS}

\item{nTop}{min fragments per cell to be immediately filtered}

\item{normMethod}{normalization method for footprint plot relative to bias}

\item{threads}{number of therads for parallel execution}

\item{verboseHeader}{verbose sections}

\item{verboseAll}{verbose sections and subsections}

\item{...}{additional args}
}
\description{
This function will create an Arrow Files from input files
for downstream analysis
}
