% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Footprinting.R
\name{plotFootprints}
\alias{plotFootprints}
\title{Plot footprints for an ArchRProject}
\usage{
plotFootprints(input = NULL, positions = NULL,
  plotName = "Plot-Footprints", groupBy = "Clusters",
  useGroups = NULL, pal = NULL, flank = 250, flankNorm = 50,
  smoothWindow = 10, minCells = 25, nTop = NULL,
  normMethod = "none", height = 6, width = 4, addDOC = TRUE,
  useSink = TRUE, threads = 1, verboseHeader = TRUE,
  verboseAll = FALSE, ...)
}
\arguments{
\item{input}{An ArchRProject object or Footprint Summarized Experiment}

\item{positions}{QQQ A GenomicRangesList, a list, or a SimpleList object containing the positions to incorporate into the footprint. Each position should be QQQ.}

\item{groupBy}{QQQ The column name in sampleColData to use for grouping multiple samples together prior to footprinting.}

\item{useGroups}{QQQ}

\item{pal}{The name or numeric index of a custom palette from ArchR_palettes to use for plotting the lines corresponding to the footprints.}

\item{flank}{QQQ The number of basepairs from the position center to consider as the flank.}

\item{flankNorm}{QQQ The number of basepairs to consider at the edge of the flank region to be used for footprint normalization.}

\item{smoothWindow}{QQQ The size in basepairs of the sliding window to be used for smoothing of the footprint signal.}

\item{nTop}{QQQ The number of positions to consider. Only the top nTop positions based on QQQ will be considered for the footprint.}

\item{normMethod}{QQQ The name of the normalization method to use to normalize the footprint relative to the Tn5 insertion bias. Options include QQQ.}

\item{threads}{The number threads to be used for parallel computing.}

\item{verboseHeader}{A boolean value that determines whether standard output includes verbose sections.}

\item{verboseAll}{A boolean value that determines whether standard output includes verbose subsections.}

\item{...}{additional args}
}
\description{
This function will plot footprints for all samples in a given ArchRProject or a properly-formatted Summarized Experiment
}
