% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{plotFragmentSizes}
\alias{plotFragmentSizes}
\title{Plot the fragment size distribution for each sample}
\usage{
plotFragmentSizes(
  ArchRProj = NULL,
  nbp = 750,
  returnDF = FALSE,
  threads = getArchRThreads()
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{nbp}{QQQ I FEEL LIKE NBP ISNT A GREAT NAME BUT IM OK WITH IT The axis limit in basepairs to use when plotting the fragment size distribution.}

\item{returnDF}{A boolean value that indicates whether to return a \code{data.frame} containing the plot information
instead of plotting the fragment size distribution.}

\item{threads}{An integer specifying the number of threads to use for calculation. By default this uses the number of threads set by \code{addArchRThreads()}.}
}
\description{
This function will plot a fragment size distribution for each sample. Only cells in the \code{ArchRProject} are used when making this plot.
}
