% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeData.R
\name{plotGroups}
\alias{plotGroups}
\title{Visualize Groups from ArchR Project}
\usage{
plotGroups(
  ArchRProj = NULL,
  groupBy = "Sample",
  colorBy = "colData",
  name = "TSSEnrichment",
  imputeWeights = NULL,
  log2Norm = NULL,
  pal = NULL,
  ylim = NULL,
  size = 0.5,
  baseSize = 6,
  ratioYX = NULL
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{groupBy}{The name of the column in \code{cellColData} to use for grouping cells together for summarizing and plotting.}

\item{colorBy}{A string indicating whether the numeric values to be used in the violin plot should be from a column in \code{cellColData} ("cellColData") or by a data matrix in the ArrowFiles (i.e. "GeneScoreMatrix", "MotifMatrix", "PeakMatrix").}

\item{name}{The name of the column in \code{cellColData} or the featureName/rowname of the data matrix to be used for plotting.
For example if colorBy is \code{cellColData} then name refers to a column name in the cellcoldata (see \code{getCellcoldata()}), if colorBy is "GeneScoreMatrix" then name refers to a gene name which can be listed by \code{getFeatures(ArchRProj, useMatrix = "GeneScoreMatrix")}.}

\item{pal}{A custom palette (see \code{paletteDiscrete} or \code{ArchRPalettes}) used to override discreteSet/continuousSet for coloring vector.}

\item{ylim}{A vector of two numeric values indicating the lower and upper bounds of the y-axis on the plot.}

\item{size}{The numeric size of the points to be plotted.}

\item{baseSize}{The base font size to use in the plot.}

\item{ratioYX}{The aspect ratio of the x and y axes on the plot.}
}
\description{
This function will group, summarize and then plot data from an ArchRProject for visual comparison.
}
