% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkerFeatures.R
\name{plotMarkerHeatmap}
\alias{plotMarkerHeatmap}
\title{Plot a Heatmap of Identified Marker Features}
\usage{
plotMarkerHeatmap(
  seMarker = NULL,
  cutOff = "FDR <= 0.01 & Log2FC >= 0.5",
  log2Norm = TRUE,
  scaleTo = 10^4,
  scaleRows = TRUE,
  plotLog2FC = FALSE,
  limits = c(-2, 2),
  grepExclude = NULL,
  pal = NULL,
  binaryClusterRows = TRUE,
  clusterCols = TRUE,
  subsetMarkers = NULL,
  labelMarkers = NULL,
  nLabel = 15,
  nPrint = 15,
  labelRows = FALSE,
  returnMatrix = FALSE,
  transpose = FALSE,
  invert = FALSE,
  logFile = createLogFile("plotMarkerHeatmap")
)
}
\arguments{
\item{seMarker}{A \code{SummarizedExperiment} object returned by \code{getMarkerFeatures()}.}

\item{cutOff}{A valid-syntax logical statement that defines which marker features from \code{seMarker} will be plotted
in the heatmap. \code{cutoff} can contain any of the \code{assayNames} from \code{seMarker}.}

\item{log2Norm}{A boolean value indicating whether a log2 transformation should be performed on the values in
\code{seMarker} prior to plotting. Should be set to \code{TRUE} for counts-based assays (but not assays like "DeviationsMatrix").
This only applies if \code{plotLog2FC = FALSE}.}

\item{scaleTo}{Each column in the assay Mean from \code{seMarker} will be normalized to a column sum designated by \code{scaleTo}
prior to log2 normalization. If \code{log2Norm = FALSE} this option has no effect.}

\item{scaleRows}{A boolean value that indicates whether the heatmap should display row-wise z-scores instead of raw values.}

\item{plotLog2FC}{A boolean value that indicates whether the \verb{log2(fold change)} values should be plotted on the heatmap or not.
If \code{TRUE}, the \verb{log2(fold change)} value (stored in the assay of \code{seMarker} labeled "Log2FC") is plotted.
If \code{FALSE}, the mean value (stored in the assay of \code{seMarker} labeled "Mean") is plotted instead, with optional log normalization
based on the \code{log2Norm} parameter.}

\item{limits}{A numeric vector of two numbers that represent the lower and upper limits of the heatmap color scheme.}

\item{grepExclude}{A character vector or string that indicates the \code{rownames} or a specific pattern that identifies
rownames from \code{seMarker} to be excluded from the heatmap.}

\item{pal}{A custom continuous palette from \code{ArchRPalettes} (see \code{paletteContinuous()}) used to override the default continuous palette for the heatmap.}

\item{binaryClusterRows}{A boolean value that indicates whether a binary sorting algorithm should be used for fast clustering of heatmap rows.}

\item{clusterCols}{A boolean value that indicates whether the columns of the marker heatmap should be clustered.}

\item{subsetMarkers}{A vector of rownames from seMarker to use for subsetting of seMarker to only plot specific features on the heatmap.
Note that these rownames are expected to be integers that come from \code{rownames(rowData(seMarker))}. If this parameter is used for
subsetting, then the values provided to \code{cutOff} are effectively ignored.}

\item{labelMarkers}{A character vector listing the \code{rownames} of \code{seMarker} that should be labeled on the side of the heatmap.}

\item{nLabel}{An integer value that indicates how many of the top \code{n} features for each column in \code{seMarker} should be labeled on the side of the heatmap.
To remove all feature labels, set \code{nLabel = 0}.}

\item{nPrint}{If provided \code{seMarker} is from "GeneScoreMatrix" print the top \code{n} genes for each group based on how uniquely up-regulated the gene is.}

\item{labelRows}{A boolean value that indicates whether all rows should be labeled on the side of the heatmap.}

\item{returnMatrix}{A boolean value that indicates whether the final heatmap matrix should be returned in lieu of plotting the actual heatmap.}

\item{transpose}{A boolean value that indicates whether the heatmap should be transposed prior to plotting or returning.}

\item{invert}{A boolean value that indicates whether the heatmap will display the features with the
lowest \verb{log2(fold change)}. In this case, the heatmap will display features that are specifically lower in the given cell
group compared to all other cell groups. Additionally, the color palette is inverted for visualization. This is useful when
looking for down-regulated markers (\verb{log2(fold change) < 0}) instead of up-regulated markers (\verb{log2(fold change) > 0}).}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
This function will plot a heatmap of the results from markerFeatures
}
\examples{

#Get Test Project
proj <- getTestProject()

#Get Markers
seMarker <- getMarkerFeatures(
  ArchRProj = proj, 
  useMatrix = "PeakMatrix", 
  testMethod = "binomial", 
  binarize = TRUE
)

#Plot Markers
p <- plotMarkerHeatmap(seMarker)

#PDF
plotPDF(p, name = "Marker-Heatmap", ArchRProj = proj)

}
