% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkerFeatures.R
\name{plotMarkers}
\alias{plotMarkers}
\title{Plot Differential Markers}
\usage{
plotMarkers(
  seMarker = NULL,
  name = NULL,
  cutOff = "FDR <= 0.01 & abs(Log2FC) >= 0.5",
  plotAs = "Volcano",
  scaleTo = 10^4,
  rastr = TRUE
)
}
\arguments{
\item{seMarker}{A \code{SummarizedExperiment} object returned by \code{getMarkerFeatures()}.}

\item{name}{The name of a column in \code{seMarker} (i.e. cell grouping in \code{groupBy} or \code{useGroups} for \code{getMarkerFeatures()}) to be plotted.
To see available options try \code{colnames(seMarker)}.}

\item{cutOff}{A valid-syntax logical statement that defines which marker features from \code{seMarker} will be plotted.
\code{cutoff} can contain any of the \code{assayNames} from \code{seMarker}.}

\item{plotAs}{A string indicating whether to plot a volcano plot ("Volcano") or an MA plot ("MA").}

\item{rastr}{A boolean value that indicates whether the plot should be rasterized using \code{ggrastr}. This does not rasterize
lines and labels, just the internal portions of the plot.}
}
\description{
This function will plot one group/column of a differential markers as an MA or Volcano plot.
}
