% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trajectory.R
\name{plotTrajectory}
\alias{plotTrajectory}
\title{QQQ Visualize a Trajectory from ArchR Project}
\usage{
plotTrajectory(ArchRProj = NULL, embedding = "UMAP",
  trajectory = "Trajectory", colorBy = "colData",
  name = "Trajectory", log2Norm = NULL, pal = NULL, size = 0.5,
  rastr = TRUE, quantCut = c(0.05, 0.95), quantHex = 0.5,
  discreteSet = NULL, continuousSet = NULL, randomize = TRUE,
  keepAxis = FALSE, baseSize = 6, addArrow = TRUE, plotAs = NULL,
  plotParams = list(), ...)
}
\arguments{
\item{ArchRProj}{An `ArchRProject` object.}

\item{embedding}{The name of the embedding to use to visualize the given `trajectory`. See `ArchR::computeEmbedding()` for more information.}

\item{trajectory}{The name of the trajectory as a column in `cellColData` to plot.}

\item{colorBy}{QQQ UNCLEAR A string indicating whether points/hexes in the plot should be colored by cellColData ("cellColData") or by the gene activity score matrix ("GeneScoreMatrix").}

\item{name}{The name of the trajectory as a column in `cellColData` to plot.}

\item{log2Norm}{QQQ A boolean value indicating whether a log2 transformation should be performed on the values in QQQ plotting.}

\item{pal}{QQQ The name or numeric index of a custom palette from `ArchRPalettes` to use for QQQ.}

\item{size}{QQQ A number indicating the size of the points to plot if `plotAs` is set to "points".}

\item{rastr}{A boolean value that indicates whether the plot should be rasterized. This does not rasterize lines and labels, just the internal portions of the plot.}

\item{quantCut}{If this is not null, a quantile cut is performed to threshold the top and bottom of the distribution. This prevents skewed color scales caused by strong outliers. The format of this should be c(x,y) where x is the upper threshold and y is the lower threshold. For example, quantileCut = c(0.975,0.025) will take the top and bottom 2.5% of values and set them to the value of the 97.5th and 2.5th percentile values respectively.}

\item{quantHex}{QQQ UNCLEAR quantile evaluation for each hex in geom_hex}

\item{discreteSet}{QQQ The name or numeric index of a discrete palette from `ArchRPalettes` for visualizing QQQ in the embedding.}

\item{continuousSet}{QQQ The name or numeric index of a continuous palette from `ArchRPalettes` for visualizing QQQ in the embedding.}

\item{randomize}{A boolean value that indicates whether to randomize points prior to plotting to prevent cells from one cluster being present at the front of the plot.}

\item{keepAxis}{A boolean value that indicates whether the x and y axis ticks and labels should be plotted.}

\item{baseSize}{QQQ The base font size to use in the plot.}

\item{addArrow}{QQQ A boolean value that indicates whether QQQ.}

\item{plotAs}{A string that indicates whether points ("points") should be plotted or a hexplot ("hex") should be plotted.}

\item{plotParams}{Additional parameters to pass to `ggPoint()` or `ggHex()`.}

\item{...}{additional args}
}
\description{
QQQ This function will plot a trajectory that was created from QQQ
}
