% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_from_FASTA.R
\name{prepare_data_from_FASTA}
\alias{prepare_data_from_FASTA}
\title{Title}
\usage{
prepare_data_from_FASTA(inputFastaFilename)
}
\arguments{
\item{inputFastaFilename}{}
}
\value{
Returns one-hot-encoded sequences as matrix with dimensions #Features
x #Sequences
}
\description{
Title
}
\examples{

}
