% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BulkProjection.R
\name{projectData}
\alias{projectData}
\title{Project single cell data into single cell subspace}
\usage{
projectData(
  projector = NULL,
  projectee = NULL,
  reducedDims = "IterativeLSI",
  embedding = "UMAP",
  verbose = TRUE,
  threads = getArchRThreads(),
  logFile = createLogFile("projectBulkATAC")
)
}
\arguments{
\item{projector}{An \code{ArchRProject} object containing the dimensionality reduction matrix passed by \code{reducedDims}.}

\item{projectee}{a single cell Summarized Experiment containing the cells to be .}

\item{reducedDims}{A string specifying the reducedDims.}

\item{embedding}{A string specifying embedding.}

\item{verbose}{A boolean value indicating whether to use verbose output during execution of this function. Can be set to FALSE for a cleaner output.}

\item{threads}{The number of threads used for parallel execution}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
This function will Project Bulk ATAC-seq data into single cell subspace.
}
