% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentSemanticIndexing.R
\name{projectLSI}
\alias{projectLSI}
\title{Project LSI}
\usage{
projectLSI(mat, LSI, returnModel = FALSE, verbose = TRUE,
  tstart = NULL, ...)
}
\arguments{
\item{mat}{sparseMatrix (dgcMatrix) for LSI}

\item{LSI}{previous LSI transform to project into}

\item{returnModel}{return projection information}

\item{verbose}{verbose}

\item{tstart}{time stamp to pass}

\item{...}{additional args}
}
\description{
This function will compute a LSI Projection (TF-IDF followed by SVD projection)
}
