% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{recoverArchRProject}
\alias{recoverArchRProject}
\title{Recover ArchRProject if broken sampleColData/cellColData}
\usage{
recoverArchRProject(ArchRProj)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}
}
\description{
This function will recover an ArchRProject if it has broken sampleColData or cellColData due to different versions of bioconductor s4vectors.
}
\examples{

# Get Test Project
proj <- getTestProject()

# Try to Recover ArchR Project
proj <- recoverArchRProject(proj)

}
