% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperUtils.R
\name{reformatFragmentFiles}
\alias{reformatFragmentFiles}
\title{Reformat Fragment Files to be Tabix and Chr Sorted JJJ}
\usage{
reformatFragmentFiles(fragmentFiles = NULL, seqnamesIsChr = TRUE)
}
\arguments{
\item{fragmentFiles}{a character vector of paths to fragment files to be reformatted}

\item{seqnamesIsChr}{a boolean describing to check if seqnames containt "chr".}
}
\description{
This function provides help in reformatting Fragment Files for reading in createArrowFiles.
It will handle weird anomalies found that cause errors in reading tabix bgzip'd fragment files.
}
