% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archR_auxiliary_functionsI.R
\name{reorder_archRresult}
\alias{reorder_archRresult}
\title{Reorder archR raw clustering at the chosen iteration}
\usage{
reorder_archRresult(
  archRresult,
  iteration = 3,
  clustMethod = "hc",
  linkage = "average",
  distMethod = "euclid",
  regularize = TRUE,
  topN = 10,
  returnOrder = FALSE,
  position_agnostic_dist = FALSE,
  config
)
}
\arguments{
\item{archRresult}{The archRresult object.}

\item{iteration}{Specify clusters at which iteration of archR are to be reordered.}

\item{clustMethod}{Specify 'hc' for hierarchical clustering.}

\item{linkage}{One of linkage values as specified for hierarchical clustering.}

\item{distMethod}{Distance measure to be used with hierarchical clustering.}

\item{regularize}{Specify TRUE if regularization is to be performed before 
comparison. See argument 'topN'.}

\item{topN}{Keep only the topN positions for comparing basis vectors.}

\item{returnOrder}{Specify TRUE if only the computed order for hierarchical 
clustering is to be returned}

\item{position_agnostic_dist}{If position agnostic distance measure is to be 
used.}

\item{config}{Pass the configuration of the archR result object.}
}
\value{
Returns ordering returned from hclust or the re-ordered clusters
}
\description{
We use hierarchical clustering for reordering archR raw
clusters
}
