% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represent_matrix_of_acgt.R
\name{represent_matrix_of_acgt}
\alias{represent_matrix_of_acgt}
\title{Represent or Image a matrix of sequences given a data matrix  of size
\#Features x \#Sequences.}
\usage{
represent_matrix_of_acgt(givenMat, positions,
  mark_seq_at_every = as.integer(ncol(givenMat) * 0.1),
  mark_pos_at_every = 10, plot.title = "DNA Sequences",
  savePDFfilename = NA, verbose = 0)
}
\arguments{
\item{givenMat}{The data matrix describing the set of sequences to be
visualized as an image. The current version is applicable only for DNA
sequences.}

\item{positions}{Labels of the positions in the sequences.}

\item{mark_seq_at_every}{Tickmarks for sequences, i.e., these are y-ticks.}

\item{mark_pos_at_every}{Tickmarks for sequence positions, i.e., these are
x-ticks. The tickmarks can be specified for every N-th position or every
percentage N-th position. For every N-th position, simply specify the absolute
positive value. For every percentage N-th position, specify 0.01 for 10\%.}

\item{plot.title}{The title of the plot.}

\item{savePDFfilename}{Name of the file which will be saved as PDF.}

\item{verbose}{Default \code{0} which will not print any messages, or can be
set to \code{1} which will print messages.}
}
\value{

}
\description{
Represent or Image a matrix of sequences given a data matrix  of size
\#Features x \#Sequences.
}
\examples{

}
