% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part4_functions.R
\name{sensitivity_ij}
\alias{sensitivity_ij}
\title{Get total sensitivity of decision i given decision j}
\usage{
sensitivity_ij(data, decision_i, decision_j, value_j, metric, notj = FALSE)
}
\arguments{
\item{data}{data.frame}

\item{decision_i}{string: name of decision variable i}

\item{decision_j}{string: name of decision variable j}

\item{metric}{string: name of performance metric under evaluation}

\item{notj}{logical: pipes direct to the \code{me_ij()} function.}
}
\description{
Get total sensitivity of decision i given decision j
}
