% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalDefaults.R
\name{setArchRLB}
\alias{setArchRLB}
\title{Set the globally-applied load balancing setup}
\usage{
setArchRLB(lb = TRUE)
}
\arguments{
\item{lb}{A boolean to set load balancing. If set to \code{FALSE} then the computation is first divided
to (at most) as many jobs are there are cores and then the jobs are started (aka prescheduled),
each job possibly covering more than one value. If set to \code{TRUE} then one job is forked for each
value of the input \code{X}. The former is better for short computations or large number of values in
the inut \code{X}, the latter is better for jobs that have high variance of completion time and not too
many values of the input \code{X} compared to ArchR Threads. Default is \code{TRUE}, i.e. with load balancing.}
}
\description{
This function will set the load balancing setup when \code{mclapply} (parallelised version of lapply)
is used in ArchR functions.
}
\examples{

# Disable load balancing (prescheduled)
setArchRLB(lb = FALSE)

}
