% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalDefaults.R
\name{setArchRLocking}
\alias{setArchRLocking}
\title{Set a globally-applied H5 file locking setup}
\usage{
setArchRLocking()
}
\description{
This function will set the default H5 file locking parameters to the system
}
