% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalDefaults.R
\name{setArchRPathFormat}
\alias{setArchRPathFormat}
\title{Set the globally-applied outputDirectory path format}
\usage{
setArchRPathFormat(fullPath = TRUE)
}
\arguments{
\item{fullPath}{A boolean value indicating whether full (absolute) path format of the
\code{outputDirectory} is used in ArchR functions.  Default is \code{TRUE}.}
}
\description{
This function will set the path format (absolute vs. relative paths) of the \code{outputDirectory}
used in ArchR functions.
}
\examples{

# Store and use relative path format
setArchRPathFormat(fullPath = FALSE)

}
