% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesUtils.R
\name{shuffleGRanges}
\alias{shuffleGRanges}
\title{Shuffle Genomic Ranges}
\usage{
shuffleGRanges(subject, genome, n, shuffleChr = TRUE)
}
\arguments{
\item{subject}{see validGRanges}

\item{genome}{see validBSgenome}

\item{n}{nPermutations}

\item{shuffleChr}{shuffle across chromosomes randomly vs using previous knowledge of chromosome distribution}
}
\description{
Shuffle Genomic Ranges
}
