% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsify_mat.R
\name{sparsify_mat}
\alias{sparsify_mat}
\title{Sparsify a matrix}
\usage{
sparsify_mat(given_mat, threshold = 0.5)
}
\arguments{
\item{given_mat}{The matrix to be sparsified.}

\item{threshold}{A value between 0.1 and 0.9 inclusive.}
}
\value{
Sparsified matrix of same dimension
}
\description{
This function sparsifies a given matrix below a given a
threshold.
}
