% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{[.ArchRProject}
\alias{[.ArchRProject}
\title{Subset cells directly from ArchRProject}
\usage{
\method{[}{ArchRProject}(x, i, j)
}
\description{
This function will allow adding directly to cellColData with a \code{$} accessor.
}
