% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterCells.R
\name{subsetCells}
\alias{subsetCells}
\title{Subset cells in an ArchRProject.}
\usage{
subsetCells(ArchRProj = NULL, cellNames = NULL)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{cellNames}{A character vector of \code{cellNames} that will be subsetted of the current \code{ArchRProject}.}
}
\description{
This function returns an ArchRProject object that contains a specified subset of cells.
}
