% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesUtils.R
\name{subsetSeqnames}
\alias{subsetSeqnames}
\title{Subset by Seqnames}
\usage{
subsetSeqnames(gr, seqNames, useNames = FALSE)
}
\arguments{
\item{gr}{grange}

\item{seqnames}{seqnames to subset}
}
\description{
Subset by Seqnames
}
