% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesUtils.R
\name{subsetSeqnamesGR}
\alias{subsetSeqnamesGR}
\title{Subset a Genomic Ranges object by the provided seqnames}
\usage{
subsetSeqnamesGR(gr = NULL, names = NULL)
}
\arguments{
\item{gr}{A \code{GRanges} object to be subsetted.}

\item{names}{A character vector containing the \code{seqnames} to keep from the provided \code{GRanges} object.}
}
\description{
This function returns a subsetted Genomic Ranges object based on a vector of provided seqnames
}
