% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trajectory.R
\name{trajectoryHeatmap}
\alias{trajectoryHeatmap}
\title{Plot a Heatmap of Features across a Trajectory}
\usage{
trajectoryHeatmap(seTrajectory = NULL, scaleRows = TRUE,
  limits = c(-2, 2), grepExclude = NULL, pal = NULL,
  labelMarkers = NULL, labelTop = 50, labelRows = FALSE,
  returnMat = FALSE, ...)
}
\arguments{
\item{seTrajectory}{A `SummarizedExperiment` object that results from `markerFeatures()`.}

\item{scaleRows}{A boolean value that indicates whether row-wise z-scores should be computed on matrix in `seTrajectory`.}

\item{limits}{A numeric vector of two numbers that represent the lower and upper color limits of the heatmap color scheme.}

\item{grepExclude}{QQQ A character vector or string that indicates the `rownames` or a specific pattern that identifies rownames from `seTrajectory` to be excluded from the heatmap.}

\item{pal}{QQQ The name or numeric index of a custom palette from `ArchRPalettes` to use for the heatmap}

\item{labelMarkers}{A character vector listing the `rownames` of `seTrajectory` that should be labeled on the side of the heatmap.}

\item{labelTop}{A boolean value that indicates whether the top features for each column in `seTrajectory` should be labeled on the side of the heatmap.}

\item{labelRows}{A boolean value that indicates whether all rows should be labeled on the side of the heatmap.}

\item{returnMat}{QQQ A boolean value that indicates whether the final heatmap matrix should be returned in lieu of plotting the actual heatmap.}

\item{...}{additional args}
}
\description{
This function will plot a heatmap of the results from getTrajectory
}
