% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperUtils.R
\name{validBSgenome}
\alias{validBSgenome}
\title{Get/Validate BSgenome}
\usage{
validBSgenome(genome = NULL, masked = FALSE)
}
\arguments{
\item{genome}{This option must be one of the following: (i) the name of a valid genome (for example "hg38", "hg19", or "mm10"), (ii) the name of a \code{BSgenome} package (for ex. "BSgenome.Hsapiens.UCSC.hg19"), or (iii) a \code{BSgenome} object.}

\item{masked}{A boolean describing whether or not to access the masked version of the selected genome. See \code{BSgenome::getBSgenome()}.}
}
\description{
This function will attempt to get or validate an input as a BSgenome.
}
