% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trajectory.R
\name{VisualizeTrajectory}
\alias{VisualizeTrajectory}
\title{Visualize Embedding from ArchR Project}
\usage{
VisualizeTrajectory(ArchRProj = NULL, embedding = "UMAP",
  trajectory = "Trajectory", colorBy = "colData",
  name = "Trajectory", log2Norm = NULL, pal = NULL, size = 0.5,
  rastr = TRUE, quantCut = c(0.05, 0.95), quantHex = 0.5,
  discreteSet = NULL, continuousSet = NULL, randomize = TRUE,
  keepAxis = FALSE, baseSize = 6, addArrow = TRUE, plotAs = NULL,
  plotParams = list(), ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{embedding}{embedding to visualize (see computeEmbedding)}

\item{colorBy}{colorBy cellColData or Arrays in Arrows (ie GeneScoreMatrix)}

\item{name}{name of column in cellColData or Feature in Array in Arrows}

\item{log2Norm}{log2 Normalize features if they are continuous}

\item{pal}{custom palette to use for plotting}

\item{size}{size of points in plot}

\item{rastr}{rastr points in plot}

\item{quantCut}{quantile cut of continuous features}

\item{quantHex}{quantile evaluation for each hex in geom_hex}

\item{discreteSet}{discrete palette for visualizing embedding}

\item{continuousSet}{continuous palette for visualizing embedding}

\item{randomize}{randomize points prior to plotting}

\item{keepAxis}{keep x and y axis for plot}

\item{baseSize}{base size for text in plot}

\item{plotAs}{how to plot (points vs hex)}

\item{plotParams}{additional params to pass to ggPoint/ggHex}

\item{...}{additional args}
}
\description{
This function will plot an embedding that was created from
computeEmbedding
}
