% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_basis_vectors.R
\name{viz_basis_vectors_in_combined_heatmaps_seqlogos}
\alias{viz_basis_vectors_in_combined_heatmaps_seqlogos}
\alias{viz_basis_vectors_as_seqlogo}
\alias{viz_basis_vectors_as_heatmap}
\title{Visualize the NMF basis vectors in a paired heatmap and sequence logo
plot}
\usage{
viz_basis_vectors_in_combined_heatmaps_seqlogos(
  feat_mat,
  method = "custom",
  pos_lab = NULL,
  add_pseudo_counts = FALSE,
  pdf_name = NULL,
  sinuc_or_dinuc = "sinuc"
)

viz_basis_vectors_as_seqlogo(
  feat_mat,
  method = "custom",
  pos_lab = NULL,
  add_pseudo_counts = FALSE,
  pdf_name = NULL,
  sinuc_or_dinuc = "sinuc"
)

viz_basis_vectors_as_heatmap(
  feat_mat,
  pos_lab = NULL,
  add_pseudo_counts = FALSE,
  pdf_name = NULL,
  sinuc_or_dinuc = "sinuc"
)
}
\arguments{
\item{feat_mat}{The features matrix output from NMF. Expected
dimensionality: number of columns represent the number of factors from NMF,
and the number of rows is 4 times the length of the sequences in the
collection.}

\item{method}{For \code{ggseqlogo} -- either of "custom", "bits", or
"probability". Default is "custom".}

\item{pos_lab}{Labels for sequence positions, should be of same
length as that of the sequences. Default value is NULL, when the 
positions are labeled from 1 to the length of the sequences.}

\item{add_pseudo_counts}{Boolean, taking values TRUE/T or FALSE/F, default
set to FALSE. Setting it to TRUE will enable adding pseudo-counts to the
features matrix.}

\item{pdf_name}{Name of the file which will be saved as PDF
(also provide the extension).}

\item{sinuc_or_dinuc}{"sinuc" or "dinuc"}
}
\value{
nothing

nothing.

nothing.
}
\description{
The given features matrix (with 4 rows) is represented as a
heatmap followed by a sequence logo where the positions are aligned for
better visualization.
}
\section{Functions}{
\itemize{
\item \code{viz_basis_vectors_as_seqlogo}: Visualize the 
NMF basis vectors as a sequence logo

\item \code{viz_basis_vectors_as_heatmap}: Visualize the 
NMF basis vectors as a heatmap
}}

\examples{
res <- readRDS(system.file("extdata", "example_archRresult.rds", 
         package = "archR", mustWork = TRUE))

viz_basis_vectors_in_combined_heatmaps_seqlogos(
                  feat_mat = get_clBasVec_m(res,iter=1))


res <- readRDS(system.file("extdata", "example_archRresult.rds", 
         package = "archR", mustWork = TRUE))

viz_basis_vectors_as_seqlogo(feat_mat = get_clBasVec_m(res,iter=1))
         

res <- readRDS(system.file("extdata", "example_archRresult.rds", 
         package = "archR", mustWork = TRUE))

viz_basis_vectors_as_seqlogo(feat_mat = get_clBasVec_m(res,iter=1))
         

}
\seealso{
Other visualization functions: 
\code{\link{plot_ggheatmap}()},
\code{\link{plot_ggseqlogo}()},
\code{\link{viz_seqs_as_acgt_mat_from_seqs}()}
}
\concept{visualization functions}
