% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_matrix_of_acgt.R
\name{viz_matrix_of_acgt}
\alias{viz_matrix_of_acgt}
\title{Visualize a Matrix of Sequences as an Image.}
\usage{
viz_matrix_of_acgt(
  givenMat,
  position_labels,
  sinuc_or_dinuc = "sinuc",
  choose_colors = c(A = "darkgreen", C = "blue", G = "orange", T = "red", N = "white"),
  annClusters = NULL,
  mark_seq_at_every = as.integer(ncol(givenMat) * 0.1),
  mark_pos_at_every = 10,
  plot.title = "DNA Sequences",
  saveFilename = NULL,
  fileType = "png",
  verbose = 0
)
}
\arguments{
\item{givenMat}{The data matrix describing the set of sequences to be
visualized as an image. The current version is applicable only for DNA
sequences.}

\item{position_labels}{Labels of the positions in the sequences.}

\item{sinuc_or_dinuc}{Specify whether single- or di-nucleotide profiles are
to be visualized.}

\item{choose_colors}{Specify a color scheme for the alphabet as a list.
Default is the color scheme for DNA sequences, i.e., c(A = "darkgreen",
C = "blue", G = "orange", T = "red")}

\item{annClusters}{Provide cluster information for annotation on the matrix.
Format is same as the `$ix` variable from output of `base::sort`.}

\item{mark_seq_at_every}{Tickmarks for sequences, i.e., these are y-ticks.}

\item{mark_pos_at_every}{Tickmarks for sequence positions, i.e., these are
x-ticks. The tickmarks can be specified for every N-th position or every
percentage N-th position. For every N-th position, simply specify the
absolute positive value. For every percentage N-th position, specify 0.01 for
 10\%.}

\item{plot.title}{The title of the plot.}

\item{saveFilename}{Name of the file in which the plot will be saved.
Default `NULL`, shows plot on screen.}

\item{fileType}{Specify file type. Default filetype is 'png'. This is ignored
if `saveFilename` is set to `NULL`.}

\item{verbose}{Default \code{0} which will not print any messages, or can be
set to \code{1} which will print messages.}
}
\value{
nothing.
}
\description{
The data matrix is expected to have features alogn the rows and
sequences along the columns.
}
\seealso{
Other visualization functions: 
\code{\link{plot_ggheatmap}()},
\code{\link{plot_ggseqlogo}()},
\code{\link{viz_basis_vectors_as_heatmap}()},
\code{\link{viz_basis_vectors_as_seqlogo}()},
\code{\link{viz_basis_vectors_in_combined_heatmaps_seqlogos}()}
}
\concept{visualization functions}
