% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_matrix_of_acgt.R
\name{viz_seqs_as_acgt_mat_from_ohe}
\alias{viz_seqs_as_acgt_mat_from_ohe}
\title{Visualize one-hot encoded DNA sequences as an image}
\usage{
viz_seqs_as_acgt_mat_from_ohe(
  givenMat,
  position_labels,
  sinuc_or_dinuc = "sinuc",
  choose_colors = c(A = "darkgreen", C = "blue", G = "orange", T = "red", N = "white"),
  annClusters = NULL,
  mark_seq_at_every = as.integer(ncol(givenMat) * 0.1),
  mark_pos_at_every = 10,
  plot.title = "DNA Sequences",
  saveFilename = NULL,
  fileType = "png",
  verbose = FALSE
)
}
\arguments{
\item{givenMat}{The data matrix describing the set of sequences to be
visualized as an image. The current version is applicable only for DNA
sequences.}

\item{position_labels}{Labels of the positions in the sequences.}

\item{sinuc_or_dinuc}{Specify whether single- or di-nucleotide profiles are
to be visualized.}

\item{choose_colors}{Specify a color scheme for the alphabet as a list.
Default is the color scheme for DNA sequences, i.e., c(A = "darkgreen",
C = "blue", G = "orange", T = "red")}

\item{annClusters}{Provide cluster information for annotation on the matrix.
Format is same as the `$ix` variable from output of `base::sort`.}

\item{mark_seq_at_every}{Tickmarks for sequences, i.e., these are y-ticks.}

\item{mark_pos_at_every}{Tickmarks for sequence positions, i.e., these are
x-ticks. The tickmarks can be specified for every N-th position or every
percentage N-th position. For every N-th position, simply specify the
absolute positive value. For every percentage N-th position, specify 0.01 for
 10\%.}

\item{plot.title}{The title of the plot.}

\item{saveFilename}{Name of the file in which the plot will be saved.
Default `NULL`, shows plot on screen.}

\item{fileType}{Specify file type. Default filetype is 'png'. This is ignored
if `saveFilename` is set to `NULL`.}

\item{verbose}{Logical. Use TRUE to print messages, FALSE otherwise.}
}
\value{
nothing.
}
\description{
Given the one-hot encoded matrix, visualize each sequence as an image. 
The matrix is expected to have features along the rows and
sequences along the columns.
}
\seealso{
\code{\link{viz_seqs_as_acgt_mat_from_seqs}} for visualizing DNA sequences
 passed as a DNAStringSet object.

Other visualization functions: 
\code{\link{plot_ggheatmap}()},
\code{\link{plot_ggseqlogo}()},
\code{\link{viz_basis_vectors_as_heatmap}()},
\code{\link{viz_basis_vectors_as_seqlogo}()},
\code{\link{viz_basis_vectors_in_combined_heatmaps_seqlogos}()},
\code{\link{viz_seqs_as_acgt_mat_from_seqs}()}
}
\concept{visualization functions}
