% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_matrix_of_acgt_image.R
\name{viz_seqs_as_acgt_mat_from_seqs}
\alias{viz_seqs_as_acgt_mat_from_seqs}
\title{Visualize raw DNA sequences as an image}
\usage{
viz_seqs_as_acgt_mat_from_seqs(
  rawSeqs,
  position_labels = NULL,
  xt_freq = 5,
  yt_freq = 100,
  col = c("darkgreen", "blue", "orange", "red"),
  savefilename = NULL,
  filetype = "PNG",
  fwidth = 450,
  fheight = 900,
  funits = "px"
)
}
\arguments{
\item{rawSeqs}{The sequences as a DNAStringSet object.}

\item{position_labels}{The labels to be used for the sequence positions.
Default: Sequence positions are labeled from 1 to the length of the
sequences.}

\item{xt_freq}{The x-axis tick frequency.}

\item{yt_freq}{The y-axis tick frequency.}

\item{col}{A vector of four colors used for the DNA bases A, C, G, and T (in
that order)}

\item{savefilename}{Specify the filename (with extension) for saving the
plot to disk.}

\item{filetype}{Specify the file type, namely PNG, JPEG, TIFF.}

\item{fwidth}{Specify the width for the plot. This depends on the length of
sequences.}

\item{fheight}{Specify the height for the plot. This depends on the number of
sequences.}

\item{funits}{Specify the units in which the height and width are given.}
}
\value{
Nothing returned to the R interpreter.
}
\description{
This function plots the collection of sequences as an image
matrix.
}
\seealso{
Other visualization functions: 
\code{\link{plot_ggheatmap}()},
\code{\link{plot_ggseqlogo}()},
\code{\link{viz_basis_vectors_as_heatmap}()},
\code{\link{viz_basis_vectors_as_seqlogo}()},
\code{\link{viz_basis_vectors_in_combined_heatmaps_seqlogos}()}
}
\concept{visualization functions}
