% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationPeaks.R
\name{addArchRAnnotations}
\alias{addArchRAnnotations}
\title{Add ArchR annotations to an ArchRProject}
\usage{
addArchRAnnotations(
  ArchRProj = NULL,
  db = "ArchR",
  collection = "EncodeTFBS",
  name = collection,
  force = FALSE,
  logFile = createLogFile("addArchRAnnotations")
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{db}{A string indicating the database or a path to a database to use for peak annotation. Options include ArchR,
LOLA, and a valid path to a file of class \code{ArchRAnno}.}

\item{collection}{A string indicating which collection within the database to collect for annotation.
For ArchR, options are "ATAC", "EncodeTFBS", "CistromeTFBS", or "Codex".
For LOLA, options include "EncodeTFBS" "CistromeTFBS", "CistromeEpigenome", "Codex", or "SheffieldDnase".
If supplying a custom \code{ArchRAnno} file please select a valid collection from within that database.}

\item{name}{The name of the \code{peakAnnotation} object to be stored in the \code{ArchRProject}.}

\item{force}{A boolean value indicating whether to force the \code{peakAnnotation} object indicated by \code{name} to be
overwritten if it already exists in the given \code{ArchRProject}.}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
This function adds information about which peaks in the ArchR database contain input regions to a given ArchRProject.
For each peak, a binary value is stored indicating whether each region is observed within the peak region.
}
\examples{

# Get Test ArchR Project
proj <- getTestProject()

# Add Motif Annotations
proj <- addArchRAnnotations(proj, name = "test")

}
