% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{addArchRGenome}
\alias{addArchRGenome}
\title{Add a globally defined genome to all ArchR functions. JJJ}
\usage{
addArchRGenome(genome = NULL)
}
\arguments{
\item{genome}{The default genome to be used for all ArchR functions. This value is stored as a global environment variable, not part of the \code{ArchRProject}. This can be overwritten on a per-function basis using the given function's \code{geneAnnotation} and  \code{genomeAnnotation} parameter.}
}
\description{
This function will set the genome across all ArchR functions.
}
