% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalDefaults.R
\name{addArchRLocking}
\alias{addArchRLocking}
\title{Add a globally-applied H5 file locking setup}
\usage{
addArchRLocking(locking = FALSE)
}
\arguments{
\item{locking}{The default value for H5 File Locking}
}
\description{
This function will set the default H5 file locking parameters
}
\examples{

# Disable/Add ArchR H5 Locking Globally
addArchRLocking(locking=FALSE)

}
