% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{addArchRThreads}
\alias{addArchRThreads}
\title{Add global number of threads for default parallel computing.}
\usage{
addArchRThreads(threads = floor(parallel::detectCores()/2))
}
\arguments{
\item{threads}{default number of threads to be used for parallel execution in ArchR functions by default.}
}
\description{
This function will set the global number of threads to be used for ArchR functions.
}
