% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixDeviations.R
\name{addBgdPeaks}
\alias{addBgdPeaks}
\title{QQQ}
\usage{
addBgdPeaks(ArchRProj, niterations = 50, w = 0.1, binSize = 50,
  seed = 1, outFile = file.path(getOutputDirectory(ArchRProj),
  "Background-Peaks.rds"), force = FALSE, ...)
}
\arguments{
\item{ArchRProj}{QQQ}

\item{niterations}{QQQ}

\item{w}{QQQ}

\item{binSize}{QQQ}

\item{seed}{QQQ}

\item{outFile}{QQQ}

\item{force}{QQQ}
}
\description{
QQQ
}
