% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{addClusters}
\alias{addClusters}
\title{Add cluster information to an ArchRProject}
\usage{
addClusters(input = NULL, reducedDims = "IterativeLSI",
  name = "Clusters", sampleCells = NULL, seed = 1,
  method = "Seurat", dimsToUse = NULL, corCutOff = 0.75,
  knnAssign = 10, nOutlier = 20, verbose = TRUE, tstart = NULL,
  force = FALSE, ...)
}
\arguments{
\item{input}{Either (i) an \code{ArchRProject} object containing the dimensionality reduction matrix passed by \code{reducedDims} or (ii) a dimensionality reduction matrix. This object will be used for cluster identification.}

\item{reducedDims}{The name of the \code{reducedDims} object (i.e. IterativeLSI) to retrieve from the designated \code{ArchRProject}. Not required if input is a matrix.}

\item{name}{The column name of the column to be added to \code{cellColData} if \code{input} is an \code{ArchRProject} object.}

\item{sampleCells}{An integer specifying number of cells to subset perform clustering and assign the remainder cells by euclidean distance.}

\item{seed}{A number to be used as the seed for random number generation required in cluster determination. It is recommended to keep track of the seed used so that you can reproduce results downstream.}

\item{method}{A string indicated the clustering method to be used. Supported methods are "Seurat" and "Scran".}

\item{dimsToUse}{A vector containing the dimensions from the \code{reducedDims} object to use in clustering.}

\item{corCutOff}{A numeric cutoff for the correlation of each dimension to the sequencing depth. If the dimension has a correlation to sequencing depth that is greater than the corCutOff, it will be excluded from analysis.}

\item{knnAssign}{The number of nearest neighbors to be used during clustering for assignment of outliers (clusters with less than nOutlier cells).}

\item{nOutlier}{The minimum number of cells required for a group of cells to be called as a cluster. If a group of cells does not reach this threshold, then the cells will be considered outliers and assigned to nearby clusters.}

\item{verbose}{A boolean value indicating whether to use verbose output during execution of this function. Can be set to FALSE for a cleaner output.}

\item{tstart}{The time at which the function run was started. Useful for keeping track of how long clustering takes relative to a start time.}

\item{force}{A boolean value that indicates whether or not to overwrite data in a given column when the value passed to \code{name} already exists as a column name in \code{cellColData}.}

\item{...}{Additional arguments to be provided to Seurat::FindClusters or scran::buildSNNGraph (for example, knn = 50, jaccard = TRUE)}
}
\description{
This function will identify clusters from a reduced dimensions object in an ArchRProject or from a supplied reduced dimensions matrix.
}
