% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZZ-CoAccessibility.R
\name{addCoAccessibility}
\alias{addCoAccessibility}
\title{Extend Filter then Normalize Scores for Summits}
\usage{
addCoAccessibility(ArchRProj, reducedDims = "IterativeLSI", k = 100,
  knnIteration = 5000, overlapCutoff = 0.8, seed = 1,
  knnMethod = "nabor", maxDist = 250000, scaleTo = 10^4,
  log2Norm = TRUE, ...)
}
\description{
Extend Filter then Normalize Scores for Summits
}
