% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoAccessibility.R
\name{addCoAccessibility}
\alias{addCoAccessibility}
\title{Add Peak Co-Accessibility to ArchR Project}
\usage{
addCoAccessibility(ArchRProj, reducedDims = "IterativeLSI", k = 100,
  knnIteration = 5000, overlapCutoff = 0.8, maxDist = 250000,
  scaleTo = 10^4, log2Norm = TRUE, seed = 1, knnMethod = "nabor",
  ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{reducedDims}{reduced dimensions for KNN groupings}

\item{k}{k-nearest neighbors}

\item{knnIteration}{number of KNN groupings to test overlapCutoff}

\item{overlapCutoff}{overlap maximum between group and previous groups to be added to group list}

\item{maxDist}{maximum distance in bp between peaks for co-accessibility}

\item{scaleTo}{scale group accessibility to prior to computing correlations}

\item{log2Norm}{log2 normalize prior to computing correlations}

\item{seed}{seed for sampling}

\item{knnMethod}{method for KNN computations}

\item{...}{additional args}

\item{threads}{number of threads}
}
\description{
This function will randomly group cells and compute correlations of knn groupings
}
