% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoAccessibility.R
\name{addCoAccessibility}
\alias{addCoAccessibility}
\title{Add Peak Co-Accessibility to an ArchRProject}
\usage{
addCoAccessibility(ArchRProj, reducedDims = "IterativeLSI", k = 100,
  knnIteration = 5000, overlapCutoff = 0.8, maxDist = 250000,
  scaleTo = 10^4, log2Norm = TRUE, seed = 1, knnMethod = "nabor",
  threads = getArchRThreads(), ...)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{reducedDims}{The name of the \code{reducedDims} object (i.e. "IterativeLSI") to retrieve from the designated \code{ArchRProject}.}

\item{k}{The number of k-nearest neighbors to use for creating single cell groups for correlation.}

\item{knnIteration}{The number of KNN groupings to test for passing the supplied \code{overlapCutoff}.}

\item{overlapCutoff}{The maximum allowable overlap between the current group and all previous groups to permit the current group be added to the group list during k-nearest neighbor calculations.}

\item{maxDist}{The maximum allowable distance in basepairs between two peaks to consider for co-accessibility.}

\item{scaleTo}{A numeric value indicating how to scale the accessibility of a single cell group prior to computing co-accessibility correlations.}

\item{log2Norm}{A boolean value indicating whether to log2 transform the single cell groups prior to computing co-accessibility correlations.}

\item{seed}{A number to be used as the seed for random number generation required in cluster determination. It is recommended to keep track of the seed used so that you can reproduce results downstream.}

\item{knnMethod}{The method to be used for k-nearest neighbor computations. Options are "nabor", "RANN", and "FNN" and the corresponding package is required.}

\item{threads}{The number of threads to be used for parallel computing.}

\item{...}{additional args}
}
\description{
This function will add co-accessibility scores to peaks in a given ArchRProject
}
