% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrativeAnalysis.R
\name{addCoAccessibility}
\alias{addCoAccessibility}
\title{Add Peak Co-Accessibility to an ArchRProject}
\usage{
addCoAccessibility(
  ArchRProj = NULL,
  reducedDims = "IterativeLSI",
  dimsToUse = 1:30,
  scaleDims = NULL,
  corCutOff = 0.75,
  k = 100,
  knnIteration = 500,
  overlapCutoff = 0.8,
  maxDist = 1e+05,
  scaleTo = 10^4,
  log2Norm = TRUE,
  seed = 1,
  threads = getArchRThreads(),
  verbose = TRUE,
  logFile = createLogFile("addCoAccessibility")
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{reducedDims}{The name of the \code{reducedDims} object (i.e. "IterativeLSI") to retrieve from the designated \code{ArchRProject}.}

\item{dimsToUse}{A vector containing the dimensions from the \code{reducedDims} object to use in clustering.}

\item{scaleDims}{A boolean value that indicates whether to z-score the reduced dimensions for each cell. This is useful for minimizing
the contribution of strong biases (dominating early PCs) and lowly abundant populations. However, this may lead to stronger sample-specific
biases since it is over-weighting latent PCs. If set to \code{NULL} this will scale the dimensions based on the value of \code{scaleDims} when the
\code{reducedDims} were originally created during dimensionality reduction. This idea was introduced by Timothy Stuart.}

\item{corCutOff}{A numeric cutoff for the correlation of each dimension to the sequencing depth. If the dimension has a correlation to
sequencing depth that is greater than the \code{corCutOff}, it will be excluded from analysis.}

\item{k}{The number of k-nearest neighbors to use for creating single-cell groups for correlation analyses.}

\item{knnIteration}{The number of k-nearest neighbor groupings to test for passing the supplied \code{overlapCutoff}.}

\item{overlapCutoff}{The maximum allowable overlap between the current group and all previous groups to permit the current group be
added to the group list during k-nearest neighbor calculations.}

\item{maxDist}{The maximum allowable distance in basepairs between two peaks to consider for co-accessibility.}

\item{scaleTo}{The total insertion counts from the designated group of single cells is summed across all relevant peak regions from
the \code{peakSet} of the \code{ArchRProject} and normalized to the total depth provided by \code{scaleTo}.}

\item{log2Norm}{A boolean value indicating whether to log2 transform the single-cell groups prior to computing co-accessibility correlations.}

\item{seed}{A number to be used as the seed for random number generation required in cluster determination. It is recommended to keep track
of the seed used so that you can reproduce results downstream.}

\item{threads}{The number of threads to be used for parallel computing.}
}
\description{
This function will add co-accessibility scores to peaks in a given ArchRProject
}
