% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixDeviations.R
\name{addDeviationsMatrix}
\alias{addDeviationsMatrix}
\title{Add a matrix of transcription factor deviations to Arrow Files in ArchRProject}
\usage{
addDeviationsMatrix(ArchRProj, annotations = NULL, matches = NULL,
  bgdPeaks = getBgdPeaks(ArchRProj), matrixName = NULL, out = c("z",
  "deviations"), binarize = FALSE, threads = 1, parallelParam = NULL,
  force = FALSE, ...)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{annotations}{QQQ UNCLEAR WHAT ANNOTATIONS MEANS. The name of the annotations object name stored in the \code{ArchRProject}.}

\item{matrixName}{The name to be used for storage of the deviations matrix in the provided \code{ArchRProject}.}

\item{out}{A string or character vector that indicates whether to save the ouptut matrices as deviations ("deviations") z-scores ("z"), or both (c("deviations","z")).}

\item{binarize}{A boolean value indicating whether the input matrix should be binarized before calculating deviations. This is often desired when working with insertion counts.}

\item{threads}{The number of threads to be used for parallel computing.}

\item{parallelParam}{A list of parameters to be passed for biocparallel/batchtools parallel computing.}

\item{force}{QQQ A boolean value indicating whether to force the matrix indicated by \code{matrixName} to be overwritten if it already exist in the given \code{ArchRProject}.}
}
\description{
This function for each QQQ will independently compute transcription factor deviations.
}
