% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixDeviations.R
\name{addDeviationsMatrix}
\alias{addDeviationsMatrix}
\title{Add a matrix of deviations for a given peakAnnotation to Arrow Files in ArchRProject}
\usage{
addDeviationsMatrix(
  ArchRProj = NULL,
  peakAnnotation = NULL,
  matches = NULL,
  bgdPeaks = getBgdPeaks(ArchRProj),
  matrixName = NULL,
  out = c("z", "deviations"),
  binarize = FALSE,
  threads = getArchRThreads(),
  verbose = TRUE,
  parallelParam = NULL,
  force = FALSE,
  logFile = createLogFile("addDeviationsMatrix")
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{peakAnnotation}{The name of the \code{peakAnnotation} stored in the \code{ArchRProject}.}

\item{matches}{A custom \code{peakAnnotation} matches object used as input for the hypergeometric test. See
\code{motifmatchr::matchmotifs()} for additional information.}

\item{bgdPeaks}{A \code{SummarizedExperiment} that contains for each peak a set of background peaks matched by biases such as total accessibility
and GC nucleotide content. This can be computed using \code{addBgdPeaks} and accessed by \code{getBgdPeaks}.}

\item{matrixName}{The name to be used for storage of the deviations matrix in the provided \code{ArchRProject}.}

\item{out}{A string or character vector that indicates whether to save the ouptut matrices as deviations ("deviations")
z-scores ("z"), or both (c("deviations","z")).}

\item{binarize}{A boolean value indicating whether the input matrix should be binarized before calculating deviations.
This is often desired when working with insertion counts.}

\item{threads}{The number of threads to be used for parallel computing.}

\item{verbose}{A boolean value that determines whether standard output includes verbose sections.}

\item{parallelParam}{A list of parameters to be passed for biocparallel/batchtools parallel computing.}

\item{force}{A boolean value indicating whether to force the matrix indicated by \code{matrixName} to be overwritten if it
already exists in the ArrowFiles associated with the given \code{ArchRProject}.}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
This function will compute peakAnnotation deviations for each ArrowFiles independently while controlling for global biases (low-memory requirement).
}
