% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoubletsScores.R
\name{addDoubletScores}
\alias{addDoubletScores}
\title{Add Doublet Scores to a collection of Arrow files or an ArchRProject}
\usage{
addDoubletScores(input = NULL, useMatrix = "TileMatrix", k = 10,
  nTrials = 5, dimsToUse = 1:25, corCutOff = 0.75,
  knnMethod = "UMAP", UMAPParams = list(),
  LSIParams = list(sampleCells = NULL), outDir = "QualityControl",
  threads = getArchRThreads(), parallelParam = NULL,
  verboseHeader = TRUE, verboseAll = FALSE, ...)
}
\arguments{
\item{input}{An \code{ArchRProject} object or a character vector containing the names of ArrowFiles to be used.}

\item{useMatrix}{The name of the matrix to be used for performing doublet identification analyses. Options include "TileMatrix" and "PeakMatrix".}

\item{k}{The number of cells neighboring a simulated doublet to be considered as putative doublets.}

\item{nTrials}{The number of times to simulate nCell (number of cells in the sample) doublets to use for doublet simulation when calculating doublet scores.}

\item{dimsToUse}{A vector containing the dimensions from the \code{reducedDims} object to use in clustering.}

\item{corCutOff}{A numeric cutoff for the correlation of each dimension to the sequencing depth. If the dimension has a correlation to sequencing depth that is greater than the \code{corCutOff}, it will be excluded from analysis.}

\item{knnMethod}{The name of the dimensionality reduction method to be used for k-nearest neighbors calculation. Possible values are "UMAP" or "SVD".}

\item{UMAPParams}{The list of parameters to pass to the UMAP function if "UMAP" is designated to \code{knnMethod}. See the function \code{umap} in the uwot package.}

\item{LSIParams}{The list of parameters to pass to the \code{IterativeLSI()} function. See \code{IterativeLSI()}.}

\item{outDir}{The relative path to the output directory for relevant plots/results from doublet identification.}

\item{threads}{The number of threads to be used for parallel computing.}

\item{parallelParam}{A list of parameters to be passed for biocparallel/batchtools parallel computing.}

\item{verboseHeader}{A boolean value that determines whether standard output includes verbose sections.}

\item{verboseAll}{A boolean value that determines whether standard output includes verbose subsections.}

\item{...}{additional args}
}
\description{
For each sample in the ArrowFiles or ArchRProject provided, this function will independently assign inferred doublet information
to each cell. This allows for removing strong heterotypic doublet-based clusters downstream. A doublet results from a droplet that
contained two cells, causing the ATAC-seq data to be a mixture of the signal from each cell.
}
