% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoubletsScores.R
\name{addDoubletScores}
\alias{addDoubletScores}
\title{Add Doublet Scores to Arrows/ArchRProject}
\usage{
addDoubletScores(input, useMatrix = "TileMatrix", k = 200,
  nTrials = 100, knnMethod = "UMAP", UMAPParams = list(),
  LSIParams = list(), useClusters = FALSE, outDir = "QualityControl",
  threads = 1, parallelParam = NULL, verboseHeader = TRUE,
  verboseAll = FALSE, ...)
}
\arguments{
\item{input}{ArchRProject or ArrowFiles}

\item{useMatrix}{matrix name for performing analyses}

\item{k}{number of cells nearby a simulated doublet to consider}

\item{nTrials}{number of trials to simulate doublets in thousands}

\item{knnMethod}{dimension reduction to use for KNN (UMAP or SVD)}

\item{UMAPParams}{list of parameters to pass to uwot::umap}

\item{LSIParams}{list of parameters to pass to IterativeLSI}

\item{outDir}{directory to plot and store results from analyses}

\item{threads}{number of threads}

\item{parallelParam}{parallel parameters for batch style execution}

\item{verboseHeader}{verbose sections}

\item{verboseAll}{verbose sections and subsections}

\item{...}{additional args}

\item{threads}{number of therads for parallel execution}

\item{parallelParam}{parallel parameters for batch style execution}
}
\description{
This function for each sample will independently assign inferred doublet information
to each cell for removing strong heterotropic doublet based clusters downstream.
}
