% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Embedding.R
\name{addEmbedding}
\alias{addEmbedding}
\title{Add an embedding of a reduced dimensions object to an ArchRProject}
\usage{
addEmbedding(ArchRProj = NULL, reducedDims = "IterativeLSI",
  embedding = "UMAP", embeddingOut = NULL, dimsToUse = NULL,
  corCutOff = 0.75, saveModel = TRUE, seed = 1, force = FALSE,
  threads = getArchRThreads(), embeddingParams = list(), ...)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{reducedDims}{The name of the \code{reducedDims} object (i.e. IterativeLSI) to use from the designated \code{ArchRProject}.}

\item{embedding}{The type of the embedding to add to the \code{ArchRProject} object. Possible options include "UMAP", "TUMAP", "RTSNE", and "FFRTSNE".}

\item{embeddingOut}{The name for the embedding to be stored as in the \code{ArchRProject} object.}

\item{dimsToUse}{A vector containing the dimensions from the \code{reducedDims} object to use in computing the embedding.}

\item{corCutOff}{A numeric cutoff for the correlation of each dimension to the sequencing depth. If the dimension has a correlation to sequencing depth that is greater than the corCutOff, it will be excluded from analysis.}

\item{saveModel}{A boolean value indicating whether to save UMAP model for later usage such as projection. Only relevant for UMAP.}

\item{seed}{A number to be used as the seed for random number generation required in cluster determination. It is recommended to keep track of the seed used so that you can reproduce results downstream.}

\item{force}{A boolean value that indicates whether or not to overwrite the relevant data in the \code{ArchRProject} object if the given \code{embedding} already exists.}

\item{threads}{The number of threads to use for embedding generation computations.}

\item{embeddingParams}{A list of extra parameters to pass to the designated \code{embedding} function.}

\item{...}{additional args}
}
\description{
This function will compute an embedding and add it to an ArchRProject.
}
