% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeEmbedding.R
\name{addEmbedding}
\alias{addEmbedding}
\title{Add an embedding of a reduced dimensions object to an ArchRProject}
\usage{
addEmbedding(ArchRProj = NULL, reducedDims = "IterativeLSI",
  embedding = "UMAP", embeddingOut = NULL, dimsToUse = NULL,
  corCutOff = 0.75, saveModel = TRUE, seed = 1, force = FALSE,
  threads = floor(detectCores()/2), embeddingParams = list(), ...)
}
\arguments{
\item{ArchRProj}{An `ArchRProject` object.}

\item{reducedDims}{QQQ The name of the `reducedDims` object to use from the designated `ArchRProject`. Possible options include "IterativeLSI", QQQ.}

\item{embedding}{QQQ The name of the embedding to add to the `ArchRProject` object. Possible options include "UMAP", "TUMAP", "RTSNE", and "FFRTSNE".}

\item{embeddingOut}{QQQ}

\item{dimsToUse}{QQQ A vector containing the dimensions from the `reducedDims` object to use in clustering.}

\item{corCutOff}{QQQ A numeric cutoff for the correlation of each dimension to the sequencing depth. If the dimension has a correlation to sequencing depth that is QQQ greater than the corCutOff, it will be excluded from analysis.}

\item{saveModel}{QQQ A boolean value indicating whether QQQ.}

\item{seed}{QQQ A number to be used as the seed for random number generation required in cluster determination. It is recommended to keep track of the seed used so that you can reproduce results downstream.}

\item{force}{A boolean value that indicates whether or not to overwrite the relevant data in the `ArchRProject` object if the given `embedding` already exists.}

\item{threads}{The number of threads to use for embedding generation computations.}

\item{embeddingParams}{A list of extra parameters to pass to the designated `embedding` function.}

\item{...}{additional args}
}
\description{
This function will compute an embedding and add it to an ArchRProject.
}
