% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeEmbedding.R
\name{addEmbedding}
\alias{addEmbedding}
\title{Compute Embedding from Reduced Dimensions in ArchR Project}
\usage{
addEmbedding(ArchRProj = NULL, reducedDims = "IterativeLSI",
  embedding = "UMAP", embeddingOut = NULL, saveModel = TRUE,
  seed = 1, force = FALSE, embeddingParams = list(), ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{reducedDims}{reduced dimensions to use}

\item{embedding}{embedding type (umap, tumap, rtsne, fftrtsne)}

\item{...}{additional args}

\item{colorBy}{colorBy cellColData or Arrays in Arrows (ie GeneScoreMatrix)}

\item{name}{name of column in cellColData or Feature in Array in Arrows}

\item{log2Norm}{log2 Normalize features if they are continuous}

\item{pal}{custom palette to use for plotting}

\item{size}{size of points in plot}

\item{rastr}{rastr points in plot}

\item{quantCut}{quantile cut of continuous features}

\item{quantHex}{quantile evaluation for each hex in geom_hex}

\item{discreteSet}{discrete palette for visualizing embedding}

\item{continuousSet}{continuous palette for visualizing embedding}

\item{randomize}{randomize points prior to plotting}

\item{keepAxis}{keep x and y axis for plot}

\item{baseSize}{base size for text in plot}

\item{plotContinuous}{how to plot continuous features (points and hex)}

\item{plotParams}{additional params to pass to ggPoint/ggHex}

\item{plotWidth}{plot width used for creating a consistent plot independent of legend size}

\item{plotHeight}{plot height used for creating a consistent plot independent of legend size}
}
\description{
This function will plot an embedding that was created from
computeEmbedding
}
