% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{addFeatureCounts}
\alias{addFeatureCounts}
\title{This function will add total counts of scATAC cells in provided features into ArchRProject.}
\usage{
addFeatureCounts(
  ArchRProj = NULL,
  features = NULL,
  name = NULL,
  addRatio = TRUE,
  threads = getArchRThreads()
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{features}{A \code{GRanges} object of features to count scATAC-seq data in.}

\item{name}{A character defining the name of the features. "\code{name}Counts" and "\code{name}Ratio" will be added to the \code{ArchRProject}.}

\item{addRatio}{A boolean indicating whether to add the "\code{name}Ratio" to the \code{ArchRProject}.}

\item{threads}{The number of threads to use for parallel execution.}
}
\description{
This function will add total counts of scATAC cells in provided features into ArchRProject.
}
