% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixFeatures.R
\name{addFeatureMatrix}
\alias{addFeatureMatrix}
\title{Add a feature matrix to an ArchRProject or a set of ArrowFiles}
\usage{
addFeatureMatrix(input = NULL, features = NULL,
  matrixName = "FeatureMatrix", ceiling = 10^9, binarize = FALSE,
  threads = getArchRThreads(), parallelParam = NULL, force = TRUE)
}
\arguments{
\item{input}{An \code{ArchRProject} object or character vector of paths to ArrowFiles.}

\item{features}{A \code{GRanges} object containing the regions (aka features) to use for counting insertions for each cell.}

\item{matrixName}{The name to be used for storage of the feature matrix in the provided \code{ArchRProject} or ArrowFiles.}

\item{ceiling}{The maximum counts per feature allowed. This is used to prevent large biases in feature counts.}

\item{binarize}{A boolean value indicating whether the feature matrix should be binarized prior to storage. This can be useful for
downstream analyses when working with insertion counts.}

\item{threads}{The number of threads to be used for parallel computing.}

\item{parallelParam}{A list of parameters to be passed for biocparallel/batchtools parallel computing.}

\item{force}{A boolean value indicating whether to force the matrix indicated by \code{matrixName} to be overwritten if it already exists in the \code{input}.}
}
\description{
This function for each sample will independently compute counts for each feature per cell in the provided ArchRProject or set of ArrowFiles.
}
