% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAIntegration.R
\name{addGeneIntegrationMatrix}
\alias{addGeneIntegrationMatrix}
\title{Add GeneIntegrationMatrix to ArrowFiles or an ArchRProject JJJ}
\usage{
addGeneIntegrationMatrix(
  ArchRProj = NULL,
  seRNA = NULL,
  groupBy = NULL,
  reducedDims = "IterativeLSI",
  sampleList = SimpleList(),
  sampleRNA = 5000,
  nGenes = 2000,
  useImputation = TRUE,
  reduction = "cca",
  addToArrow = TRUE,
  threads = getArchRThreads(),
  transferParams = list(),
  force = FALSE,
  verboseHeader = TRUE,
  verboseAll = FALSE,
  ...
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{seRNA}{A scRNA \code{SummarizedExperiment} or \code{SeuratObject} to be integrated with.}

\item{groupBy}{A column name in either \code{colData} or \code{metadata} of seRNA to compute prediction scores.}

\item{reducedDims}{The name of the \code{reducedDims} object (i.e. "IterativeLSI") to retrieve from the designated \code{ArchRProject}.}

\item{sampleList}{A named list of sampleNames from \code{ArchRProj} containing cell names in \code{seRNA} that are used for constraining data integration.
For example when integrating hematopoiesis it is useful to partition integration for "CD34", "BMMC" and "PBMC" single cells.}

\item{sampleRNA}{Number of cells from seRNA to be used for integration. This is useful for lowering memory and speeding
up results of \code{Seurat::FindTransferAnchors} when integrating datasets.}

\item{nGenes}{Number of variable genes determined by \code{Seurat::FindVariableGenes} to use for integration.}

\item{useImputation}{Use imputation for Gene Score Matrix prior to integration.}

\item{reduction}{Seurat reduction method for integrating modalities. See \code{Seurat::FindTransferAnchors} for reduction methods.}

\item{addToArrow}{Add integrated matched RNA to ArrowFiles afterwards.}

\item{threads}{The number of threads to be used for parallel computing.}

\item{force}{A boolean value indicating whether to force the matrix indicated by \code{matrixName} to be overwritten if it already exist in the given \code{input}.}

\item{parallelParam}{A list of parameters to be passed for biocparallel/batchtools parallel computing.}

\item{subThreading}{A boolean determining whether possible use threads within each multi-threaded subprocess if greater than the number of input samples.}
}
\description{
This function, for each sample, will independently compute counts for each tile
per cell and then infer gene activity scores.
}
