% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixGeneScores.R
\name{addGeneScoreMatrix}
\alias{addGeneScoreMatrix}
\title{Add GeneScoreMatrix to Arrows/ArchRProject}
\usage{
addGeneScoreMatrix(input = NULL, genes = NULL,
  geneModel = "max(exp(-abs(x)/5000), exp(-1))",
  matrixName = "GeneScoreMatrix", upstream = c(5000, 1e+05),
  downstream = c(5000, 1e+05), tileSize = 500, ceiling = 4,
  useGeneBoundaries = TRUE, scaleTo = 10000, excludeChr = c("chrY",
  "chrM"), blacklist = NULL, threads = 1, parallelParam = NULL,
  force = FALSE, ...)
}
\arguments{
\item{input}{ArchRProject or ArrowFiles}

\item{genes}{genes as a GRanges object}

\item{geneModel}{gene model as a string for weighting peaks for gene score calculation (function of x)}

\item{upstream}{upstream the Gene Start to consider for calculation}

\item{downstream}{downstream the Gene Start to consider for calculation}

\item{tileSize}{tileSize for binning counts prior to gene score calculation}

\item{ceiling}{ceiling of read counts per tile (prevent huge biases)}

\item{scaleTo}{scale gene scores to}

\item{excludeChr}{exclude chromosomes from this analysis}

\item{blacklist}{blacklist GRanges used to remove tiles prior to calculation}

\item{threads}{number of threads}

\item{parallelParam}{parallel parameters for batch style execution}

\item{force}{force overwriting previous TileMatrix in ArrowFile}
}
\description{
This function for each sample will independently compute counts for each tile
per cell and then infer gene activity scores.
}
