% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupCoverages.R
\name{addGroupCoverages}
\alias{addGroupCoverages}
\title{Add Group Coverages to ArchR Project}
\usage{
addGroupCoverages(ArchRProj, groupBy = "Clusters", useLabels = TRUE,
  minCells = 40, maxCells = 500, maxFragments = 25 * 10^6,
  minReplicates = 2, maxReplicates = 5, sampleRatio = 0.8,
  kmerLength = 6, threads = 16, parallelParam = "mclapply",
  force = FALSE, verboseHeader = TRUE, verboseAll = FALSE, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{groupBy}{group cells by this column in cellColData}

\item{useLabels}{use sample labels to create sample guided subgroupings as pseudo replicates}

\item{minCells}{minimum cells per group for coverage files}

\item{maxCells}{maximum cells per group for coverage files}

\item{maxFragments}{maximum fragments per group for coverage files (this prevents large files created for optimizing memory)}

\item{minReplicates}{minimum replicates for group for coverage files}

\item{maxReplicates}{maximum replicates for group for coverage files}

\item{sampleRatio}{sampling ratio for pseudo replicates when needed}

\item{kmerLength}{kmer length for adding Tn5 bias estimation}

\item{threads}{number of threads}

\item{parallelParam}{parallel parameters for batch style execution}

\item{force}{force creating coverage files if existed}

\item{verboseHeader}{verbose sections}

\item{verboseAll}{verbose sections and subsections}

\item{...}{additional args}
}
\description{
This function will merge cells within each group into an insertion
coverage file
}
