% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Harmony.R
\name{addHarmony}
\alias{addHarmony}
\title{Add Harmony Batch Corrected Reduced Dims to an ArchRProject}
\usage{
addHarmony(
  ArchRProj = NULL,
  reducedDims = "IterativeLSI",
  dimsToUse = NULL,
  scaleDims = NULL,
  corCutOff = 0.75,
  name = "Harmony",
  groupBy = "Sample",
  verbose = TRUE,
  force = FALSE,
  ...
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object containing the dimensionality reduction matrix passed by \code{reducedDims}.}

\item{reducedDims}{The name of the \code{reducedDims} object (i.e. "IterativeLSI") to retrieve from the designated \code{ArchRProject}.}

\item{dimsToUse}{A vector containing the dimensions from the \code{reducedDims} object to use in clustering.}

\item{scaleDims}{A boolean that indicates whether to z-score the reduced dimensions for each cell. This is useful forminimizing the contribution
of strong biases (dominating early PCs) and lowly abundant populations. However, this may lead to stronger sample-specific biases since
it is over-weighting latent PCs. If set to \code{NULL} this will scale the dimensions based on the value of \code{scaleDims} when the \code{reducedDims} were
originally created during dimensionality reduction. This idea was introduced by Timothy Stuart.}

\item{corCutOff}{A numeric cutoff for the correlation of each dimension to the sequencing depth. If the dimension has a correlation
to sequencing depth that is greater than the \code{corCutOff}, it will be excluded from analysis.}

\item{name}{The name to store harmony output as a \code{reducedDims} in the \code{ArchRProject} object.}

\item{groupBy}{The name of the column in \code{cellColData} to use for grouping cells together for vars in harmony batch correction.
The value of \code{groupBy} is passed to the \code{vars_use} parameter in \code{harmony::HarmonyMatrix()}. When run through ArchR, this parameter
defines which variables to correct for during batch correction. See \code{harmony::HarmonyMatrix()} for more information.}

\item{verbose}{A boolean value indicating whether to use verbose output during execution of this function. Can be set to FALSE for a cleaner output.}

\item{force}{A boolean value that indicates whether or not to overwrite data in a given column when the value passed to \code{name} already
exists as a column name in \code{cellColData}.}

\item{...}{Additional arguments to be provided to harmony::HarmonyMatrix}
}
\description{
This function will add the Harmony batch-corrected reduced dimensions to an ArchRProject.
}
