% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{addImputeWeights}
\alias{addImputeWeights}
\title{QQQ}
\usage{
addImputeWeights(ArchRProj = NULL, reducedDims = "IterativeLSI",
  dimsToUse = NULL, td = 3, ka = 4, sampleCells = max(5000,
  floor(nCells(ArchRProj)/10)), k = 15, epsilon = 1, weighted = TRUE)
}
\arguments{
\item{ArchRProj}{An `ArchRProject` object.}

\item{reducedDims}{QQQ The name of the `reducedDims` object to retrieve from the designated `ArchRProject`. Options include QQQ. QQQ Not required if input is a matrix.}

\item{dimsToUse}{QQQ A vector containing the dimensions from the `reducedDims` object to use in clustering.}

\item{td}{QQQ}

\item{ka}{QQQ}

\item{sampleCells}{A vector containing the names (QQQ or indicies??) of the cells to which cluster information should be added in `cellColData`.}

\item{k}{QQQ}

\item{epsilon}{QQQ}

\item{weighted}{QQQ}
}
\description{
This function QQQ
}
