% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{addImputeWeights}
\alias{addImputeWeights}
\title{Add Imputation Weights to an ArchRProject}
\usage{
addImputeWeights(ArchRProj = NULL, reducedDims = "IterativeLSI",
  dimsToUse = NULL, scaleDims = NULL, corCutOff = 0.75, td = 3,
  ka = 4, sampleCells = max(5000, floor(nCells(ArchRProj)/50)),
  k = 15, epsilon = 1)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{reducedDims}{The name of the \code{reducedDims} object (i.e. "IterativeLSI") to retrieve from the designated \code{ArchRProject}.}

\item{dimsToUse}{A vector containing the dimensions from the \code{reducedDims} object to use.}

\item{scaleDims}{A boolean that indicates whether to z-score the reduced dimensions for each cell. This is useful forminimizing the contribution
of strong biases (dominating early PCs) and lowly abundant populations. However, this may lead to stronger sample-specific biases since
it is over-weighting latent PCs. If set to \code{NULL} this will scale the dimensions based on the value of \code{scaleDims} when the \code{reducedDims} were
originally created during dimensionality reduction. This idea was introduced by Timothy Stuart.}

\item{corCutOff}{A numeric cutoff for the correlation of each dimension to the sequencing depth. If the dimension has a correlation to
sequencing depth that is greater than the \code{corCutOff}, it will be excluded.}

\item{td}{The diffusion time parameter determines the number of smoothing iterations to be performed (see MAGIC from van Dijk et al Cell 2018).}

\item{ka}{The k-nearest neighbors autotune parameter to equalize the effective number of neighbors for each cell, thereby diminishing
the effect of differences in density. (see MAGIC from van Dijk et al Cell 2018).}

\item{sampleCells}{The number of cells to sub-sample to compute an imputation block. An imputation block is a cell x cell matrix that
describes the linear combination for imputation for numerical values within these cells. ArchR creates many blocks to keep this
cell x cell matrix sparse for memory concerns.}

\item{k}{The number of nearest neighbors for smoothing to use for MAGIC (see MAGIC from van Dijk et al Cell 2018).}

\item{epsilon}{The value for the standard deviation of the kernel for MAGIC (see MAGIC from van Dijk et al Cell 2018).}
}
\description{
This function computes imputations weights that describe each cell as a linear combination of many cells based on a MAGIC diffusion matrix.
}
