% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{addImputeWeights}
\alias{addImputeWeights}
\title{Add Imputation Weights to ArchRProject}
\usage{
addImputeWeights(ArchRProj = NULL, reducedDims = "IterativeLSI",
  dimsToUse = NULL, corCutOff = 0.75, td = 3, ka = 4,
  sampleCells = max(5000, floor(nCells(ArchRProj)/10)), k = 15,
  epsilon = 1, ...)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{reducedDims}{The name of the \code{reducedDims} object (i.e. IterativeLSI) to retrieve from the designated \code{ArchRProject}.}

\item{dimsToUse}{A vector containing the dimensions from the \code{reducedDims} object to use in clustering.}

\item{corCutOff}{A numeric cutoff for the correlation of each dimension to the sequencing depth. If the dimension has a correlation to sequencing depth that is greater than the \code{corCutOff}, it will be excluded.}

\item{td}{diffusion time (number of iterations) for MAGIC}

\item{ka}{kNN autotune parameter for MAGIC}

\item{sampleCells}{number of cells to sample per block of estimated imputation matrix}

\item{k}{number of nearest neighbors to use for MAGIC}

\item{epsilon}{a value for the standard deviation of the kernel for MAGIC}

\item{...}{additional params}
}
\description{
This function computes imputations weights that describe each cell as a linear combination of many cells based on MAGIC diffusion matrix.
}
