% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{addImputeWeights}
\alias{addImputeWeights}
\title{Add TileMatrix to Arrows/ArchRProject}
\usage{
addImputeWeights(ArchRProj = NULL, reducedDims = "IterativeLSI",
  dimsToUse = NULL, td = 3, ka = 4, sampleCells = max(5000,
  floor(nCells(ArchRProj)/10)), k = 15, weighted = TRUE)
}
\arguments{
\item{input}{ArchRProject or ArrowFiles}

\item{chromSizes}{chromomosome sizes used for identifying number of tiles to count}

\item{windowSize}{size for each window to break up each chromosome}

\item{binarize}{save as a Sparse.Binary.Matrix or Sparse.Integer.Matrix}

\item{excludeChr}{exclude chromosomes from this analysis}

\item{threads}{number of threads}

\item{parallelParam}{parallel parameters for batch style execution}

\item{force}{force overwriting previous TileMatrix in ArrowFile}
}
\description{
This function for each sample will independently compute counts for each tile
per cell in the Arrow File
}
