% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeLSI.R
\name{addIterativeLSI}
\alias{addIterativeLSI}
\title{Add an Iterative LSI-based dimensionality reduction to an ArchRProject}
\usage{
addIterativeLSI(ArchRProj = NULL, useMatrix = "TileMatrix",
  name = "IterativeLSI", iterations = 2,
  clusterParams = list(resolution = 0.3, sampleCells = 10000, n.start =
  25), dimsToUse = 1:30, LSIMethod = 2, scaleDims = TRUE,
  corCutOff = 0.75, binarize = TRUE, sampleCells = NULL,
  varFeatures = 50000, selectionMethod = "var", scaleTo = 10000,
  totalFeatures = 5e+05, filterQuantile = 0.995,
  saveIterations = TRUE, outDir = getOutputDirectory(ArchRProj),
  runHarmony = FALSE, harmonyParams = list(),
  threads = getArchRThreads(), seed = 1, verboseHeader = TRUE,
  verboseAll = FALSE, force = FALSE)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{useMatrix}{The name of the data matrix to retrieve from the ArrowFiles associated with the \code{ArchRProject}. Valid options are "TileMatrix" or "PeakMatrix".}

\item{name}{The name to use for storage of the IterativeLSI dimensionality reduction in the \code{ArchRProject} as a \code{reducedDims} object.}

\item{iterations}{The number of LSI iterations to perform.}

\item{clusterParams}{Additional parameters to be passed to \code{addClusters()} for clustering within each iteration. These must be either length 1 or the total number of \code{iterations} - 1.}

\item{dimsToUse}{A vector containing the dimensions from the \code{reducedDims} object to use in clustering.}

\item{LSIMethod}{A number or string indicating the order of operations in the TF-IDF normalization.
Possible values are: 1 or "tf-logidf", 2 or "log(tf-idf)", and 3 or "logtf-logidf".}

\item{scaleDims}{A boolean describing whether to rescale the total variance for each principal component. This is useful for minimizing the contribution of strong biases (dominating early PCs) and lowly abundant populations. However, this may lead to stronger sample-specific biases since it is over-weighting latent PCs.}

\item{corCutOff}{A numeric cutoff for the correlation of each dimension to the sequencing depth. If the dimension has a correlation to sequencing depth that is greater than the \code{corCutOff}, it will be excluded from analysis.}

\item{binarize}{A boolean value indicating whether the matrix should be binarized before running LSI. This is often desired when working with insertion counts.}

\item{sampleCells}{An integer specifying the number of cells to sample in order to perform a sub-sampled LSI and sub-sampled clustering.}

\item{varFeatures}{The number of N variable features to use for LSI. The top N features will be used based on the \code{selectionMethod}.}

\item{selectionMethod}{The selection method to be used for identifying the top variable features. Valid options are "var" for log-variability or "vmr" for variance-to-mean ratio.}

\item{scaleTo}{Each column in the matrix designated by \code{useMatrix} will be normalized to a column sum designated by \code{scaleTo} prior to variance calculation and TF-IDF normalization.}

\item{totalFeatures}{The number of features to consider for use in LSI after ranking the features by the total number of insertions. These features are the only ones used throught the variance identification and LSI. These are an equivalent when using a \code{TileMatrix} to a defined peakSet.}

\item{filterQuantile}{A number \link{0,1} that indicates the quantile above which features should be removed based on insertion counts prior to the first iteration of the iterative LSI paradigm. For example, if \code{filterQuantile = 0.99}, any features above the 99th percentile in insertion counts will be ignored for the first LSI iteration.}

\item{saveIterations}{A boolean value indicating whether the results of each LSI iterations should be saved as compressed \code{.rds} files in the designated \code{outDir}.}

\item{outDir}{The output directory for saving LSI iterations if desired. Default is in the \code{outputDirectory} of the \code{ArchRProject}.}

\item{runHarmony}{A boolean value indicating whether harmony-based batch correction should be run during the LSI iterations.}

\item{harmonyParams}{Additional parameters to be passed to \code{harmony::HarmonyMatrix()}.}

\item{threads}{The number of threads to be used for parallel computing.}

\item{seed}{A number to be used as the seed for random number generation. It is recommended to keep track of the seed used so that you can reproduce results downstream.}

\item{verboseHeader}{A boolean value that determines whether standard output includes verbose sections.}

\item{verboseAll}{A boolean value that determines whether standard output includes verbose subsections.}

\item{force}{A boolean value that indicates whether or not to overwrite relevant data in the \code{ArchRProject} object.}
}
\description{
This function will compute an iterative LSI dimensionality reduction on an ArchRProject.
}
