% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentSemanticIndexing.R
\name{addIterativeLSI}
\alias{addIterativeLSI}
\title{Add an Iterative LSI-based dimensionality reduction to an ArchRProject}
\usage{
addIterativeLSI(ArchRProj = NULL, useMatrix = "TileMatrix",
  reducedDimsOut = "IterativeLSI", iterations = 3, dimsToUse = 1:25,
  corCutOff = 0.75, LSIMethod = 1, binarize = TRUE,
  sampleCells = max(c(floor(nCells(ArchRProj)/4), 5000)),
  varFeatures = 50000, selectionMethod = "var", scaleTo = 10000,
  totalFeatures = 5e+05, filterQuantile = 0.99,
  saveIterations = TRUE, outDir = getOutputDirectory(ArchRProj),
  clusterParams = list(), runHarmony = FALSE, harmonyParams = list(),
  threads = 1, seed = 1, verboseHeader = TRUE, verboseAll = FALSE,
  force = FALSE, ...)
}
\arguments{
\item{ArchRProj}{An `ArchRProject` object.}

\item{useMatrix}{QQQ The name of the data matrix to retrieve from the given ArrowFile. Options include "TileMatrix", "GeneScoreMatrix", QQQ.}

\item{reducedDimsOut}{QQQ The name to use for storage of the IterativeLSI dimensionality reduction in the `ArchRProject` as a `reducedDims` object.}

\item{iterations}{The number of LSI iterations to perform.}

\item{dimsToUse}{QQQ A vector containing the dimensions from the `reducedDims` object to use in clustering.}

\item{corCutOff}{QQQ A numeric cutoff for the correlation of each dimension to the sequencing depth. If the dimension has a correlation to sequencing depth that is QQQ greater than the corCutOff, it will be excluded from analysis.}

\item{LSIMethod}{QQQ}

\item{binarize}{A boolean value indicating whether the matrix should be binarized before running LSI. This is often desired when working with insertion counts.}

\item{sampleCells}{A vector containing the names (QQQ or indicies??) of the cells to which cluster information should be added in `cellColData`.}

\item{varFeatures}{The number of N variable features to use for LSI. The top N features will be used based on the `selectionMethod`.}

\item{selectionMethod}{QQQ The selection method to be used for identifying the top variable features. Valid options are "var" or "vmr" (QQQ what is vmr?).}

\item{scaleTo}{QQQ scaleTo for Cluster Averages for variance calculation}

\item{totalFeatures}{The number of features to consider for use in LSI ranked by the total number of insertion counts.}

\item{filterQuantile}{QQQ Remove features that are above this quantile based on insertion counts prior to initial LSI.}

\item{saveIterations}{A boolean value indicating whether the different LSI iterations should be saved as compressed `.rds` files in the designated `outDir`.}

\item{outDir}{The output directory for saving LSI iterations if desired.}

\item{clusterParams}{Additional parameters to be passed to `ArchR::addClusters()`.}

\item{runHarmony}{QQQ A boolean value indicating whether harmony-based batch correction should be run during the LSI iterations.}

\item{harmonyParams}{Additional parameters to be passed to `harmony::HarmonyMatrix()`.}

\item{threads}{The number of threads to be used for parallel computing.}

\item{seed}{QQQ A number to be used as the seed for random number generation required in cluster determination. It is recommended to keep track of the seed used so that you can reproduce results downstream.}

\item{verboseHeader}{A boolean value that determines whether standard output includes verbose sections.}

\item{verboseAll}{A boolean value that determines whether standard output includes verbose subsections.}

\item{force}{A boolean value that indicates whether or not to overwrite relevant data in the `ArchRProject` object.}

\item{...}{additional args}
}
\description{
This function will compute an iterative LSI dimensionality reduction
on an ArchRProject.
}
