% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrativeAnalysis.R
\name{addPeak2GeneLinks}
\alias{addPeak2GeneLinks}
\title{Add Peak2GeneLinks to an ArchRProject}
\usage{
addPeak2GeneLinks(
  ArchRProj = NULL,
  reducedDims = "IterativeLSI",
  useMatrix = "GeneIntegrationMatrix",
  dimsToUse = 1:30,
  scaleDims = NULL,
  corCutOff = 0.75,
  cellsToUse = NULL,
  k = 100,
  knnIteration = 500,
  overlapCutoff = 0.8,
  maxDist = 250000,
  scaleTo = 10^4,
  log2Norm = TRUE,
  predictionCutoff = 0.4,
  addEmpiricalPval = FALSE,
  seed = 1,
  threads = max(floor(getArchRThreads()/2), 1),
  verbose = TRUE,
  logFile = createLogFile("addPeak2GeneLinks")
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{reducedDims}{The name of the \code{reducedDims} object (i.e. "IterativeLSI") to retrieve from the designated \code{ArchRProject}.}

\item{useMatrix}{The name of the matrix containing gene expression information to be used for determining peak-to-gene links. See \code{getAvailableMatrices(ArchRProj)}}

\item{dimsToUse}{A vector containing the dimensions from the \code{reducedDims} object to use in clustering.}

\item{scaleDims}{A boolean value that indicates whether to z-score the reduced dimensions for each cell. This is useful for minimizing
the contribution of strong biases (dominating early PCs) and lowly abundant populations. However, this may lead to stronger sample-specific
biases since it is over-weighting latent PCs. If set to \code{NULL} this will scale the dimensions based on the value of \code{scaleDims} when the
\code{reducedDims} were originally created during dimensionality reduction. This idea was introduced by Timothy Stuart.}

\item{corCutOff}{A numeric cutoff for the correlation of each dimension to the sequencing depth. If the dimension has a
correlation to sequencing depth that is greater than the \code{corCutOff}, it will be excluded from analysis.}

\item{cellsToUse}{A character vector of cellNames to compute coAccessibility on if desired to run on a subset of the total cells.}

\item{k}{The number of k-nearest neighbors to use for creating single-cell groups for correlation analyses.}

\item{knnIteration}{The number of k-nearest neighbor groupings to test for passing the supplied \code{overlapCutoff}.}

\item{overlapCutoff}{The maximum allowable overlap between the current group and all previous groups to permit the current
group be added to the group list during k-nearest neighbor calculations.}

\item{maxDist}{The maximum allowable distance in basepairs between two peaks to consider for co-accessibility.}

\item{scaleTo}{The total insertion counts from the designated group of single cells is summed across all relevant peak regions
from the \code{peakSet} of the \code{ArchRProject} and normalized to the total depth provided by \code{scaleTo}.}

\item{log2Norm}{A boolean value indicating whether to log2 transform the single-cell groups prior to computing co-accessibility correlations.}

\item{predictionCutoff}{A numeric describing the cutoff for RNA integration to use when picking cells for groupings.}

\item{addEmpiricalPval}{Add empirical p-values based on randomly correlating peaks and genes not on the same seqname.}

\item{seed}{A number to be used as the seed for random number generation required in knn determination. It is recommended
to keep track of the seed used so that you can reproduce results downstream.}

\item{threads}{The number of threads to be used for parallel computing.}

\item{verbose}{A boolean value that determines whether standard output should be printed.}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
This function will add peak-to-gene links to a given ArchRProject
}
