% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationPeaks.R
\name{addPeakAnnotations}
\alias{addPeakAnnotations}
\title{Add peak annotations to an ArchRProject}
\usage{
addPeakAnnotations(
  ArchRProj = NULL,
  regions = NULL,
  name = "Region",
  force = FALSE,
  logFile = createLogFile("addPeakAnnotations")
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{regions}{A named \code{list} of \code{GRanges} that are to be overlapped with the \code{peakSet} in the \code{ArchRProject}.}

\item{name}{The name of \code{peakAnnotation} object to be stored as in \code{ArchRProject}.}

\item{force}{A boolean value indicating whether to force the \code{peakAnnotation} object indicated by \code{name} to be overwritten
if it already exists in the given \code{ArchRProject}.}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
This function adds information about which peaks contain input regions to a given ArchRProject. For each peak, a
binary value is stored indicating whether each region is observed within the peak region.
}
