% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixFeatures.R
\name{addPeakMatrix}
\alias{addPeakMatrix}
\title{Add a Peak Matrix to the ArrowFiles of an ArchRProject}
\usage{
addPeakMatrix(
  ArchRProj = NULL,
  ceiling = 4,
  binarize = FALSE,
  verbose = TRUE,
  threads = getArchRThreads(),
  parallelParam = NULL,
  force = TRUE,
  logFile = createLogFile("addPeakMatrix")
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{ceiling}{The maximum counts per feature allowed. This is used to prevent large biases in peak counts.}

\item{binarize}{A boolean value indicating whether the peak matrix should be binarized prior to storage. This can be useful
for downstream analyses when working with insertion counts.}

\item{verbose}{A boolean value that determines whether standard output includes verbose sections.}

\item{threads}{The number of threads to be used for parallel computing.}

\item{parallelParam}{A list of parameters to be passed for biocparallel/batchtools parallel computing.}

\item{force}{A boolean value indicating whether to force the "PeakMatrix" to be overwritten if it already exist in the given \code{ArchRProject}.}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
This function, for each sample, will independently compute counts for each peak
per cell in the provided ArchRProject using the "PeakMatrix".
}
\examples{

# Get Test ArchR Project
proj <- getTestProject()

# Add Peak Matrix
proj <- addPeakMatrix(proj)

}
