% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixFeatures.R
\name{addPeakMatrix}
\alias{addPeakMatrix}
\title{Add PeakMatrix to Arrows in ArchRProject}
\usage{
addPeakMatrix(ArchRProj, ceiling = 4, binarize = FALSE,
  parallelParam = NULL, threads = 1, force = FALSE, ...)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{ceiling}{QQQ The maximum counts per feature allowed. QQQ WHY?}

\item{binarize}{QQQ A boolean value indicating whether the feature matrix should be binarized QQQ prior to storage. This is often desired when working with insertion counts.}

\item{parallelParam}{A list of parameters to be passed for biocparallel/batchtools parallel computing.}

\item{threads}{The number of threads to be used for parallel computing.}

\item{force}{QQQ A boolean value indicating whether to force the matrix indicated by \code{matrixName} to be overwritten if it already exist in the given \code{ArchRProject} or ArrowFiles.}
}
\description{
This function for each sample will independently compute counts for each peak
per cell in the provided ArchRProject using the "PeakMatrix".
}
