% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{addPeakSet}
\alias{addPeakSet}
\title{Add a peak set to an ArchRProject}
\usage{
addPeakSet(ArchRProj, peakSet, force = FALSE, ...)
}
\arguments{
\item{ArchRProj}{An ArchRProject object.}

\item{peakSet}{A GRanges object containing the set of regions that define all peaks in the desired peak set.}

\item{force}{If a peakSet object has already been added to the given ArchRProject, the value of "force" determines whether or not to overwrite this peakSet.}

\item{...}{additional args}
}
\description{
This function adds a peak set as a GRanges object to a given ArchRProject.
}
