% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{addProjectSummary}
\alias{addProjectSummary}
\title{Add projectSummary tp ArchRProject}
\usage{
addProjectSummary(ArchRProj, name, summary, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{name}{name of summary input}

\item{summary}{summary vector}

\item{...}{additional args}
}
\description{
This function adds info to the projectSummary from an ArchRProject
}
