% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReproduciblePeakSet.R
\name{addReproduciblePeakSet}
\alias{addReproduciblePeakSet}
\title{Add Reproducible Peak Set to ArchR Project}
\usage{
addReproduciblePeakSet(ArchRProj = NULL, groupBy = "Clusters",
  reproducibility = "2", peaksPerCell = 500, maxPeaks = 150000,
  excludeChr = c("chrM", "chrY"), pathToMacs2 = findMacs2(),
  genomeSize = NULL, shift = -75, extsize = 150, method = "q",
  cutOff = 0.05, extendSummits = 250, promoterDist = 500,
  genomeAnno = getGenomeAnnotation(ArchRProj),
  geneAnno = getGeneAnnotation(ArchRProj),
  additionalParams = "--nomodel --nolambda", threads = 1,
  parallelParam = "mclapply", force = FALSE, verboseHeader = TRUE,
  verboseAll = FALSE, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{groupBy}{use groupings for peak calling matching group coverage files}

\item{reproducibility}{reproducibility for peak calling (string that is a function of n)}

\item{peaksPerCell}{number of peaks that can be identified per cell on average}

\item{excludeChr}{exclude chromosomes from peak calling}

\item{pathToMacs2}{path to macs2 executable (see Macs2)}

\item{genomeSize}{genome size for peak calling (see Macs2)}

\item{shift}{shift of Tn5 insertions (<- |   ) (see Macs2)}

\item{extsize}{extension of Tn5 insertions (|<- ->|) (see Macs2)}

\item{method}{significance method for Macs2 (see Macs2)}

\item{cutOff}{significance cutoff for Macs2 (see Macs2)}

\item{extendSummits}{extend peak summits for final fixed-width peaks}

\item{promoterDist}{promoter distance from TSS for annotating peaks}

\item{genomeAnno}{genome annotation for ArchRProject}

\item{geneAnno}{gene annotation for ArchRProject}

\item{additionalParams}{additional parameters to pass to Macs2 (see Macs2)}

\item{threads}{number of threads for parallel execution}

\item{parallelParam}{parallel parameters for batch style execution}

\item{force}{force creating peakSet if existed}

\item{verboseHeader}{verbose sections}

\item{verboseAll}{verbose sections and subsections}

\item{...}{additional args}
}
\description{
This function will get insertions from coverage files call peaks
and merge to get a Union Reproducible Peak Set
}
