% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{addSampleColData}
\alias{addSampleColData}
\title{Add information to sampleColData to an ArchRProject}
\usage{
addSampleColData(ArchRProj, data = NULL, name = NULL, samples = NULL,
  force = FALSE)
}
\arguments{
\item{ArchRProj}{An `ArchRProject` object.}

\item{data}{The data to add to `sampleColData`.}

\item{name}{The column header name to be used for this new data in `sampleColData`. If a column with this name already exists, you may set `force` equal to TRUE to overwrite the data in this column.}

\item{samples}{The names of the samples corresponding to `data`. Typically new data is added to all samples but you may use this argument to only add data to a subset of samples. Samples where `data` is not added are set to `NA`.}

\item{force}{A boolean value that indicates whether or not to overwrite data in a given column when the value passed to `name` already exists as a column name in `sampleColData`.}

\item{...}{additional args}
}
\description{
This function adds new data to sampleColData in ArchRProject.
}
