% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixTiles.R
\name{addTileMatrix}
\alias{addTileMatrix}
\title{Add TileMatrix to Arrows/ArchRProject}
\usage{
addTileMatrix(input = NULL, chromSizes = ifelse(inherits(input,
  "ArchRProject"), chromSizes(input), NULL),
  blacklist = ifelse(inherits(input, "ArchRProject"),
  getBlacklist(input), NULL), tileSize = 500, binarize = TRUE,
  excludeChr = c("chrM", "chrY"), threads = getArchRThreads(),
  parallelParam = NULL, force = FALSE, ...)
}
\arguments{
\item{input}{An \code{ArchRProject} object or character vector of ArrowFiles.}

\item{chromSizes}{A named numeric vector containing the chromsome names and lengths. The default behavior is to retrieve this from the \code{ArchRProject} using \code{ArchR::getChromSizes()}.}

\item{blacklist}{A \code{GRanges} object containing genomic regions to blacklist counting in these tiles. The default behavior is to retrieve this from the \code{ArchRProject} using \code{ArchR::getBlacklist()}.}

\item{tileSize}{The size of the tiles used for binning counts in the \code{TileMatrix}.}

\item{binarize}{A boolean value indicating whether the \code{TileMatrix} should be binarized prior to storage.}

\item{excludeChr}{A character vector containing the \code{seqnames} of the chromosomes that should be excluded from CNV analysis.}

\item{threads}{The number of threads to be used for parallel computing.}

\item{parallelParam}{A list of parameters to be passed for biocparallel/batchtools parallel computing.}

\item{force}{A boolean value indicating whether to force the \code{TileMatrix} to be overwritten if it already exist in the given ArrowFiles.}
}
\description{
This function for each sample will independently compute counts for each tile
per cell in the Arrow File
}
