% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixTiles.R
\name{addTileMatrix}
\alias{addTileMatrix}
\title{Add TileMatrix to ArrowFiles or an ArchRProject}
\usage{
addTileMatrix(
  input = NULL,
  chromSizes = if (inherits(input, "ArchRProject")) getChromSizes(input) else NULL,
  blacklist = if (inherits(input, "ArchRProject")) getBlacklist(input) else NULL,
  tileSize = 500,
  binarize = TRUE,
  excludeChr = c("chrM", "chrY"),
  threads = getArchRThreads(),
  parallelParam = NULL,
  force = FALSE,
  logFile = createLogFile("addTileMatrix")
)
}
\arguments{
\item{input}{An \code{ArchRProject} object or character vector of ArrowFiles.}

\item{chromSizes}{A named numeric vector containing the chromsome names and lengths. The default behavior is to retrieve
this from the \code{ArchRProject} using \code{getChromSizes()}.}

\item{blacklist}{A \code{GRanges} object containing genomic regions to blacklist counting in these tiles. The default behavior
is to retrieve this from the \code{ArchRProject} using \code{getBlacklist()}.}

\item{tileSize}{The size of the tiles used for binning counts in the "TileMatrix".}

\item{binarize}{A boolean value indicating whether the "TileMatrix" should be binarized prior to storage.}

\item{excludeChr}{A character vector containing the \code{seqnames} of the chromosomes that should be excluded from the "TileMatrix".}

\item{threads}{The number of threads to be used for parallel computing.}

\item{parallelParam}{A list of parameters to be passed for biocparallel/batchtools parallel computing.}

\item{force}{A boolean value indicating whether to force the "TileMatrix' to be overwritten if it already exist in the given \code{input}.}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
This function, for each sample, will independently compute counts for each tile
}
