% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trajectory.R
\name{addTrajectory}
\alias{addTrajectory}
\title{Add Supervised Trajectory to an ArchR Project}
\usage{
addTrajectory(ArchRProj, name = "Trajectory", trajectory = NULL,
  groupBy = "Clusters", reducedDims = "IterativeLSI",
  preFilter = 0.1, postFilter = 0.1, dof = 250, spar = 1,
  force = FALSE, ...)
}
\arguments{
\item{ArchRProj}{An `ArchRProject` object.}

\item{name}{A string indicating the name of the fitted trajectory to be added in `cellColData`.}

\item{trajectory}{QQQ UNCLEAR trajectory of groups to constrain supervised fitting (in order)}

\item{groupBy}{initial group column in cellColData to constrain supervised fit}

\item{reducedDims}{A string indicating the name of the `reducedDims` object from the `ArchRProject` that should be used for distance computation.}

\item{preFilter}{QQQ UNCLEAR pre filtering quantile for supervised trajectory fit}

\item{postFilter}{QQQ UNCLEAR post filtering quantile for supervised trajectory fit}

\item{dof}{QQQ The number of degrees of freedom to be used in QQQ.}

\item{spar}{QQQ sparsity}

\item{force}{QQQ A boolean value indicating whether to force the trajactory indicated by `name` to be overwritten if it already exist in the given `ArchRProject`.}

\item{...}{additional args}
}
\description{
This function will fit a supervised trajectory in a lower dimensional space that 
can then be used for downstream analyses.
}
