% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archR.R, R/archR_main.R
\docType{package}
\name{archR}
\alias{archR}
\alias{archR-package}
\title{archR: A package for discovering different architectures of sequence
elements}
\usage{
archR(config, seqsMat, thresholdItr = 2)
}
\arguments{
\item{config}{archR configuration object}

\item{seqsMat}{A matrix. A matrix of one-hot encoded sequences with sequences
along columns}

\item{thresholdItr}{numeric. Specify the number of iterations to perform.
Default is 2}
}
\value{
Object of class archR (or at the moment, a list of lists)
}
\description{
Given a set of DNA sequences, the \code{archR} package enables discovery of
clusters of sequence with characteristic sequence architectures.

Call this function to process a data set using archR.
}
\details{
The archR package provides three categories of important functions:
related to data preparation and manipulation, performing non-negative matrix
factorization, performing clustering, and visualization-related functions.
}
\section{Functions for data preparation and manipulation}{

\itemize{
\item \code{\link{prepare_data_from_FASTA}}
\item \code{\link{get_one_hot_encoded_seqs}}
}
}

\section{Functions for visualizations}{

\itemize{
\item \code{\link{plot_ggseqlogo}}
\item \code{\link{plot_ggheatmap}}
\item \code{\link{viz_basis_vectors_as_heatmap}}
\item \code{\link{viz_basis_vectors_as_seqlogo}}
\item \code{\link{viz_basis_vectors_in_combined_heatmaps_seqlogos}}
\item \code{\link{plot_arch_for_clusters}}
\item \code{\link{viz_matrix_of_acgt}}
}
}

