% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archR.R, R/archR_main.R
\docType{package}
\name{archR}
\alias{archR}
\alias{archR-package}
\title{archR: A package for discovering different architectures of sequence
elements}
\usage{
archR(config, tss.seqs, thresholdItr = 2)
}
\arguments{
\item{config}{archR configuration object}

\item{tss.seqs}{sequences}

\item{thresholdItr}{number of iterations to perform}
}
\value{
Object of class archR (or at the moment, a list of lists)
}
\description{
Given a set of DNA sequences, the \code{archR} package enables discovery of
clusters of sequence with characteristic sequence architectures.

Set archR configuration and call this function to process a
data set using archR.
}
\details{
The archR package provides three categories of important functions:
related to data preparation and manipulation, performing non-negative matrix
factorization, performing clustering, and visualization-related functions.
}
\section{Functions for data preparation and manipulation}{

\itemize{
\item \code{\link{prepare_data_from_FASTA}}
\item \code{\link{make_sinuc_PWMs}}
}
}

\section{Functions for visualizations}{

\itemize{
\item \code{\link{plot_ggseqlogo}}
\item \code{\link{plot_ggheatmap}}
\item \code{\link{viz_all_factors_as_heatmap}}
\item \code{\link{viz_all_factors_as_seqlogo}}
\item \code{\link{viz_all_factors_in_combined_heatmaps_seqlogos}}
\item \code{\link{plot_arch_for_clusters}}
\item \code{\link{represent_matrix_of_acgt_with_ggplot2}}
}
}

