% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRBrowser.R
\name{ArchRBrowser}
\alias{ArchRBrowser}
\title{Launch ArchR Genome Browser}
\usage{
ArchRBrowser(
  ArchRProj = NULL,
  features = getPeakSet(ArchRProj),
  loops = getCoAccessibility(ArchRProj),
  minCells = 25,
  baseSize = 10,
  borderWidth = 0.5,
  tickWidth = 0.5,
  facetbaseSize = 12,
  geneAnnotation = getGeneAnnotation(ArchRProj),
  browserTheme = "cosmo",
  threads = getArchRThreads(),
  verbose = TRUE,
  logFile = createLogFile("ArchRBrowser")
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{features}{A \code{GRanges} object containing the "features" to be plotted via the "featureTrack". This should be thought of as a
bed track. i.e. the set of peaks obtained using \verb{getPeakSet(ArchRProj))}.}

\item{loops}{A \code{GRanges} object containing the "loops" to be plotted via the "loopTrack".
This \code{GRanges} object start represents the center position of one loop anchor and the end represents the center position of another loop anchor.
A "loopTrack" draws an arc between two genomic regions that show some type of interaction. This type of track can be used
to display chromosome conformation capture data or co-accessibility links obtained using \code{getCoAccessibility()}.}

\item{minCells}{The minimum number of cells contained within a cell group to allow for this cell group to be plotted. This argument
can be used to exclude pseudo-bulk replicates generated from low numbers of cells.}

\item{baseSize}{The numeric font size to be used in the plot. This applies to all plot labels.}

\item{borderWidth}{The numeric line width to be used for plot borders.}

\item{tickWidth}{The numeric line width to be used for axis tick marks.}

\item{facetbaseSize}{The numeric font size to be used in the facets (gray boxes used to provide track labels) of the plot.}

\item{geneAnnotation}{The \code{geneAnnotation} object to be used for plotting the "geneTrack" object. See \code{createGeneAnnotation()} for more info.}

\item{browserTheme}{A \code{shinytheme} from shinythemes for viewing the ArchR Browser. If not installed this will be NULL.
To install try devtools::install_github("rstudio/shinythemes").}

\item{threads}{The number of threads to use for parallel execution.}

\item{verbose}{A boolean value that determines whether standard output should be printed.}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
This function will open an interactive shiny session in style of a browser track. It allows for normalization of the signal which
enables direct comparison across samples. Note that the genes displayed in this browser are derived from your \code{geneAnnotation}
(i.e. the \code{BSgenome} object you used) so they may not match other online genome browsers that use different gene annotations.
}
