% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorPalettes.R
\docType{data}
\name{ArchRPalettes}
\alias{ArchRPalettes}
\title{List of color palettes that can be used in plots}
\format{
An object of class \code{list} of length 30.
}
\usage{
ArchRPalettes
}
\description{
A collection of some original and some borrowed color palettes to provide appealing color aesthetics for plots in ArchR
}
\keyword{datasets}
