% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{ArchRProject}
\alias{ArchRProject}
\title{Create ArchRProject from ArrowFiles}
\usage{
ArchRProject(ArrowFiles = NULL, outputDirectory = "ArchR_Output",
  copyArrows = FALSE, geneAnnotation = NULL, genomeAnnotation = NULL,
  showLogo = TRUE)
}
\arguments{
\item{ArrowFiles}{A character vector containing the names of ArrowFiles to be used.}

\item{outputDirectory}{A name for the relative path of the outputDirectory for ArchR results}

\item{copyArrows}{A boolean indicating whether ArrowFiles should be copied into outputDirectory}

\item{showLogo}{A boolean indicating whether to show ArchR Logo after successful creation of an ArchRProject.}

\item{geneAnno}{The geneAnnotation (see createGeneAnnotation) is used for downstream analyses such as calculate TSS Enrichment Scores, Gene Scores, etc.}

\item{genomeAnno}{The genomeAnnotation (see createGenomeAnnotation) is used for downstream analyses for genome information such as nucleotide information or chromosome sizes.}
}
\description{
This function will create an ArchRProject with given ArrowFiles.
}
