% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{ArchRProject}
\alias{ArchRProject}
\title{Create ArchRProject from ArrowFiles}
\usage{
ArchRProject(ArrowFiles = NULL, outputDirectory = "ArchROutput",
  copyArrows = TRUE, geneAnnotation = getGeneAnnotation(),
  genomeAnnotation = getGenomeAnnotation(), showLogo = TRUE,
  threads = getArchRThreads())
}
\arguments{
\item{ArrowFiles}{A character vector containing the relative paths to the ArrowFiles to be used.}

\item{outputDirectory}{A name for the relative path of the outputDirectory for ArchR results. Relative to the current working directory.}

\item{copyArrows}{A boolean value indicating whether ArrowFiles should be copied into \code{outputDirectory}.}

\item{geneAnnotation}{The \code{geneAnnotation} object (see \code{createGeneAnnotation()}) to be used for downstream analyses such as calculating
TSS Enrichment Scores, Gene Scores, etc.}

\item{genomeAnnotation}{The \code{genomeAnnotation} object (see \code{createGenomeAnnotation()}) to be used for downstream analyses requiring
genome information such as nucleotide information or chromosome sizes.}

\item{showLogo}{A boolean value indicating whether to show the ascii ArchR logo after successful creation of an \code{ArchRProject}.}

\item{threads}{The number of threads to use for parallel execution.}
}
\description{
This function will create an ArchRProject from the provided ArrowFiles.
}
