% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRBrowser.R
\name{ArchRRegionTrack}
\alias{ArchRRegionTrack}
\title{Plot ArchR Region Track}
\usage{
ArchRRegionTrack(ArchRProj, region = NULL, groupBy = "Clusters",
  useGroups = NULL, useCoverages = FALSE,
  plotSummary = c("bulkTrack", "featureTrack", "geneTrack"),
  sizes = c(10, 0.5, 4), features = getPeakSet(ArchRProj),
  geneSymbol = NULL, upstream = 50000, downstream = 50000,
  tileSize = 100, normMethod = "ReadsInTSS", threads = 1,
  ylim = NULL, baseSize = 7, borderWidth = 0.4, tickWidth = 0.4,
  facetbaseSize = 7, geneAnno = getGeneAnnotation(ArchRProj),
  title = "", ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{region}{GRanges region that will be plotted in (if more that one first will be selected)}

\item{groupBy}{use groupings for bulk/scTrack}

\item{useGroups}{select a subset of groups for plotting}

\item{useCoverages}{use group coverages for track plotting}

\item{plotSummary}{summary of region track to be plotted}

\item{sizes}{sizes corresponding to plotSummary}

\item{features}{GRanges features to be plotted (ie getPeakSet(ArchRProj))}

\item{geneSymbol}{if region is null plotting can be centered at gene start site corresponding to the gene symbol}

\item{upstream}{bp upstream of geneStart to extend}

\item{downstream}{bp downstream of geneStart to extend}

\item{tileSize}{with of tiles to plot bulk/scTrack}

\item{normMethod}{normMethod normalization column in cellColData to normalize bulkTrack}

\item{threads}{number of threads for parallel execution}

\item{ylim}{y-limits for bulkTrack}

\item{baseSize}{size of font in plot}

\item{borderWidth}{border width in plot}

\item{tickWidth}{axis tick width in plot}

\item{geneAnno}{geneAnnotation for geneTrack}

\item{title}{verbose sections}

\item{...}{additional args}
}
\description{
This function will plot the coverage at an input region
}
