% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archR_helper_functions.R
\name{archRSetConfig}
\alias{archRSetConfig}
\title{Set archR run configuration in this function}
\usage{
archRSetConfig(innerChunkSize = 500, kMin = 2, kMax = 8,
  cvFolds = 5, parallelize = TRUE, nCoresUse = 32,
  nIterationsUse = 200, seedVal = 10208090, alphaBase = 0,
  alphaPow = 1, minSeqs = 25, modSelLogFile = "log.txt",
  flags = list(debugFlag = FALSE, timeFlag = FALSE, verboseFlag = TRUE,
  plotVerboseFlag = FALSE))
}
\arguments{
\item{innerChunkSize}{give}

\item{kMin}{give}

\item{kMax}{give}

\item{cvFolds}{give}

\item{parallelize}{give}

\item{nCoresUse}{give}

\item{nIterationsUse}{give}

\item{seedVal}{give}

\item{alphaBase}{give}

\item{alphaPow}{give}

\item{minSeqs}{give}

\item{modSelLogFile}{give}

\item{flags}{give}
}
\description{
this will set the configuration
}
