% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archR_auxiliary_functionsII.R
\name{archRSetConfig}
\alias{archRSetConfig}
\title{Set archR run configuration}
\usage{
archRSetConfig(
  innerChunkSize = 500,
  kMin = 2,
  kMax = 20,
  modSelType = "stability",
  tol = 10^-3,
  bound = 10^-8,
  cvFolds = 5,
  parallelize = FALSE,
  nCoresUse = NA,
  nIterationsUse = 500,
  alphaBase = 0,
  alphaPow = 1,
  minSeqs = 25,
  checkpointing = TRUE,
  flags = list(debugFlag = FALSE, timeFlag = FALSE, verboseFlag = TRUE, plotVerboseFlag
    = FALSE)
)
}
\arguments{
\item{innerChunkSize}{Numeric. Specify the size of the inner chunks of
sequences.}

\item{kMin}{Numeric. Specify the minimum of the range of values to be tested
for number of NMF basis vectors.}

\item{kMax}{Numeric. Specify the maximum of the range of values to be tested
for number of NMF basis vectors.}

\item{modSelType}{Character. Specify the model selection strategy to be used.
Default is 'stability'. Another option is 'cv' short for cross-validation.
Warning: The cross-validation approach can be time consuming and
computationally expensive than the stability-based approach. The CV approach
may be deprecate in the near future.}

\item{tol}{Numeric. Specify the tolerance value as criterion for choosing the
most appropriate number of NMF factors. Default is 1e-03.}

\item{bound}{Numeric. Specify the lower bound value as criterion for choosing
the most appropriate number of NMF factors. Default is 1e-08.}

\item{cvFolds}{Numeric. Specify the number of cross-validation folds used for
model selection. Only used when modSelType is set to 'cv'.}

\item{parallelize}{Logical. Specify whether to parallelize the procedure.
Note that running archR serially can be time consuming. Consider
parallelizing with at least 2 or 4 cores. If Slurm is available, archR's
graphical user interface, accessed with \code{\link{run_archR_UI}}, enables 
provide all input data, set archR configuration, and directly
submit/monitor slurm jobs.}

\item{nCoresUse}{The number of cores to be used when `parallelize` is set
to TRUE. If `parallelize` is FALSE, nCoresUse is ignored.}

\item{nIterationsUse}{Numeric. Specify the number of bootstrapped iterations
to be performed with NMF.}

\item{alphaBase, alphaPow}{Specify the base value and the power for computing
'alpha' in performing model selection for NMF. alpha = alphaBase^alphaPow.
Alpha specifies the regularization for NMF. Default: 0 and 1 respectively.
Warning: For future.}

\item{minSeqs}{Numeric. Specify the minimum number of sequences, such that
any cluster/chunk of size less than or equal to it will not be further
processed/clustered.}

\item{checkpointing}{Logical. Specify whether to write intermediate checkpoints
to disk as RDS files. Default is TRUE. Checkpoints and the final result are
saved to disk provided the oDir argument is set in \code{\link{archR}}.}

\item{flags}{List with four Logical elements as detailed.
\describe{
  \item{debugFlag}{Whether debug information for the run is printed}
  \item{verboseFlag}{Whether verbose information for the run is printed}
  \item{plotVerboseFlag}{Whether verbose plotting is performed for the run}
  \item{timeFlag}{Whether timing information is printed for the run}
}}
}
\value{
a list with all params for archR set
}
\description{
This function sets the configuration for `archR`.
}
