% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archR_auxiliary_functions.R
\name{collect_cluster_labels}
\alias{collect_cluster_labels}
\title{Get cluster labels for chosen level/iteration}
\usage{
collect_cluster_labels(given_seqsClustLabels, choose_levels = 1)
}
\arguments{
\item{given_seqsClustLabels}{from archR result object}

\item{choose_levels}{choose a level/iteration}
}
\value{
A numeric vector of the same saize as seqsClustLabels, with labels
only up to the chosen iteration
}
\description{
Given a seqsClustLabels, collect cluster labels for sequences at
 the chosen iteration/level.
}
