% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesUtils.R
\name{columnOverlaps}
\alias{columnOverlaps}
\title{QQQ Retreive column metadata for overlapping Genomic Ranges regions}
\usage{
columnOverlaps(query, subject, colname = "score", ignore.strand = TRUE,
  decreasing = TRUE)
}
\arguments{
\item{query}{A \code{GRanges} object to be used as the query in \code{findOverlaps()}.}

\item{subject}{A \code{GRanges} object to be used as the subject in \code{findOverlaps()}.}

\item{colname}{The column name from \code{mcols(gr)} to return for the overlapping regions. Mandatory parameter; cannot be \code{NULL}.}

\item{ignore.strand}{A boolean value indicating whether strandedness should be ignored in \code{findOverlaps()}.}

\item{decreasing}{A boolean value indicating whether the returned \code{data.frame} should be ordered in decreasing order based on \code{colname}.}
}
\description{
QQQ This function returns a data.frame containing the overlapping regions and designated metadata columns from mcols(gr)
}
