% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesUtils.R
\name{columnOverlaps}
\alias{columnOverlaps}
\title{Instead of counting overlaps get columns like max score or etc in query}
\usage{
columnOverlaps(query, subject, colname = "score", ignore.strand = TRUE,
  decreasing = TRUE)
}
\arguments{
\item{query}{granges query}

\item{subject}{granges subject}

\item{colname}{mcols(gr)[[colname]] cannot be null}

\item{decreasing}{for order}
}
\description{
Instead of counting overlaps get columns like max score or etc in query
}
