% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyNMF_model_selection_functions.R
\name{compute_q2}
\alias{compute_q2}
\title{Compute \eqn{Q^2} Value}
\usage{
compute_q2(A, recA)
}
\arguments{
\item{A}{The original matrix}

\item{recA}{The reconstructed matrix}
}
\value{
Q^2, the reconstruction accuracy
}
\description{
Computes the reconstruction accuracy, \eqn{Q^2}, given the
original matrix and the recontructed matrix to check against.
}
