% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperUtils.R
\name{confusionMatrix}
\alias{confusionMatrix}
\title{Create a Confusion Matrix based on two value vectors}
\usage{
confusionMatrix(i = NULL, j = NULL)
}
\arguments{
\item{i}{A character/numeric value vector to see concordance with j.}

\item{j}{A character/numeric value vector to see concordance with i.}
}
\description{
This function creates a confusion matrix based on two value vectors.
}
