% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArrow.R
\name{createArrowFiles}
\alias{createArrowFiles}
\title{Create Arrow Files}
\usage{
createArrowFiles(inputFiles = NULL, sampleNames = NULL,
  outputNames = paste0("./", sampleNames), validBaroces = NULL,
  geneAnno = NULL, genomeAnno = NULL, filterFrags = 1000,
  filterTSS = 4, removeFilteredCells = TRUE, minFrags = 500,
  outDir = "QualityControl", nucLength = 147, TSSParams = list(),
  excludeChr = c("chrM", "chrY"), nChunk = 5, bcTag = "qname",
  gsubExpression = NULL, bamFlag = NULL, offsetPlus = 4,
  offsetMinus = -5, addTileMat = TRUE, TileMatParams = list(),
  addGeneScoreMat = TRUE, GeneScoreMatParams = list(), force = FALSE,
  threads = 1, parallelParam = NULL, verboseHeader = TRUE,
  verboseAll = FALSE, ...)
}
\arguments{
\item{inputFiles}{The names of the input files to use to generate the arrow files. These files can be in any of the following formats: tabix QQQ, QQQ BAM,  or a fragments file). The precise format of each file type QQQ...}

\item{sampleNames}{The names to assign to the samples that correspond to the "inputFiles". Each input file should receive a unique sample name. This list should be in the same order as "inputFiles".}

\item{outputNames}{The prefix to use for output files. Each input file should receive a unique output file name. This list should be in the same order as "inputFiles". For example, if the predix is "PBMC" the output file will be named "PBMC.arrow"}

\item{geneAnno}{QQQ The geneAnnotation in QQQ format to associate with these arrow files. This is used downstream to calculate TSS Scores etc.}

\item{genomeAnno}{QQQ The genomeAnnotation in QQQ format to associate with these arrow files. This is used downstream to collect chromosome sizes and nucleotide information etc.}

\item{filterFrags}{The minimum number of mapped ATAC-seq fragments required per cell to pass filtering for use in downstream analyses.}

\item{filterTSS}{The minimum numeric transcription start site (TSS) enrichment score required for a cell to pass filtering for use in downstream analyses. TSS enrichment score is a measurement of signal-to-background in ATAC-seq.}

\item{removeFilteredCells}{A boolean value that determines whether to remove fragments corresponding to cells that do not pass filterFrags and filterTSS.}

\item{minFrags}{QQQ min fragments per cell to be immediately filtered}

\item{outDir}{The name or path for the output directory for QC-level information and plots for each sample/arrow.}

\item{nucLength}{The length in basepairs that wraps around a nucleosome. This number is used for identifying fragments as sub-nucleosome, mono-nucleosome, or multi-nucleosome spanning}

\item{TSSParams}{QQQ TSS parameters for computing TSS scores}

\item{excludeChr}{The names of chromosomes to be excluded from downstream analyses. In most human/mouse analyses, this includes the mitochondrial DNA (chrM) and the male sex chromosome (chrY). This does, however, not exclude the corresponding fragments from being stored in the .arrow file.}

\item{nChunk}{The number of chunks to divide each chromosome into reading in input files. Higher numbers reduce memory usage but increase compute time.}

\item{bcTag}{The name of the field in the input bam file containing the barcode tag information. See ScanBam in Rsamtools.}

\item{bamFlag}{QQQ A list of bam flags to be used for reading in fragments from input bam files. Fromat should be QQQ. See ScanBam in Rsamtools.}

\item{offsetPlus}{The numeric offset to apply to a "+" stranded Tn5 insertion to account for the precise Tn5 binding site. See Buenrostro et al. Nature Methods 2013.}

\item{offsetMinus}{The numeric offset to apply to a "-" stranded Tn5 insertion to account for the precise Tn5 binding site. Ssee Buenrostro et al. Nature Methods 2013.}

\item{addTileMat}{A boolean value indicating whether to add a "Tile Matrix" to each Arrow file. A Tile Matrix is a counts matrix that, instead of using peaks, uses a fixed-width sliding window of bins across the whole genome.}

\item{TileMatParams}{A list of parameters to pass to the addTileMatrix function. See addTileMatrix for options.}

\item{addGeneScoreMat}{A boolean value indicating whether to add a Gene-Score Matrix to each Arrow file. A Gene-Score Matrix uses ATAC-seq signal proximal to the TSS to estimate gene activity.}

\item{GeneScoreMatParams}{A list of parameters to pass to the addGeneScoreMatrix function. See addGeneScoreMatrix for options.}

\item{force}{A bollean value indicating whether to force arrow files to be overwritten if already exist in outDir.}

\item{threads}{The number threads to be used for parallel computing.}

\item{parallelParam}{QQQ A list of parameters to be used for batch-style parallel computing.}

\item{...}{additional args}
}
\description{
This function will create Arrow Files from input files. These Arrow Files are the main constituent for downstream analysis in ArchR.
}
