% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArrow.R
\name{createArrowFiles}
\alias{createArrowFiles}
\title{Create Arrow Files}
\usage{
createArrowFiles(inputFiles = NULL, sampleNames = NULL,
  outputNames = paste0("./", sampleNames), validBarcodes = NULL,
  geneAnno = NULL, genomeAnno = NULL, filterFrags = 1000,
  filterTSS = 4, removeFilteredCells = TRUE, minFrags = 500,
  outDir = "QualityControl", nucLength = 147, TSSParams = list(),
  excludeChr = c("chrM", "chrY"), nChunk = 5, bcTag = "qname",
  gsubExpression = NULL, bamFlag = NULL, offsetPlus = 4,
  offsetMinus = -5, addTileMat = TRUE, TileMatParams = list(),
  addGeneScoreMat = TRUE, GeneScoreMatParams = list(), force = FALSE,
  threads = 1, parallelParam = NULL, verboseHeader = TRUE,
  verboseAll = FALSE, ...)
}
\arguments{
\item{inputFiles}{The names of the input files to use to generate the arrow files. These files can be in any of the following formats: scATAC tabix fragment files or a bam file).}

\item{sampleNames}{The names to assign to the samples that correspond to the "inputFiles". Each input file should receive a unique sample name. This list should be in the same order as "inputFiles".}

\item{outputNames}{The prefix to use for output files. Each input file should receive a unique output file name. This list should be in the same order as "inputFiles". For example, if the predix is "PBMC" the output file will be named "PBMC.arrow"}

\item{validBarcodes}{A list of validBarcode strings to be used for filtering cells read from each input file (see getValidBarcodes for 10x fragment files).}

\item{geneAnno}{The geneAnnotation (see createGeneAnnotation) to associate with these arrow files. This is used downstream to calculate TSS Enrichment Scores etc.}

\item{genomeAnno}{The genomeAnnotation (see createGenomeAnnotation) format to associate with these arrow files. This is used downstream to collect chromosome sizes and nucleotide information etc.}

\item{filterFrags}{The minimum number of mapped ATAC-seq fragments required per cell to pass filtering for use in downstream analyses.}

\item{filterTSS}{The minimum numeric transcription start site (TSS) enrichment score required for a cell to pass filtering for use in downstream analyses. TSS enrichment score is a measurement of signal-to-background in ATAC-seq.}

\item{removeFilteredCells}{A boolean value that determines whether to remove fragments corresponding to cells that do not pass filterFrags and filterTSS.}

\item{minFrags}{The minimum fragments per cell to be filtered immediately before any QC calculations (such as TSS Enrichment).}

\item{outDir}{The name or path for the output directory for QC-level information and plots for each sample/arrow.}

\item{nucLength}{The length in basepairs that wraps around a nucleosome. This number is used for identifying fragments as sub-nucleosome, mono-nucleosome, or multi-nucleosome spanning}

\item{TSSParams}{TSS parameters for computing TSS Enrichment scores. This includes \code{window} which describes the window centered at each TSS (default 101), the \code{flank} which describes the +/- window size to compute TSS enrichment (default 2000) , the \code{norm} which describes the normalization window size at the flanks to compute TSS enrichment (default 100 i.e. -2000:-1901 and 1901:2000).}

\item{excludeChr}{The names of chromosomes to be excluded from downstream analyses. In most human/mouse analyses, this includes the mitochondrial DNA (chrM) and the male sex chromosome (chrY). This does, however, not exclude the corresponding fragments from being stored in the .arrow file.}

\item{nChunk}{The number of chunks to divide each chromosome into reading in input files. Higher numbers reduce memory usage but increase compute time.}

\item{bcTag}{The name of the field in the input bam file containing the barcode tag information. See ScanBam in Rsamtools.}

\item{gsubExpression}{A regular expression to clean up the barcode tag read in from a bam file. For example if the barcode is appended to the qname (read name) like for Shendure mouse data the gsubExpression would be ":.*" for getting the string after the colon in the qname.}

\item{bamFlag}{A list of bam flags to be used for reading in fragments from input bam files. Format should be scanBamFlag for ScanBam in Rsamtools.}

\item{offsetPlus}{The numeric offset to apply to a "+" stranded Tn5 insertion to account for the precise Tn5 binding site. See Buenrostro et al. Nature Methods 2013.}

\item{offsetMinus}{The numeric offset to apply to a "-" stranded Tn5 insertion to account for the precise Tn5 binding site. See Buenrostro et al. Nature Methods 2013.}

\item{addTileMat}{A boolean value indicating whether to add a "Tile Matrix" to each Arrow file. A Tile Matrix is a counts matrix that, instead of using peaks, uses a fixed-width sliding window of bins across the whole genome.}

\item{TileMatParams}{A list of parameters to pass to the \code{addTileMatrix()} function. See \code{ArchR::addTileMatrix()} for options.}

\item{addGeneScoreMat}{A boolean value indicating whether to add a Gene-Score Matrix to each Arrow file. A Gene-Score Matrix uses ATAC-seq signal proximal to the TSS to estimate gene activity.}

\item{GeneScoreMatParams}{A list of parameters to pass to the \code{\link[=addGeneScoreMatrix]{addGeneScoreMatrix()}} function. See \code{ArchR::addGeneScoreMatrix()} for options.}

\item{force}{A boolean value indicating whether to force arrow files to be overwritten if they already exist in \code{outDir}.}

\item{threads}{The number of threads to be used for parallel computing.}

\item{parallelParam}{A list of parameters to be passed for biocparallel/batchtools parallel computing.}

\item{verboseHeader}{A boolean value that determines whether standard output includes verbose sections.}

\item{verboseAll}{A boolean value that determines whether standard output includes verbose subsections.}

\item{...}{additional args}
}
\description{
This function will create Arrow Files from input files. These Arrow Files are the main constituent for downstream analysis in ArchR.
}
