% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationGenome.R
\name{createGeneAnnotation}
\alias{createGeneAnnotation}
\title{Create a gene annotation object for ArchR}
\usage{
createGeneAnnotation(
  genome = NULL,
  TxDb = NULL,
  OrgDb = NULL,
  genes = NULL,
  exons = NULL,
  TSS = NULL,
  annoStyle = NULL,
  singleStrand = FALSE
)
}
\arguments{
\item{genome}{A string that specifies the genome (ie "hg38", "hg19", "mm10", "mm9"). If \code{genome} is not supplied,
\code{TxDb} and \code{OrgDb} are required. If genome is supplied, \code{TxDb} and \code{OrgDb} will be ignored.}

\item{TxDb}{A \code{TxDb} object (transcript database) from Bioconductor which contains information for gene/transcript coordinates.
For example, from \code{txdb <- TxDb.Hsapiens.UCSC.hg38.knownGene}.}

\item{OrgDb}{An \code{OrgDb} object (organism database) from Bioconductor which contains information for gene/transcript symbols from ids.
For example, from \code{orgdb <- org.Hs.eg.db}.}

\item{genes}{A \code{GRanges} object containing gene coordinates (start to end). Must have a symbols column matching the symbols column of \code{exons}.}

\item{exons}{A \code{GRanges} object containing gene exon coordinates. Must have a symbols column matching the symbols column of \code{genes}.}

\item{TSS}{A \code{GRanges} object containing standed transcription start site coordinates for computing TSS enrichment scores downstream.}

\item{annoStyle}{annotation style to map between gene names and various gene identifiers e.g. "ENTREZID", "ENSEMBL".}

\item{singleStrand}{A boolean for GenomicFeatures::genes(\code{single.strand.genes.only}) parameter}
}
\description{
This function will create a gene annotation object that can be used for creating ArrowFiles or an ArchRProject, etc.
}
\examples{

if (!require("TxDb.Hsapiens.UCSC.hg19.knownGene", quietly = TRUE)) BiocManager::install("TxDb.Hsapiens.UCSC.hg19.knownGene")
if (!require("org.Hs.eg.db", quietly = TRUE)) BiocManager::install("org.Hs.eg.db")
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
library(org.Hs.eg.db)

# Get Txdb
TxDb <- TxDb.Hsapiens.UCSC.hg19.knownGene

# Get OrgDb
OrgDb <- org.Hs.eg.db

# Create Genome Annotation
geneAnno <- createGeneAnnotation(TxDb=TxDb, OrgDb=OrgDb)

# Also can create from a string if BSgenome exists
geneAnno <- createGeneAnnotation("hg19")

}
