% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{createGenomeAnnotation}
\alias{createGenomeAnnotation}
\title{Create Genome Annotation for ArchR}
\usage{
createGenomeAnnotation(genome = NULL, chromSizes = NULL,
  blacklist = NULL, filter = TRUE)
}
\arguments{
\item{genome}{A string that points to a BSgenome or a BSgenome object (ie hg38, BSgenome.Hsapiens.UCSC.hg38).}

\item{chromSizes}{A GRanges of chromosome start and end coordinates.}

\item{blacklist}{A GRanges of regions that should be excluded from analyses due to unwanted biases.}

\item{filter}{A boolean indicating whether non-normal chromosome scaffolds should be excluded.}
}
\description{
This function will create a genome annotation that can be used for createArrowFiles, ArchRProject, etc.
}
