% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationGenome.R
\name{createGenomeAnnotation}
\alias{createGenomeAnnotation}
\title{Create a genome annotation object for ArchR}
\usage{
createGenomeAnnotation(
  genome = NULL,
  chromSizes = NULL,
  blacklist = NULL,
  filter = TRUE,
  filterChr = c("chrM")
)
}
\arguments{
\item{genome}{Either (i) a string that is a valid \code{BSgenome} or (ii) a \code{BSgenome} object (ie "hg38" or "BSgenome.Hsapiens.UCSC.hg38").}

\item{chromSizes}{A \code{GRanges} object containing chromosome start and end coordinates.}

\item{blacklist}{A \code{GRanges} object containing regions that should be excluded from analyses due to unwanted biases.}

\item{filter}{A boolean value indicating whether non-standard chromosome scaffolds should be excluded.
These "non-standard" chromosomes are defined by \code{filterChrGR()} and by manual annotation using the \code{filterChr} parameter.}

\item{filterChr}{A character vector indicating the seqlevels that should be removed if manual removal is desired for certain seqlevels.
If no manual removal is desired, \code{filterChr} should be set to \code{NULL}. If \code{filter} is set to \code{TRUE} but \code{filterChr} is set to \code{NULL},
non-standard chromosomes will still be removed as defined in \code{filterChrGR()}.}
}
\description{
This function will create a genome annotation object that can be used for creating ArrowFiles or an ArchRProject, etc.
}
