% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archR_auxiliary_functionsI.R
\name{.get_clusters_from_hc_using_cutree}
\alias{.get_clusters_from_hc_using_cutree}
\title{Get clusters out of hierarchical clustering object using cutree. 
Used internally by archR.}
\usage{
.get_clusters_from_hc_using_cutree(
  hcObj,
  distMat,
  hStep = 0.05,
  parentChunks = NULL,
  verbose = FALSE
)
}
\arguments{
\item{hcObj}{The hierarchical clustering object as returned by 
\code{\link[stats]{hclust}}.}

\item{distMat}{The distance matrix that was used by hclust, a 
\code{\link[stats]{dist}} object.}

\item{hStep}{Numeric. The step size used to increment height values for 
cutree. Default value is 0.05.}

\item{parentChunks}{List. Specify the factor numbers in the previous 
iteration of archR that factors in the current iteration resulted from. 
Default value is NULL.}

\item{verbose}{Logical. Specify TRUE for verbose output.}
}
\description{
Clusters from a hierarchical clustering object are obtained 
by using cutree at different heights of the tree.
}
