% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakAnnotation.R
\name{enrichHeatmap}
\alias{enrichHeatmap}
\title{Heatmap of Peak Annotation Hypergeometric Enrichment in Marker Peaks.}
\usage{
enrichHeatmap(seEnrich = NULL, pal = paletteContinuous(set =
  "white_blue_purple", n = 100), limits = c(0, 60), n = 10,
  clusterCols = TRUE, clusterRows = TRUE, labelRows = TRUE, ...)
}
\arguments{
\item{pal}{A custom continuous palette (see paletteContinuous) used to override the continuous palette for the heatmap.}

\item{limits}{A numeric vector of two numbers that represent the lower and upper color limits of the heatmap color scheme.}

\item{n}{number of top enriched peakAnnotations per column to keep for visibility purposes}

\item{clusterCols}{A boolean indicating whether or not to cluster columns in heatmap.}

\item{clusterRows}{A boolean indicating whether or not to cluster rows in heatmap.}

\item{labelRows}{A boolean indicating whether or not to label rows in heatmap.}

\item{...}{additional args}

\item{seMarker}{A \code{SummarizedExperiment} object returned by \code{ArchR::peakAnnoEnrichment()}.}
}
\description{
This function will plot a heatmap of hypergeometric enrichment of peakAnnotation within the defined Marker Peaks (see peakAnnoEnrichment).
}
