% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part3_functions.R
\name{enumerate}
\alias{enumerate}
\title{Enumerate Your Decision Tree}
\usage{
enumerate(d = NULL, c = NULL, .id = FALSE)
}
\arguments{
\item{d}{data.frame of decision-alternative items}

\item{c}{data.frame of constraints}
}
\description{
A function to enumerate a simple decision tree, under very specific formatting. Soon to be deprecated.
Encouraged to use new functions like \code{enumerate_binary()}, \code{enumerate_ds()}, etc.
}
\author{
Tim Fraser
}
