% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data matrix}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}
}
\usage{
example_data
}
\description{
An example data matrix generated from example_data.fa FASTA file. This is a
collection of 1000 sequences. There are four clusters present in them. The
vignette explains this further.
}
\keyword{datasets}
