% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesUtils.R
\name{extendGRanges}
\alias{extendGRanges}
\title{QQQ Extend regions from a Genomic Ranges object}
\usage{
extendGRanges(x, upstream, downstream)
}
\arguments{
\item{x}{A \code{GRanges} object.}

\item{upstream}{The number of basepairs upstream (5') to extend each region in \code{x}. Strand-aware.}

\item{downstream}{The number of basepairs downstream (3') to extend each region in \code{x}. Strand-aware.}
}
\description{
QQQ This function extends each region in a Genomic Ranges object by a designated upstream and downstream extension in a strand-aware fashion
}
