% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterCells.R
\name{filterCells}
\alias{filterCells}
\title{Filter cells in an ArchRProject JJJ}
\usage{
filterCells(ArchRProj = NULL, cellNames = NULL)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{cellNames}{A character vector of cellNames that will be subsetted of the current \code{ArchRProject}.}
}
\description{
This function returns an ArchRProject object with a subset of cells desired.
}
