% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterCells.R
\name{FilterCells}
\alias{FilterCells}
\title{Extend Filter then Normalize Scores for Summits}
\usage{
FilterCells(ArchRProj, filterList)
}
\arguments{
\item{df}{dataframe where first column is sample names 2nd column is group information and 3rd column is MACS2 summit files}

\item{genome}{mm9, hg19 character or BSgenome object}

\item{blacklist}{regions to blacklist}

\item{extend}{how to extend summits (summit +- extend)}

\item{scorePerMillion}{normalized Score-per-million minimum to keep}

\item{selectionRules}{string with a formula containing n (majority = (n+1)/2, multiple samples = 2)}
}
\description{
Extend Filter then Normalize Scores for Summits
}
