% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterCells.R
\name{filterCells}
\alias{filterCells}
\title{Filter cells in an ArchRProject}
\usage{
filterCells(ArchRProj, filterList, ...)
}
\arguments{
\item{ArchRProj}{an ArchR Project object}

\item{filterList}{list of filters for filtering cells from cellColData}

\item{...}{additional params}
}
\description{
This function plots a list of filters to see which cells would pass filter.
}
