% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterCells.R
\name{filterDoublets}
\alias{filterDoublets}
\title{Filter Doublets From an ArchRProject}
\usage{
filterDoublets(ArchRProj, cutEnrich = 1, cutScore = -Inf,
  filterRatio = 1, ...)
}
\arguments{
\item{ArchRProj}{An `ArchRProject` object}

\item{cutEnrich}{The minimum numeric cutoff for `DoubletEnrichment`. This number is equivalent to the number of simulated doublets identified as a nearest neighbor to the cell divided by the expected number given a random uniform distribution.}

\item{cutScore}{The minimum numeric cutoff for `DoubletScore` which represents the -log10(binomial adjusted p-value) for the `DoubletEnrichment`.}

\item{filterRatio}{JJJ The maximum ratio of predicted doublets to filter based on the number of pass-filter cells. If there are 10,000 cells the maximum would be filterRatio * 10,000^2 / (100,000).}

\item{...}{additional params}
}
\description{
This function wil filter doublets from an ArchRProject after addDoubletScores has been run
}
