% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterCells.R
\name{filterPlot}
\alias{filterPlot}
\title{Filter plot for cells in an ArchRProject}
\usage{
filterPlot(ArchRProj = NULL, filterList = NULL, sampleNames = NULL, ...)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{filterList}{A list of filters based on \code{cellColData} to apply when filtering cells.
Format should be a named numeric list where the name corresponds to a column name in \code{cellColData}
and the value corresponds to the filter criteria. A lower threshold is expected, below which
cells are filtered (a higher threshold is optional and can be passed as a numeric vector along with the lower threshold).
Only the first 2 filters will be plotted.}

\item{sampleNames}{The sample names corresponding to the subset of samples to plot. If \code{NULL}, all samples are included.}

\item{...}{additional params}
}
\description{
This function plots a list of attributes with filter criteria to visualize which cells would pass filter.
}
