% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part4_functions.R
\name{get_arm}
\alias{get_arm}
\title{Get Associational Rule Mining (ARM) Metrics}
\usage{
get_arm(data, p = "front", .id = "id")
}
\arguments{
\item{p}{name of binary variable containing pareto front}
}
\description{
This function takes a data.frame of binary features
and an integer number p containing the
index of the feature among the ones in the dataset
that we consider the good feature for the purposes of the
analysis and computes...
\itemize{
\item support (F),
\item confidence (F-> P),
\item conf (P -> F) and
\item lift (F & P)
for all features in the dataset
}
}
\author{
Tim Fraser
}
