% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_from_FASTA.R
\name{get_one_hot_encoded_seqs}
\alias{get_one_hot_encoded_seqs}
\title{Title}
\usage{
get_one_hot_encoded_seqs(givenFastaSeqs)
}
\arguments{
\item{givenFastaSeqs}{}
}
\value{

}
\description{
Title
}
\examples{

}
