% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archR_auxiliary_functionsI.R
\name{get_seqs_clusters_in_a_list}
\alias{get_seqs_clusters_in_a_list}
\title{Retrieve sequence clusters as a list}
\usage{
get_seqs_clusters_in_a_list(seqsClustLabels)
}
\arguments{
\item{seqsClustLabels}{Sequences with cluster labels as in the archR result
object.}
}
\value{
A list holding sequence clusters.
}
\description{
Given the sequence cluster labels from a archR result object,
returns the clusters separated as a list.
}
\examples{

clustLabels <- sample(seq_len(4), 50, replace = TRUE)

get_seqs_clusters_in_a_list(clustLabels)

}
