% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archR_auxiliary_functionsI.R
\name{get_seqs_clusters_in_a_list}
\alias{get_seqs_clusters_in_a_list}
\title{Retrieve sequence clusters as a list}
\usage{
get_seqs_clusters_in_a_list(seqsClustLabels, chooseLevel = 1)
}
\arguments{
\item{seqsClustLabels}{Sequences with cluster labels as in the archR result
object}

\item{chooseLevel}{Specify the level (archR iteration) at which sequence
clusters are to be reported. Default is 1.}
}
\value{
A list holding sequence clusters
}
\description{
Retrieve sequence clusters as a list
}
