% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represent_matrix_of_acgt_with_ggplot2.R
\name{get_tick_labels}
\alias{get_tick_labels}
\title{Getter function for labels of x- or y-ticks.}
\usage{
get_tick_labels(givenVec, use_asis = TRUE, at_every)
}
\value{
A label vector that can be directly used for setting the tickmarks.
}
\description{
A convenience function to provide the right set of x-tick labels
 for the image matrix representing the sequences.
}
