% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getAnnotation}
\alias{getAnnotation}
\title{Get Embedding from ArchRProject}
\usage{
getAnnotation(ArchRProj, name = NULL, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{name}{name of annotations}

\item{...}{additional args}
}
\description{
This function gets an embedding from an ArchRProject
}
