% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getArchRThreads}
\alias{getArchRThreads}
\title{Get global number of threads for default parallel computing.}
\usage{
getArchRThreads()
}
\description{
This function will get the global number of threads to be used for ArchR functions.
}
