% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getArrowFiles}
\alias{getArrowFiles}
\title{Get ArrowFiles from an ArchRProject}
\usage{
getArrowFiles(ArchRProj, ...)
}
\arguments{
\item{ArchRProj}{An ArchRProject object.}

\item{...}{additional args}
}
\description{
This function gets the names of all ArrowFiles associated with a given ArchRProject.
}
