% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getArrowFiles}
\alias{getArrowFiles}
\title{Get ArrowFiles in ArchRProject}
\usage{
getArrowFiles(ArchRProj, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{...}{additional args}
}
\description{
This function gets ArrowFiles in ArchRProject
}
