% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getBlacklist}
\alias{getBlacklist}
\title{Get blacklist from ArchRProject}
\usage{
getBlacklist(ArchRProj, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{...}{additional args}
}
\description{
This function gets the blacklist as a GRanges from genomeAnnotation in ArchRProject
}
