% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getBlacklist}
\alias{getBlacklist}
\title{Get the blacklist from an ArchRProject}
\usage{
getBlacklist(ArchRProj, ...)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{...}{additional args}
}
\description{
This function gets the blacklist (the regions to be excluded from analysis) as a GRanges object from the genomeAnnotation of a given ArchRProject.
}
