% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectMethods.R
\name{getCellColData}
\alias{getCellColData}
\title{Get cellColData from an ArchRProject}
\usage{
getCellColData(ArchRProj = NULL, select = NULL, drop = FALSE)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{select}{A character vector of column names to select from \code{cellColData} if you would like to subset the returned data.}

\item{drop}{A boolean value that indicates whether to drop the \code{dataframe} structure and convert to a vector if selecting only one column.}
}
\description{
This function gets the cellColData from a given ArchRProject.
}
\examples{

# Get Test ArchR Project
proj <- getTestProject()

# Get Cell Column Data
getCellColData(proj)

}
