% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getCellColData}
\alias{getCellColData}
\title{Get cellColData from an ArchRProject}
\usage{
getCellColData(ArchRProj, select = NULL, drop = FALSE, ...)
}
\arguments{
\item{ArchRProj}{An `ArchRProject` object.}

\item{select}{A character vector of column names to select from `cellColData` if you would like to subset the returned data.}

\item{drop}{A boolean value that indicates whether to drop the `dataframe` structure and convert to a vector if selecting only one column.}

\item{...}{additional args}
}
\description{
This function gets the cellColData from a given ArchRProject.
}
