% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getCellColData}
\alias{getCellColData}
\title{Get cellColData in ArchRProject}
\usage{
getCellColData(ArchRProj, select = NULL, drop = FALSE, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{select}{select a subset of column names from cellColData can put in a string function}

\item{drop}{drop if selecting only one column name}

\item{...}{additional args}
}
\description{
This function gets sampleColData in ArchRProject
}
