% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getCellNames}
\alias{getCellNames}
\title{Get cellNames from an ArchRProject}
\usage{
getCellNames(ArchRProj, ...)
}
\arguments{
\item{ArchRProj}{An `ArchRProject` object.}

\item{...}{additional args}
}
\description{
This function gets the cellNames from a given ArchRProject object.
}
