% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getCellNames}
\alias{getCellNames}
\title{Get cellNames in ArchRProject}
\usage{
getCellNames(ArchRProj, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{...}{additional args}
}
\description{
This function gets cellNames in ArchRProject
}
