% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectMethods.R
\name{getChromLengths}
\alias{getChromLengths}
\title{Get chromLengths from ArchRProject}
\usage{
getChromLengths(ArchRProj = NULL)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}
}
\description{
This function gets the chromosome lengths as a vector from the genomeAnnotation of a given ArchRProject.
}
