% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getChromLengths}
\alias{getChromLengths}
\title{Get chromLengths from ArchRProject}
\usage{
getChromLengths(ArchRProj, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{...}{additional args}
}
\description{
This function gets chromosome lengths as a vector from genomeAnnotation in ArchRProject
}
