% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getChromSizes}
\alias{getChromSizes}
\title{Get chromSizes from ArchRProject}
\usage{
getChromSizes(ArchRProj, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{...}{additional args}
}
\description{
This function gets chromosome lengths as GRanges from genomeAnnotation in ArchRProject
}
