% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getChromSizes}
\alias{getChromSizes}
\title{Get chromSizes from ArchRProject}
\usage{
getChromSizes(ArchRProj, ...)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{...}{additional args}
}
\description{
This function gets the chromosome lengths as a GRanges onject from the genomeAnnotation of a given ArchRProject.
}
