% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrativeAnalysis.R
\name{getCoAccessibility}
\alias{getCoAccessibility}
\title{Get the peak co-accessibility from an ArchRProject}
\usage{
getCoAccessibility(
  ArchRProj = NULL,
  corCutOff = 0.5,
  resolution = 1,
  returnLoops = TRUE
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{corCutOff}{A numeric describing the minimum numeric peak-to-peak correlation to return.}

\item{resolution}{A numeric describing the bp resolution to use when returning loops. This helps with overplotting of correlated regions.
This only takes affect if \code{returnLoops = TRUE}.}

\item{returnLoops}{A boolean indicating to return the co-accessibility signal as a \code{GRanges} "loops" object designed for use with
the \code{ArchRBrowser()} or as an \code{ArchRBrowserTrack()}.}
}
\description{
This function obtains co-accessibility data from an ArchRProject.
}
