% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getEmbedding}
\alias{getEmbedding}
\title{Get Embedding from ArchRProject}
\usage{
getEmbedding(ArchRProj, embedding = "IterativeLSI", return = "df", ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{embedding}{embedding name in ArchRProject}

\item{return}{return embedding as df or all info}

\item{...}{additional args}
}
\description{
This function gets an embedding from an ArchRProject
}
