% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getEmbedding}
\alias{getEmbedding}
\title{Get embedding information stored in an ArchRProject}
\usage{
getEmbedding(ArchRProj, embedding = "UMAP", returnDF = TRUE, ...)
}
\arguments{
\item{ArchRProj}{An `ArchRProject` object.}

\item{embedding}{QQQ The name of the `embeddings` object to retrieve from the designated `ArchRProject`. Options include QQQ.}

\item{returnDF}{A boolean value indicating whether to return the embedding object as a `data.frame`. Otherwise, it will return the full embedding object.}

\item{...}{additional args}
}
\description{
QQQ This function gets an embedding (i.e. UMAP) from a given ArchRProject.
}
