% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getEmbedding}
\alias{getEmbedding}
\title{Get embedding information stored in an ArchRProject}
\usage{
getEmbedding(ArchRProj, embedding = "UMAP", returnDF = TRUE, ...)
}
\arguments{
\item{ArchRProj}{An ArchRProject object.}

\item{embedding}{QQQ The name of the embedding object to retrieve from the designated ArchRProject. Options include QQQ.}

\item{...}{additional args}

\item{return}{If set to "df", the function will return the embedding object as a data.frame. Otherwise, it will return the full embedding object.}
}
\description{
QQQ This function gets an embedding (i.e. QQQ) from a given ArchRProject.
}
