% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getEmbedding}
\alias{getEmbedding}
\title{Get embedding information stored in an ArchRProject}
\usage{
getEmbedding(ArchRProj = NULL, embedding = "UMAP", returnDF = TRUE,
  ...)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{embedding}{The name of the \code{embeddings} object (i.e. UMAP, TSNE see \code{embeddingOut} of the \code{addEmbeddings()} function) to retrieve from the designated \code{ArchRProject}.}

\item{returnDF}{A boolean value indicating whether to return the embedding object as a \code{data.frame}. Otherwise, it will return the full embedding object.}

\item{...}{additional args}
}
\description{
This function gets an embedding (i.e. UMAP) from a given ArchRProject.
}
