% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getExons}
\alias{getExons}
\title{Get Exons from ArchRProject}
\usage{
getExons(ArchRProj, symbols = NULL, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{...}{additional args}
}
\description{
This function gets exons from geneAnnotation in ArchRProject
}
