% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectMethods.R
\name{getExons}
\alias{getExons}
\title{Get the exons from an ArchRProject}
\usage{
getExons(ArchRProj = NULL, symbols = NULL)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{symbols}{A character vector containing the gene symbols for the genes where exons should be extracted.}
}
\description{
This function gets the exons coordinates as a GRanges object from the geneAnnotation of a given ArchRProject.
}
\examples{

# Get Test ArchR Project
proj <- getTestProject()

# Get Exons in ArchRProj
getExons(proj)

# Get Exons globally
getExons()

}
