% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getFeatures}
\alias{getFeatures}
\title{Return Available Features for a given Matrix in ArrowFiles within an ArchRProject}
\usage{
getFeatures(ArchRProj, useMatrix = "GeneScoreMatrix", select = NULL,
  ignore.case = TRUE, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{useMatrix}{Matrix Name as in Arrow Files (ie TileMatrix, GeneScoreMatrix, ...)}

\item{select}{select a specific name with grep}

\item{ignore.case}{ignore case when searching with select}

\item{...}{additional args}
}
\description{
This function will identify available features for a matrix and return them for downstream
plotting utils.
}
