% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getFeatures}
\alias{getFeatures}
\title{Get the features that could be selected from a given data matrix within an ArchRProject}
\usage{
getFeatures(
  ArchRProj = NULL,
  useMatrix = "GeneScoreMatrix",
  select = NULL,
  ignore.case = TRUE
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{useMatrix}{The name of the data matrix as stored in the ArrowFiles of the \code{ArchRProject}. Options include "TileMatrix", "GeneScoreMatrix", etc.}

\item{select}{A string specifying a specific feature name (or rowname) to be found with \code{grep}.}

\item{ignore.case}{QQQ SHOULD BE ignoreCase FOR CONSISTENCY. A boolean value indicating whether to ignore the case (upper-case / lower-case) when searching via grep for the string passed to \code{select}.}
}
\description{
This function will identify available features from a given data matrix  (i.e. "GeneScoreMatrix", or "TileMatrix") and return them for downstream plotting utilities.
}
