% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Footprinting.R
\name{getFootprints}
\alias{getFootprints}
\title{Calculate footprints from an ArchRProject}
\usage{
getFootprints(
  ArchRProj = NULL,
  positions = NULL,
  plotName = "Plot-Footprints",
  groupBy = "Clusters",
  useGroups = NULL,
  flank = 250,
  minCells = 25,
  nTop = NULL,
  threads = getArchRThreads(),
  verbose = TRUE,
  logFile = createLogFile("getFootprints")
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{positions}{A \code{list} or \code{GRangesList} of \code{GRanges} containing the positions to incorporate into the footprint. Each position should be stranded.}

\item{plotName}{The prefix to add to the file name for the output PDF file containing the footprint plots.}

\item{groupBy}{The name of the column in \code{cellColData} used in the \code{addGroupCoverages()} function for grouping multiple cells together.}

\item{useGroups}{A character vector that is used to select a subset of groups by name from the designated \code{groupBy} column in \code{cellColData}.
This limits the groups used to perform footprinting.}

\item{flank}{The number of basepairs from the position center (+/-) to consider as the flank.}

\item{minCells}{The minimum number of cells required in a given cell group to permit footprint generation.}

\item{nTop}{The number of genomic regions to consider. Only the top \code{nTop} genomic regions based on the "score" column in the \code{GRanges}
object will be considered for the footprint.}

\item{threads}{The number of threads to be used for parallel computing.}

\item{verbose}{A boolean value that determines whether standard output includes verbose sections.}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
This function will get footprints for all samples in a given ArchRProject and return a summarized experiment object that can be used for downstream analyses
}
