% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrowRead.R
\name{getFragmentsFromArrow}
\alias{getFragmentsFromArrow}
\title{Read Fragments from Arrow}
\usage{
getFragmentsFromArrow(ArrowFile, chr = NULL, cellNames = NULL,
  method = "fast", verbose = TRUE, ...)
}
\arguments{
\item{ArrowFile}{ArchRProject or ArrowFiles}

\item{chr}{GRanges to count for each cell}

\item{cellNames}{matrix output name in ArrowFiles cannot be a protected matrix name}

\item{method}{ceiling for the number of counts per feature}

\item{verbose}{binarize matrix}

\item{...}{additional params}
}
\description{
This function for each sample will independently compute counts for each feature
per cell in the Arrow File
}
