% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrowRead.R
\name{getFragmentsFromArrow}
\alias{getFragmentsFromArrow}
\title{Get the fragments from an ArrowFile}
\usage{
getFragmentsFromArrow(
  ArrowFile = NULL,
  chr = NULL,
  cellNames = NULL,
  verbose = TRUE
)
}
\arguments{
\item{ArrowFile}{The path to the ArrowFile from which fragments should be obtained.}

\item{chr}{A name of a chromosome to be used to subset the fragments \code{GRanges} object to a specific chromsome if desired.}

\item{cellNames}{A character vector indicating the cell names of a subset of cells from which fragments whould be extracted. This allows for extraction of fragments from only a subset of selected cells. By default, this function will extract all cells from the provided ArrowFile using \code{getCellNames()}.}

\item{verbose}{A boolean value indicating whether to use verbose output during execution of this function. Can be set to \code{FALSE} for a cleaner output.}
}
\description{
This function retrieves the fragments from a given ArrowFile as a GRanges object.
}
