% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectMethods.R
\name{getGeneAnnotation}
\alias{getGeneAnnotation}
\title{Get geneAnnotation from an ArchRProject}
\usage{
getGeneAnnotation(ArchRProj = NULL)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}
}
\description{
This function gets the geneAnnotation from a given ArchRProject
}
\examples{

# Get Test ArchR Project
proj <- getTestProject()

# Get Gene Annotation in ArchRProj
getGeneAnnotation(proj)

# Get Gene Annotation loaded globally
getGeneAnnotation()

}
