% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getGeneAnnotation}
\alias{getGeneAnnotation}
\title{Get geneAnnotation from an ArchRProject}
\usage{
getGeneAnnotation(ArchRProj = NULL)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}
}
\description{
This function gets the geneAnnotation from a given ArchRProject
}
