% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getGeneAnnotation}
\alias{getGeneAnnotation}
\title{Get geneAnnotation from an ArchRProject}
\usage{
getGeneAnnotation(ArchRProj, ...)
}
\arguments{
\item{ArchRProj}{An `ArchRProject` object.}

\item{...}{additional args}
}
\description{
This function gets the geneAnnotation (see createGeneAnnotation) from a given ArchRProject
}
