% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getGenes}
\alias{getGenes}
\title{Get the genes from an ArchRProject}
\usage{
getGenes(ArchRProj, symbols = NULL, ...)
}
\arguments{
\item{ArchRProj}{An `ArchRProject` object.}

\item{symbols}{A character vector containing the gene symbols to subset from the `geneAnnotation`.}

\item{...}{additional args}
}
\description{
This function gets the genes start to end coordinates as a GRanges from the geneAnnotation of a given ArchRProject.
}
