% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getGenes}
\alias{getGenes}
\title{Get the genes from an ArchRProject}
\usage{
getGenes(ArchRProj = NULL, symbols = NULL)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{symbols}{A character vector containing the gene symbols to subset from the \code{geneAnnotation}.}
}
\description{
This function gets the genes start and end coordinates as a GRanges object from the geneAnnotation of a given ArchRProject.
}
