% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getGenome}
\alias{getGenome}
\title{Get genome from ArchRProject}
\usage{
getGenome(ArchRProj, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{...}{additional args}
}
\description{
This function gets the genome from genomeAnnotation in ArchRProject
}
