% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getGenome}
\alias{getGenome}
\title{Get the genome used by an ArchRProject}
\usage{
getGenome(ArchRProj, ...)
}
\arguments{
\item{ArchRProj}{An ArchRProject object.}

\item{...}{additional args}
}
\description{
This function gets the name of the genome from the genomeAnnotation used by a given ArchRProject.
}
