% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getGenomeAnnotation}
\alias{getGenomeAnnotation}
\title{Get genomeAnnotation from an ArchRProject}
\usage{
getGenomeAnnotation(ArchRProj, ...)
}
\arguments{
\item{ArchRProj}{An ArchRProject object.}

\item{...}{additional args}
}
\description{
This function gets the genomeAnnotation (in format QQQ) from a given ArchRProject.
}
