% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectMethods.R
\name{getGenomeAnnotation}
\alias{getGenomeAnnotation}
\title{Get the genomeAnnotation from an ArchRProject}
\usage{
getGenomeAnnotation(ArchRProj = NULL)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}
}
\description{
This function gets the genomeAnnotation from a given ArchRProject.
}
