% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getGenomeAnnotation}
\alias{getGenomeAnnotation}
\title{Get the genomeAnnotation from an ArchRProject}
\usage{
getGenomeAnnotation(ArchRProj, ...)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{...}{additional args}
}
\description{
This function gets the genomeAnnotation from a given ArchRProject.
}
