% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupExport.R
\name{getGroupBW}
\alias{getGroupBW}
\title{Export Group BigWigs}
\usage{
getGroupBW(
  ArchRProj = NULL,
  groupBy = "Sample",
  normMethod = "ReadsInTSS",
  tileSize = 100,
  maxCells = 1000,
  ceiling = 4,
  verbose = TRUE,
  threads = getArchRThreads(),
  logFile = createLogFile("getGroupBW")
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{groupBy}{A string that indicates how cells should be grouped. This string corresponds to one of the standard or
user-supplied \code{cellColData} metadata columns (for example, "Clusters"). Cells with the same value annotated in this metadata
column will be grouped together and the average signal will be plotted.}

\item{normMethod}{The name of the column in \code{cellColData} by which normalization should be performed. The recommended and default value
is "ReadsInTSS" which simultaneously normalizes tracks based on sequencing depth and sample data quality. Accepted values are
"None", "ReadsInTSS", "nCells", "ReadsInPromoter", or "nFrags".}

\item{tileSize}{The numeric width of the tile/bin in basepairs for plotting ATAC-seq signal tracks. All insertions in a single bin will be summed.}

\item{maxCells}{Maximum number of cells used for each bigwig.}

\item{ceiling}{Maximum contribution of accessibility per cell in each tile.}

\item{verbose}{A boolean specifying to print messages during computation.}

\item{threads}{An integer specifying the number of threads for parallel.}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
This function will group, summarize and export a bigwig for each group in an ArchRProject.
}
