% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectMethods.R
\name{getGroupSummary}
\alias{getGroupSummary}
\title{Get summary for Groups in ArchRProject}
\usage{
getGroupSummary(
  ArchRProj = NULL,
  groupBy = "Sample",
  select = "TSSEnrichment",
  summary = "median",
  removeNA = TRUE
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{groupBy}{The name of the column in \code{cellColData} to use for grouping multiple cells together for summarizing information.}

\item{select}{A character vector containing the column names to select from \code{cellColData}.}

\item{summary}{A character vector describing which method for summarizing across group. Options include "median", "mean", or "sum".}

\item{removeNA}{Remove NA's from summary method.}
}
\description{
This function summarizes a numeric cellColData entry across groupings in a ArchRProject.
}
