% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{getImputeWeights}
\alias{getImputeWeights}
\title{Get outputDirectory in ArchRProject}
\usage{
getImputeWeights(ArchRProj, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{...}{additional args}
}
\description{
This function gets outputDirectory from ArchRProject
}
