% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{getImputeWeights}
\alias{getImputeWeights}
\title{Get Imputation Weights from ArchRProject}
\usage{
getImputeWeights(ArchRProj = NULL)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}
}
\description{
This function gets imputation weights from an ArchRProject to impute numeric values.
}
