% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getInputFiles}
\alias{getInputFiles}
\title{Get Input Files from paths to create arrows}
\usage{
getInputFiles(paths, ...)
}
\arguments{
\item{paths}{A character vector of paths to search for usable input files.}

\item{...}{additional args}
}
\description{
This function will look for fragment files and bam files in the input paths and return the full path and sample names
}
