% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getMatches}
\alias{getMatches}
\title{Get Annotation Matches from ArchRProject}
\usage{
getMatches(ArchRProj, name = NULL, annoName = NULL, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{name}{name of annotations}

\item{annoName}{name to subset with annotations}

\item{...}{additional args}
}
\description{
This function gets annotation matches from an ArchRProject
}
