% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getMatches}
\alias{getMatches}
\title{Get annotation matches from an ArchRProject}
\usage{
getMatches(ArchRProj, name = NULL, annoName = NULL, ...)
}
\arguments{
\item{ArchRProj}{An ArchRProject object.}

\item{name}{name of annotations}

\item{annoName}{name to subset with annotations}

\item{...}{additional args}
}
\description{
This function gets annotation matches from a given ArchRProject.
}
