% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getMatches}
\alias{getMatches}
\title{Get annotation matches from an ArchRProject}
\usage{
getMatches(ArchRProj, name = NULL, annoName = NULL, ...)
}
\arguments{
\item{ArchRProj}{An `ArchRProject` object.}

\item{name}{QQQ The name of the annotation object to retrieve from the designated `ArchRProject`. Options include QQQ.}

\item{annoName}{QQQ ??? name to subset with annotations}

\item{...}{additional args}
}
\description{
This function gets annotation matches from a given ArchRProject.
}
