% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationPeaks.R
\name{getMatches}
\alias{getMatches}
\title{Get peak annotation matches from an ArchRProject}
\usage{
getMatches(ArchRProj = NULL, name = NULL, annoName = NULL)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{name}{The name of the \code{peakAnnotation} object (i.e. Motifs) to retrieve from the designated \code{ArchRProject}.}

\item{annoName}{The name of a specific annotation to subset within the \code{peakAnnotation}.}
}
\description{
This function gets peak annotation matches from a given ArchRProject. The peaks in the returned object are in the
same order as the peaks returned by \code{getPeakSet()}.
}
