% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrowRead.R
\name{getMatrixFromArrow}
\alias{getMatrixFromArrow}
\title{Read Fragments from Arrow}
\usage{
getMatrixFromArrow(ArrowFile = NULL, useMatrix = "GeneScoreMatrix",
  useSeqnames = NULL, cellNames = NULL, ArchRProj = NULL,
  verbose = TRUE, binarize = FALSE, ...)
}
\arguments{
\item{ArrowFile}{ArchRProject or ArrowFiles}

\item{useMatrix}{matrix name to get from Arrow}

\item{useSeqnames}{use a subset of seqnames for matrix}

\item{cellNames}{ceiling for the number of counts per feature}

\item{verbose}{binarize matrix}

\item{...}{additional params}
}
\description{
This function for each sample will independently compute counts for each feature
per cell in the Arrow File
}
