% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrowRead.R
\name{getMatrixFromArrow}
\alias{getMatrixFromArrow}
\title{Get a data matrix stored in an ArrowFile}
\usage{
getMatrixFromArrow(ArrowFile = NULL, useMatrix = "GeneScoreMatrix",
  useSeqnames = NULL, cellNames = NULL, ArchRProj = NULL,
  verbose = TRUE, binarize = FALSE)
}
\arguments{
\item{ArrowFile}{The path to an ArrowFile from which the selected data matrix should be obtained.}

\item{useMatrix}{The name of the data matrix to retrieve from the given ArrowFile. Options include "TileMatrix", "GeneScoreMatrix", etc.}

\item{useSeqnames}{A character vector of chromosome names to be used to subset the data matrix being obtained.}

\item{cellNames}{A character vector indicating the cell names of a subset of cells from which fragments whould be extracted. This allows for extraction of fragments from only a subset of selected cells. By default, this function will extract all cells from the provided ArrowFile using \code{getCellNames()}.}

\item{ArchRProj}{JJJ An \code{ArchRProject} object to be used for getting additional information for cells in \code{cellColData}. This is useful when you want to keep information created when analyzing an ArchRProject that is not present in the ArrowFile.}

\item{verbose}{A boolean value indicating whether to use verbose output during execution of  this function. Can be set to FALSE for a cleaner output.}

\item{binarize}{A boolean value indicating whether the matrix should be binarized before return. This is often desired when working with insertion counts.}
}
\description{
This function gets a given data matrix from an individual ArrowFile.
}
