% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrowRead.R
\name{getMatrixFromArrow}
\alias{getMatrixFromArrow}
\title{Get a data matrix stored in an ArrowFile}
\usage{
getMatrixFromArrow(
  ArrowFile = NULL,
  useMatrix = "GeneScoreMatrix",
  useSeqnames = NULL,
  excludeChr = NULL,
  cellNames = NULL,
  ArchRProj = NULL,
  verbose = TRUE,
  binarize = FALSE,
  logFile = createLogFile("getMatrixFromArrow")
)
}
\arguments{
\item{ArrowFile}{The path to an ArrowFile from which the selected data matrix should be obtained.}

\item{useMatrix}{The name of the data matrix to retrieve from the given ArrowFile. Options include "TileMatrix", "GeneScoreMatrix", etc.}

\item{useSeqnames}{A character vector of chromosome names to be used to subset the data matrix being obtained.}

\item{excludeChr}{A character vector containing the \code{seqnames} of the chromosomes that should be excluded from this analysis.}

\item{cellNames}{A character vector indicating the cell names of a subset of cells from which fragments whould be extracted.
This allows for extraction of fragments from only a subset of selected cells. By default, this function will extract all cells from
the provided ArrowFile using \code{getCellNames()}.}

\item{ArchRProj}{An \code{ArchRProject} object to be used for getting additional information for cells in \code{cellColData}.
In some cases, data exists within the \code{ArchRProject} object that does not exist within the ArrowFiles. To access this data, you can
provide the \code{ArchRProject} object here.}

\item{verbose}{A boolean value indicating whether to use verbose output during execution of  this function. Can be set to FALSE for a cleaner output.}

\item{binarize}{A boolean value indicating whether the matrix should be binarized before return. This is often desired when working with insertion counts.}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
This function gets a given data matrix from an individual ArrowFile.
}
\examples{

#Get Test Arrow
arrow <- getTestArrow()

# Get Fragments
se <- getMatrixFromArrow(arrow)

}
