% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrowRead.R
\name{getMatrixFromProject}
\alias{getMatrixFromProject}
\title{Get a data matrix stored in an ArchRProject}
\usage{
getMatrixFromProject(
  ArchRProj = NULL,
  useMatrix = "GeneScoreMatrix",
  useSeqnames = NULL,
  verbose = TRUE,
  binarize = FALSE,
  threads = getArchRThreads(),
  logFile = createLogFile("getMatrixFromProject")
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object to get data matrix from.}

\item{useMatrix}{The name of the data matrix to retrieve from the given ArrowFile. Options include "TileMatrix", "GeneScoreMatrix", etc.}

\item{useSeqnames}{A character vector of chromosome names to be used to subset the data matrix being obtained.}

\item{verbose}{A boolean value indicating whether to use verbose output during execution of  this function. Can be set to FALSE for a cleaner output.}

\item{binarize}{A boolean value indicating whether the matrix should be binarized before return.
This is often desired when working with insertion counts. Note that if the matrix has already been binarized previously, this should be set to \code{TRUE}.}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
This function gets a given data matrix from an \code{ArchRProject} and returns it as a \code{SummarizedExperiment}.
This function will return the matrix you ask it for, without altering that matrix unless you tell it to.
For example, if you added your \code{PeakMatrix} using \code{addPeakMatrix()} with \code{binarize = TRUE}, then
\code{getMatrixFromProject()} will return a binarized \code{PeakMatrix}. Alternatively, you could set \code{binarize = TRUE}
in the parameters passed to \code{getMatrixFromProject()} and the \code{PeakMatrix} will be binarized as you pull
it out. No other normalization is applied to the matrix by this function.
}
