% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getOutputDirectory}
\alias{getOutputDirectory}
\title{Get outputDirectory in ArchRProject}
\usage{
getOutputDirectory(ArchRProj, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{...}{additional args}
}
\description{
This function gets outputDirectory from ArchRProject
}
