% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakAnnotation.R
\name{getPeakAnnotation}
\alias{getPeakAnnotation}
\title{Get peakAnnotation from an ArchRProject}
\usage{
getPeakAnnotation(ArchRProj, name = NULL, ...)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{name}{The name of the peakAnnotation object (i.e. Motifs) to retrieve from the designated \code{ArchRProject}.}

\item{...}{additional args}
}
\description{
This function gets a peakAnnotation from a given ArchRProject.
}
