% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getPeakSet}
\alias{getPeakSet}
\title{Get the peak set from an ArchRProject}
\usage{
getPeakSet(ArchRProj, ...)
}
\arguments{
\item{ArchRProj}{An ArchRProject object.}

\item{...}{additional args}
}
\description{
This function gets the peakSet as a GRanges object from an ArchRProject.
}
