% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectMethods.R
\name{getPeakSet}
\alias{getPeakSet}
\title{Get the peak set from an ArchRProject}
\usage{
getPeakSet(ArchRProj = NULL)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}
}
\description{
This function gets the peak set as a GRanges object from an ArchRProject.
}
