% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakAnnotation.R
\name{getPositions}
\alias{getPositions}
\title{Get peak annotation positions from an ArchRProject}
\usage{
getPositions(ArchRProj = NULL, name = NULL, annoName = NULL)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{name}{The name of the \code{peakAnnotation} object (i.e. Motifs) to retrieve from the designated \code{ArchRProject}.}

\item{annoName}{The name of a specific annotation to subset within the \code{peakAnnotation}.}
}
\description{
This function gets the peak annotation positions (i.e. Motifs) from a given ArchRProject.
}
