% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakAnnotation.R
\name{getPositions}
\alias{getPositions}
\title{Get peakAnnotation positions from an ArchRProject}
\usage{
getPositions(ArchRProj, name = NULL, annoName = NULL, ...)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{name}{The name of the peakAnnotation object (i.e. Motifs) to retrieve from the designated \code{ArchRProject}.}

\item{annoName}{The name of a specific annotation to subset within the peakAnnotation.}

\item{...}{additional args}
}
\description{
This function gets the peakAnnotation positions (i.e. Motifs) from a given ArchRProject.
}
