% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakAnnotation.R
\name{getPositions}
\alias{getPositions}
\title{Get annotation positions from an ArchRProject}
\usage{
getPositions(ArchRProj, name = NULL, annoName = NULL, ...)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{name}{QQQ The name of the annotation object to retrieve from the designated \code{ArchRProject}. Options include QQQ.}

\item{annoName}{QQQ ??? name to subset with annotations}

\item{...}{additional args}
}
\description{
This function gets the annotation positions from a given ArchRProject.
}
