% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectMethods.R
\name{getProjectSummary}
\alias{getProjectSummary}
\title{Get projectSummary from an ArchRProject}
\usage{
getProjectSummary(ArchRProj = NULL, returnSummary = FALSE)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{returnSummary}{A boolean value indicating whether to return a summary of the \code{ArchRProject} or to just print the summary.}
}
\description{
This function prints the projectSummary from an ArchRProject
}
\examples{

# Get Test ArchR Project
proj <- getTestProject()

# Get Project Summary
getProjectSummary(proj)

}
