% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getProjectSummary}
\alias{getProjectSummary}
\title{Get projectSummary from an ArchRProject}
\usage{
getProjectSummary(ArchRProj = NULL, returnSummary = FALSE, ...)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{returnSummary}{A boolean value indicating whether to return a summary of the \code{ArchRProject} or to just print the summary.}

\item{...}{additional args}
}
\description{
This function prints the projectSummary from an ArchRProject
}
