% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getReducedDims}
\alias{getReducedDims}
\title{Get Reduced Dimensions from ArchRProject}
\usage{
getReducedDims(ArchRProj, reducedDims = "TileLSI", return = "matrix",
  ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{reducedDims}{reduced dimensions name in ArchRProject}

\item{return}{return reduced dimensions as matrix or all info}

\item{...}{additional args}
}
\description{
This function gets an embedding from an ArchRProject
}
