% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getSampleColData}
\alias{getSampleColData}
\title{Get sampleColData from an ArchRProject}
\usage{
getSampleColData(ArchRProj, select = NULL, drop = FALSE, ...)
}
\arguments{
\item{ArchRProj}{An `ArchRProject` object.}

\item{select}{QQQ A character vector containing the column names to select from sampleColData.}

\item{drop}{A boolean value that indicates whether to drop the `dataframe` structure and convert to a vector if selecting only one column.}

\item{...}{additional args}
}
\description{
This function gets the `sampleColData` from a given `ArchRProject`.
}
