% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectMethods.R
\name{getSampleColData}
\alias{getSampleColData}
\title{Get sampleColData from an ArchRProject}
\usage{
getSampleColData(ArchRProj = NULL, select = NULL, drop = FALSE)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{select}{A character vector containing the column names to select from \code{sampleColData}.}

\item{drop}{A boolean value that indicates whether to drop the \code{dataframe} structure and convert to a vector if selecting only one column.}
}
\description{
This function gets the sampleColData from a given ArchRProject.
}
