% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getSampleColData}
\alias{getSampleColData}
\title{Get sampleColData in ArchRProject}
\usage{
getSampleColData(ArchRProj, select = NULL, drop = FALSE, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{select}{select a subset of column names from sampleColData}

\item{drop}{drop if selecting only one column name}

\item{...}{additional args}
}
\description{
This function gets sampleColData in ArchRProject
}
