% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectMethods.R
\name{getSampleNames}
\alias{getSampleNames}
\title{Get the sample names from an ArchRProject}
\usage{
getSampleNames(ArchRProj = NULL)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}
}
\description{
This function gets the names of all samples from a given ArchRProject.
}
