% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getSampleNames}
\alias{getSampleNames}
\title{Get sampleNames in ArchRProject}
\usage{
getSampleNames(ArchRProj, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{...}{additional args}
}
\description{
This function gets sampleNames in ArchRProject
}
