% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getSampleNames}
\alias{getSampleNames}
\title{Get sampleNames from an ArchRProject}
\usage{
getSampleNames(ArchRProj, ...)
}
\arguments{
\item{ArchRProj}{An `ArchRProject` object.}

\item{...}{additional args}
}
\description{
This function gets the sampleNames from a given ArchRProject.
}
