% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputData.R
\name{getTestFragments}
\alias{getTestFragments}
\title{Get PBMC Small Test Fragments}
\usage{
getTestFragments(version = 2)
}
\arguments{
\item{version}{version of test fragments to return}
}
\description{
V1 : This function will download fragments for a small PBMC test dataset.
V2 : This function will return test fragments for a small PBMC test dataset in your cwd.
}
\examples{

# Get Test Fragments
fragments <- getTestFragments()

}
