% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputData.R
\name{getTestProject}
\alias{getTestProject}
\title{Get PBMC Small Test Project}
\usage{
getTestProject()
}
\description{
This function will download an ArchRProject for a small PBMC test dataset (2k Cells) spanning chr1 and 2 (~2-300MB).
}
