% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trajectory.R
\name{getTrajectory}
\alias{getTrajectory}
\title{Get Supervised Trajectory from an ArchR Project}
\usage{
getTrajectory(
  ArchRProj = NULL,
  name = "Trajectory",
  useMatrix = "GeneScoreMatrix",
  groupEvery = 1,
  log2Norm = TRUE,
  scaleTo = 10000,
  smoothWindow = 11,
  threads = getArchRThreads()
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{name}{A string indicating the name of the fitted trajectory in \code{cellColData} to retrieve from the given \code{ArchRProject}.}

\item{useMatrix}{The name of the data matrix from the \code{ArrowFiles} to get numerical values for each cell from. Recommended
matrices are "GeneScoreMatrix", "PeakMatrix", or "MotifMatrix".}

\item{groupEvery}{The number of sequential percentiles to group together when generating a trajectory. This is similar to smoothing
via a non-overlapping sliding window across pseudo-time. If \code{groupEvery = 2}, the values for percentiles \link{1 and 2}, \link{3 and 4},
\link{5 and 6}, etc. will be grouped together.}

\item{log2Norm}{A boolean value that indicates whether the summarized trajectory matrix should be log2 transformed. If you are using
a "MotifMatrix" set to FALSE.}

\item{scaleTo}{Once the sequential trajectory matrix is created, each column in that matrix will be normalized to a column sum
indicated by \code{scaleTo}. Setting this to \code{NULL} will prevent any normalization and should be done in certain circumstances
(for ex. if you are using a "MotifMatrix").}

\item{smoothWindow}{An integer value indicating the smoothing window in size (relaive to \code{groupEvery}) for the sequential
trajectory matrix to better reveal temporal dynamics.}

\item{threads}{The number of threads to be used for parallel computing.}
}
\description{
This function will get a supervised trajectory from an \code{ArchRProject} (see \code{addTrajectory}), get data
from a desired matrix, and smooth each value across the input trajectory.
}
