% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trajectory.R
\name{getTrajectory}
\alias{getTrajectory}
\title{Get Supervised Trajectory from an ArchR Project}
\usage{
getTrajectory(ArchRProj = NULL, name = "Trajectory",
  useMatrix = "GeneScoreMatrix", varCutOff = 0.1,
  maxFeatures = 25000, groupEvery = 2, threads = getArchRThreads(),
  scaleTo = 10000, log2Norm = TRUE, smooth = TRUE,
  smoothFormula = "y ~ s(x, bs = 'cs')", ...)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{name}{A string indicating the name of the fitted trajectory in \code{cellColData} to retrieve from the given \code{ArchRProject}.}

\item{useMatrix}{The name of the data matrix from the \code{ArrowFiles} to get numerical values for each cell from.}

\item{varCutOff}{Variance Quantile Cutoff for identifying top variable features across trajectory.}

\item{maxFeatures}{The maximum number of features (ordered by variance) to consider from \code{useMatrix} when generating a trajectory.}

\item{groupEvery}{The number of sequential percentiles (0-100 every x percentile) to group together when generating a trajectory (similar to smoothing).}

\item{threads}{The number of threads to be used for parallel computing.}

\item{scaleTo}{Scale group data matrix to this value for normalizaiton.}

\item{log2Norm}{A boolean value that indicates whether the summarized trajectory matrix should be log2 transformed.}

\item{smooth}{A boolean value indicating whether the summarized trajectory matrix should be smoothed in a row-wise fashion.}

\item{smoothFormula}{The smoothing formula to use in the generalized additive model. See the \code{formula} parameter in \code{mgcv::gam()}.}

\item{...}{additional args}
}
\description{
This function will fit get a supervised trajectory from an ArchRProject and aggregate signal
from a matrix and smooth across the trajectory
}
