% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getTSS}
\alias{getTSS}
\title{Get TSS from ArchRProject}
\usage{
getTSS(ArchRProj, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{...}{additional args}
}
\description{
This function gets TSS from geneAnnotation in ArchRProject
}
