% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{getTutorialData}
\alias{getTutorialData}
\title{Get Tutorial Data For ArchR}
\usage{
getTutorialData(tutorial = "hematopoiesis", ...)
}
\arguments{
\item{tutorial}{The name of the available tutorial for which to retreive the tutorial data. Options are "Hematopoiesis", "PBMC", "FreshFrozen". "Hematopoiesis" refers to QQQ. "PBMC" refers to QQQ. "FreshFrozen" refers to QQQ data from Granja et al. Nature Biotechnology 2019.}

\item{...}{additional args}
}
\description{
This function will download data for a given tutorial and return the input files required for ArchR
}
