% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixDeviations.R
\name{getVarDeviations}
\alias{getVarDeviations}
\title{QQQ}
\usage{
getVarDeviations(ArchRProj, name = "MotifMatrix", plot = TRUE,
  n = 25)
}
\arguments{
\item{ArchRProj}{QQQ}

\item{name}{QQQ}

\item{plot}{QQQ}

\item{n}{QQQ}
}
\description{
QQQ
}
