% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgplotHelper.R
\name{ggAlignPlots}
\alias{ggAlignPlots}
\title{Align GG Plots}
\usage{
ggAlignPlots(..., sizes, type = "v", plotList = NULL,
  grobList = NULL)
}
\arguments{
\item{...}{ggplots}

\item{sizes}{sizes are a vector or list of values for each ggplot ie c(1,1) for two plots}

\item{type}{v,vertical or h,horizontal}
}
\description{
Align GG Plots
}
