% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgplotHelper.R
\name{ggAlignPlots}
\alias{ggAlignPlots}
\title{Align Ggplots vertically or horizontally}
\usage{
ggAlignPlots(..., sizes, type = "v", plotList = NULL,
  grobList = NULL, draw = TRUE)
}
\arguments{
\item{...}{ggplots}

\item{sizes}{sizes are a vector or list of values for each ggplot ie c(1,1) for two plots}

\item{type}{v,vertical or h,horizontal}

\item{plotList}{add a list of plots to be aligned}

\item{grobList}{add a list of grobs to be aligned}
}
\description{
This function aligns ggplots vertically or horizontally
}
