% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgplotHelper.R
\name{ggGroup}
\alias{ggGroup}
\title{A ggplot-based ridge/violin plot wrapper function}
\usage{
ggGroup(
  x = NULL,
  y = NULL,
  xlabel = NULL,
  ylabel = NULL,
  groupOrder = NULL,
  groupSort = FALSE,
  size = 1,
  baseSize = 10,
  ridgeScale = 1,
  ratioYX = NULL,
  alpha = 1,
  title = "",
  pal = paletteDiscrete(values = x, set = "stallion"),
  addBoxPlot = TRUE,
  plotAs = "ridges",
  ...
)
}
\arguments{
\item{x}{A character vector containing the categorical x-axis values for each y-axis value.}

\item{y}{A numeric vector containing the y-axis values for each point.}

\item{xlabel}{The label to plot for the x-axis.}

\item{ylabel}{The label to plot for the y-axis.}

\item{groupOrder}{A character vector indicating a custom order for plotting x-axis categorical values. Should contain all possible
values of \code{x} in the desired order.}

\item{groupSort}{A boolean indicating whether to sort groups based on the average value of the group.}

\item{size}{The line width for boxplot/summary lines.}

\item{baseSize}{The base font size (in points) to use in the plot.}

\item{ridgeScale}{A numeric indicating the relative size for each ridge in the ridgeplot.}

\item{ratioYX}{The aspect ratio of the x and y axes on the plot.}

\item{alpha}{A number indicating the transparency to use for each point. See \code{ggplot2} for more details.}

\item{title}{The title of the plot.}

\item{pal}{A named custom palette (see \code{paletteDiscrete()} and \code{ArchRPalettes}) for discrete coloring.}

\item{addBoxPlot}{A boolean indicating whether to add a boxplot to the plot if \code{plotAs="violin"}.}

\item{plotAs}{A string indicating how the groups should be plotted. Acceptable values are "ridges" (for a \code{ggrides}-style plot) or "violin" (for a violin plot).}

\item{...}{Additional parameters to pass to \code{ggplot2} for plotting.}
}
\description{
This function is a wrapper around ggplot geom_density_ridges or geom_violin to allow for plotting group distribution plots in ArchR.
}
