% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgplotHelper.R
\name{ggHex}
\alias{ggHex}
\title{GG Violin Plot}
\usage{
ggHex(x, y, color, extend = 0.05, ratioYX = 1, xlim = NULL,
  ylim = NULL, bins = 150, pal = paletteContinuous(set =
  "solar_extra"), title = "", baseSize = 12, xlabel = "x",
  ylabel = "y", fun = "median", ...)
}
\arguments{
\item{x}{categorical values to each y value}

\item{y}{numeric values}

\item{pal}{color palette see paletteDiscrete for examples}

\item{xlabel}{xlabel}

\item{ylabel}{ylabel}

\item{base_size}{base_size of theme}

\item{size}{size of barplot lines}
}
\description{
GG Violin Plot
}
