% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgplotHelper.R
\name{ggHex}
\alias{ggHex}
\title{A ggplot-based Hexplot wrapper function summary of points in a standardized manner}
\usage{
ggHex(x = NULL, y = NULL, color = NULL, pal = paletteContinuous(set
  = "solar_extra"), bins = 200, xlim = NULL, ylim = NULL,
  extend = 0.05, xlabel = "x", ylabel = "y", title = "",
  colorTitle = "values", baseSize = 6, ratioYX = 1, FUN = "mean",
  quantCut = c(0.01, 0.99), addPoints = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector containing the x-axis values for each point.}

\item{y}{A numeric vector containing the y-axis values for each point.}

\item{color}{QQQ A vector of color values to be plotted. This vector must be the same length as QQQ `length(x) * length(y)`}

\item{pal}{QQQ The name or numeric index of a custom palette from `ArchRPalettes` to use for QQQ.}

\item{bins}{QQQ NOT CLEAR WHAT IS A BIN. The number of bins to be used for plotting the hexplot. QQQ `bins` indicates the total number of hexagons that will fit within the surface area of the plot.}

\item{xlim}{A vector of two numeric values indicating the lower and upper bounds of the x-axis on the plot.}

\item{ylim}{A vector of two numeric values indicating the lower and upper bounds of the y-axis on the plot.}

\item{extend}{QQQ A numeric value indicating the fraction to extend the x-axis and y-axis beyond the maximum and minimum values if `xlim` and `ylim` are not provided. For example, 0.05 will extend the x-axis and y-axis by 5% on each end.}

\item{xlabel}{The label to plot for the x-axis.}

\item{ylabel}{The label to plot for the y-axis.}

\item{title}{The title of the plot.}

\item{colorTitle}{The label to use for the legend corresponding to `color`.}

\item{baseSize}{QQQ The base font size to use in the plot.}

\item{ratioYX}{The aspect ratio of the x and y axes on the plot.}

\item{FUN}{The function to use for summarizing data into hexagons. Typically "mean" or something similar.}

\item{quantCut}{If this is not null, a quantile cut is performed to threshold the top and bottom of the distribution. This prevents skewed color scales caused by strong outliers. The format of this should be c(x,y) where x is the upper threshold and y is the lower threshold. For example, quantileCut = c(0.975,0.025) will take the top and bottom 2.5% of values and set them to the value of the 97.5th and 2.5th percentile values respectively.}

\item{addPoints}{QQQ UNCLEAR WHY POINTS WOULD BE ADDED TO A HEXPLOT. ALSO, IN ggViolin THIS IS CALLED "points". A boolean value indicating whether individual points should be shown on the hexplot.}

\item{...}{additional params to pass}
}
\description{
This function will plot x,y coordinates values summarized in hexagons in a standardized manner
}
