% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgplotUtils.R
\name{ggHex}
\alias{ggHex}
\title{A ggplot-based Hexplot wrapper function summary of points in a standardized manner}
\usage{
ggHex(
  x = NULL,
  y = NULL,
  color = NULL,
  pal = paletteContinuous(set = "solarExtra"),
  bins = 200,
  xlim = NULL,
  ylim = NULL,
  extend = 0.05,
  xlabel = "x",
  ylabel = "y",
  title = "",
  colorTitle = "values",
  baseSize = 6,
  ratioYX = 1,
  FUN = "median",
  hexCut = c(0.02, 0.98),
  addPoints = FALSE,
  ...
)
}
\arguments{
\item{x}{A numeric vector containing the x-axis values for each point.}

\item{y}{A numeric vector containing the y-axis values for each point.}

\item{color}{A numeric/categorical vector containing coloring information for each point.}

\item{pal}{A custom continuous palette from \code{ArchRPalettes} for coloration of hexes.}

\item{bins}{The number of bins to be used for plotting the hexplot. \code{bins} indicates the total number of hexagons that will fit within the surface area of the plot.}

\item{xlim}{A numeric vector of two values indicating the lower and upper bounds of the x-axis on the plot.}

\item{ylim}{A numeric vector of two values indicating the lower and upper bounds of the y-axis on the plot.}

\item{extend}{A numeric value indicating the fraction to extend the x-axis and y-axis beyond the maximum and minimum values if \code{xlim} and \code{ylim} are not provided. For example, 0.05 will extend the x-axis and y-axis by 5 percent on each end.}

\item{xlabel}{The label to plot for the x-axis.}

\item{ylabel}{The label to plot for the y-axis.}

\item{title}{The title of the plot.}

\item{colorTitle}{The label to use for the legend corresponding to \code{color}.}

\item{baseSize}{The base font size (in points) to use in the plot.}

\item{ratioYX}{The aspect ratio of the x and y axes on the plot.}

\item{FUN}{The function to use for summarizing data into hexagons. Typically "mean" or something similar.}

\item{hexCut}{If this is not null, a quantile cut is performed to threshold the top and bottom of the distribution of values.
This prevents skewed color scales caused by strong outliers. The format of this should be c(a,b) where \code{a} is the upper threshold
and \code{b} is the lower threshold. For example, hexCut = c(0.025,0.975) will take the top and bottom 2.5 percent of values and set
them to the value of the 97.5th and 2.5th percentile values respectively.}

\item{addPoints}{A boolean value indicating whether individual points should be shown on the hexplot.}

\item{...}{Additional params for plotting}
}
\description{
This function will plot x,y coordinate values summarized in hexagons in a standardized manner
}
\examples{

# Create Random Data
m <- data.frame(matrix(rnorm(300, 2),ncol=3))

# Plot
p <- ggHex(x = m[,1], y = m[,2], color = m[,3])

# To PDF
pdf("test.pdf", width = 4, height = 4)
p
dev.off()

}
