% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgplotHelper.R
\name{ggOneToOne}
\alias{ggOneToOne}
\title{A ggplot-based one-to-one dot plot}
\usage{
ggOneToOne(x = NULL, y = NULL, size = 2, alpha = 1, xlabel = "x",
  ylabel = "y", title = "Correlation", min = 0.05, max = 0.9999,
  nPlot = 100 * 10^3, nKernel = 100, densityMax = 0.95,
  extend = 0.05, baseSize = 6, rastr = TRUE,
  pal = paletteContinuous(set = "blue_yellow"), ...)
}
\arguments{
\item{x}{A numeric vector containing the x-axis values for each point.}

\item{y}{A numeric vector containing the y-axis values for each point.}

\item{size}{The numeric size of the points to plot.}

\item{alpha}{geom_point alpha}

\item{xlabel}{The label to plot for the x-axis.}

\item{ylabel}{The label to plot for the y-axis.}

\item{title}{The title of the plot.}

\item{min}{xmin quantile [0,1]}

\item{max}{xmax quantile [0,1]}

\item{nPlot}{The number of points to plot. When this value is less than the total points, sample is used to extract random data points to plot.}

\item{nKernel}{The value of n to use the kde2d from the MASS package.}

\item{baseSize}{base_font size default is 12}

\item{pal}{continuous color palette to use}
}
\description{
This function is a wrapper around ggplot geom_point to allow for plotting one-to-one sample comparisons in ArchR.
}
