% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgplotHelper.R
\name{ggOneToOne}
\alias{ggOneToOne}
\title{GG Plot One to One Heatscatter}
\usage{
ggOneToOne(x = NULL, y = NULL, size = 2, alpha = 1, xlabel = "x",
  ylabel = "y", title = "Correlation", min = 0.05, max = 0.9999,
  nPlot = 100 * 10^3, nKernel = 100, densityMax = 0.95,
  extend = 0.05, baseSize = 6, rastr = TRUE,
  pal = paletteContinuous(set = "viridis"), ...)
}
\arguments{
\item{x}{x vector of data to be plot}

\item{y}{y vector of data to be plot}

\item{size}{plot point size}

\item{alpha}{alpha transperancy of points}

\item{xlabel}{xlabel}

\item{ylabel}{ylabel}

\item{title}{ggtitle}

\item{min}{x/y min quantile [0,1]}

\item{max}{x/y max quantile [0,1]}

\item{nPlot}{number of points to plot (correlations computed prior)}

\item{nKernel}{n for MASS::kde2d default = 100}

\item{baseSize}{base size of fonts in plot}

\item{pal}{color palette to use for density}
}
\description{
GG Plot One to One Heatscatter
}
