% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgplotUtils.R
\name{ggPoint}
\alias{ggPoint}
\title{A ggplot-based dot plot wrapper function}
\usage{
ggPoint(
  x = NULL,
  y = NULL,
  color = NULL,
  discrete = TRUE,
  discreteSet = "stallion",
  continuousSet = "solarExtra",
  labelMeans = TRUE,
  pal = NULL,
  defaultColor = "lightGrey",
  colorDensity = FALSE,
  size = 1,
  xlim = NULL,
  ylim = NULL,
  extend = 0.05,
  xlabel = "x",
  ylabel = "y",
  title = "",
  randomize = FALSE,
  seed = 1,
  colorTitle = NULL,
  colorOrder = NULL,
  colorLimits = NULL,
  alpha = 1,
  baseSize = 10,
  legendSize = 3,
  ratioYX = 1,
  labelAsFactors = TRUE,
  fgColor = "black",
  bgColor = "white",
  bgWidth = 1,
  labelSize = 3,
  addFit = NULL,
  rastr = FALSE,
  dpi = 300,
  ...
)
}
\arguments{
\item{x}{A numeric vector containing the x-axis values for each point.}

\item{y}{A numeric vector containing the y-axis values for each point.}

\item{color}{A numeric/categorical vector used to determine the coloration for each point.}

\item{discrete}{A boolean value indicating whether the supplied data is discrete (\code{TRUE}) or continuous (\code{FALSE}).}

\item{discreteSet}{The name of a custom palette from \code{ArchRPalettes} to use for categorical/discrete color.
This argument is only used if \code{discrete} is set to \code{TRUE}.}

\item{continuousSet}{The name of a custom palette from \code{ArchRPalettes} to use for numeric color.
This argument is only used if \code{discrete} is set to \code{FALSE}.}

\item{labelMeans}{A boolean value indicating whether the mean of each categorical/discrete color should be labeled.}

\item{pal}{A custom palette used to override discreteSet/continuousSet for coloring vector.}

\item{defaultColor}{The default color for points that do not have another color applied (i.e. \code{NA} values).}

\item{colorDensity}{A boolean value indicating whether the density of points on the plot should be indicated by color.
If \code{TRUE}, continuousSet is used as the color palette.}

\item{size}{The numeric size of the points to be plotted.}

\item{xlim}{A numeric vector of two values indicating the lower and upper bounds of the x-axis on the plot.}

\item{ylim}{A numeric vector of two values indicating the lower and upper bounds of the y-axis on the plot.}

\item{extend}{A numeric value indicating the fraction to extend the x-axis and y-axis beyond the maximum and minimum
values if \code{xlim} and \code{ylim} are not provided. For example, 0.05 will extend the x-axis and y-axis by 5 percent on each end.}

\item{xlabel}{The label to plot for the x-axis.}

\item{ylabel}{The label to plot for the y-axis.}

\item{title}{The title of the plot.}

\item{randomize}{A boolean value indicating whether to randomize the order of the points when plotting.}

\item{seed}{A numeric seed number for use in randomization.}

\item{colorTitle}{A title to be added to the legend if \code{color} is supplied.}

\item{colorOrder}{A vector that allows you to control the order of palette colors associated with the values in \code{color}.
For example if you have \code{color} as \code{c("a","b","c")} and want to have the first color selected from the palette be used for
"c", the second color for "b", and the third color for "a", you would supply the \code{colorOrder} as \code{c("c", "b", "a")}.}

\item{colorLimits}{A numeric vector of two values indicating the lower and upper bounds of colors if numeric. Values
beyond these limits are thresholded.}

\item{alpha}{A number indicating the transparency to use for each point. See \code{ggplot2} for more details.}

\item{baseSize}{The base font size (in points) to use in the plot.}

\item{legendSize}{The size in inches to use for plotting the color legend.}

\item{ratioYX}{The aspect ratio of the x and y axes on the plot.}

\item{labelAsFactors}{A boolean indicating whether to label the \code{color} input as a numeric factor (\code{TRUE}) or with a character string (\code{FALSE}).}

\item{fgColor}{The foreground color of the plot.}

\item{bgColor}{The background color of the plot.}

\item{bgWidth}{The background relative width size of the halos in the labeling.}

\item{labelSize}{The numeric font size of labels.}

\item{addFit}{A string indicating the method to use for adding a fit/regression line to the plot (see \code{ggplot2::geom_smooth()} methods).
If set to \code{NULL}, no fit/regression line is added.}

\item{rastr}{A boolean value that indicates whether the plot should be rasterized using \code{ggrastr}. This does not rasterize
lines and labels, just the internal portions of the plot.}

\item{dpi}{The resolution in dots per inch to use for the plot.}
}
\description{
This function is a wrapper around ggplot geom_point to allow for a more intuitive plotting of ArchR data.
}
