% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgplotHelper.R
\name{ggViolin}
\alias{ggViolin}
\title{GG Violin Plot}
\usage{
ggViolin(x = NULL, y = NULL, xlabel = NULL, ylabel = NULL,
  xOrder = NULL, points = FALSE, size = 1, baseSize = 6,
  ratioYX = 1, sampleRatio = 0.1, title = "",
  pal = paletteDiscrete(values = x, set = "stallion"), ...)
}
\arguments{
\item{x}{categorical values to each y value}

\item{y}{numeric values}

\item{xlabel}{xlabel}

\item{ylabel}{ylabel}

\item{xOrder}{custom order of x for plotting}

\item{points}{add points using ggrastr geom_quasirandom?}

\item{size}{size of barplot lines}

\item{baseSize}{base size of fonts in plot}

\item{pal}{color palette see paletteDiscrete for examples}
}
\description{
GG Violin Plot
}
