% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgplotHelper.R
\name{ggViolin}
\alias{ggViolin}
\title{GG Violin Plot}
\usage{
ggViolin(x, y, base_size = 12, xlabel = NULL, ylabel = NULL,
  points = FALSE, baseSize = 6, ratioYX = 1, sampleRatio = 0.1,
  size = 1, title = "", pal = paletteDiscrete(values = x, set =
  "stallion"))
}
\arguments{
\item{x}{categorical values to each y value}

\item{y}{numeric values}

\item{base_size}{base_size of theme}

\item{xlabel}{xlabel}

\item{ylabel}{ylabel}

\item{size}{size of barplot lines}

\item{pal}{color palette see paletteDiscrete for examples}
}
\description{
GG Violin Plot
}
