% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperUtils.R
\name{\%bcin\%}
\alias{\%bcin\%}
\title{QQQ Generic matching function for S4Vector objects primarily used for run length encodings}
\usage{
x \%bcin\% table
}
\arguments{
\item{x}{An `S4Vector` object search for in `table`.}

\item{table}{The set of `S4Vector` objects to serve as the base for the match function.}
}
\description{
QQQ This function provides a general matching function for S4Vector objects primarily used for run length encodings
}
