% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperUtils.R
\name{\%bcni\%}
\alias{\%bcni\%}
\title{QQQ Negated matching function for S4Vector objects primarily used for run length encodings}
\usage{
x \%bcni\% table
}
\arguments{
\item{x}{An `S4Vector` object search for in `table`.}

\item{table}{The set of `S4Vector` objects to serve as the base for the match function.}
}
\description{
QQQ This function provides the reciprocal of %bcin% for S4Vector objects primarily used for run length encodings
}
