% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperUtils.R
\name{\%ni\%}
\alias{\%ni\%}
\title{Negated Value Matching}
\usage{
x \%ni\% table
}
\arguments{
\item{x}{The value to search for in \code{table}.}

\item{table}{The set of values to serve as the base for the match function.}
}
\description{
This function is the reciprocal of %in%. See the match funciton in base R
}
