% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archR_auxiliary_functionsII.R
\name{handle_dir_creation}
\alias{handle_dir_creation}
\title{Handle directory creation}
\usage{
handle_dir_creation(
  givenODir,
  flags = list(debugFlag = FALSE, verboseFlag = FALSE, plotFlag = FALSE, timeFlag =
    FALSE)
)
}
\arguments{
\item{givenODir}{Specify the output directory name with its complete path.}

\item{flags}{List with four Logical elements as detailed.
\describe{
  \item{debugFlag}{Whether debug information for the run is printed}
  \item{verboseFlag}{Whether verbose information for the run is printed}
  \item{plotVerboseFlag}{Whether verbose plotting is performed for the run}
  \item{timeFlag}{Whether timing information is printed for the run}
}}
}
\description{
Given the output directory name with its complete path, this
function checks if a directory of same name exists the given location. If
yes, then it adds a suffix (a number) to the given directory name, and
proceeds to create the directory.
}
