% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{imputeMatrix}
\alias{imputeMatrix}
\title{Impute a matrix with impute weights}
\usage{
imputeMatrix(mat = NULL, imputeWeights = NULL,
  threads = getArchRThreads(), verbose = FALSE,
  logFile = createLogFile("imputeMatrix"))
}
\arguments{
\item{mat}{A matrix or sparseMatrix of class dgCMatrix to be imputed.}

\item{imputeWeights}{An R object containing impute weights as returned by \code{getImputeWeights(ArchRProj)}. See \code{addImputeWeights()} for more details.}

\item{threads}{The number of threads to be used for parallel computing.}

\item{verbose}{A boolean value indicating whether to use verbose output during execution of this function. Can be set to FALSE for a cleaner output.}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
This function gets imputation weights from an ArchRProject to impute a numerical matrix
}
