% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{loadArchRProject}
\alias{loadArchRProject}
\title{Load Previous ArchRProject into R}
\usage{
loadArchRProject(path = "./", force = FALSE, showLogo = TRUE)
}
\arguments{
\item{path}{A character path to an ArchRProject directory that was previously saved.}

\item{force}{A boolean indicating when re-normalizing paths if an annotation/bdgPeaks is not found ignore and continue}

\item{showLogo}{show ArchRLogo upon completion.}
}
\description{
This function will load a previously saved ArchRProject and re-normalize paths for usage.
}
