% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperUtils.R
\name{mapLabels}
\alias{mapLabels}
\title{Re-map a character vector of labels from an old set of labels to a new set of labels}
\usage{
mapLabels(labels = NULL, newLabels = NULL, oldLabels = names(newLabels))
}
\arguments{
\item{labels}{A character vector containing lables to map.}

\item{newLabels}{A character vector (same length as oldLabels) to map labels to from oldLabels.}

\item{oldLabels}{A character vector (same length as newLabels) to map labels from to newLabels}
}
\description{
This function takes a character vector of labels and uses a set of old and new labels
to re-map from the old label set to the new label set.
}
\examples{

# Get Test ArchR Project
proj <- getTestProject()

# Get Peak Annotations
proj$ClusterLabels <- mapLabels(proj$Clusters, c("T", "B", "M"), c("C1", "C2", "C3"))

}
