% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkerFeatures.R
\name{markerAnnoEnrich}
\alias{markerAnnoEnrich}
\title{QQQ}
\usage{
markerAnnoEnrich(seMarker = NULL, ArchRProj = NULL,
  annotations = NULL, matches = NULL,
  cutOff = "FDR <= 0.1 & Log2FC >= 0.5", background = "bgdPeaks", ...)
}
\arguments{
\item{seMarker}{QQQ A `SummarizedExperiment` object returned by `ArchR::markerFeatures()`.}

\item{ArchRProj}{An `ArchRProject` object.}

\item{annotations}{QQQ}

\item{matches}{QQQ}

\item{cutOff}{A valid-syntax logical statement that defines which marker features from `seMarker` will be plotted in the heatmap. `cutoff` can contain any of the `assayNames` from `seMarker`.}

\item{background}{QQQ}

\item{...}{additional args}
}
\description{
This function QQQ
}
