% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkerFeatures.R
\name{markerFeatures}
\alias{markerFeatures}
\title{Identify Marker Features for each Group}
\usage{
markerFeatures(ArchRProj = NULL, groupBy = "Clusters",
  useGroups = NULL, bdgGroups = NULL, useMatrix = "GeneScoreMatrix",
  bias = c("TSSEnrichment", "log10(nFrags)"), normBy = NULL,
  testMethod = "wilcoxon", minCells = 50, maxCells = 500,
  threads = 1, k = 100, bufferRatio = 0.8, binarize = FALSE,
  useSeqnames = NULL, method = "ArchR", verboseHeader = TRUE,
  verboseAll = FALSE, ...)
}
\arguments{
\item{ArchRProj}{ArchR Project}

\item{groupBy}{group cells by this column in cellColData}

\item{useGroups}{use subset of groups in group column in cellColData for comparisons}

\item{bdgGroups}{use subset of groups in group column in cellColData for background}

\item{useMatrix}{matrix name in Arrow Files that will be used for identifying features}

\item{bias}{biases to account for in selecting null group using info from cellColData}

\item{normBy}{normalize by column in cellColData prior to test}

\item{testMethod}{pairwise test method group vs null}

\item{minCells}{minimum cells per group for testing}

\item{maxCells}{maximum cells per group for testing}

\item{k}{knn for matching cell biases}

\item{bufferRatio}{buffering ratio for matching cell biases}

\item{binarize}{binarize prior to testing}

\item{useSeqnames}{specific seqnames to use only}

\item{method}{marker identification method}

\item{verboseHeader}{verbose sections}

\item{verboseAll}{verbose sections and subsections}

\item{...}{additional args}
}
\description{
This function will identify a null set of cells that match biases per cell
while maintaining the input group proportions. Then it will compute a pairwise
test of the group vs the null set.
}
