% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkerFeatures.R
\name{markerFeatures}
\alias{markerFeatures}
\title{Identify Marker Features for each cell grouping}
\usage{
markerFeatures(ArchRProj = NULL, groupBy = "Clusters",
  useGroups = NULL, bdgGroups = NULL, useMatrix = "GeneScoreMatrix",
  bias = c("TSSEnrichment", "log10(nFrags)"), normBy = NULL,
  testMethod = "wilcoxon", minCells = 50, maxCells = 500,
  scaleTo = 10^4, threads = 1, k = 100, bufferRatio = 0.8,
  binarize = FALSE, useSeqnames = NULL, method = "ArchR",
  verboseHeader = TRUE, verboseAll = FALSE, ...)
}
\arguments{
\item{ArchRProj}{An `ArchRProject` object.}

\item{groupBy}{QQQ The name of the column in `cellColData` to use for grouping cells together for marker feature identification.}

\item{useGroups}{QQQ A character vector that is used to select a subset of groups by name from the designated `groupBy` column in `cellColData`. This limits the groups used to perform marker feature identification.}

\item{bdgGroups}{QQQ A character vector that is used to select a subset of groups by name from the designated `groupBy` column in `cellColData` to be used for background calculations in marker feature identification.}

\item{useMatrix}{QQQ The name of the matrix to be used for performing doublet identification analyses. Options include "TileMatrix", QQQ.}

\item{bias}{QQQ A character vector indicating the potential bias variables to account for in selecting a matched null group for marker feature identification. These should be column names from `cellColData`.}

\item{normBy}{QQQ The name of a column in `cellColData` that should be normalized across cells prior to performing marker feature identification.}

\item{testMethod}{QQQ The name of the pairwise test method to use in comparing cell groupings to the null cell grouping during marker feature identification. Valid options include "wilcoxon", "ttest", and "binomial".}

\item{minCells}{QQQ The minimum number of cells per cell group required to perform marker feature identification.}

\item{maxCells}{QQQ The maximum number of cells to consider from a single cell group when performing marker feature identification.}

\item{scaleTo}{QQQ}

\item{threads}{The number of threads to be used for parallel computing.}

\item{k}{QQQ UNCLEAR knn for matching cell biases}

\item{bufferRatio}{QQQ UNCLEAR buffering ratio for matching cell biases}

\item{binarize}{A boolean value indicating whether to binarize the matrix indicated by `useMatrix` prior to marker feature identification. This is often used when dealing with an insertion counts matrix such as a `TileMatrix`.}

\item{useSeqnames}{A character vector that indicates which seqnames should be used in marker feature identification. Features from seqnames that are not listed will be ignored.}

\item{method}{The name of the method to be used  for marker feature identification. Valid options are "ArchR" which will use the default ArchR method or "Venice" which will use the `Signac::VeniceMarker()` fucntion.}

\item{verboseHeader}{A boolean value that determines whether standard output includes verbose sections.}

\item{verboseAll}{A boolean value that determines whether standard output includes verbose subsections.}

\item{...}{additional args}
}
\description{
This function will identify features that are definitional of each provided cell grouping where possible
}
