% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkerHeatmap.R
\name{markerHeatmap}
\alias{markerHeatmap}
\title{Plot a Heatmap of Identified Marker Features}
\usage{
markerHeatmap(seMarker, FDR = 0.001, log2FC = 0.1, log2Norm = TRUE,
  scaleTo = 10^4, scaleRows = TRUE, limits = c(-2, 2),
  grepExclude = NULL, pal = NULL, binaryClusterRows = TRUE,
  labelMarkers = NULL, labelTop = NULL, labelRows = FALSE,
  returnMat = FALSE, ...)
}
\arguments{
\item{seMarker}{Summarized Experiment result from markerFeatures}

\item{FDR}{False-Discovery Rate Cutoff to Be called a Marker}

\item{log2FC}{Log2 Fold Change Cutoff to Be called a Marker}

\item{log2Norm}{log2 Normalization prior to plotting set true for counting assays (not DeviationsMatrix!)}

\item{scaleTo}{scale to prior to log2 Normalization, if log2Norm is FALSE this does nothing}

\item{scaleRows}{compute row z-scores on matrix}

\item{limits}{heatmap color limits}

\item{grepExclude}{remove features by grep}

\item{pal}{palette for heatmap, default will use solar_extra}

\item{binaryClusterRows}{fast clustering implementation for row clustering by binary sorting}

\item{labelMarkers}{label specific markers by name on heatmap (matches rownames of seMarker)}

\item{labelTop}{label the top features for each column in seMarker}

\item{labelRows}{label all rows}

\item{returnMat}{return final matrix that is used for plotting heatmap}

\item{...}{additional args}
}
\description{
This function will plot a heatmap of the results from markerFeatures
}
