% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkerFeatures.R, R/MarkerHeatmap.R
\name{markerHeatmap}
\alias{markerHeatmap}
\title{Plot a Heatmap of Identified Marker Features}
\usage{
markerHeatmap(seMarker, cutOff = "FDR <= 0.001 & Log2FC >= 0.1",
  log2Norm = TRUE, scaleTo = 10^4, scaleRows = TRUE, limits = c(-2,
  2), grepExclude = NULL, pal = NULL, binaryClusterRows = TRUE,
  labelMarkers = NULL, labelTop = NULL, labelRows = FALSE,
  returnMat = FALSE, ...)

markerHeatmap(seMarker, cutOff = "FDR <= 0.001 & Log2FC >= 0.1",
  log2Norm = TRUE, scaleTo = 10^4, scaleRows = TRUE, limits = c(-2,
  2), grepExclude = NULL, pal = NULL, binaryClusterRows = TRUE,
  labelMarkers = NULL, labelTop = NULL, labelRows = FALSE,
  returnMat = FALSE, ...)
}
\arguments{
\item{seMarker}{QQQ A `SummarizedExperiment` object returned by `ArchR::markerFeatures()`.}

\item{cutOff}{A valid-syntax logical statement that defines which marker features from `seMarker` will be plotted in the heatmap. `cutoff` can contain any of the `assayNames` from `seMarker`.}

\item{log2Norm}{A boolean value indicating whether a log2 transformation whould be performed on the values in `seMarker` prior to plotting. Should be set to `TRUE` for counts-based assays (but not assays like `DeviationsMatrix`).}

\item{scaleTo}{QQQ scale to prior to log2 Normalization, if log2Norm is FALSE this does nothing}

\item{scaleRows}{A boolean value that indicates whether the heatmap should display row-wise z-scores instead of raw values.}

\item{limits}{A numeric vector of two numbers that represent the lower and upper color limits of the heatmap color scheme.}

\item{grepExclude}{QQQ A character vector or string that indicates the `rownames` or a specific pattern that identifies rownames from `seMarker` to be excluded from the heatmap.}

\item{pal}{QQQ The name or numeric index of a custom palette from `ArchRPalettes` to use for the heatmap}

\item{binaryClusterRows}{A boolean value that indicates whether a binary sorting algorithm should be used for fast clustering of heatmap rows.}

\item{labelMarkers}{A character vector listing the `rownames` of `seMarker` that should be labeled on the side of the heatmap.}

\item{labelTop}{A boolean value that indicates whether the top features for each column in `seMarker` should be labeled on the side of the heatmap.}

\item{labelRows}{A boolean value that indicates whether all rows should be labeled on the side of the heatmap.}

\item{returnMat}{QQQ A boolean value that indicates whether the final heatmap matrix should be returned in lieu of plotting the actual heatmap.}

\item{...}{additional args}

\item{invert}{QQQ}

\item{seMarker}{QQQ A `SummarizedExperiment` object returned by `ArchR::markerFeatures()`.}

\item{cutOff}{A valid-syntax logical statement that defines which marker features from `seMarker` will be plotted in the heatmap. `cutoff` can contain any of the `assayNames` from `seMarker`.}

\item{log2Norm}{A boolean value indicating whether a log2 transformation should be performed on the values in `seMarker` prior to plotting. Should be set to `TRUE` for counts-based assays (but not assays like `DeviationsMatrix`).}

\item{scaleTo}{QQQ scale to prior to log2 Normalization, if log2Norm is FALSE this does nothing}

\item{scaleRows}{A boolean value that indicates whether the heatmap should display row-wise z-scores instead of raw values.}

\item{limits}{A numeric vector of two numbers that represent the lower and upper color limits of the heatmap color scheme.}

\item{grepExclude}{QQQ A character vector or string that indicates the `rownames` or a specific pattern that identifies rownames from `seMarker` to be excluded from the heatmap.}

\item{pal}{QQQ The name or numeric index of a custom palette from `ArchRPalettes` to use for the heatmap}

\item{binaryClusterRows}{A boolean value that indicates whether a binary sorting algorithm should be used for fast clustering of heatmap rows.}

\item{labelMarkers}{A character vector listing the `rownames` of `seMarker` that should be labeled on the side of the heatmap.}

\item{labelTop}{A boolean value that indicates whether the top features for each column in `seMarker` should be labeled on the side of the heatmap.}

\item{labelRows}{A boolean value that indicates whether all rows should be labeled on the side of the heatmap.}

\item{returnMat}{QQQ A boolean value that indicates whether the final heatmap matrix should be returned in lieu of plotting the actual heatmap.}

\item{invert}{QQQ}

\item{...}{additional args}
}
\description{
This function will plot a heatmap of the results from markerFeatures

This function will plot a heatmap of the results from markerFeatures
}
