% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkerFeatures.R
\name{markerHeatmap}
\alias{markerHeatmap}
\title{Plot a Heatmap of Identified Marker Features}
\usage{
markerHeatmap(seMarker = NULL, cutOff = "FDR <= 0.01 & Log2FC >= 0.5",
  log2Norm = TRUE, scaleTo = 10^4, scaleRows = TRUE,
  plotLog2FC = FALSE, limits = c(-2, 2), grepExclude = NULL,
  pal = NULL, binaryClusterRows = TRUE, labelMarkers = NULL,
  nLabel = NULL, nPrint = 20, labelRows = FALSE, returnMat = FALSE,
  invert = FALSE, ...)
}
\arguments{
\item{seMarker}{A \code{SummarizedExperiment} object returned by \code{markerFeatures()}.}

\item{cutOff}{A valid-syntax logical statement that defines which marker features from \code{seMarker} will be plotted in the heatmap. \code{cutoff} can contain any of the \code{assayNames} from \code{seMarker}.}

\item{log2Norm}{A boolean value indicating whether a log2 transformation should be performed on the values in \code{seMarker} prior to plotting. Should be set to \code{TRUE} for counts-based assays (but not assays like "DeviationsMatrix").}

\item{scaleTo}{This numerical value to depth-normalize to prior to log2 Normalization, if log2Norm is FALSE this does nothing}

\item{scaleRows}{A boolean value that indicates whether the heatmap should display row-wise z-scores instead of raw values.}

\item{limits}{A numeric vector of two numbers that represent the lower and upper limits of the heatmap color scheme.}

\item{grepExclude}{A character vector or string that indicates the \code{rownames} or a specific pattern that identifies rownames from \code{seMarker} to be excluded from the heatmap.}

\item{pal}{A custom continuous palette from \code{ArchRPalettes} (see \code{paletteContinuous()}) used to override the default continuous palette for the heatmap.}

\item{binaryClusterRows}{A boolean value that indicates whether a binary sorting algorithm should be used for fast clustering of heatmap rows.}

\item{labelMarkers}{A character vector listing the \code{rownames} of \code{seMarker} that should be labeled on the side of the heatmap.}

\item{nLabel}{An integer value that indicates whether the top \code{n} features for each column in \code{seMarker} should be labeled on the side of the heatmap.}

\item{labelRows}{A boolean value that indicates whether all rows should be labeled on the side of the heatmap.}

\item{returnMat}{A boolean value that indicates whether the final heatmap matrix should be returned in lieu of plotting the actual heatmap.}

\item{invert}{A boolean value that indicates whether the heatmap will be inverted when sorting features (rows) and color. For example, this is useful when looking for down-regulated markers (Log2FC < 0) instead of up-regulated markers (Log2FC > 0).}

\item{...}{additional args}
}
\description{
This function will plot a heatmap of the results from markerFeatures
}
