% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkerFeatures.R
\name{markerPlot}
\alias{markerPlot}
\title{Plot Differential Markers}
\usage{
markerPlot(seMarker = NULL, name = NULL,
  cutOff = "FDR <= 0.01 & abs(Log2FC) >= 0.5", plotAs = "Volcano",
  scaleTo = 10^4, ...)
}
\arguments{
\item{seMarker}{A \code{SummarizedExperiment} object returned by \code{ArchR::markerFeatures()}.}

\item{name}{column/group name of seMarker to be plotted.}

\item{cutOff}{A valid-syntax logical statement that defines which marker features from \code{seMarker} will be plotted. \code{cutoff} can contain any of the \code{assayNames} from \code{seMarker}.}

\item{plotAs}{plot as "Volcano" or "MA" plot.}

\item{...}{additional args}
}
\description{
This function will plot one group/column of a differential markers se as a MA or Volcano plot.
}
