% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkerFeatures.R
\name{markerRanges}
\alias{markerRanges}
\title{QQQ}
\usage{
markerRanges(seMarker, cutOff = "FDR <= 0.01 & Log2FC >= 0.5", ...)
}
\arguments{
\item{seMarker}{QQQ A `SummarizedExperiment` object returned by `ArchR::markerFeatures()`.}

\item{cutOff}{A valid-syntax logical statement that defines which marker features from `seMarker` will be plotted in the heatmap. `cutoff` can contain any of the `assayNames` from `seMarker`.}

\item{...}{additional args}
}
\description{
This function QQQ
}
