% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesUtils.R
\name{mergeGRanges}
\alias{mergeGRanges}
\title{Merge Genomic Ranges}
\usage{
mergeGRanges(gr, ignore.strand = TRUE)
}
\arguments{
\item{gr}{see validGRanges}

\item{ignore.strand}{ignore strandedness for merging}
}
\description{
Merge Genomic Ranges
}
