% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesUtils.R
\name{mergeGRanges}
\alias{mergeGRanges}
\title{QQQ Merge regions within a single Genomic Ranges object}
\usage{
mergeGRanges(gr, ignore.strand = TRUE)
}
\arguments{
\item{gr}{A \code{GRanges} object.}

\item{ignore.strand}{A boolean value indicating whether strandedness should be ignored in \code{findOverlaps()}.}
}
\description{
QQQ This function merges overlapping regions within a single Genomic Ranges object
}
