% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesUtils.R
\name{nonOverlappingGR}
\alias{nonOverlappingGR}
\title{Retreive a non-overlapping set of regions from a Genomic Ranges object}
\usage{
nonOverlappingGR(gr = NULL, by = "score", decreasing = TRUE,
  verbose = FALSE)
}
\arguments{
\item{gr}{A \code{GRanges} object.}

\item{by}{The name of a column in \code{mcols(gr)} that should be used to determine how overlapping regions should be resolved. The resolution of overlapping regions also depends on \code{decreasing}. For example, if a column named "score" is used for \code{by}, \code{decreasing = TRUE} means that the highest "score" in the overlap will be retained and \code{decreasing = FALSE} means that the lowest "score" in the overlap will be retained.}

\item{decreasing}{A boolean value indicating whether the values in the column indicated via \code{by} should be ordered in decreasing order. If \code{TRUE}, the higher value in \code{by} will be retained.}

\item{verbose}{A boolean value that determines whether the output should include extra reporting.}
}
\description{
This function returns a GRanges object containing a non-overlapping set regions derived from a supplied Genomic Ranges object.
}
