% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesUtils.R
\name{nonOverlappingGRanges}
\alias{nonOverlappingGRanges}
\title{QQQ Retreive a non-overlapping set of regions from a Genomic Ranges object}
\usage{
nonOverlappingGRanges(gr, by = "score", decreasing = TRUE,
  verbose = FALSE)
}
\arguments{
\item{gr}{A `GRanges` object.}

\item{by}{QQQ The name of a column in `mcols(gr)` that should be used to determine how overlapping regions should be resolved. The resolution of overlapping regions also depends on `decreasing`. For example, if a column named "score" is used for `by`, `decreasing = TRUE` means that the highest "score" in the overlap will be retained and `decreasing = FALSE` means that the lowest "score" in the overlap will be retained.}

\item{decreasing}{QQQ A boolean value indicating whether the values in the column indicated via `by` should be ordered in decreasing order. If `TRUE`, the higher value in `by` will be retained.}

\item{verbose}{A boolean value that determines whether the output should include extra reporting.}
}
\description{
QQQ This function returns a GRanges object containing a non-overlapping set regions derived from a supplied Genomic Ranges object.
}
