% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesUtils.R
\name{overlappingBP}
\alias{overlappingBP}
\title{Merge Genomic Ranges}
\usage{
overlappingBP(query, subject, ignore.strand = TRUE)
}
\arguments{
\item{query}{see validGRanges}

\item{subject}{see validGRanges}

\item{ignore.strand}{ignore strandedness for overlaps}
}
\description{
Merge Genomic Ranges
}
