% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesUtils.R
\name{overlappingBP}
\alias{overlappingBP}
\title{QQQ Identify the number of bases that overlap two Genomic Ranges objects}
\usage{
overlappingBP(query, subject, ignore.strand = TRUE)
}
\arguments{
\item{query}{A `GRanges` object to be used as the query in `findOverlaps()`.}

\item{subject}{A `GRanges` object to be used as the subject in `findOverlaps()`.}

\item{ignore.strand}{A boolean value indicating whether strandedness should be ignored in `findOverlaps()`.}
}
\description{
QQQ This function returns a data.frame describing how many basepairs overlap the provided query and subject Genomic Ranges objects
}
