% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesUtils.R
\name{overlapsMany}
\alias{overlapsMany}
\title{Overlaps Many includes information from mcols(gr)}
\usage{
overlapsMany(query, subject, by, ignore.strand = TRUE)
}
\arguments{
\item{query}{see validGRanges}

\item{subject}{see validGRanges}

\item{by}{column in subject to split overlaps by}

\item{ignore.strand}{ignore strandedness for overlaps}
}
\description{
Overlaps Many includes information from mcols(gr)
}
