% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesUtils.R
\name{overlapsManyGR}
\alias{overlapsManyGR}
\title{Overlap with many genomic regions.}
\usage{
overlapsManyGR(query = NULL, subject = NULL, by = NULL,
  ignoreStrand = TRUE)
}
\arguments{
\item{query}{A \code{GRanges} object to be used as the query in \code{GenomicRanges::findOverlaps()}.}

\item{subject}{A \code{GRanges} object containing a column in \code{mcols(subject)} sub-grouping to be used as the subject in \code{GenomicRanges::findOverlaps()}.}

\item{by}{The name of a column in \code{mcols(subject)} that should be used to determine how overlapping regions should be sub-grouped.}

\item{ignoreStrand}{A boolean value indicating whether strandedness should be ignored in \code{GenomicRanges::findOverlaps()}.}
}
\description{
This function returns a sparse matrix that describes the overlap with each sub-grouped genomic region as specified in the "by" column.
}
