% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorPalettes.R
\name{paletteContinuous}
\alias{paletteContinuous}
\title{Continuous Color Palette}
\usage{
paletteContinuous(set = "solarExtra", n = 256, reverse = FALSE)
}
\arguments{
\item{set}{The name of a color palette provided in the \code{ArchRPalettes} list object.}

\item{n}{The number of unique colors to generate as part of this continuous color palette.}

\item{reverse}{A boolean variable that indicates whether to return the palette colors in reverse order.}
}
\description{
Continuous Color Palette
}
