% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorPalettes.R
\name{paletteDiscrete}
\alias{paletteDiscrete}
\title{Optimized discrete color palette}
\usage{
paletteDiscrete(set = "stallion", values, reverse = FALSE,
  returnStructure = FALSE, ...)
}
\arguments{
\item{set}{continuous palette name or number}

\item{values}{is a vector containing the sample names used in the plot which will be given a color}

\item{reverse}{return reversed values}

\item{returnStructure}{return structure palette}
}
\description{
This function assesses the number of inputs and returns a tailored color palette for aesthetics
}
