% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_arch_for_clusters.R
\name{plot_arch_for_clusters}
\alias{plot_arch_for_clusters}
\title{Plot the sequence architectures of the sequence clusters.}
\usage{
plot_arch_for_clusters(
  givenSamplesMatrix,
  givenFeaturesMatrix,
  nCluster,
  clustering_sol,
  seqs = NULL,
  position_labels = NA,
  plotMethod = "custom",
  add_pseudo_counts = FALSE,
  sinuc_or_dinuc = "sinuc"
)
}
\arguments{
\item{givenSamplesMatrix}{The samples matrix resulting for NMF.}

\item{givenFeaturesMatrix}{The features matrix resulting from NMF.}

\item{nCluster}{The number of clusters.}

\item{clustering_sol}{The clustering solution as returned by
\code{get_clusters} function.}

\item{seqs}{For representing the sequences as an image
(TO-DO: May not be needed).
Default value is NULL when no image gets printed.}

\item{position_labels}{Labels of the positions in the sequences. Used for
visualization with function
\code{viz_basis_vectors_in_combined_heatmaps_seqlogos}.}

\item{plotMethod}{'custom' or 'bits', passed to ggseqlogo}

\item{add_pseudo_counts}{OK}

\item{sinuc_or_dinuc}{'sinuc' or 'dinuc'}
}
\value{
nothing.
}
\description{
The different architectures characteristic of the different
sequence clusters are visualized. Therefore, the function takes as arguments
the samplesMatrix and featuresMatrix output by NMF, the number of clusters,
the clustering solution obtained and the position labels for the sequences.
}
