% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_arch_for_clusters.R
\name{plot_arch_for_clusters}
\alias{plot_arch_for_clusters}
\title{Plot cluster architectures as sequence logos.}
\usage{
plot_arch_for_clusters(
  seqs,
  clust_list,
  pos_lab,
  xt_freq = 5,
  set_titles = TRUE,
  pdf_width = 11,
  pdf_height = 2,
  pdf_name = "archR_sequence_architectures.pdf",
  ...
)
}
\arguments{
\item{seqs}{Sequences as a \code{\link{Biostrings::DNAStringSet}}.}

\item{clust_list}{Clusters as a list of sequence IDs in each cluster.}

\item{pos_lab}{Labels for sequence positions, should be of same
length as length of the sequences.}

\item{xt_freq}{Frequency of x-axis ticks.}

\item{pdf_width, pdf_height}{Width and height in inches of the PDF file. 
Default values are 11 and 2.}

\item{pdf_name}{Specify the PDF filename.}

\item{...}{Additional args passed to \code{\link{plot_ggseqlogo_of_seqs}}.}

\item{method}{Specify either 'bits' for information content or 
'prob' for probability.}
}
\value{
A list of (ggplot2-based) sequence logo plots is returned. When a 
valid file name is specified, the list of plots is also written to the PDF 
file (one plot per page).
}
\description{
Given a collection of FASTA sequences as a DNAStringSet object,
and the clusters information, this function plots the architectures for all
clusters. If a name for the PDF file is provided, the resulting set of
architecture sequence logos are saved as a multi-page PDF.
}
