% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ggheatmap.R
\name{plot_ggseqlogo}
\alias{plot_ggseqlogo}
\title{Visualize a given (PWM) matrix as a sequence logo.}
\usage{
plot_ggseqlogo(pwm_mat, method = "custom", pos_lab = NULL, pdf_name = NULL)
}
\arguments{
\item{pwm_mat}{Matrix (usually a PWM, but can be any non-normalized matrix) 
to be represented as a sequence logo.}

\item{method}{For \code{ggseqlogo}; either of 'custom', 'bits', or
'probability'. Default is 'custom'.}

\item{pos_lab}{Labels of the positions in the sequences.}

\item{pdf_name}{Name of the file which will be saved as PDF.}
}
\value{
A ggplot2 object so you can simply call \code{print} or \code{save}
on it later. If \code{pdf_name} is given, it is also saved in addition to 
returning the ggplot object.
}
\description{
Visualize a given (PWM) matrix as a sequence logo.
}
\seealso{
\code{\link{plot_ggheatmap}} for plotting PWMs as heatmaps,
\code{\link{plot_ggseqlogo_of_seqs}} for visualizing a collection of 
sequences by their sequence logo.

Other visualization functions: 
\code{\link{plot_ggheatmap}()},
\code{\link{viz_basis_vectors_as_heatmap}()},
\code{\link{viz_basis_vectors_as_seqlogo}()},
\code{\link{viz_basis_vectors_in_combined_heatmaps_seqlogos}()},
\code{\link{viz_seqs_as_acgt_mat_from_seqs}()}
}
\concept{visualization functions}
