% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ggseqlogo.R
\name{plot_ggseqlogo}
\alias{plot_ggseqlogo}
\title{Visualize a given (PWM) matrix as a sequence logo.}
\usage{
plot_ggseqlogo(pwmMat, position_labels = NULL, savePDFfilename = NULL)
}
\arguments{
\item{pwmMat}{Matrix (usually a PWM, but can be non-normalized/any matrix)
to be represented as a sequence logo.}

\item{position_labels}{Labels of the positions in the sequences.}

\item{savePDFfilename}{Name of the file which will be saved as PDF.}
}
\value{
A ggplot object so you can simply call \code{print} or \code{save}
on it later. If \code{savePDFfilename} is given, it is also saved and the
\code{ggplot} object returned.
}
\description{
Visualize a given (PWM) matrix as a sequence logo.
}
\examples{


}
