% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ggseqlogo.R
\name{plot_ggseqlogo}
\alias{plot_ggseqlogo}
\title{Visualize a given (PWM) matrix as a sequence logo}
\usage{
plot_ggseqlogo(
  pwmMat,
  plotMethod = "custom",
  position_labels = NULL,
  savePDFfilename = NULL
)
}
\arguments{
\item{pwmMat}{Matrix (usually a PWM, but can be any non-normalized matrix) to
be represented as a sequence logo.}

\item{plotMethod}{For \code{ggseqlogo}; either of 'custom', 'bits', or
'probability'. Default is 'custom'.}

\item{position_labels}{Labels of the positions in the sequences.}

\item{savePDFfilename}{Name of the file which will be saved as PDF.}
}
\value{
A ggplot object so you can simply call \code{print} or \code{save}
on it later. If \code{savePDFfilename} is given, it is also saved and the
\code{ggplot} object returned.
}
\description{
Visualize a given (PWM) matrix as a sequence logo
}
\seealso{
\code{\link{plot_ggheatmap}} for plotting PWMs as heatmaps

Other visualization functions: 
\code{\link{plot_ggheatmap}()},
\code{\link{viz_basis_vectors_as_heatmap}()},
\code{\link{viz_basis_vectors_as_seqlogo}()},
\code{\link{viz_basis_vectors_in_combined_heatmaps_seqlogos}()},
\code{\link{viz_matrix_of_acgt}()}
}
\concept{visualization functions}
