% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_arch_for_clusters.R
\name{plot_ggseqlogo_of_seqs}
\alias{plot_ggseqlogo_of_seqs}
\title{Plot sequence logo of a collection of sequences}
\usage{
plot_ggseqlogo_of_seqs(seqs, position_labels, xt_freq = 1, title = "TITLE")
}
\arguments{
\item{seqs}{Collection of sequences as a Biostrings::DNAStringSet object.}

\item{position_labels}{Numeric vector of labels for sequence positions.
This should be the samelength as the width of the given sequences.}

\item{xt_freq}{Specify the frequency of the x-axis ticks.}

\item{title}{The title for the plot}
}
\description{
A wrapper to ggseqlogo plotting. Given a collection of
sequences, this function plots the sequence logo.
}
