% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeData.R
\name{plotEmbedding}
\alias{plotEmbedding}
\title{Visualize Embedding from ArchR Project}
\usage{
plotEmbedding(ArchRProj = NULL, embedding = "UMAP",
  colorBy = "colData", name = "Sample", log2Norm = NULL,
  imputeWeights = getImputeWeights(ArchRProj), pal = NULL,
  size = 0.5, rastr = TRUE, quantCut = c(0.01, 0.99),
  discreteSet = NULL, continuousSet = NULL, randomize = TRUE,
  keepAxis = FALSE, baseSize = 6, plotAs = NULL,
  plotParams = list(), ...)
}
\arguments{
\item{ArchRProj}{An `ArchRProject` object.}

\item{embedding}{The name of the embedding to plot. See `ArchR::computeEmbedding()` for more information.}

\item{colorBy}{QQQ UNCLEAR A string indicating whether points in the plot should be colored by a column in cellColData ("cellColData") or by the gene activity score matrix ("GeneScoreMatrix").}

\item{name}{QQQ UNCLEAR The name of the column in `cellColData` or `GeneScoreMatrix` to use for QQQ.}

\item{log2Norm}{QQQ A boolean value indicating whether a log2 transformation should be performed on the values in QQQ plotting.}

\item{imputeWeights}{QQQ}

\item{pal}{QQQ The name or numeric index of a custom palette from `ArchRPalettes` to use for QQQ.}

\item{size}{QQQ A number indicating the size of the points to plot if `plotAs` is set to "points".}

\item{rastr}{A boolean value that indicates whether the plot should be rasterized. This does not rasterize lines and labels, just the internal portions of the plot.}

\item{quantCut}{If this is not null, a quantile cut is performed to threshold the top and bottom of the distribution. This prevents skewed color scales caused by strong outliers. The format of this should be c(x,y) where x is the upper threshold and y is the lower threshold. For example, quantileCut = c(0.975,0.025) will take the top and bottom 2.5% of values and set them to the value of the 97.5th and 2.5th percentile values respectively.}

\item{discreteSet}{QQQ The name or numeric index of a discrete palette from `ArchRPalettes` for visualizing QQQ in the embedding.}

\item{continuousSet}{QQQ The name or numeric index of a continuous palette from `ArchRPalettes` for visualizing QQQ in the embedding.}

\item{randomize}{A boolean value that indicates whether to randomize points prior to plotting to prevent cells from one cluster being present at the front of the plot.}

\item{keepAxis}{A boolean value that indicates whether the x and y axis ticks and labels should be plotted.}

\item{baseSize}{QQQ The base font size to use in the plot.}

\item{plotAs}{A string that indicates whether points ("points") should be plotted or a hexplot ("hex") should be plotted.}

\item{plotParams}{Additional parameters to pass to `ggPoint()` or `ggHex()`.}

\item{...}{additional args}

\item{quantHex}{QQQ UNCLEAR quantile evaluation for each hex in geom_hex}
}
\description{
This function will plot an embedding that was created from
computeEmbedding
}
