% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeData.R
\name{plotEmbedding}
\alias{plotEmbedding}
\title{Visualize an Embedding from ArchR Project}
\usage{
plotEmbedding(
  ArchRProj = NULL,
  embedding = "UMAP",
  colorBy = "cellColData",
  name = "Sample",
  log2Norm = NULL,
  imputeWeights = if (!grepl("coldata", tolower(colorBy[1])))
    getImputeWeights(ArchRProj),
  pal = NULL,
  size = 0.1,
  sampleCells = NULL,
  rastr = TRUE,
  quantCut = c(0.001, 0.999),
  discreteSet = NULL,
  continuousSet = NULL,
  randomize = TRUE,
  keepAxis = FALSE,
  baseSize = 10,
  plotAs = NULL,
  threads = getArchRThreads(),
  ...
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{embedding}{The name of the embedding stored in the \code{ArchRProject} to be plotted. See \code{computeEmbedding()} for more information.}

\item{colorBy}{A string indicating whether points in the plot should be colored by a column in \code{cellColData} ("cellColData") or by
a data matrix in the corresponding ArrowFiles (i.e. "GeneScoreMatrix", "MotifMatrix", "PeakMatrix").}

\item{name}{The name of the column in \code{cellColData} or the featureName/rowname of the data matrix to be used for plotting.
For example if colorBy is "cellColData" then \code{name} refers to a column name in the \code{cellcoldata} (see \code{getCellcoldata()}). If \code{colorBy}
is "GeneScoreMatrix" then \code{name} refers to a gene name which can be listed by \code{getFeatures(ArchRProj, useMatrix = "GeneScoreMatrix")}.}

\item{log2Norm}{A boolean value indicating whether a log2 transformation should be performed on the values (if continuous) in plotting.}

\item{imputeWeights}{The weights to be used for imputing numerical values for each cell as a linear combination of other cells values.
See \code{addImputationWeights()} and \code{getImutationWeights()} for more information.}

\item{pal}{A custom palette (see \code{paletteDiscrete} or \code{ArchRPalettes}) used to override discreteSet/continuousSet for coloring vector.}

\item{size}{A number indicating the size of the points to plot if \code{plotAs} is set to "points".}

\item{rastr}{A boolean value that indicates whether the plot should be rasterized. This does not rasterize lines and labels, just the
internal portions of the plot.}

\item{quantCut}{If this is not \code{NULL}, a quantile cut is performed to threshold the top and bottom of the distribution of numerical values.
This prevents skewed color scales caused by strong outliers. The format of this should be c(x,y) where x is the lower threshold and y is
the upper threshold. For example, quantileCut = c(0.025,0.975) will take the 2.5th percentile and 97.5 percentile of values and set
values below/above to the value of the 2.5th and 97.5th percentile values respectively.}

\item{discreteSet}{The name of a discrete palette from \code{ArchRPalettes} for visualizing \code{colorBy} in the embedding if a discrete color set is desired.}

\item{continuousSet}{The name of a continuous palette from \code{ArchRPalettes} for visualizing \code{colorBy} in the embedding if a continuous color set is desired.}

\item{randomize}{A boolean value that indicates whether to randomize points prior to plotting to prevent cells from one cluster being
uniformly present at the front of the plot.}

\item{keepAxis}{A boolean value that indicates whether the x- and y-axis ticks and labels should be plotted.}

\item{baseSize}{The base font size to use in the plot.}

\item{plotAs}{A string that indicates whether points ("points") should be plotted or a hexplot ("hex") should be plotted. By default
if \code{colorBy} is numeric, then \code{plotAs} is set to "hex".}

\item{...}{Additional parameters to pass to \code{ggPoint()} or \code{ggHex()}.}
}
\description{
This function will plot an embedding stored in an ArchRProject
}
