% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeData.R
\name{plotEmbedding}
\alias{plotEmbedding}
\title{Visualize Embedding from ArchR Project}
\usage{
plotEmbedding(ArchRProj = NULL, embedding = "UMAP",
  colorBy = "colData", name = "Sample", log2Norm = NULL,
  imputeWeights = NULL, pal = NULL, size = 0.5, rastr = TRUE,
  quantCut = c(0.01, 0.99), discreteSet = NULL, continuousSet = NULL,
  randomize = TRUE, keepAxis = FALSE, baseSize = 6, plotAs = NULL,
  plotParams = list(), ...)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{embedding}{The name of the embedding to plot. See \code{ArchR::computeEmbedding()} for more information.}

\item{colorBy}{A string indicating whether points in the plot should be colored by a column in cellColData ("cellColData") or by a data matrix in the ArrowFiles (i.e. "GeneScoreMatrix", "MotifMatrx", "PeakMatrix").}

\item{name}{The name of the column in \code{cellColData} or the featureName in the data matrix.}

\item{log2Norm}{A boolean value indicating whether a log2 transformation should be performed on the values (if continuous) in plotting.}

\item{imputeWeights}{imputation weights for imputing numerical values for each cell as a linear combination of other cells values (see add/getImutationWeights).}

\item{pal}{A custom palette used to override discreteSet/continuousSet for coloring vector.}

\item{size}{A number indicating the size of the points to plot if \code{plotAs} is set to "points".}

\item{rastr}{A boolean value that indicates whether the plot should be rasterized with ggrastr. This does not rasterize lines and labels, just the internal portions of the plot.}

\item{quantCut}{If this is not null, a quantile cut is performed to threshold the top and bottom of the distribution of numerical values.
This prevents skewed color scales caused by strong outliers. The format of this should be c(x,y) where x is the upper threshold and y is
the lower threshold. For example, quantileCut = c(0.975,0.025) will take the top and bottom 2.5% of values and set them to the value of
the 97.5th and 2.5th percentile values respectively.}

\item{discreteSet}{The name of a discrete palette from \code{ArchRPalettes} for visualizing colorBy in the embedding.}

\item{continuousSet}{The name of a continuous palette from \code{ArchRPalettes} for visualizing colorBy in the embedding.}

\item{randomize}{A boolean value that indicates whether to randomize points prior to plotting to prevent cells from one cluster being present at the front of the plot.}

\item{keepAxis}{A boolean value that indicates whether the x and y axis ticks and labels should be plotted.}

\item{baseSize}{The base font size to use in the plot.}

\item{plotAs}{A string that indicates whether points ("points") should be plotted or a hexplot ("hex") should be plotted.}

\item{plotParams}{Additional parameters to pass to \code{ggPoint()} or \code{ggHex()}.}

\item{...}{additional args}
}
\description{
This function will plot an embedding that was created from
computeEmbedding
}
