% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Footprinting.R
\name{plotFootprints}
\alias{plotFootprints}
\title{Plot Footprints}
\usage{
plotFootprints(
  seFoot = NULL,
  names = NULL,
  pal = NULL,
  flank = 250,
  flankNorm = 50,
  normMethod = "Subtract",
  smoothWindow = NULL,
  baseSize = 6,
  plot = TRUE,
  ArchRProj = NULL,
  plotName = paste0("Plot-Footprints-", normMethod),
  height = 6,
  width = 4,
  addDOC = TRUE,
  logFile = createLogFile("plotFootprints")
)
}
\arguments{
\item{seFoot}{JJJ A summarized experiment object containing information on footprints returned by the \code{getFootprints()} function.}

\item{names}{JJJ A character vector containing the names of the transcription factors to be plotted.}

\item{pal}{The name of a custom palette from \code{ArchRPalettes} to use for plotting the lines corresponding to the footprints.}

\item{flank}{The number of basepairs from the position center (+/-) to consider as the flank.}

\item{flankNorm}{The number of basepairs to consider at the edge of the flank region (+/-) to be used for footprint normalization.}

\item{normMethod}{The name of the normalization method to use to normalize the footprint relative to the Tn5 insertion bias. Options
include "none", "subtract", "divide". "Subtract" means subtracting the normalized Tn5 Bias. "Divide" means dividing the normalized Tn5 Bias.}

\item{smoothWindow}{The size in basepairs of the sliding window to be used for smoothing of the footprint signal.}

\item{baseSize}{JJJ}

\item{plot}{JJJ A boolean value indicating whether or not the footprints should be plotted (\code{TRUE}) or JJJ (\code{FALSE}).}

\item{ArchRProj}{JJJ An \code{ArchRProject} object to be used for JJJ.}

\item{plotName}{JJJ A string indicating the name of the file to be used for output plots.}

\item{height}{The height in inches to be used for the output PDF file.}

\item{width}{The width in inches to be used for the output PDF file.}

\item{addDOC}{A boolean variable that determines whether to add the date of creation to end of the PDF file name. This is useful for
preventing overwritting of old plots.}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
This function will get footprints for all samples in a given ArchRProject or a properly-formatted Summarized Experiment
}
