% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Footprinting.R
\name{plotFootprints}
\alias{plotFootprints}
\title{Plot Footprints}
\usage{
plotFootprints(
  seFoot = NULL,
  names = NULL,
  pal = NULL,
  flank = 250,
  flankNorm = 50,
  normMethod = "Subtract",
  smoothWindow = NULL,
  baseSize = 6,
  plot = TRUE,
  ArchRProj = NULL,
  plotName = paste0("Plot-Footprints-", normMethod),
  height = 6,
  width = 4,
  addDOC = TRUE,
  force = FALSE,
  logFile = createLogFile("plotFootprints")
)
}
\arguments{
\item{seFoot}{A summarized experiment object containing information on footprints returned by the \code{getFootprints()} function.}

\item{names}{A character vector containing the names of the transcription factors to be plotted. These should match colnames of \code{seFoot}.}

\item{pal}{The name of a custom palette from \code{ArchRPalettes} to use for plotting the lines corresponding to the footprints.}

\item{flank}{The number of basepairs from the position center (+/-) to consider as the flank.}

\item{flankNorm}{The number of basepairs to consider at the edge of the flank region (+/-) to be used for footprint normalization.}

\item{normMethod}{The name of the normalization method to use to normalize the footprint relative to the Tn5 insertion bias. Options
include "none", "subtract", "divide". "Subtract" means subtracting the normalized Tn5 Bias. "Divide" means dividing the normalized Tn5 Bias.}

\item{smoothWindow}{The size in basepairs of the sliding window to be used for smoothing of the footprint signal.}

\item{baseSize}{A numeric specifying the baseSize of font in the plots.}

\item{plot}{A boolean value indicating whether or not the footprints should be plotted (\code{TRUE}) or returned as grob objects (\code{FALSE}).}

\item{ArchRProj}{An \code{ArchRProject} object to be used for plotting directory in \code{getOutputDirectory}.}

\item{plotName}{A string indicating the name/prefix of the file to be used for output plots.}

\item{height}{The height in inches to be used for the output PDF file.}

\item{width}{The width in inches to be used for the output PDF file.}

\item{addDOC}{A boolean variable that determines whether to add the date of creation to end of the PDF file name. This is useful for
preventing overwritting of old plots.}

\item{force}{If many footprints are requested when plot = FALSE, please set force = TRUE.
This prevents large amount of footprint plots stored as an object.}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
This function will get footprints for all samples in a given ArchRProject or a properly-formatted Summarized Experiment
}
\examples{

# Get Test ArchR Project
proj <- getTestProject()

# Get Motif Positions
positions <- getPositions(proj)

# Get Footprints
seFoot <- getFootprints(ArchRProj = proj, positions = positions, groupBy = "Clusters", minCells = 10)

# Plot Footprints
plotFootprints(seFoot, smoothWindow = 11)

}
