% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Footprinting.R
\name{plotFootprints}
\alias{plotFootprints}
\title{Plot footprints for an ArchRProject}
\usage{
plotFootprints(ArchRProj = NULL, positions = NULL,
  plotName = "Plot-Footprints", groupBy = "Clusters",
  useGroups = NULL, pal = NULL, flank = 250, flankNorm = 50,
  smoothWindow = 10, minCells = 25, nTop = NULL,
  normMethod = "none", inputSE = NULL, height = 6, width = 4,
  addDOC = TRUE, useSink = TRUE, threads = 1, verboseHeader = TRUE,
  verboseAll = FALSE, ...)
}
\arguments{
\item{ArchRProj}{An `ArchRProject` object}

\item{positions}{QQQ A GenomicRangesList, a list, or a SimpleList object containing the positions to incorporate into the footprint. Each position should be QQQ.}

\item{plotName}{QQQ The prefix to add to the file name for the output PDF file.}

\item{groupBy}{QQQ The name of the column in `sampleColData` to use for grouping multiple samples together prior to footprinting.}

\item{useGroups}{A character vector that is used to select a subset of groups by name from the designated `groupBy` column in `sampleColData`. This limits the groups used to perform footprinting.}

\item{pal}{The name or numeric index of a custom palette from `ArchRPalettes` to use for plotting the lines corresponding to the footprints.}

\item{flank}{QQQ The number of basepairs from the position center to consider as the flank.}

\item{flankNorm}{QQQ The number of basepairs to consider at the edge of the flank region to be used for footprint normalization.}

\item{smoothWindow}{QQQ The size in basepairs of the sliding window to be used for smoothing of the footprint signal.}

\item{minCells}{QQQ The minimum number of cells required in a given cell group to permit footprint generation.}

\item{nTop}{QQQ The number of genomic regions to consider. Only the top `nTop` genomic regions based on QQQ will be considered for the footprint.}

\item{normMethod}{QQQ The name of the normalization method to use to normalize the footprint relative to the Tn5 insertion bias. Options include QQQ.}

\item{inputSE}{QQQ The name of the normalization method to use to normalize the footprint relative to the Tn5 insertion bias. Options include QQQ.}

\item{height}{The height in inches to be used for the output PDF.}

\item{width}{The width in inches to be used for the output PDF file.}

\item{addDOC}{A boolean variable that determines whether to add the date of creation to end of the PDF file name. This is useful for preventing overwritting of old plots.}

\item{useSink}{QQQ ???}

\item{threads}{The number of threads to be used for parallel computing.}

\item{verboseHeader}{A boolean value that determines whether standard output includes verbose sections.}

\item{verboseAll}{A boolean value that determines whether standard output includes verbose subsections.}

\item{...}{additional args}
}
\description{
This function will plot footprints for all samples in a given ArchRProject or a properly-formatted Summarized Experiment
}
