% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{plotFragmentSizes}
\alias{plotFragmentSizes}
\title{Plot fragment size distribution for each sample JJJ}
\usage{
plotFragmentSizes(ArchRProj = NULL, nbp = 750, returnDF = FALSE,
  threads = getArchRThreads())
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{nbp}{Number of basepairs to plot fragment size distribution.}

\item{returnDF}{A boolean for return plot as a \code{data.frame} instead.}

\item{threads}{An integer specifying the number of threads to use for calculation. By default this uses threads set by \code{addArchRThreads}.}
}
\description{
This function will plot a fragment size distribution for each sample. Cells in \code{ArchRProject} are the only ones
used when making this plot.
}
