% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeData.R
\name{plotGroups}
\alias{plotGroups}
\title{Visualize Groups from ArchR Project}
\usage{
plotGroups(ArchRProj, groupBy = "Sample", colorBy = "colData",
  name = "TSSEnrichment", imputeWeights = getImputeWeights(ArchRProj),
  log2Norm = NULL, pal = NULL, ylim = NULL, size = 0.5,
  baseSize = 6, ratioYX = 0.5, points = FALSE, ...)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{groupBy}{use groupings in cellColData for summarizing and plotting}

\item{colorBy}{A string indicating whether numeric values in violin plot should be from a column in cellColData ("cellColData") or by a data matrix in the ArrowFiles (i.e. "GeneScoreMatrix", "MotifMatrx", "PeakMatrix").}

\item{name}{The name of the column in \code{cellColData} or the featureName in the data matrix.}

\item{pal}{A custom palette used to override discreteSet/continuousSet for coloring vector.}

\item{ylim}{A vector of two numeric values indicating the lower and upper bounds of the y-axis on the plot.}

\item{size}{The numeric size of the points to be plotted.}

\item{baseSize}{The base font size to use in the plot.}

\item{ratioYX}{The aspect ratio of the x and y axes on the plot.}

\item{points}{A boolean value that indicates whether points should be added to the plot using \code{geom_quasirandom()}}

\item{...}{additional args}
}
\description{
This function will plot an embedding that was created from computeEmbedding
}
