% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeData.R
\name{plotGroups}
\alias{plotGroups}
\title{Visualize Groups from ArchR Project}
\usage{
plotGroups(ArchRProj, groupBy = "Sample", colorBy = "colData",
  name = "TSSEnrichment", imputeWeights = getImputeWeights(ArchRProj),
  log2Norm = NULL, pal = NULL, ylim = NULL, size = 0.5,
  baseSize = 6, ratioYX = 0.75, points = FALSE, ...)
}
\arguments{
\item{ArchRProj}{ArchRProject}

\item{groupBy}{use groupings in cellColData for summarizing and plotting}

\item{colorBy}{colorBy cellColData or Arrays in Arrows (ie GeneScoreMatrix)}

\item{name}{name of column in cellColData or Feature in Array in Arrows}

\item{pal}{custom palette to use for plotting}

\item{ylim}{limits for features in plot}

\item{size}{size of points in ggplot}

\item{baseSize}{rastr points in ggplot}

\item{ratioYX}{ratio of Y axis to X axis}

\item{points}{add points to plot using quasirandom}

\item{...}{additional args}
}
\description{
This function will plot an embedding that was created from
computeEmbedding
}
