% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeData.R
\name{plotGroups}
\alias{plotGroups}
\title{Visualize Groups from ArchR Project}
\usage{
plotGroups(ArchRProj, groupBy = "Sample", colorBy = "colData",
  name = "TSSEnrichment", imputeWeights = getImputeWeights(ArchRProj),
  log2Norm = NULL, pal = NULL, ylim = NULL, size = 0.5,
  baseSize = 6, ratioYX = 0.5, points = FALSE, ...)
}
\arguments{
\item{ArchRProj}{An `ArchRProject` object.}

\item{groupBy}{use groupings in cellColData for summarizing and plotting}

\item{colorBy}{QQQ UNCLEAR A string indicating whether points/hexes in the plot should be colored by cellColData ("cellColData") or by the gene activity score matrix ("GeneScoreMatrix").}

\item{name}{QQQ UNCLEAR The name of the column in `cellColData` or `GeneScoreMatrix` to use for QQQ.}

\item{pal}{QQQ The name or numeric index of a custom palette from `ArchRPalettes` to use for QQQ.}

\item{ylim}{A vector of two numeric values indicating the lower and upper bounds of the y-axis on the plot.}

\item{size}{The numeric size of the points to be plotted.}

\item{baseSize}{QQQ The base font size to use in the plot.}

\item{ratioYX}{The aspect ratio of the x and y axes on the plot.}

\item{points}{A boolean value that indicates whether points should be added to the plot using `geom_quasirandom()`}

\item{...}{additional args}
}
\description{
QQQ This function will plot an embedding that was created from computeEmbedding
}
