% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeData.R
\name{plotGroups}
\alias{plotGroups}
\title{Visualize Groups from ArchR Project}
\usage{
plotGroups(
  ArchRProj = NULL,
  groupBy = "Sample",
  colorBy = "colData",
  name = "TSSEnrichment",
  imputeWeights = if (!grepl("coldata", tolower(colorBy[1])))
    getImputeWeights(ArchRProj),
  maxCells = 1000,
  quantCut = c(0.002, 0.998),
  log2Norm = NULL,
  pal = NULL,
  discreteSet = "stallion",
  ylim = NULL,
  size = 0.5,
  baseSize = 6,
  ratioYX = NULL,
  ridgeScale = 2,
  plotAs = "ridges",
  threads = getArchRThreads(),
  ...
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{groupBy}{The name of the column in \code{cellColData} to use for grouping cells together for summarizing and plotting.}

\item{colorBy}{A string indicating whether the numeric values to be used in the violin plot should be from a column in
\code{cellColData} ("cellColData") or from a data matrix in the ArrowFiles (i.e. "GeneScoreMatrix", "MotifMatrix", "PeakMatrix").}

\item{name}{The name of the column in \code{cellColData} or the featureName/rowname of the data matrix to be used for plotting.
For example if \code{colorBy} is "cellColData" then \code{name} refers to a column name in the cellcoldata (see \code{getCellcoldata()}). If \code{colorBy}
is "GeneScoreMatrix" then \code{name} refers to a gene name which can be listed by \code{getFeatures(ArchRProj, useMatrix = "GeneScoreMatrix")}.}

\item{imputeWeights}{The weights to be used for imputing numerical values for each cell as a linear combination of other cells values. See \code{addImputationWeights()} and \code{getImutationWeights()} for more information.}

\item{maxCells}{JJJ The maximum cells to consider when making the plot.}

\item{quantCut}{JJJ If this is not null, a quantile cut is performed to threshold the top and bottom of the distribution of values.
This prevents skewed color scales caused by strong outliers. The format of this should be c(a,b) where \code{a} is the upper threshold and
\code{b} is the lower threshold. For example, quantCut = c(0.025,0.975) will take the top and bottom 2.5 percent of values and set them
to the value of the 97.5th and 2.5th percentile values respectively.}

\item{log2Norm}{A boolean value indicating whether a log2 transformation should be performed on the values (if continuous) in plotting.}

\item{pal}{A custom palette (see \code{paletteDiscrete} or \code{ArchRPalettes}) used to override discreteSet/continuousSet for coloring vector.}

\item{discreteSet}{JJJ The name of a discrete palette from \code{ArchRPalettes} for visualizing \code{colorBy} if a discrete color set is desired.}

\item{ylim}{A vector of two numeric values indicating the lower and upper bounds of the y-axis on the plot.}

\item{size}{The numeric size of the points to be plotted.}

\item{baseSize}{The base font size to use in the plot.}

\item{ratioYX}{The aspect ratio of the x and y axes on the plot.}

\item{ridgeScale}{The scale factor for the relative heights of each ridge when making a ridgeplot with \code{ggridges}.}

\item{plotAs}{A string that indicates whether a rigdge plot ("ridges") should be plotted or a violin plot ("violin") should be plotted.}

\item{threads}{The number of threads to be used for parallel computing.}

\item{...}{Additional parameters to pass to \code{ggGroup()}.}
}
\description{
This function will group, summarize and then plot data from an ArchRProject for visual comparison.
}
