% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{plotPDF}
\alias{plotPDF}
\title{Plot PDF in outputDirectory of an ArchRProject}
\usage{
plotPDF(..., name = "Plot", width = 6, height = 6,
  ArchRProj = NULL, addDOC = TRUE, useDingbats = FALSE,
  plotList = NULL, useSink = TRUE)
}
\arguments{
\item{...}{additional args to pdf}

\item{name}{The file name to be used for the output PDF file.}

\item{width}{The width in inches to be used for the output PDF file.}

\item{height}{The height in inches to be used for the output PDF.}

\item{ArchRProj}{An ArchRProject object.}

\item{addDOC}{A boolean variable that determines whether to add the date of creation to end of the PDF file name. This is useful for preventing overwritting of old plots.}

\item{useDingbats}{A boolean variable that determines wheter to use dingbats characters for plotting points.}
}
\description{
This function will save a plot or set of plots as a PDF file in the output directory of a given ArchRProject.
}
