% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{plotPDF}
\alias{plotPDF}
\title{Plot PDF in outputDirectory of ArchRProject}
\usage{
plotPDF(..., name = "Plot", width = 6, height = 6,
  ArchRProj = NULL, addDOC = TRUE, useDingbats = FALSE,
  plotList = NULL, useSink = TRUE)
}
\arguments{
\item{...}{additional args to pdf}

\item{name}{name of PDF file}

\item{width}{width of PDF in inches}

\item{height}{height of PDF in inches}

\item{ArchRProj}{ArchRProject}

\item{addDOC}{add date of creation to end of plot file name}

\item{useDingbats}{use dingbats characters for plotting}
}
\description{
This function will plot PDF in output directory of an ArchRProject
}
