% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRProjectMethods.R
\name{plotPDF}
\alias{plotPDF}
\title{Plot PDF in outputDirectory of an ArchRProject}
\usage{
plotPDF(..., name = "Plot", width = 6, height = 6,
  ArchRProj = NULL, addDOC = TRUE, useDingbats = FALSE,
  plotList = NULL, useSink = TRUE)
}
\arguments{
\item{...}{vector of plots to be plotted (if input is a list use plotList instead)}

\item{name}{The file name to be used for the output PDF file.}

\item{width}{The width in inches to be used for the output PDF file.}

\item{height}{The height in inches to be used for the output PDF.}

\item{ArchRProj}{QQQ WHAT IS PLOT DIRECTORY? An \code{ArchRProject} object to be used for getting plotDirectory in outputDirectory.}

\item{addDOC}{A boolean variable that determines whether to add the date of creation to the end of the PDF file name. This is useful for preventing overwritting of old plots.}

\item{useDingbats}{A boolean variable that determines wheter to use dingbats characters for plotting points.}

\item{plotList}{A \code{list} of plots to be printed to the output PDF file. Each element of \code{plotList} should be a printable plot formatted object (ggplot2, plot, heatmap, etc).}

\item{useSink}{QQQ A boolean value indicating whether sink should be used to hide output messages that result from plotting.}
}
\description{
This function will save a plot or set of plots as a PDF file in the outputDirectory of a given ArchRProject.
}
