% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrativeAnalysis.R
\name{plotPeak2GeneHeatmap}
\alias{plotPeak2GeneHeatmap}
\title{Plot Peak2Gene Heatmap from an ArchRProject}
\usage{
plotPeak2GeneHeatmap(
  ArchRProj = NULL,
  corCutOff = 0.45,
  FDRCutOff = 0.0001,
  varCutOffATAC = 0.25,
  varCutOffRNA = 0.25,
  k = 25,
  nPlot = 25000,
  limitsATAC = c(-2, 2),
  limitsRNA = c(-2, 2),
  groupBy = "Clusters",
  palGroup = NULL,
  palATAC = paletteContinuous("solarExtra"),
  palRNA = paletteContinuous("blueYellow"),
  verbose = TRUE,
  returnMatrices = FALSE,
  seed = 1,
  logFile = createLogFile("plotPeak2GeneHeatmap")
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{corCutOff}{A numeric describing the minimum numeric peak-to-gene correlation to return.}

\item{FDRCutOff}{A numeric describing the maximum numeric peak-to-gene false discovery rate to return.}

\item{varCutOffATAC}{A numeric describing the minimum variance quantile of the ATAC peak accessibility when selecting links.}

\item{varCutOffRNA}{A numeric describing the minimum variance quantile of the RNA gene expression when selecting links.}

\item{k}{An integer describing the number of k-means clusters to group peak-to-gene links prior to plotting heatmaps.}

\item{nPlot}{An integer describing the maximum number of peak-to-gene links to plot in heatmap.}

\item{limitsATAC}{An integer describing the maximum number of peak-to-gene links to plot in heatmap.}

\item{limitsRNA}{An integer describing the maximum number of peak-to-gene links to plot in heatmap.}

\item{groupBy}{The name of the column in \code{cellColData} to use for labeling KNN groupings. The maximum group appeared in the KNN groupings is used.}

\item{palGroup}{A color palette describing the colors in \code{groupBy}. For example, if groupBy = "Clusters" try paletteDiscrete(ArchRProj$Clusters) for a color palette.}

\item{palATAC}{A color palette describing the colors to be used for the ATAC heatmap. For example, paletteContinuous("solarExtra").}

\item{palRNA}{A color palette describing the colors to be used for the RNA heatmap. For example, paletteContinuous("blueYellow").}

\item{verbose}{A boolean value that determines whether standard output should be printed.}

\item{returnMatrices}{A boolean value that determines whether the matrices should be returned with kmeans id versus plotting.}

\item{seed}{A number to be used as the seed for random number generation. It is recommended to keep track of the seed used so that you can
reproduce results downstream.}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
This function plots side by side heatmaps of linked ATAC and Gene regions from \code{addPeak2GeneLinks}.
}
\examples{

# Get Test ArchR Project
proj <- getTestProject()

# Add P2G Links
proj <- addPeak2GeneLinks(proj, k = 20)

# Get P2G Links
p2g <- getPeak2GeneLinks(proj)

# Plot P2G
p <- plotPeak2GeneHeatmap(proj)
plotPDF(p, name = "P2G-Heatmap", ArchRProj = proj)

}
