% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trajectory.R
\name{plotTrajectory}
\alias{plotTrajectory}
\title{Visualize a Trajectory from ArchR Project}
\usage{
plotTrajectory(ArchRProj = NULL, embedding = "UMAP",
  trajectory = "Trajectory", colorBy = "colData",
  name = "Trajectory", log2Norm = NULL, imputeWeights = NULL,
  pal = NULL, size = 0.5, rastr = TRUE, quantCut = c(0.05, 0.95),
  quantHex = 0.5, discreteSet = NULL, continuousSet = NULL,
  randomize = TRUE, keepAxis = FALSE, baseSize = 6,
  addArrow = TRUE, plotAs = NULL, plotParams = list(), ...)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{embedding}{The name of the embedding to use to visualize the given \code{trajectory}. See \code{addEmbedding()} for more information.}

\item{trajectory}{The column name in \code{cellColData} that refers the trajectory to be plotted. See \code{addTrajectory()} for more information.}

\item{colorBy}{A string indicating whether points in the plot should be colored by a column in \code{cellColData} ("cellColData") or by a data matrix in the associated ArrowFiles (i.e. "GeneScoreMatrix", "MotifMatrix", "PeakMatrix").}

\item{name}{The name of the column in \code{cellColData} or the featureName/rowname of the data matrix to be used for plotting.
For example if colorBy is \code{cellColData} then name refers to a column name in the cellcoldata (see \code{getCellcoldata()}), if colorBy is \code{GeneScoreMatrix} then name refers to a gene name which can be listed by \code{getFeatures(ArchRProj, useMatrix = "GeneScoreMatrix")}.}

\item{log2Norm}{A boolean value indicating whether a log2 transformation should be performed on the values from \code{colorBy}.}

\item{imputeWeights}{The weights to be used for imputing numerical values for each cell as a linear combination of other cells values. See \code{addImputationWeights()} and \code{getImutationWeights()} for more information.}

\item{pal}{The name of a custom palette from \code{ArchRPalettes} to use for coloring cells.}

\item{size}{A number indicating the size of the points to plot if \code{plotAs} is set to "points".}

\item{rastr}{A boolean value that indicates whether the plot should be rasterized. This does not rasterize lines and labels, just the internal portions of the plot.}

\item{quantCut}{If this is not \code{NULL}, a quantile cut is performed to threshold the top and bottom of the distribution of numerical values.
This prevents skewed color scales caused by strong outliers. The format of this should be c(x,y) where x is the lower threshold and y is
the upper threshold. For example, quantileCut = c(0.025,0.975) will take the 2.5th percentile and 97.5 percentile of values and set values below/above to the value of
the 2.5th and 97.5th percentile values respectively.}

\item{quantHex}{The xth quantile of all dots within each individual hexagon will determine the the numerical value for coloring to be displayed. This occurs when \code{plotAs} = "hex" or NULL (if numerical values by default).}

\item{discreteSet}{The name of a discrete palette from \code{ArchRPalettes} for visualizing \code{colorBy} in the embedding if a discrete color set is desired.}

\item{continuousSet}{The name of a continuous palette from \code{ArchRPalettes} for visualizing \code{colorBy} in the embedding if a continuous color set is desired.}

\item{randomize}{A boolean value that indicates whether to randomize points prior to plotting to prevent cells from one cluster being present at the front of the plot.}

\item{keepAxis}{A boolean value that indicates whether the x and y axis ticks and labels should be plotted.}

\item{baseSize}{The base font size to use in the plot.}

\item{addArrow}{A boolean value that indicates whether to add a smoothed arrow in the embedding based on the aligned trajectory.}

\item{plotAs}{A string that indicates whether points ("points") should be plotted or a hexplot ("hex") should be plotted. By default if \code{colorBy} is numeric this is "hex".}

\item{plotParams}{Additional parameters to pass to \code{ggPoint()} or \code{ggHex()}.}

\item{...}{additional args}
}
\description{
This function will plot a trajectory that was created onto an embedding.
}
