% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{plotTSSEnrichment}
\alias{plotTSSEnrichment}
\title{Plot fragment size distribution for each sample JJJ}
\usage{
plotTSSEnrichment(ArchRProj = NULL, TSS = getTSS(ArchRProj),
  flank = 2000, norm = 100, smooth = 11, returnDF = FALSE,
  threads = getArchRThreads())
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{TSS}{A GRanges object containing stranded TSS ranges. By default will try to get this info from geneAnnotation stored in \code{ArchRProj}.}

\item{flank}{A numeric specifying how far in bp (+/-) to extend TSS for plotting.}

\item{norm}{A numeric specifying the number of base pairs from the ends of the flanks to be used for normalization.
For example if \code{flank=2000} and \code{norm=100}, the TSS insertions will be normalized by +/- 1900-2000 bp from the TSS.}

\item{smooth}{A numeric describing the smoothing window to be applied to TSS plot.}

\item{returnDF}{A boolean for return plot as a \code{data.frame} instead.}

\item{threads}{An integer specifying the number of threads to use for calculation. By default this uses threads set by \code{addArchRThreads}.}
}
\description{
This function will plot a TSS enrichment plot for each sample. Cells in \code{ArchRProject} are the only ones
used when making this plot.
}
