% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_from_FASTA.R
\name{prepare_data_from_FASTA}
\alias{prepare_data_from_FASTA}
\title{Given a FASTA file, this function return a matrix with one-hot encoded
sequences along the columns or the simply the raw sequences.}
\usage{
prepare_data_from_FASTA(inputFastaFilename, rawSeq = FALSE,
  sinuc_or_dinuc = "sinuc")
}
\arguments{
\item{inputFastaFilename}{Provide the name (with complete path) of the input
FASTA file.}

\item{rawSeq}{TRUE or FALSE, set this to TRUE if you want the raw sequences.}

\item{sinuc_or_dinuc}{character string, 'sinuc' or 'dinuc' to select for
single or dinucleotide profiles.}
}
\value{
A matrix of sequences represented with one-hot-encoding. Dimensions
of the matrix: 4*(sequence length) x number of sequences.
}
\description{
In the data matrix, the features are along rows, and sequences along columns.
Currently, only mono-character features for DNA sequences are supported.
Therefore, the length of the feature vector is four times the
length of the sequences (since the DNA alphabet is four characters).
}
