% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperUtils.R
\name{reformatFragmentFiles}
\alias{reformatFragmentFiles}
\title{Reformat Fragment Files to be Tabix and Chr Sorted}
\usage{
reformatFragmentFiles(
  fragmentFiles = NULL,
  checkChrPrefix = getArchRChrPrefix()
)
}
\arguments{
\item{fragmentFiles}{A character vector the paths to fragment files to be reformatted}

\item{checkChrPrefix}{A boolean value that determines whether seqnames should be checked to contain
"chr". IF set to \code{TRUE}, any seqnames that do not contain "chr" will be removed from the fragment files.}
}
\description{
This function provides help in reformatting Fragment Files for reading in createArrowFiles.
It will handle weird anomalies found that cause errors in reading tabix bgzip'd fragment files.
}
