% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archR_auxiliary_functionsI.R
\name{reorder_archRresult}
\alias{reorder_archRresult}
\title{Reorder archR clustering at the current last level}
\usage{
reorder_archRresult(
  archRresult,
  iteration = 3,
  clustMethod = "hc",
  linkage = "average",
  distMethod = "euclid",
  regularize = TRUE,
  topN = 10,
  returnOrder = FALSE,
  position_agnostic_dist = FALSE,
  config
)
}
\arguments{
\item{archRresult}{The archRresult object}
}
\value{
Ordering returned from hclust
}
\description{
We use hclust/hierarchical clustering for reordering archR
clusters
}
\details{
Keep this internal or external function?
}
