% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archR_auxiliary_functionsI.R
\name{reorder_archRresult}
\alias{reorder_archRresult}
\title{Reorder archR raw clustering at the chosen iteration}
\usage{
reorder_archRresult(
  archRresult,
  iteration = length(archRresult$seqsClustLabels),
  clustMethod = "hc",
  linkage = "ward.D",
  distMethod = "euclid",
  regularize = TRUE,
  topN = floor(0.5 * archRresult$rawSeqs@ranges@width),
  decisionToReorder = TRUE,
  config,
  ...
)
}
\arguments{
\item{archRresult}{The archRresult object.}

\item{iteration}{Specify clusters at which iteration of archR are to be 
reordered.}

\item{clustMethod}{Specify 'hc' for hierarchical clustering. Currently, only
hierarchical clustering is supported.}

\item{linkage}{One of linkage values as specified for hierarchical clustering.
Default is `ward.D'.}

\item{distMethod}{Distance measure to be used with hierarchical clustering. 
Available options are "euclid" (default), "cor" for correlation, "cosangle" 
for cosine angle, "modNW" for modified Needleman-Wunsch similarity (see 
\code{\link[TFBSTools]{PFMSimilarity}}).}

\item{regularize}{Logical. Specify TRUE if regularization is to be performed 
before comparison. See argument 'topN'.}

\item{topN}{Use only the 'topN' dimensions of each basis vector for comparing 
them. Note that since each basis vector has 4L or 16L (mono- or 
dinucleotides) dimensions, each dimension is a combination of nucleotide and 
its position in the sequence. This argument selects the top N dimensions of 
the basis vector.}

\item{decisionToReorder}{Logical. Specify TRUE if reordering using 
hierarchical agglomerative clustering is to be performed, otherwise FALSE.}

\item{config}{Pass the configuration of the archR result object.}

\item{...}{ignored}
}
\value{
Returns the re-ordered clusters. When `decisionToReorder' is FALSE, 
it returns the already existing basis vectors, each as singleton clusters. 
The sequence cluster labels and sequence clusters are also handled 
accordingly.
}
\description{
We use hierarchical clustering for reordering/collating raw
clusters from archR's given iteration.
}
