% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{saveArchRProject}
\alias{saveArchRProject}
\title{Save ArchRProject for Later Usage}
\usage{
saveArchRProject(
  ArchRProj = NULL,
  outputDirectory = getOutputDirectory(ArchRProj),
  overwrite = TRUE,
  load = TRUE,
  dropCells = FALSE,
  logFile = createLogFile("saveArchRProject"),
  threads = getArchRThreads()
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{outputDirectory}{A directory path to save all ArchR output and \code{ArchRProject} to. Default is outputDirectory of the \code{ArchRProject}.}

\item{overwrite}{When writing to outputDirectory, overwrite existing files with new files.}

\item{dropCells}{A boolean indicating whether to drop cells that are not in \code{ArchRProject} from corresponding Arrow Files.}

\item{logFile}{The path to a file to be used for logging ArchR output.}

\item{threads}{The number of threads to use for parallel execution.}
}
\description{
This function will organize arrows and project output into a directory and save the ArchRProject for later usage.
}
