% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesUtils.R
\name{shuffleGRanges}
\alias{shuffleGRanges}
\title{QQQ Randomly shuffle a Genomic Ranges object}
\usage{
shuffleGRanges(subject, genome, n, shuffleChr = TRUE)
}
\arguments{
\item{subject}{A \code{GRanges} object.}

\item{genome}{The name of a valid genome (for example "hg38", "hg19", or "mm10"). See \code{ArchR::validBSgenome()}.}

\item{n}{QQQ The number of permutations to perform during shuffling.}

\item{shuffleChr}{QQQ A boolean value indicating whether to shuffle across chromosomes randomly or to QQQ (WHAT?) use previous knowledge of chromosome distribution}
}
\description{
QQQ This function randomly shuffles a Genomic Ranges object.
}
