% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{subsetArchRProject}
\alias{subsetArchRProject}
\title{Subset an ArchRProject for downstream analysis}
\usage{
subsetArchRProject(
  ArchRProj = NULL,
  cells = getCellNames(ArchRProj),
  outputDirectory = "ArchRSubset"
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{cells}{A vector of cells to subset \code{ArchRProject} by. Alternatively can provide a subset \code{ArchRProject}.}

\item{outputDirectory}{A directory path to save all ArchR output and the subsetted \code{ArchRProject} to.}
}
\description{
This function will subset and ArchRProject by cells and save the output to a new directory and re-load the subsetted ArchRProject.
}
