% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesUtils.R
\name{subsetSeqnames}
\alias{subsetSeqnames}
\title{Subset a Genomic Ranges object by the provided seqnames}
\usage{
subsetSeqnames(gr, seqNames, useNames = FALSE)
}
\arguments{
\item{gr}{A `GRanges` object to be subsetted.}

\item{seqNames}{A character vector containing the `seqnames` to keep from the provided `GRanges` object.}

\item{useNames}{QQQ A boolean value indicating whether QQQ.}
}
\description{
This function returns a subsetted Genomic Ranges object based on a vector of provided seqnames
}
