% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgplotHelper.R
\name{theme_ArchR}
\alias{theme_ArchR}
\title{ggplot2 default theme for ArchR}
\usage{
theme_ArchR(color = "black", baseSize = 6, baseFamily = "",
  baseLineSize = 0.5, baseRectSize = 0.5, plotMarginCm = 1,
  legendPosition = "bottom", legendTextSize = 5, axisTickCm = 0.1,
  xText90 = FALSE, yText90 = FALSE, ...)
}
\arguments{
\item{color}{color of theme}

\item{baseSize}{is the size of the font for the axis text and title}

\item{baseFamily}{is family for font}

\item{baseLineSize}{is the size of line}

\item{baseRectSize}{is the size of rectangle boxes}

\item{plotMarginCm}{plot margin in cm}

\item{legendPosition}{where is the legend default bottom}

\item{legendTextSize}{0.75*base_size}

\item{axisTickCm}{axis tick length in cm}

\item{xText90}{rotate x axis text 90 degrees}

\item{yText90}{rotate y axis text 90 degrees}
}
\description{
This function returns a ggplot2 theme that is black borded with black font.
}
