% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgplotUtils.R
\name{theme_ArchR}
\alias{theme_ArchR}
\title{ggplot2 default theme for ArchR}
\usage{
theme_ArchR(
  color = "black",
  textFamily = "sans",
  baseSize = 10,
  baseLineSize = 0.5,
  baseRectSize = 0.5,
  plotMarginCm = 1,
  legendPosition = "bottom",
  legendTextSize = 5,
  axisTickCm = 0.1,
  xText90 = FALSE,
  yText90 = FALSE
)
}
\arguments{
\item{color}{The color to be used for text, lines, ticks, etc for the plot.}

\item{textFamily}{The font default family to be used for the plot.}

\item{baseSize}{The base font size (in points) to use in the plot.}

\item{baseLineSize}{The base line width (in points) to be used throughout the plot.}

\item{baseRectSize}{The base line width (in points) to use for rectangular boxes throughout the plot.}

\item{plotMarginCm}{The width in centimeters of the whitespace margin around the plot.}

\item{legendPosition}{The location to put the legend. Valid options are "bottom", "top", "left", and "right.}

\item{legendTextSize}{The base text size (in points) for the legend text.}

\item{axisTickCm}{The length in centimeters to be used for the axis ticks.}

\item{xText90}{A boolean value indicating whether the x-axis text should be rotated 90 degrees counterclockwise.}

\item{yText90}{A boolean value indicating whether the y-axis text should be rotated 90 degrees counterclockwise.}
}
\description{
This function returns a ggplot2 theme that is black borded with black font.
}
