% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trajectory.R
\name{trajectoryHeatmap}
\alias{trajectoryHeatmap}
\title{Plot a Heatmap of Features across a Trajectory}
\usage{
trajectoryHeatmap(
  seTrajectory = NULL,
  varCutOff = 0.9,
  maxFeatures = 25000,
  scaleRows = TRUE,
  limits = c(-1.5, 1.5),
  grepExclude = NULL,
  pal = NULL,
  labelMarkers = NULL,
  labelTop = 50,
  labelRows = FALSE,
  rowOrder = NULL,
  returnMat = FALSE
)
}
\arguments{
\item{seTrajectory}{A \code{SummarizedExperiment} object that results from calling \code{markerFeatures()}.}

\item{scaleRows}{A boolean value that indicates whether row-wise z-scores should be computed on the matrix provided by \code{seTrajectory}.}

\item{limits}{A numeric vector of two numbers that represent the lower and upper limits of the heatmap color scheme.}

\item{grepExclude}{A character vector or string that indicates the \code{rownames} or a specific pattern that identifies
rownames from \code{seTrajectory} to be excluded from the heatmap.}

\item{pal}{A custom continuous palette (see \code{paletteContinuous()}) used to override the default continuous palette for the heatmap.}

\item{labelMarkers}{A character vector listing the \code{rownames} of \code{seTrajectory} that should be labeled on the side of the heatmap.}

\item{labelTop}{A number indicating how many of the top N features, based on variance, in \code{seTrajectory} should be labeled on the side of the heatmap.}

\item{labelRows}{A boolean value that indicates whether all rows should be labeled on the side of the heatmap.}

\item{returnMat}{A boolean value that indicates whether the final heatmap matrix should be returned in lieu of plotting the actual heatmap.}
}
\description{
This function will plot a heatmap of the results from getTrajectory
}
