% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trajectory.R
\name{trajectoryHeatmap}
\alias{trajectoryHeatmap}
\title{Plot a Heatmap of Features across a Trajectory}
\usage{
trajectoryHeatmap(seTrajectory = NULL, scaleRows = TRUE,
  limits = c(-2, 2), grepExclude = NULL, pal = NULL,
  labelMarkers = NULL, labelTop = 50, labelRows = FALSE,
  returnMat = FALSE, ...)
}
\arguments{
\item{seTrajectory}{Summarized Experiment result from markerFeatures}

\item{scaleRows}{compute row z-scores on matrix}

\item{limits}{heatmap color limits}

\item{pal}{palette for heatmap, default will use solar_extra}

\item{labelMarkers}{label specific markers by name on heatmap (matches rownames of seTrajectory)}

\item{labelTop}{label the top features for each column in seTrajectory}

\item{labelRows}{label all rows}

\item{returnMat}{return final matrix that is used for plotting heatmap}

\item{...}{additional args}

\item{varCutOff}{variance cut off for selecting features varying across trajectory}
}
\description{
This function will plot a heatmap of the results from getTrajectory
}
