% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trajectory.R
\name{trajectoryHeatmap}
\alias{trajectoryHeatmap}
\title{Plot a Heatmap of Features across a Trajectory}
\usage{
trajectoryHeatmap(seTrajectory = NULL, scaleRows = TRUE,
  limits = c(-2, 2), grepExclude = NULL, pal = NULL,
  labelMarkers = NULL, labelTop = 50, labelRows = FALSE,
  returnMat = FALSE, ...)
}
\arguments{
\item{seTrajectory}{A \code{SummarizedExperiment} object that results from \code{markerFeatures()}.}

\item{scaleRows}{A boolean value that indicates whether row-wise z-scores should be computed on matrix in \code{seTrajectory}.}

\item{limits}{A numeric vector of two numbers that represent the lower and upper color limits of the heatmap color scheme.}

\item{grepExclude}{A character vector or string that indicates the \code{rownames} or a specific pattern that identifies rownames from \code{seTrajectory} to be excluded from the heatmap.}

\item{pal}{A custom continuous palette (see paletteContinuous) used to override the continuous palette for the heatmap.}

\item{labelMarkers}{A character vector listing the \code{rownames} of \code{seTrajectory} that should be labeled on the side of the heatmap.}

\item{labelTop}{A boolean value that indicates whether the top features for each column in \code{seTrajectory} should be labeled on the side of the heatmap.}

\item{labelRows}{A boolean value that indicates whether all rows should be labeled on the side of the heatmap.}

\item{returnMat}{A boolean value that indicates whether the final heatmap matrix should be returned in lieu of plotting the actual heatmap.}

\item{...}{additional args}
}
\description{
This function will plot a heatmap of the results from getTrajectory
}
