% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_all_factors_in_combined_heatmaps_seqlogos.R
\name{viz_all_factors_in_combined_heatmaps_seqlogos}
\alias{viz_all_factors_in_combined_heatmaps_seqlogos}
\title{Visualize the features matrix in a paired heatmap and sequence logo plot.}
\usage{
viz_all_factors_in_combined_heatmaps_seqlogos(featuresMatrix,
  position_labels = NA, add_pseudo_counts = F,
  savePDFfilename = NULL)
}
\arguments{
\item{featuresMatrix}{The features matrix output from NMF. Expected
dimensionality: number of columns represent the number of factors from NMF,
and the number of rows is 4 times the length of the sequences in the collection.}

\item{position_labels}{Labels of the positions in the sequences.}

\item{add_pseudo_counts}{Boolean, taking values TRUE/T or FALSE/F, default set
to FALSE. Setting it to TRUE will enable adding pseudo-counts to the features
matrix.}

\item{savePDFfilename}{Name of the file which will be saved as PDF
(also provide the extension).}
}
\value{
nothing
}
\description{
The given features matrix (with 4 rows) is represented as a
heatmap followed by a sequence logo where the positions are aligned for
better visualization.
}
\examples{


}
