% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/viz_basis_vectors_in_combined_heatmaps_seqlogos.R
\name{viz_basis_vectors_in_combined_heatmaps_seqlogos}
\alias{viz_basis_vectors_in_combined_heatmaps_seqlogos}
\title{Visualize the NMF basis vectors in a paired heatmap and sequence logo
plot}
\usage{
viz_basis_vectors_in_combined_heatmaps_seqlogos(
  featuresMatrix,
  plotMethod = "custom",
  position_labels = NA,
  add_pseudo_counts = FALSE,
  savePDFfilename = NULL,
  sinuc_or_dinuc = "sinuc"
)
}
\arguments{
\item{featuresMatrix}{The features matrix output from NMF. Expected
dimensionality: number of columns represent the number of factors from NMF,
and the number of rows is 4 times the length of the sequences in the
collection.}

\item{plotMethod}{For \code{ggseqlogo}; either of "custom", "bits", or
"probability".
Default is "custom".}

\item{position_labels}{Labels of the positions in the sequences.}

\item{add_pseudo_counts}{Boolean, taking values TRUE/T or FALSE/F, default
set
 to FALSE. Setting it to TRUE will enable adding pseudo-counts to the
 features
 matrix.}

\item{savePDFfilename}{Name of the file which will be saved as PDF
(also provide the extension).}

\item{sinuc_or_dinuc}{"sinuc" or "dinuc"}
}
\value{
nothing
}
\description{
The given features matrix (with 4 rows) is represented as a
heatmap followed by a sequence logo where the positions are aligned for
better visualization.
}
\seealso{
\code{\link{viz_basis_vectors_as_heatmap}} for plotting only as
heatmaps, \code{\link{viz_basis_vectors_as_seqlogo}} for plotting only as
sequence logos.

Other visualization functions: 
\code{\link{plot_ggheatmap}()},
\code{\link{plot_ggseqlogo}()},
\code{\link{viz_basis_vectors_as_heatmap}()},
\code{\link{viz_basis_vectors_as_seqlogo}()},
\code{\link{viz_matrix_of_acgt}()}
}
\concept{visualization functions}
